Cardano.Constitution.Validator.Data.Sorted

Documentation

constitutionValidator ∷ ConstitutionConfig → ConstitutionValidator

Expects a constitution-configuration, statically *OR* at runtime via Tx.liftCode

defaultConstitutionValidator ∷ ConstitutionValidator

Statically configure the validator with the defaultConstitutionConfig.

mkConstitutionCode ∷ ConstitutionConfig → CompiledCode ConstitutionValidator

Make a constitution code by supplied the config at runtime.

See Note [Manually constructing a Configuration value]

defaultConstitutionCode ∷ CompiledCode ConstitutionValidator

The code of the constitution statically configured with the defaultConstitutionConfig.