Codec.Extras.SerialiseViaFlat

Documentation

newtype SerialiseViaFlat a

Newtype to provide Serialise instances for types with a Flat instance that just encodes the flat-serialized value as a CBOR bytestring

Constructors

SerialiseViaFlat

Fields

unSerialiseViaFlat ∷ a

decodeViaFlatWith ∷ Get a → Decoder s a

data DeserialiseFailureInfo

Similar to DeserialiseFailure, with the difference that plain string reason messages are turned into the datatype: DeserialiseFailureReason.

Constructors

DeserialiseFailureInfo

Fields

dfOffset ∷ ByteOffset

dfReason ∷ DeserialiseFailureReason

data DeserialiseFailureReason

The reason of the cbor failure as a datatype, not as a plain string.

Constructors

EndOfInput
Not enough input provided

ExpectedBytes
The bytes inside the input are malformed.

OtherReason String
This is either a cbor failure that we (plutus) are not aware of, or an underlying flat failure. We use whatever message cborg or flat returns.

readDeserialiseFailureInfo ∷ DeserialiseFailure → DeserialiseFailureInfo

The errors returned by cborg are plain strings (untyped). With this function we try to map onto datatypes, those cborg error messages that we are interested in.

Currently we are only interested in error messages returned by the decodeBytes decoder; see scriptCBORDecoder.