Data.Aeson.Extra

Documentation

buildObject ∷ (Object → Object) → Value

Build a JSON object omitting optional keys if a corresponding value is Nothing.

Example: buildObject $ requiredField "field1" a . requiredField "field2" c . optionalField "field3" (Just "hello") . optionalField "field4" Nothing builds this JSON object: { "field1": a, "field2": c, "field3": "hello" } omitting optional field4.

optionalField ∷ ToJSON a ⇒ Key → Maybe a → Object → Object

requiredField ∷ ToJSON a ⇒ Key → a → Object → Object

stripPrefix

Arguments

∷ String
Field prefix to strip

→ String
Field name

→ String

A field label modifier that strips a prefix from the camelCased field name; >>> stripPrefix "preamble" "preambleTitle" "title"