Data.Aeson.Flatten

Documentation

flattenObject ∷ Text → Object → Object

Turn a nested object into a "flat" object where the keys represent paths into the original object. The keys in the result will be the keys in the original path, separated by sep. The inverse of unflattenObject.

unflattenObject ∷ Text → Object → Object

Turn a "flat" object whose keys represent paths into an unflattened object. The keys in the result will be the resulting path, separated by sep. The inverse of flattenObject.

mergeObject ∷ Object → Object → Object

Merge two objects, merging the values where both sides have an entry for a key rather than taking the first.

mergeValue ∷ Value → Value → Value

Merge two values, merging the objects using mergeObject. Can't merge anything else.

objToHm ∷ Object → HashMap Text Value

Convert an object to a hashmap. Compatibility shim for pre-/post-aeson-2.

hmToObj ∷ HashMap Text Value → Object

Convert a hashmap to an object. Compatibility shim for pre-/post-aeson-2.