Data.Either.Extras

Documentation

unsafeFromEither ∷ Exception e ⇒ Either e a → a

If argument is Left throw an exception, otherwise return the value inside Right.

fromRightM ∷ Monad m ⇒ (a → m b) → Either a b → m b

If argument is Right return its value, otherwise apply the given action to the value of Left and return its result.