Data.RandomAccessList.Class

Documentation

class RandomAccessList e where

Typeclass for various types implementing the "signature" of a random-access list.

A true random-access list should have good asymptotic behaviour for these methods also, but for convenience we also provide implementations for e.g. '[a]', which has bad lookup performance.

Minimal complete definition

empty, cons, uncons, length

Associated Types

type Element e ∷ Type

The type of elements in the list.

Methods

empty ∷ e

The empty list.

cons ∷ Element e → e → e

Prepend an element to the list.

uncons ∷ e → Maybe (Element e, e)

Un-prepend an element to the list.

length ∷ e → Word64

Get the length of the list. May have linear complexity, but useful.

consSlab ∷ NonEmptyVector (Element e) → e → e

Prepend many elements to the list. Has a default implementation, but implementations can provide more efficient ones.

indexZero ∷ e → Word64 → Maybe (Element e)

Lookup an element in the list. 0-based index.

indexOne ∷ e → Word64 → Maybe (Element e)

Lookup an element in the list. 1-based index.

unsafeIndexZero ∷ e → Word64 → Element e

Lookup an element in the list, partially.

unsafeIndexOne ∷ e → Word64 → Element e

Lookup an element in the list, partially.

head ∷ (RandomAccessList e, a ~ Element e) ⇒ e → a

tail ∷ RandomAccessList e ⇒ e → e

newtype AsRAL a

Constructors

AsRAL a