Data.SatInt

Description

Adapted from SafeInt to perform saturating arithmetic (i.e. returning max or min bounds) instead of throwing on overflow.

This is not quite as fast as using Int or Int64 directly, but we need the safety.

Documentation

data SatInt

unsafeToSatInt ∷ Int → SatInt

Wrap an Int as a SatInt. This is unsafe because the Int can be a result of an arbitrary potentially underflowing/overflowing operation.

fromSatInt ∷ ∀ a. Num a ⇒ SatInt → a

An optimized version of fromIntegral . unSatInt.

dividedBy ∷ SatInt → Natural → SatInt

Divide a SatInt by a natural number. If the natural number is zero, return maxBound; if we're at the maximum or minimum value then leave the input unaltered. This should never throw.