Evaluation.Builtins.BLS12_381.TestClasses

Documentation

class (Eq a, Show a, Arbitrary a, ArbitraryBuiltin a) ⇒ TestableAbelianGroup a where

The code for the property tests for G1 and G2 is essentially identical, so it's worth abstracting over the common features. The blst Haskell FFI uses a phantom type to do this but unfortunately we have to hide that to stop the builtin machinery spotting it and then we have to re-abstract here.

Methods

groupName ∷ String

zeroTerm ∷ PlcTerm

addTerm ∷ PlcTerm → PlcTerm → PlcTerm

negTerm ∷ PlcTerm → PlcTerm

scalarMulTerm ∷ PlcTerm → PlcTerm → PlcTerm

eqTerm ∷ PlcTerm → PlcTerm → PlcTerm

toTerm ∷ a → PlcTerm

class TestableAbelianGroup a ⇒ HashAndCompress a where

Methods

compressedSize ∷ Int

compress ∷ a → ByteString

compressTerm ∷ PlcTerm → PlcTerm

uncompressTerm ∷ PlcTerm → PlcTerm

hashToGroupTerm ∷ PlcTerm → PlcTerm → PlcTerm

Orphan instances

Arbitrary Element
Generate an arbitrary element of G1. It's tricky to construct such an element directly without using quite low-level operations on the curve because a random point on the curve is highly unlikely to be in the subgroup G1, but fortunately hashToGroup always produces an element of the subgroup, so we can produce random elements of G1 by hasing random bytestrings.

Instance details

Methods

arbitrary ∷ Gen Element

shrink ∷ Element → [Element]

Arbitrary Element
See the comment for the Arbitrary instance for G1.

Instance details

Methods

arbitrary ∷ Gen Element

shrink ∷ Element → [Element]