Evaluation.Builtins.Bitwise.CIP0123

Description

Tests for CIP-0123(the second batch of bitwise builtins).

Documentation

shiftHomomorphism ∷ [TestTree]

There should exist a monoid homomorphism between natural number addition and function composition for shifts over a fixed bytestring argument.

rotateHomomorphism ∷ [TestTree]

There should exist a monoid homomorphism between integer addition and function composition for rotations over a fixed bytestring argument.

csbHomomorphism ∷ [TestTree]

There should exist a monoid homomorphism between bytestring concatenation and natural number addition.

shiftClear ∷ Property

Shifting by more than the bit length (either positive or negative) clears the result.

rotateRollover ∷ Property

Rotations by more than the bit length 'roll over' bits.

csbRotate ∷ Property

Rotations do not change how many set (and clear) bits there are.

shiftPosClearLow ∷ Property

Positive shifts clear low-index bits.

shiftNegClearHigh ∷ Property

Negative shifts clear high-index bits.

rotateMoveBits ∷ Property

Rotations move bits, but don't change them.

csbComplement ∷ Property

For any bytestring whose bit length is n and has k set bits, its complement should have n - k set bits.

csbInclusionExclusion ∷ Property

The inclusion-exclusion principle: specifically, for any x and y, the number of set bits in x XOR y should be the number of set bits in x OR y minus the number of set bits in x AND y.

csbXor ∷ Property

For any bytestring x, the number of set bits in x XOR x should be 0.

ffsReplicate ∷ Property

If we replicate any byte any (positive) number of times, the first set bit should be the same as in the case where we replicated it exactly once.

ffsXor ∷ Property

For any choice of bytestring, if we XOR it with itself, there should be no set bits; that is, finding the first set bit should give -1.

ffsIndex ∷ Property

If we find a valid index for the first set bit, then:

The specified index should have a set bit; and

Any valid smaller index should have a clear bit.

We hack the generator slightly here to ensure we don't end up with all-zeroes (or the empty bytestring), as otherwise, the test wouldn't be meaningful.

ffsZero ∷ Property

Finding the first set bit in a bytestring with only zero bytes should always give -1.

shiftMinBound ∷ Property

If given Int minBound as an argument, shifts behave sensibly.

rotateMinBound ∷ Property

If given Int minBound as an argument, rotations behave sensibly.

ffs6453 ∷ Property

Test for a regression found here.