Evaluation.Builtins.Costing

Documentation

eqCostStream ∷ CostStream → CostStream → Bool

Direct equality of two CostStreams. Same as deriving stock Eq. We don't want to do the latter, because the semantics of a CostStream are those of the sum of its elements and the derived Eq instance would conflict with that.

fromCostList ∷ NonEmptyList CostingInteger → CostStream

toCostList ∷ CostStream → NonEmptyList CostingInteger

toExBudgetList ∷ ExBudgetStream → NonEmptyList ExBudget

magnitudes ∷ [(SatInt, SatInt)]

A list of ranges: (0, 0), (1, 10) : (11, 100) : (101, 1000) : ... : [(10^18, maxBound)].

toRange ∷ SatInt → (SatInt, SatInt)

Return the range (in the sense of magnitudes) in which the given SatInt belongs. E.g.

>>> toRange 42
(11,100)
>>> toRange 1234
(1001,10000)

chooseSatInt ∷ (SatInt, SatInt) → Gen SatInt

Generate a SatInt in the given range.

checkEqualsVia ∷ Show a ⇒ (a → a → Bool) → a → a → Property

Same as (===) except accepts a custom equality checking function.

bottom ∷ a

A value to use in tests to make sure what's not supposed to be forced isn't forced.

test_magnitudes ∷ TestTree

Test that magnitudes has the correct bounds.

test_CostStreamDistribution ∷ TestTree

Show the distribution of generated CostStreams as a diagnostic.

test_toCostListRoundtrip ∷ TestTree

Test that fromCostList . toCostList is identity.

test_fromCostListRoundtrip ∷ TestTree

Test that toCostList . fromCostList is identity.

test_unconsCostRoundtrip ∷ TestTree

Test that uncurry reconsCost . unconsCost is identity.

test_sumCostStreamIsSum ∷ TestTree

Test that sumCostStream returns the sum of the elements of a CostStream.

test_mapCostStreamIsMap ∷ TestTree

Test that mapCostStream applies a function to each element of a CostStream.

test_addCostStreamIsAdd ∷ TestTree

Test that the sum of a stream returned by addCostStream equals the sum of the sums of its two arguments.

test_minCostStreamIsMin ∷ TestTree

Test that the sum of a stream returned by minCostStream equals the minimum of the sums of its two arguments.

test_zipCostStreamIsZip ∷ TestTree

Test that the sum of a stream returned by zipCostStream equals an ExBudget containing the sums of its two arguments.

test_mapCostStreamReasonableLength ∷ TestTree

Test that mapCostStream preserves the length of the stream.

test_addCostStreamReasonableLength ∷ TestTree

Test that the length of the stream returned by addCostStream equals the sum of the lengths of its two arguments.

test_minCostStreamReasonableLength ∷ TestTree

Test that the length of the stream returned by addCostStream is

greater than or equal to the minimum of the lengths of its two arguments

smaller than or equal to the sum of the lengths of its two arguments.

test_zipCostStreamReasonableLength ∷ TestTree

Test that the length of the stream returned by zipCostStream equals the maximum of the lengths of its two arguments.

test_mapCostStreamHandlesBottom ∷ TestTree

Test that mapCostStream preserves the laziness of its argument.

test_addCostStreamHandlesBottom ∷ TestTree

Test that mapCostStream preserves the laziness of its two arguments.

test_minCostStreamHandlesBottom ∷ TestTree

Test that minCostStream preserves the laziness of its two arguments.

postAlignWith ∷ a → [a] → [a] → ([a], [a])

Pad the shortest of the given lists by appending the given element to it until the length of the result matches the length of the other list.

>>> postAlignWith 'a' "bcd" "e"
("bcd","eaa")
>>> postAlignWith 'a' "b" "cdef"
("baaa","cdef")

test_zipCostStreamHandlesBottom ∷ TestTree

Test that zipCostStream preserves the laziness of its two arguments.

sierpinskiSize ∷ Int → Int

The size sierpinskiRose of the given depth.

sierpinskiRose ∷ Int → CostRose

Return a finite balanced tree with each node (apart from the leaves) having exactly 3 children. The parameter is the depth of the tree. Named after https://en.wikipedia.org/wiki/Sierpi%C5%84ski_triangle

test_flattenCostRoseIsLinearForSierpinskiRose ∷ Int → TestTree

Traverse a sierpinskiRose of the given depth and display the total amount of elements processed. See test_flattenCostRoseIsLinear for why we do this.

test_flattenCostRoseIsLinear ∷ TestTree

Test that traversing a larger CostRose takes _linearly_ more time. The actual test can only be done with eyes unfortunately, because the tests are way too noisy for evaluation times to be reported even remotely accurately.

genCostRose ∷ NonEmptyList SatInt → Gen CostRose

Generate a CostRose from the given list by splitting the list into sublists and generating a CostRose for each of them recursively.

fromCostRose ∷ CostRose → NonEmptyList SatInt

collectListLengths ∷ CostRose → [Int]

Return the lengths of all the forests in a CostRose.

test_CostRoseListLengthsDistribution ∷ TestTree

Show the distribution of forest lengths in generated CostRose values as a diagnostic.

test_genCostRoseSound ∷ TestTree

Test that genCostRose only takes costs from its argument when generating a CostRose.

test_flattenCostRoseSound ∷ TestTree

Test that flattenCostRose returns the elements of its argument.

test_flattenCostRoseHandlesBottom ∷ TestTree

Test that flattenCostRose is lazy.

test_costsAreNeverNegative ∷ TestTree

Test that memoryUsage called over a value of a built-in type never returns a stream containing a negative cost.

test_costing ∷ TestTree

Orphan instances

Arbitrary CostStream

Instance details

Methods

arbitrary ∷ Gen CostStream

shrink ∷ CostStream → [CostStream]

Arbitrary CostRose

Instance details

Methods

arbitrary ∷ Gen CostRose

shrink ∷ CostRose → [CostRose]

Arbitrary SatInt
Generate asymptotically bigger SatInts with exponentially lower chance. This is in order to make the generator of CostStream produce streams whose sums are more or less evenly distributed across magnitudes.

Instance details

Methods

arbitrary ∷ Gen SatInt

shrink ∷ SatInt → [SatInt]

CoArbitrary SatInt

Instance details

Methods

coarbitrary ∷ SatInt → Gen b → Gen b

Function SatInt

Instance details

Methods

function ∷ (SatInt → b) → SatInt :-> b

Foldable NonEmptyList

Instance details

Methods

fold ∷ Monoid m ⇒ NonEmptyList m → m

foldMap ∷ Monoid m ⇒ (a → m) → NonEmptyList a → m

foldMap' ∷ Monoid m ⇒ (a → m) → NonEmptyList a → m

foldr ∷ (a → b → b) → b → NonEmptyList a → b

foldr' ∷ (a → b → b) → b → NonEmptyList a → b

foldl ∷ (b → a → b) → b → NonEmptyList a → b

foldl' ∷ (b → a → b) → b → NonEmptyList a → b

foldr1 ∷ (a → a → a) → NonEmptyList a → a

foldl1 ∷ (a → a → a) → NonEmptyList a → a

toList ∷ NonEmptyList a → [a]

null ∷ NonEmptyList a → Bool

length ∷ NonEmptyList a → Int

elem ∷ Eq a ⇒ a → NonEmptyList a → Bool

maximum ∷ Ord a ⇒ NonEmptyList a → a

minimum ∷ Ord a ⇒ NonEmptyList a → a

sum ∷ Num a ⇒ NonEmptyList a → a

product ∷ Num a ⇒ NonEmptyList a → a