PlutusCore.Builtin.Case

Documentation

class AnnotateCaseBuiltin uni where

Methods

annotateCaseBuiltin ∷ UniOf term ~ uni ⇒ Type TyName uni ann → [term] → Either Text [(term, [Type TyName uni ann])]

Given a tag for a built-in type and a list of branches, annotate each of the branches with its expected argument types or fail if casing on values of the built-in type isn't supported. Note: you don't need to include the resulting type of the whole case matching in the returning list here.

class CaseBuiltin uni where

Methods

caseBuiltin ∷ UniOf term ~ uni ⇒ Some (ValueOf uni) → Vector term → HeadSpine Text term (Some (ValueOf uni))

Given a constant with its type tag and a vector of branches, choose the appropriate branch or fail if the constant doesn't correspond to any of the branches (or casing on constants of this type isn't supported at all).

data CaserBuiltin uni

A data version of CaseBuiltin. we parameterize the evaluator by a CaserBuiltin so that the caller can choose whether to use the caseBuiltin method or the always failing caser (the latter is required for earlier protocol versions when we didn't support casing on builtins).

Constructors

CaserBuiltin

Fields

unCaserBuiltin ∷ !(∀ term. UniOf term ~ uni ⇒ Some (ValueOf uni) → Vector term → HeadSpine Text term (Some (ValueOf uni)))

unavailableCaserBuiltin ∷ Int → CaserBuiltin uni