PlutusCore.Builtin.Elaborate

Documentation

type ElaborateFromTo uni i j val a = HandleHole uni i j val (TypeHole a)

Specialize each Haskell type variable in a as a type representing a PLC type variable. i is a fresh id and j is a final one as in TrySpecializeAsVar, but since HandleHole can specialize multiple variables, j can be equal to i + n for any n (including 0).