PlutusCore.Builtin.HasConstant

Documentation

data BuiltinError

The type of errors that readKnown and makeKnown can return.

Constructors

BuiltinUnliftingEvaluationError UnliftingEvaluationError

BuiltinEvaluationFailure

notAConstant ∷ BuiltinError

class HasConstant term where

Ensures that term has a Constant-like constructor to lift values to and unlift values from.

Methods

asConstant ∷ term → Either BuiltinError (Some (ValueOf (UniOf term)))

Unwrap from a Constant-like constructor throwing an UnliftingError if the provided term is not a wrapped Haskell value.

fromConstant ∷ Some (ValueOf (UniOf term)) → term

Wrap a Haskell value as a term.

type HasConstantIn uni term = (UniOf term ~ uni, HasConstant term)

Ensures that term has a Constant-like constructor to lift values to and unlift values from and connects term and its uni.

fromValueOf ∷ ∀ a term. HasConstant term ⇒ UniOf term (Esc a) → a → term

Wrap a Haskell value (given its explicit type tag) as a term.

fromValue ∷ ∀ a term. (HasConstant term, UniOf term `HasTermLevel` a) ⇒ a → term

Wrap a Haskell value (provided its type is in the universe) as a term.