PlutusCore.Builtin.KnownKind

Documentation

data SingKind k where

The type of singletonized Haskell kinds representing Plutus kinds. Indexing by a Haskell kind allows us to avoid an error call in the ToKind instance of DefaultUni and not worry about proxies in the type of knownKind.

Constructors

SingType ∷ SingKind Type

SingKindArrow ∷ SingKind k → SingKind l → SingKind (k → l) infixr 5

withSingKind ∷ Kind ann → (∀ k. SingKind k → r) → r

Feed the SingKind version of the given Kind to the given continuation.

class KnownKind k where

For reifying Haskell kinds representing Plutus kinds at the term level.

Methods

knownKind ∷ SingKind k

bringKnownKind ∷ SingKind k → (KnownKind k ⇒ r) → r

Satisfy the KnownKind constraint of a continuation using the given SingKind.

withKnownKind ∷ Kind ann → (∀ k. KnownKind k ⇒ Proxy k → r) → r

class ToKind (uni ∷ Type → Type) where

For computing the Plutus kind of a built-in type. See kindOfBuiltinType.

Methods

toSingKind ∷ uni (Esc (a ∷ k)) → SingKind k

Reify the kind of a type from the universe at the term level.

demoteKind ∷ SingKind k → Kind ()

Convert a reified Haskell kind to a Plutus kind.

kindOfBuiltinType ∷ ToKind uni ⇒ uni (Esc a) → Kind ()

Compute the kind of a type from a universe.