PlutusCore.Builtin.KnownTypeAst

Documentation

data TyNameRep (kind ∷ Type)

Representation of a type variable: its name and unique and an implicit kind.

Constructors

TyNameRep Symbol Nat

data family TyVarRep (name ∷ TyNameRep kind) ∷ kind

Representation of an intrinsically-kinded type variable: a name.

data family TyAppRep (fun ∷ dom → cod) (arg ∷ dom) ∷ cod

Representation of an intrinsically-kinded type application: a function and an argument.

data family TyForallRep (name ∷ TyNameRep kind) (a ∷ Type) ∷ Type

Representation of of an intrinsically-kinded universal quantifier: a bound name and a body.

data Hole

The kind of holes.

data family RepHole x

A hole in the Rep context.

data family TypeHole a

A hole in the Type context.

type family RunHole hole where ...

Turn a hole in the GHC.Type -> GHC.Type form into one of the Hole form. This only changes the kind of the given argument. This is a way of encoding forall a. a -> Hole at the kind level, which we don't attempt to use, because GHC apparently hates polymorphism at the kind level and chokes upon encountering it.

Equations

RunHole RepHole = RepHole

RunHole TypeHole = TypeHole

type HasTermLevel uni = Includes uni

Specifies that the given type is a built-in one and its values can be embedded into a Term.

type HasTypeLevel uni x = KnownTypeAst Void uni (ElaborateBuiltin uni x)

Specifies that the given type is a built-in one and can be embedded into a Kind.

type HasTypeAndTermLevel uni x = (uni `HasTypeLevel` x, uni `HasTermLevel` x)

The product of HasTypeLevel and HasTermLevel.

mkTyBuiltin ∷ ∀ a (x ∷ a) uni ann tyname. uni `HasTypeLevel` x ⇒ ann → Type tyname uni ann

Convert a Haskell representation of a possibly 0-ary application of a built-in type to arbitrary types implementing KnownTypeAst.

type KnownBuiltinTypeAst tyname uni x = AllBuiltinArgs uni (KnownTypeAst tyname uni) x

A constraint for "a is a KnownTypeAst by means of being included in uni".

class KnownTypeAst tyname uni x where

This class allows one to convert the type-level Haskell representation of a Plutus type into the corresponding Plutus type. Associated type families are needed to help elaboration.

Depending on the universe converting a Haskell type to a Plutus team can give different results (e.g. Int can be a built-in type instead of being encoded via built-in Integer), hence this class takes a uni argument. Plus, elaboration is universe-specific too.

Minimal complete definition

Nothing

Associated Types

type IsBuiltin uni x ∷ Bool

Whether x is a built-in type.

type IsBuiltin uni x = IsBuiltin uni (ElaborateBuiltin uni x)

type ToHoles uni (hole ∷ Type → Type) x ∷ [Hole]

Return every part of the type that can be a to-be-instantiated type variable. For example, in Integer there's no such types and in (a, b) it's the two arguments (a and b) and the same applies to a -> b (to mention a type that is not built-in).

Takes a hole in the GHC.Type -> GHC.Type form (a convention originally adopted in the elaborator, perhaps not a very helpful one), which can be turned into an actual Hole via RunHole.

type ToHoles uni hole x = ToHoles uni hole (ElaborateBuiltin uni x)

type ToBinds uni (acc ∷ [Some TyNameRep]) x ∷ [Some TyNameRep]

Collect all unique variables (a variable consists of a textual name, a unique and a kind) in an accumulator and return the accumulator once a leaf is reached.

type ToBinds uni acc x = ToBinds uni acc (ElaborateBuiltin uni x)

Methods

typeAst ∷ Type tyname uni ()

The Plutus counterpart of the x type.

default typeAst ∷ KnownTypeAst tyname uni (ElaborateBuiltin uni x) ⇒ Type tyname uni ()

toTypeAst ∷ ∀ a tyname uni (x ∷ a) proxy. KnownTypeAst tyname uni x ⇒ proxy x → Type tyname uni ()

Return the Plutus counterpart of the x type.

type family Insert x xs where ...

Insert x into xs unless it's already there.

Equations

Insert x '[] = '[x]

Insert x (x : xs) = x ': xs

Insert x (y : xs) = y ': Insert x xs

type family Delete x xs where ...

Delete the first x from a list. Which is okay since we only ever put things in once.

Equations

Delete _ '[] = '[]

Delete x (x ': xs) = xs

Delete x (y ': xs) = y ': Delete x xs