PlutusCore.Builtin.Result

Documentation

data EvaluationError structural operational

The type of errors that can occur during evaluation. There are two kinds of errors:

Structural ones -- these are errors that are indicative of the _structure_ of the program being wrong. For example, a free variable was encountered during evaluation, a non-function was applied to an argument or tailList was applied to a non-list.

Operational ones -- these are errors that are indicative of the _logic_ of the program being wrong. For example, error was executed, tailList was applied to an empty list or evaluation ran out of gas.

On the chain both of these are just regular failures and we don't distinguish between them there: if a script fails, it fails, it doesn't matter what the reason was. However in the tests it does matter why the failure occurred: a structural error may indicate that the test was written incorrectly while an operational error may be entirely expected.

In other words, structural errors are "runtime type errors" and operational errors are regular runtime errors. Which means that evaluating an (erased) well-typed program should never produce a structural error, only an operational one. This creates a sort of "runtime type system" for UPLC and it would be great to stick to it and enforce in tests etc, but we currently don't.

Constructors

StructuralError !structural

OperationalError !operational

newtype UnliftingError

The error message part of an UnliftingEvaluationError.

Constructors

MkUnliftingError

Fields

unUnliftingError ∷ Text

newtype UnliftingEvaluationError

When unlifting of a PLC term into a Haskell value fails, this error is thrown.

Constructors

MkUnliftingEvaluationError

Fields

unUnliftingEvaluationError ∷ EvaluationError UnliftingError UnliftingError

data BuiltinError

The type of errors that readKnown and makeKnown can return.

Constructors

BuiltinUnliftingEvaluationError UnliftingEvaluationError

BuiltinEvaluationFailure

data BuiltinResult a

The monad that makeKnown runs in. Equivalent to ExceptT BuiltinError (Writer (DList Text)), except optimized in two ways:

everything is strict

has the BuiltinSuccess constructor that is used for returning a value with no logs attached, which is the most common case for us, so it helps a lot not to construct and deconstruct a redundant tuple

Moving from ExceptT BuiltinError (Writer (DList Text)) to this data type gave us a speedup of 8% of total evaluation time.

Logs are represented as a DList, because we don't particularly care about the efficiency of logging, since there's no logging on the chain and builtins don't emit much anyway. Otherwise we'd have to use text-builder or text-builder-linear or something of this sort.

Constructors

BuiltinSuccess a

BuiltinSuccessWithLogs (DList Text) a

BuiltinFailure (DList Text) BuiltinError

notAConstant ∷ BuiltinError

underTypeError ∷ BuiltinError

operationalUnliftingError ∷ Text → BuiltinError

Construct a prism focusing on the *EvaluationFailure part of err by taking that *EvaluationFailure and

pretty-printing and embedding it into an UnliftingError for the setter part of the prism

returning it directly for the opposite direction (there's no other way to convert an UnliftingError to an evaluation failure, since the latter doesn't carry any content)

This is useful for providing AsUnliftingError instances for types such as CkUserError and CekUserError.

structuralUnliftingError ∷ Text → BuiltinError

emit ∷ Text → BuiltinResult ()

Add a log line to the logs.

withLogs ∷ DList Text → BuiltinResult a → BuiltinResult a

Prepend logs to a BuiltinResult computation.

throwing ∷ MonadError e m ⇒ AReview e t → t → m x

throwing_ ∷ MonadError e m ⇒ AReview e () → m x

builtinResultFailure ∷ BuiltinResult a