PlutusCore.Builtin.TestKnown

Description

A module providing a tool for making sure that each type from a particular universe implements KnownTypeAst, ReadKnown and MakeKnown. Kept separate from the modules defining those classes in order not to introduce dependencies between them and because this extra safety is scary looking and it's best to keep it separate from what is necessary.

Documentation

class (uni `Everywhere` ImplementedKnownTypeAst uni, uni `Everywhere` ImplementedReadKnownIn uni, uni `Everywhere` ImplementedMakeKnownIn uni) ⇒ TestTypesFromTheUniverseAreAllKnown uni

An instance of this class not having any constraints ensures that every type (according to Everywhere) from the universe has 'KnownTypeAst, ReadKnownIn and MakeKnownIn instances.