PlutusCore.Check.Normal

Description

This module makes sure types are normalized inside programs.

Documentation

checkProgram ∷ (HasUniApply uni, MonadError (NormCheckError tyname name uni fun ann) m) ⇒ Program tyname name uni fun ann → m ()

Ensure that all types in the Program are normalized.

checkTerm ∷ (HasUniApply uni, MonadError (NormCheckError tyname name uni fun ann) m) ⇒ Term tyname name uni fun ann → m ()

Ensure that all types in the Term are normalized.

isNormalType ∷ HasUniApply uni ⇒ Type tyname uni ann → Bool

data NormCheckError tyname name uni fun ann

Constructors

BadType !ann !(Type tyname uni ann) !Text

BadTerm !ann !(Term tyname name uni fun ann) !Text