PlutusCore.Check.Scoping

Documentation

data ScopedName

Constructors

TypeName TyName

TermName Name

isSameScope ∷ ScopedName → ScopedName → Bool

data Stays

Staying names.

Constructors

StaysOutOfScopeVariable
An out-of-scope variable does not get renamed and hence stays.

StaysFreeVariable
A free variable does not get renamed and hence stays.

data Disappears

Changing names.

Constructors

DisappearsBinding
A binding gets renamed and hence the name that it binds disappears.

DisappearsVariable
A bound variable gets renamed and hence its name disappears.

data NameAction

A name either stays or disappears.

Constructors

Stays Stays

Disappears Disappears

data NameAnn

Constructors

NameAction NameAction ScopedName

NotAName

class ToScopedName name where

Methods

toScopedName ∷ name → ScopedName

introduceBound ∷ ToScopedName name ⇒ name → NameAnn

Annotation for a binding saying "supposed to disappear".

registerBound ∷ ToScopedName name ⇒ name → NameAnn

Annotation for a bound variable saying "supposed to disappear".

registerOutOfScope ∷ ToScopedName name ⇒ name → NameAnn

Annotation for an out-of-scope variable saying "supposed to stay out of scope".

registerFree ∷ ToScopedName name ⇒ name → NameAnn

Annotation for a free variable saying "supposed to stay free".

class Reference n t where

Methods

referenceVia ∷ (∀ name. ToScopedName name ⇒ name → NameAnn) → n → t NameAnn → t NameAnn

Take a registering function, apply it to the provided name, create a type/term variable out of the resulting annotation and the original name and reference that variable in the provided type/term by prepending a constructor to it mentioning the variable.

referenceBound ∷ Reference n t ⇒ n → t NameAnn → t NameAnn

Reference the provided variable in the provided type/term as an in-scope one.

referenceOutOfScope ∷ Reference n t ⇒ n → t NameAnn → t NameAnn

Reference the provided variable in the provided type/term as an out-of-scope one.

data ScopeEntry

Each kind of old and new names.

Constructors

DisappearedBindings

DisappearedVariables

AppearedBindings

AppearedVariables

StayedOutOfScopeVariables

StayedFreeVariables

newtype ScopeInfo

A ScopeInfo is a set of ScopedNames for each of the ScopeEntry. If a ScopeEntry is not present in the map, the corresponding set of ScopeNames is considered to be empty.

Constructors

ScopeInfo (Map ScopeEntry (Set ScopedName))

unScopeInfo ∷ ScopeInfo → Map ScopeEntry (Set ScopedName)

to ∷ ScopeEntry → ScopeInfo → Set ScopedName

Extract the set stored in the provided ScopeInfo at the provided ScopeEntry.

emptyScopeInfo ∷ ScopeInfo

checkEmptyOn ∷ (Set ScopedName → Set ScopedName → Set ScopedName) → (Set ScopedName → Set ScopedName → ScopeError) → Set ScopedName → Set ScopedName → Either ScopeError ()

Check if a set is empty and report an error with the set embedded in it otherwise.

mergeScopeInfo ∷ ScopeInfo → ScopeInfo → Either ScopeError ScopeInfo

Merge two ScopeInfos checking that binders in them do not intersect along the way.

newtype ScopeErrorOrInfo

Constructors

ScopeErrorOrInfo (Either ScopeError ScopeInfo)

unScopeErrorOrInfo ∷ ScopeErrorOrInfo → Either ScopeError ScopeInfo

data BindingRemoval

Whether it's OK if the pass removes bindings. A renamer isn't supposed to do that, but for example an inliner may do it, since it's basically the entire point of an inliner.

Constructors

BindingRemovalOk

BindingRemovalNotOk

class EstablishScoping t where

Methods

establishScoping ∷ t ann → Quote (t NameAnn)

Traverse a t freshening every name (both at the binding and the use sites) and annotating the freshened names with either DisappearsBinding or StaysFreeVariable depending on whether the name occurs at the binding or the use site.

In addition to that every binder should be decorated with one out-of-scope variable (annotated with StaysOutOfScopeVariable) and one in-scope one (annotated with DisappearsVariable).

Note that no original name occurring in t should survive this procedure (and hence we don't care if any of the freshened names clashes with an original one as all original ones are supposed to be gone).

How to provide an implementation:

handle bindings with 'freshen*Name' + establishScopingBinder (or similar)

handle variables with 'freshen*Name' + registerFree

everything else is direct recursion + Applicative stuff

class CollectScopeInfo t where

Methods

collectScopeInfo ∷ t NameAnn → ScopeErrorOrInfo

Collect scoping information after scoping was established and renaming was performed.

How to provide an implementation:

handle names (both bindings and variables) with handleSname

everything else is direct recursion + Monoid stuff

type Scoping t = (EstablishScoping t, CollectScopeInfo t)

establishScopingBinder ∷ (Reference name value, ToScopedName name, EstablishScoping sort, EstablishScoping value) ⇒ (NameAnn → name → sort NameAnn → value NameAnn → value NameAnn) → name → sort ann → value ann → Quote (value NameAnn)

Take a constructor for a binder, a name bound by it, a sort (kind/type), a value of that sort (type/term) and call establishScoping on both the sort and its value and reassemble the original binder with the annotated sort and its value, but also decorate the reassembled binder with one out-of-scope variable and one in-scope one.

data ScopeError

Every kind of error thrown by the scope checking machinery at different stages.

Constructors

UnannotatedName !ScopedName

NameChangedItsScope

Fields

_oldName ∷ !ScopedName

_newName ∷ !ScopedName

NameUnexpectedlyDisappeared

Fields

_oldName ∷ !ScopedName

_newName ∷ !ScopedName

NameUnexpectedlyStayed !ScopedName

DuplicateBindersInTheInput

Fields

_duplicateBindersLeft ∷ !(Set ScopedName)

_duplicateBindersRight ∷ !(Set ScopedName)

DuplicateBindersInTheOutput !(Set ScopedName) !(Set ScopedName)

DisappearedBindingsDiscordWithBoundVariables

Fields

_disappearedBindings ∷ !(Set ScopedName)

_boundVariables ∷ !(Set ScopedName)

DisappearedBindingsDiscordWithOutOfScopeVariables

Fields

_disappearedBindings ∷ !(Set ScopedName)

_outOfScopeVariables ∷ !(Set ScopedName)

AppearedBindingsDiscordWithBoundVariables

Fields

_appearedBindings ∷ !(Set ScopedName)

_boundVariables ∷ !(Set ScopedName)

DisappearedBindingsClashWithFreeVariables

Fields

_disappearedBindings ∷ !(Set ScopedName)

_freeVariables ∷ !(Set ScopedName)

DisappearedBindingsClashWithAppearedBindings

Fields

_disppearedBindings ∷ !(Set ScopedName)

_appearedBindings ∷ !(Set ScopedName)

AppearedBindingsClashWithFreeVariabes

Fields

_appearedBindings ∷ !(Set ScopedName)

_freeVariables ∷ !(Set ScopedName)

overrideSname ∷ ScopeEntry → ScopedName → ScopeInfo → ScopeInfo

Override the set at the provided ScopeEntry to contain only the provided ScopedName.

applyStays ∷ Stays → ScopedName → ScopeInfo

Use a Stays to handle an unchanged old name.

applyDisappears ∷ Disappears → ScopedName → ScopedName → ScopeInfo

Use a Disappears to handle differing old and new names.

applyNameAction ∷ NameAction → ScopedName → ScopedName → Either ScopeError ScopeInfo

Use a NameAction to handle an old and a new name.

handleSname ∷ ToScopedName name ⇒ NameAnn → name → ScopeErrorOrInfo

Use a NameAnn to handle a new name.

symmetricDifference ∷ Ord a ⇒ Set a → Set a → Set a

checkScopeInfo ∷ BindingRemoval → ScopeInfo → Either ScopeError ()

Check that each kind of Set from ScopeInfo relates to all other ones in a certain way. We start with these three relations that are based on the assumption that for each binder we add at least one out-of-scope variable and at least one in-scope one:

disappeared bindings should be the same as stayed out of scope variables (ensures that old bindings disappear via renaming and not via removal)

disappeared bindings should be the same as disappeared variables (ensures that old names consistently disappear at the binding and use sites)

appeared bindings should be the same as appeared variables (ensures that new names consistently appear at the binding and use sites)

Once we've ensured all of that, we're left with only three sets and 3C2 equals 3, so we only need to consider three more relations:

disappeared bindings should not intersect with free variables (an internal sanity check)

appeared bindings should not intersect with disappeared bindings

appeared bindings should not intersect with free variables

The last two ensure that no new name has an old name's unique.

data ScopeCheckError t

The type of errors that the scope checking machinery returns.

Constructors

ScopeCheckError

Fields

_input ∷ !(t NameAnn)

What got fed to the scoping check pass before preparation.

_prepared ∷ !(t NameAnn)

What got fed to the scoping check pass after preparation.

_output ∷ !(t NameAnn)

What got out of it.

_error ∷ !ScopeError

The error returned by the scoping check pass.

checkRespectsScoping

Arguments

∷ Scoping t

⇒ BindingRemoval

→ (t NameAnn → t NameAnn)
For preparation before running the scoping tests. Commonly, either runQuote . rename or id.

→ (t NameAnn → t NameAnn)
The runner of the pass.

→ t ann

→ Either (ScopeCheckError t) ()

Check if a pass respects scoping.

Returns the thing that the scoping tests run on, the result of the pass and the scope checking outcome, respectively.