PlutusCore.Compiler.Erase

Documentation

eraseTerm ∷ HasUnique name TermUnique ⇒ Term tyname name uni fun ann → Term name uni fun ann

Erase a Typed Plutus Core term to its untyped counterpart.

Restricted to Plc terms with Names, because erasing a (Named-)Debruijn term will mess up its debruijn indexing and thus break scope-checking. -- FIXME: Lift this restriction of eraseTerm for (Named-)DeBruijn terms.

eraseProgram ∷ HasUnique name TermUnique ⇒ Program tyname name uni fun ann → Program name uni fun ann