PlutusCore.Core.Instance.Pretty.Default

Description

While the flexible pretty-printing infrastructure is useful when you want it, it's helpful to have an implementation of the default Pretty typeclass that does the default thing.

Orphan instances

Pretty Name

Instance details

Methods

pretty ∷ Name → Doc ann

prettyList ∷ [Name] → Doc ann

Pretty TyName

Instance details

Methods

pretty ∷ TyName → Doc ann

prettyList ∷ [TyName] → Doc ann

Pretty ann ⇒ Pretty (Kind ann)

Instance details

Methods

pretty ∷ Kind ann → Doc ann0

prettyList ∷ [Kind ann] → Doc ann0

(PrettyClassic tyname, PrettyParens (SomeTypeIn uni), Pretty ann) ⇒ Pretty (Type tyname uni ann)

Instance details

Methods

pretty ∷ Type tyname uni ann → Doc ann0

prettyList ∷ [Type tyname uni ann] → Doc ann0

(PrettyClassic tyname, PrettyClassic name, PrettyUni uni, Pretty fun, Pretty ann) ⇒ Pretty (Program tyname name uni fun ann)

Instance details

Methods

pretty ∷ Program tyname name uni fun ann → Doc ann0

prettyList ∷ [Program tyname name uni fun ann] → Doc ann0

(PrettyClassic tyname, PrettyClassic name, PrettyUni uni, Pretty fun, Pretty ann) ⇒ Pretty (Term tyname name uni fun ann)

Instance details

Methods

pretty ∷ Term tyname name uni fun ann → Doc ann0

prettyList ∷ [Term tyname name uni fun ann] → Doc ann0