PlutusCore.Core.Type

Documentation

data Kind ann

Constructors

Type ann

KindArrow ann (Kind ann) (Kind ann)

toPatFuncKind ∷ Kind () → Kind ()

The kind of a pattern functor (the first Kind argument of TyIFix) at a given kind (of the second Kind argument of TyIFix):

toPatFuncKind k = (k -> *) -> k -> *

fromPatFuncKind ∷ Kind () → Maybe (Kind ())

argsFunKind ∷ Kind ann → [Kind ann]

Extract all a_i from a_0 -> a_1 -> ... -> r.

data Type tyname uni ann

A Kind assigned to expressions.

Constructors

TyVar ann tyname
Type variable

TyFun ann (Type tyname uni ann) (Type tyname uni ann)
Function type

TyIFix ann (Type tyname uni ann) (Type tyname uni ann)
Fix-point type, for constructing self-recursive types

TyForall ann tyname (Kind ann) (Type tyname uni ann)
Polymorphic type

TyBuiltin ann (SomeTypeIn uni)
Builtin type

TyLam ann tyname (Kind ann) (Type tyname uni ann)
Type lambda

TyApp ann (Type tyname uni ann) (Type tyname uni ann)
Type application

TySOP ann [[Type tyname uni ann]]
Sum-of-products type

splitFunTyParts ∷ Type tyname uni a → NonEmpty (Type tyname uni a)

Get recursively all the domains and codomains of a type. splitFunTyParts (A->B->C) = [A, B, C] splitFunTyParts (X) = [X]

funTyArgs ∷ Type tyname uni a → [Type tyname uni a]

Get the argument types of a function type. funTyArgs (A->B->C) = [A, B]

funTyResultType ∷ Type tyname uni a → Type tyname uni a

Get the result type of a function. If not a function, then is the same as id funResultType (A->B->C) = C funResultType (X) = X

data Term tyname name uni fun ann

Constructors

Var ann name
a named variable

LamAbs ann name (Type tyname uni ann) (Term tyname name uni fun ann)
lambda abstraction

Apply ann (Term tyname name uni fun ann) (Term tyname name uni fun ann)
application

TyAbs ann tyname (Kind ann) (Term tyname name uni fun ann)
type abstraction

TyInst ann (Term tyname name uni fun ann) (Type tyname uni ann)
instantiation

IWrap ann (Type tyname uni ann) (Type tyname uni ann) (Term tyname name uni fun ann)
wrapping

Unwrap ann (Term tyname name uni fun ann)
unwrapping See Note [Constr tag type]

Constr ann (Type tyname uni ann) Word64 [Term tyname name uni fun ann]
constructor

Case ann (Type tyname uni ann) (Term tyname name uni fun ann) [Term tyname name uni fun ann]
case

Constant ann (Some (ValueOf uni))
constants

Builtin ann fun
builtin functions

Error ann (Type tyname uni ann)
fail with error

data Program tyname name uni fun ann

A Program is simply a Term coupled with a Version of the core language.

Constructors

Program

Fields

_progAnn ∷ ann

_progVer ∷ Version

_progTerm ∷ Term tyname name uni fun ann

type HasTermLevel uni = Includes uni

Specifies that the given type is a built-in one and its values can be embedded into a Term.

type family UniOf a ∷ Type → Type

Extract the universe from a type.

newtype Normalized a

Constructors

Normalized

Fields

unNormalized ∷ a

data TyVarDecl tyname ann

A "type variable declaration", i.e. a name and a kind for a type variable.

Constructors

TyVarDecl

Fields

_tyVarDeclAnn ∷ ann

_tyVarDeclName ∷ tyname

_tyVarDeclKind ∷ Kind ann

data VarDecl tyname name uni ann

A "variable declaration", i.e. a name and a type for a variable.

Constructors

VarDecl

Fields

_varDeclAnn ∷ ann

_varDeclName ∷ name

_varDeclType ∷ Type tyname uni ann

data TyDecl tyname uni ann

A "type declaration", i.e. a kind for a type.

Constructors

TyDecl

Fields

_tyDeclAnn ∷ ann

_tyDeclType ∷ Type tyname uni ann

_tyDeclKind ∷ Kind ann

tyDeclVar ∷ TyVarDecl tyname ann → TyDecl tyname uni ann

type family HasUniques a ∷ Constraint

All kinds of uniques an entity contains.

newtype Binder name

This is a wrapper to mark the place where the binder is introduced (i.e. LamAbs/TyAbs) and not where it is actually used (TyVar/Var..). This marking allows us to skip the (de)serialization of binders at LamAbs/TyAbs positions iff name is DeBruijn-encoded (level or index). See for example the instance of Flat

Constructors

Binder

Fields

unBinder ∷ name

module PlutusCore.Version

Helper functions

termAnn ∷ Term tyname name uni fun ann → ann

typeAnn ∷ Type tyname uni ann → ann

mapFun ∷ (fun → fun') → Term tyname name uni fun ann → Term tyname name uni fun' ann

Map a function over the set of built-in functions.

tyVarDeclAnn ∷ ∀ tyname ann. Lens' (TyVarDecl tyname ann) ann

tyVarDeclName ∷ ∀ tyname ann tyname. Lens (TyVarDecl tyname ann) (TyVarDecl tyname ann) tyname tyname

tyVarDeclKind ∷ ∀ tyname ann. Lens' (TyVarDecl tyname ann) (Kind ann)

varDeclAnn ∷ ∀ tyname name uni ann. Lens' (VarDecl tyname name uni ann) ann

varDeclName ∷ ∀ tyname name uni ann name. Lens (VarDecl tyname name uni ann) (VarDecl tyname name uni ann) name name

varDeclType ∷ ∀ tyname name uni ann tyname uni. Lens (VarDecl tyname name uni ann) (VarDecl tyname name uni ann) (Type tyname uni ann) (Type tyname uni ann)

tyDeclAnn ∷ ∀ tyname uni ann. Lens' (TyDecl tyname uni ann) ann

tyDeclType ∷ ∀ tyname uni ann tyname uni. Lens (TyDecl tyname uni ann) (TyDecl tyname uni ann) (Type tyname uni ann) (Type tyname uni ann)

tyDeclKind ∷ ∀ tyname uni ann. Lens' (TyDecl tyname uni ann) (Kind ann)

progAnn ∷ ∀ tyname name uni fun ann. Lens' (Program tyname name uni fun ann) ann

progVer ∷ ∀ tyname name uni fun ann. Lens' (Program tyname name uni fun ann) Version

progTerm ∷ ∀ tyname name uni fun ann tyname name uni fun. Lens (Program tyname name uni fun ann) (Program tyname name uni fun ann) (Term tyname name uni fun ann) (Term tyname name uni fun ann)