PlutusCore.Crypto.BLS12_381.Pairing

Documentation

newtype MlResult

This type represents the result of computing a pairing using the Miller loop. Values of this type are ephemeral, only created during script execution. We do not provide any means of serialising, deserialising, printing, or parsing MlResult values.

Constructors

MlResult

Fields

unMlResult ∷ PT

millerLoop ∷ Element → Element → MlResult

mulMlResult ∷ MlResult → MlResult → MlResult

finalVerify ∷ MlResult → MlResult → Bool

mlResultMemSizeBytes ∷ Int

Memory usage of an MlResult point (576 bytes)

identityMlResult ∷ MlResult

For some of the tests we need a small element of the MlResult type. We can get the identity element by pairing the zero elements of G1 and G2.