PlutusCore.DataFilePaths

Description

Various file paths used in plutus-core, currently all to do with the cost model.

Documentation

costModelDataDir ∷ FilePath

builtinCostModelFileA ∷ FilePath

builtinCostModelFileB ∷ FilePath

builtinCostModelFileC ∷ FilePath

latestBuiltinCostModelFile ∷ FilePath

cekMachineCostsFileA ∷ FilePath

cekMachineCostsFileB ∷ FilePath

cekMachineCostsFileC ∷ FilePath

latestMachineCostsFile ∷ FilePath

rModelFile ∷ FilePath

The file containing the R models: only needed for cost-model-test.

benchingResultsFile ∷ FilePath

The file containing the benchmark results for the built-in functions: only needed for cost-model-test.