PlutusCore.Default.Builtins

Documentation

data DefaultFun

Default built-in functions.

When updating these, make sure to add them to the protocol version listing! See Note [New builtins/language versions and protocol versions]

Constructors

AddInteger

SubtractInteger

MultiplyInteger

DivideInteger

QuotientInteger

RemainderInteger

ModInteger

EqualsInteger

LessThanInteger

LessThanEqualsInteger

AppendByteString

ConsByteString

SliceByteString

LengthOfByteString

IndexByteString

EqualsByteString

LessThanByteString

LessThanEqualsByteString

Sha2_256

Sha3_256

Blake2b_256

VerifyEd25519Signature

VerifyEcdsaSecp256k1Signature

VerifySchnorrSecp256k1Signature

AppendString

EqualsString

EncodeUtf8

DecodeUtf8

IfThenElse

ChooseUnit

Trace

FstPair

SndPair

ChooseList

MkCons

HeadList

TailList

NullList

ChooseData

ConstrData

MapData

ListData

IData

BData

UnConstrData

UnMapData

UnListData

UnIData

UnBData

EqualsData

SerialiseData

MkPairData

MkNilData

MkNilPairData

Bls12_381_G1_add

Bls12_381_G1_neg

Bls12_381_G1_scalarMul

Bls12_381_G1_equal

Bls12_381_G1_hashToGroup

Bls12_381_G1_compress

Bls12_381_G1_uncompress

Bls12_381_G2_add

Bls12_381_G2_neg

Bls12_381_G2_scalarMul

Bls12_381_G2_equal

Bls12_381_G2_hashToGroup

Bls12_381_G2_compress

Bls12_381_G2_uncompress

Bls12_381_millerLoop

Bls12_381_mulMlResult

Bls12_381_finalVerify

Keccak_256

Blake2b_224

IntegerToByteString

ByteStringToInteger

AndByteString

OrByteString

XorByteString

ComplementByteString

ReadBit

WriteBits

ReplicateByte

ShiftByteString

RotateByteString

CountSetBits

FindFirstSetBit

Ripemd_160

ExpModInteger

DropList

LengthOfArray

ListToArray

IndexArray

nonZeroSecondArg ∷ (Integer → Integer → Integer) → Integer → Integer → BuiltinResult Integer

Turn a function into another function that fails when its second argument is 0 or calls the original function otherwise and wraps the result in pure. Useful for correctly handling div, mod, etc.

eitherToBuiltinResult ∷ Show e ⇒ Either e r → BuiltinResult r

Turn a function returning Either into another function that fails in the Left case and wraps the result in pure in the Right case.

builtinTagWidth ∷ NumBits

Using 7 bits to encode builtin tags.

encodeBuiltin ∷ Word8 → Encoding

decodeBuiltin ∷ Get Word8