PlutusCore.Eq

Description

An internal module that defines functions for deciding equality of values of data types that encode things with binders.

Documentation

newtype LR a

From left to right.

Constructors

LR

Fields

unLR ∷ a

newtype RL a

From right to left.

Constructors

RL

Fields

unRL ∷ a

type EqRename ren = RenameT (Bilateral ren) Maybe ()

The type of a runnable equality check. Maybe () is isomorphic to Bool and we use it instead of Bool, because this unlocks the convenient and readable do-notation and allows for automatic short-circuiting, which would be tedious with Rename (Bilateral ren) Bool.

type ScopedEqRename = EqRename ScopedRenaming

runEqRename ∷ Monoid ren ⇒ EqRename ren → Bool

Run an EqRename computation.

withTwinBindings ∷ (HasRenaming ren unique, HasUnique name unique, Monad m) ⇒ name → name → RenameT (Bilateral ren) m c → RenameT (Bilateral ren) m c

Record that two names map to each other.

eqNameM ∷ (HasRenaming ren unique, HasUnique name unique, Eq unique) ⇒ name → name → EqRename ren

Check equality of two names.

eqM ∷ Eq a ⇒ a → a → EqRename ren

Check equality of things having an Eq instance.