PlutusCore.Error

Documentation

data ParserError

An error encountered during parsing.

Constructors

BuiltinTypeNotAStar !Text !SourcePos

UnknownBuiltinFunction !Text !SourcePos ![Text]

InvalidBuiltinConstant !Text !Text !SourcePos

class AsParserErrorBundle r where

Minimal complete definition

_ParserErrorBundle

Methods

_ParserErrorBundle ∷ Prism' r ParserErrorBundle

_ParseErrorB ∷ Prism' r (ParseErrorBundle Text ParserError)

data ParserErrorBundle

Constructors

ParseErrorB !(ParseErrorBundle Text ParserError)

data NormCheckError tyname name uni fun ann

Constructors

BadType !ann !(Type tyname uni ann) !Text

BadTerm !ann !(Term tyname name uni fun ann) !Text

class AsNormCheckError r tyname name uni fun ann | r → tyname name uni fun ann where

Minimal complete definition

_NormCheckError

Methods

_NormCheckError ∷ Prism' r (NormCheckError tyname name uni fun ann)

_BadType ∷ Prism' r (ann, Type tyname uni ann, Text)

_BadTerm ∷ Prism' r (ann, Term tyname name uni fun ann, Text)

data UniqueError ann

Constructors

MultiplyDefined !Unique !ann !ann

IncoherentUsage !Unique !ann !ann

FreeVariable !Unique !ann

class AsUniqueError r ann | r → ann where

Minimal complete definition

_UniqueError

Methods

_UniqueError ∷ Prism' r (UniqueError ann)

_MultiplyDefined ∷ Prism' r (Unique, ann, ann)

_IncoherentUsage ∷ Prism' r (Unique, ann, ann)

_FreeVariable ∷ Prism' r (Unique, ann)

data ExpectedShapeOr a

This is needed for nice kind/type checking error messages. In some cases the type checker knows the exact type that an expression has to have for type checking to succeed (see any of checkTypeM functions and its usages), which is what ExpectedExact is suitable for. In other cases the type checker only cares about the shape of the inferred type, e.g. the type checker knows that the type of a function has to be dom -> cod for type checking to succeed, but it doesn't yet care what dom and cod exactly are. Which is what ExpectedShape is useful for as it allows one to specify the shape of an expected type with some existential variables in it when it's impossible to provide an exact type.

Constructors

ExpectedShape

Fields

!Text

The expected shape potentially referencing existential variables.

![Text]

The list of existential variables.

ExpectedExact !a

data TypeError term uni fun ann

Constructors

KindMismatch

Fields

!ann

!(Type TyName uni ())

!(ExpectedShapeOr (Kind ()))

The expected type or the shape of a kind.

!(Kind ())

The actual kind.

TypeMismatch

Fields

!ann

!term

!(ExpectedShapeOr (Type TyName uni ()))

The expected type or the shape of a type.

!(Normalized (Type TyName uni ()))

The actual type.

TyNameMismatch !ann !TyName !TyName

NameMismatch !ann !Name !Name

FreeTypeVariableE !ann !TyName

FreeVariableE !ann !Name

UnknownBuiltinFunctionE !ann !fun

class AsTypeError r term uni fun ann | r → term uni fun ann where

Minimal complete definition

_TypeError

Methods

_TypeError ∷ Prism' r (TypeError term uni fun ann)

_KindMismatch ∷ Prism' r (ann, Type TyName uni (), ExpectedShapeOr (Kind ()), Kind ())

_TypeMismatch ∷ Prism' r (ann, term, ExpectedShapeOr (Type TyName uni ()), Normalized (Type TyName uni ()))

_TyNameMismatch ∷ Prism' r (ann, TyName, TyName)

_NameMismatch ∷ Prism' r (ann, Name, Name)

_FreeTypeVariableE ∷ Prism' r (ann, TyName)

_FreeVariableE ∷ Prism' r (ann, Name)

_UnknownBuiltinFunctionE ∷ Prism' r (ann, fun)

data FreeVariableError

We cannot do a correct translation to or from de Bruijn indices if the program is not well-scoped. So we throw an error in such a case.

Constructors

FreeUnique !Unique

FreeIndex !Index

class AsFreeVariableError r where

Minimal complete definition

_FreeVariableError

Methods

_FreeVariableError ∷ Prism' r FreeVariableError

_FreeUnique ∷ Prism' r Unique

_FreeIndex ∷ Prism' r Index

data Error uni fun ann

Constructors

ParseErrorE !ParserErrorBundle

UniqueCoherencyErrorE !(UniqueError ann)

TypeErrorE !(TypeError (Term TyName Name uni fun ()) uni fun ann)

NormCheckErrorE !(NormCheckError TyName Name uni fun ann)

FreeVariableErrorE !FreeVariableError

class AsError r uni fun ann | r → uni fun ann where

Minimal complete definition

_Error

Methods

_Error ∷ Prism' r (Error uni fun ann)

_ParseErrorE ∷ Prism' r ParserErrorBundle

_UniqueCoherencyErrorE ∷ Prism' r (UniqueError ann)

_TypeErrorE ∷ Prism' r (TypeError (Term TyName Name uni fun ()) uni fun ann)

_NormCheckErrorE ∷ Prism' r (NormCheckError TyName Name uni fun ann)

_FreeVariableErrorE ∷ Prism' r FreeVariableError

throwingEither ∷ MonadError e m ⇒ AReview e t → Either t a → m a

Lifts an Either into an error context where we can embed the Left value into the error.

class Ord a ⇒ ShowErrorComponent a where

The type class defines how to print a custom component of ParseError.

Since: megaparsec-5.0.0

Minimal complete definition

showErrorComponent

Methods

showErrorComponent ∷ a → String

Pretty-print a component of ParseError.

errorComponentLen ∷ a → Int

Length of the error component in characters, used for highlighting of parse errors in input string.

Since: megaparsec-7.0.0

data ApplyProgramError

Errors from applyProgram for PIR, PLC, UPLC.

Constructors

MkApplyProgramError Version Version

Orphan instances

Pretty SourcePos

Instance details

Methods

pretty ∷ SourcePos → Doc ann

prettyList ∷ [SourcePos] → Doc ann