PlutusCore.Evaluation.Machine.Ck

Description

The CK machine.

Documentation

data EvaluationResult a

The parameterized type of results various evaluation engines return. On the PLC side this becomes (via makeKnown) either a call to Error or a value of the PLC counterpart of type a.

Constructors

EvaluationSuccess !a

EvaluationFailure

type CkEvaluationException uni fun = EvaluationException (MachineError fun) CkUserError (Term TyName Name uni fun ())

The CK machine-specific EvaluationException.

type CkM uni fun s = ReaderT (CkEnv uni fun s) (ExceptT (CkEvaluationException uni fun) (ST s))

data CkValue uni fun

runCk ∷ BuiltinsRuntime fun (CkValue uni fun) → CaserBuiltin uni → Bool → Term TyName Name uni fun () → (Either (CkEvaluationException uni fun) (Term TyName Name uni fun ()), [Text])

splitStructuralOperational ∷ Either (EvaluationException structural operational term) a → Either (ErrorWithCause structural term) (EvaluationResult a)

Preserve the contents of an StructuralError as a Left and turn an OperationalError into a Right EvaluationFailure (thus erasing the content of the error in the latter case).

unsafeSplitStructuralOperational ∷ (PrettyPlc structural, PrettyPlc term, Typeable structural, Typeable term) ⇒ Either (EvaluationException structural operational term) a → EvaluationResult a

Throw on a StructuralError and turn an OperationalError into an EvaluationFailure (thus erasing the content of the error in the latter case).

evaluateCk ∷ BuiltinsRuntime fun (CkValue uni fun) → CaserBuiltin uni → Term TyName Name uni fun () → (Either (CkEvaluationException uni fun) (Term TyName Name uni fun ()), [Text])

Evaluate a term using the CK machine with logging enabled.

evaluateCkNoEmit ∷ BuiltinsRuntime fun (CkValue uni fun) → CaserBuiltin uni → Term TyName Name uni fun () → Either (CkEvaluationException uni fun) (Term TyName Name uni fun ())

Evaluate a term using the CK machine with logging disabled.

readKnownCk ∷ ReadKnown (Term TyName Name uni fun ()) a ⇒ BuiltinsRuntime fun (CkValue uni fun) → CaserBuiltin uni → Term TyName Name uni fun () → Either (CkEvaluationException uni fun) a

Unlift a value using the CK machine.