PlutusCore.Evaluation.Machine.CostStream

Documentation

data CostStream

A lazy stream of CostingIntegers. Basically NonEmpty CostingInteger, except the elements are stored strictly.

The semantics of a stream are those of the sum of its elements. I.e. a stream that is a reordered version of another stream is considered equal to that stream.

All costs are assumed not to be negative and functions handling CostStreams may rely on this assumption. Negative costs (a.k.a. allowing the user to forge execution units at runtime) wouldn't make sense.

Constructors

CostLast !CostingInteger

CostCons !CostingInteger CostStream

unconsCost ∷ CostStream → (CostingInteger, Maybe CostStream)

Uncons an element from a CostStream and return the rest of the stream, if not empty.

reconsCost ∷ CostingInteger → Maybe CostStream → CostStream

Cons an element to a CostStream, if given any. Otherwise create a new CostStream using CostLast.

sumCostStream ∷ CostStream → CostingInteger

Add up all the costs in a CostStream.

mapCostStream ∷ (CostingInteger → CostingInteger) → CostStream → CostStream

Map a function over a CostStream.

addCostStream ∷ CostStream → CostStream → CostStream

Add two streams by interleaving their elements (as opposed to draining out one of the streams before starting to take elements from the other one). No particular reason to prefer interleaving over draining out one of the streams first.

minCostStream ∷ CostStream → CostStream → CostStream

Calculate the minimum of two CostStreams. May return a stream that is longer than either of the two (but not more than twice).