PlutusCore.Evaluation.Machine.CostingFun.SimpleJSON

Description

A JSON representation of costing functions for Plutus Core builtins which produces a simple cost model which can be used from Agda and other executables

Documentation

data LinearFunction

Constructors

LinearFunction

Fields

intercept_ ∷ Integer

slope_ ∷ Integer

data TwoVariableLinearFunction

Constructors

TwoVariableLinearFunction

Fields

intercept'_ ∷ Integer

slope1_ ∷ Integer

slope2_ ∷ Integer

data OneVariableQuadraticFunction

Constructors

OneVariableQuadraticFunction

Fields

coeff0_ ∷ Integer

coeff1_ ∷ Integer

coeff2_ ∷ Integer

data TwoVariableQuadraticFunction

Constructors

TwoVariableQuadraticFunction

Fields

minimum ∷ Integer

coeff00_ ∷ Integer

coeff10_ ∷ Integer

coeff01_ ∷ Integer

coeff20_ ∷ Integer

coeff11_ ∷ Integer

coeff02_ ∷ Integer

data ExpModCostingFunction

Constructors

ExpModCostingFunction

Fields

emcfcoeff00_ ∷ Integer

emcfcoeff11_ ∷ Integer

emcfcoeff12_ ∷ Integer

data Model

This type reflects what is actually in the JSON. The stuff in CostingFun.Core and CostingFun.JSON is much more rigid, allowing parsing only for the model types applicable to the various ModelNArguments types; it also requires entries for everything in DefaultFun. Using the type defined here allows us to be more flexible and parse stuff that's not exactly what's expected in builtinCostModel.json.

Constructors

ConstantCost Integer

AddedSizes LinearFunction

MultipliedSizes LinearFunction

MinSize LinearFunction

MaxSize LinearFunction

LinearInX LinearFunction

LinearInY LinearFunction

LinearInZ LinearFunction

LiteralInYOrLinearInZ LinearFunction

LinearInMaxYZ LinearFunction

LinearInYAndZ TwoVariableLinearFunction

QuadraticInY OneVariableQuadraticFunction

QuadraticInZ OneVariableQuadraticFunction

QuadraticInXAndY TwoVariableQuadraticFunction

SubtractedSizes LinearFunction Integer
Linear model in x-y plus minimum value for the case x-y < 0.

ConstAboveDiagonal Integer Model

ConstBelowDiagonal Integer Model

ConstOffDiagonal Integer Model

ExpModCost ExpModCostingFunction

data CpuAndMemoryModel

A CPU usage modelling function and a memory usage modelling function bundled together

Constructors

CpuAndMemoryModel

Fields

cpuModel ∷ Model

memoryModel ∷ Model