PlutusCore.Evaluation.Machine.SimpleBuiltinCostModel

Description

A program to parse a JSON representation of costing functions for Plutus Core builtins and and produce a simple cost model which can be used from Agda and other executables

Documentation

type BuiltinCostMap = [(Text, CpuAndMemoryModel)]

type BuiltinCostKeyMap = KeyMap CpuAndMemoryModel

toSimpleBuiltinCostModel ∷ BuiltinCostKeyMap → BuiltinCostMap

defaultSimpleBuiltinCostModel ∷ BuiltinCostMap