PlutusCore.Evaluation.Result

Description

This module defines a common type various evaluation machine use to return their results.

Documentation

data EvaluationResult a

The parameterized type of results various evaluation engines return. On the PLC side this becomes (via makeKnown) either a call to Error or a value of the PLC counterpart of type a.

Constructors

EvaluationSuccess !a

EvaluationFailure

isEvaluationSuccess ∷ EvaluationResult a → Bool

Check whether an EvaluationResult is an EvaluationSuccess.

isEvaluationFailure ∷ EvaluationResult a → Bool

Check whether an EvaluationResult is an EvaluationFailure.