PlutusCore.Examples.Builtins

Documentation

size ∷ ∀ a. (Bounded a, Enum a) ⇒ Int

data ExtensionFun

Constructors

Factorial

ForallFortyTwo

SumInteger

Const

Id

IdAssumeBool

IdAssumeCheckBool

IdSomeConstantBool

IdIntegerAsBool

IdFInteger

IdList

IdRank2

ScottToMetaUnit

FailingSucc

ExpensiveSucc

FailingPlus

ExpensivePlus

IsConstant

UnsafeCoerce

UnsafeCoerceEl

Undefined

Absurd

ErrorPrime

Comma

BiconstPair

Swap

SwapEls

ExtensionVersion

TrackCosts

IntNoIntegerNoWord

newtype MetaForall name a

Normally forall in the type of a Haskell function gets detected and instantiated automatically, however there's no way of doing that for a forall that binds a never referenced type variable. Which is OK, because that would be a pretty weird builtin, however it's definable and for the purpose of testing we do introduce such a builtin, hence this definition allowing us to create an all that binds a never referenced type variable in Plutus while still using makeBuiltinMeaning.

Constructors

MetaForall a

data PlcListRep (a ∷ Type)

data BuiltinErrorCall

Constructors

BuiltinErrorCall

class Whatever a where

For the most part we don't care about costing functions of example builtins, hence this class for being explicit about not caring.

Methods

whatever ∷ a

The costing function of a builtin whose costing function doesn't matter.

Orphan instances

UniOf term ~ DefaultUni ⇒ MakeKnownIn DefaultUni term Void

Instance details

Methods

makeKnown ∷ Void → BuiltinResult term

UniOf term ~ DefaultUni ⇒ ReadKnownIn DefaultUni term Void

Instance details

Methods

readKnown ∷ term → ReadKnownM Void

tyname ~ TyName ⇒ KnownTypeAst tyname DefaultUni Void

Instance details

Associated Types

type IsBuiltin DefaultUni Void ∷ Bool

type ToHoles DefaultUni hole Void ∷ [Hole]

type ToBinds DefaultUni acc Void ∷ [Some TyNameRep]

Methods

typeAst ∷ Type tyname DefaultUni ()

(ToBuiltinMeaning uni fun1, ToBuiltinMeaning uni fun2, Default (BuiltinSemanticsVariant fun1), Default (BuiltinSemanticsVariant fun2)) ⇒ ToBuiltinMeaning uni (Either fun1 fun2)

Instance details

Associated Types

type CostingPart uni (Either fun1 fun2)

data BuiltinSemanticsVariant (Either fun1 fun2)

Methods

toBuiltinMeaning ∷ HasMeaningIn uni val ⇒ BuiltinSemanticsVariant (Either fun1 fun2) → Either fun1 fun2 → BuiltinMeaning val (CostingPart uni (Either fun1 fun2))

(Default (BuiltinSemanticsVariant fun1), Default (BuiltinSemanticsVariant fun2)) ⇒ Default (BuiltinSemanticsVariant (Either fun1 fun2))

Instance details

Methods

def ∷ BuiltinSemanticsVariant (Either fun1 fun2)

(Bounded a, Bounded b) ⇒ Bounded (Either a b)

Instance details

Methods

minBound ∷ Either a b

maxBound ∷ Either a b

(Eq a, Eq b, Bounded a, Bounded b, Enum a, Enum b) ⇒ Enum (Either a b)

Instance details

Methods

succ ∷ Either a b → Either a b

pred ∷ Either a b → Either a b

toEnum ∷ Int → Either a b

fromEnum ∷ Either a b → Int

enumFrom ∷ Either a b → [Either a b]

enumFromThen ∷ Either a b → Either a b → [Either a b]

enumFromTo ∷ Either a b → Either a b → [Either a b]

enumFromThenTo ∷ Either a b → Either a b → Either a b → [Either a b]

(Bounded a, Bounded b, Ix a, Ix b) ⇒ Ix (Either a b)

Instance details

Methods

range ∷ (Either a b, Either a b) → [Either a b]

index ∷ (Either a b, Either a b) → Either a b → Int

unsafeIndex ∷ (Either a b, Either a b) → Either a b → Int

inRange ∷ (Either a b, Either a b) → Either a b → Bool

rangeSize ∷ (Either a b, Either a b) → Int

unsafeRangeSize ∷ (Either a b, Either a b) → Int