PlutusCore.Executable.AstIO

Description

Reading and writing ASTs with various name types in flat format.

Documentation

type UplcProgDB ann = Program DeBruijn DefaultUni DefaultFun ann

type UplcProgNDB ann = Program NamedDeBruijn DefaultUni DefaultFun ann

type PlcProgDB ann = Program TyDeBruijn DeBruijn DefaultUni DefaultFun ann

type PlcProgNDB ann = Program NamedTyDeBruijn NamedDeBruijn DefaultUni DefaultFun ann

type UplcTermDB ann = Term DeBruijn DefaultUni DefaultFun ann

type UplcTermNDB ann = Term NamedDeBruijn DefaultUni DefaultFun ann

type PlcTermDB ann = Term TyDeBruijn DeBruijn DefaultUni DefaultFun ann

type PlcTermNDB ann = Term NamedTyDeBruijn NamedDeBruijn DefaultUni DefaultFun ann

type PlcTypeDB ann = Type TyDeBruijn DefaultUni ann

type PlcTypeNDB ann = Type NamedTyDeBruijn DefaultUni ann

serialisePirProgramFlat ∷ Flat ann ⇒ AstNameType → PirProg ann → ByteString

serialisePlcProgramFlat ∷ Flat ann ⇒ AstNameType → PlcProg ann → ByteString

serialiseUplcProgramFlat ∷ Flat ann ⇒ AstNameType → UplcProg ann → ByteString

loadPirASTfromFlat ∷ Flat a ⇒ AstNameType → Input → IO (PirProg a)

loadPlcASTfromFlat ∷ Flat a ⇒ AstNameType → Input → IO (PlcProg a)

Read and deserialise a Flat-encoded PIR/PLC AST

loadUplcASTfromFlat ∷ Flat ann ⇒ AstNameType → Input → IO (UplcProg ann)

Read and deserialise a Flat-encoded UPLC AST

loadUplcASTfromSerialised ∷ Input → IO (UplcProg ())

loadUplcASTfromHex ∷ Input → IO (UplcProg ())

fromNamedDeBruijnUPLC ∷ UplcProgNDB ann → UplcProg ann

Convert an untyped program with named de Bruijn indices to one with textual names.

toDeBruijnTermPLC ∷ PlcTerm ann → PlcTermDB ann

Convert a typed term to one where the name type is de Bruijn indices.

toDeBruijnTermUPLC ∷ UplcTerm ann → UplcTermDB ann

Convert an untyped term to one where the name type is de Bruijn indices.

toDeBruijnUPLC ∷ UplcProg ann → UplcProgDB ann

Convert an untyped program to one where the name type is de Bruijn indices.

toDeBruijnTypePLC ∷ PlcType ann → PlcTypeDB ann

Convert a type to one where the tyname type is de Bruijn indices.

toNamedDeBruijnUPLC ∷ UplcProg ann → UplcProgNDB ann

Convert an untyped program to one where the name type is textual names with de Bruijn indices.

decodeUplcHex ∷ Text → UplcProg ()

encodeUplcHex ∷ UplcProg ann → Text