PlutusCore.Executable.Common

Documentation

module PlutusCore.Executable.Types

class PrintBudgetState cost

Minimal complete definition

printBudgetState

getInput ∷ Input → IO Text

getInteresting ∷ IO [(ExampleName, Term TyName Name DefaultUni DefaultFun ())]

getPlcExamples ∷ IO [(ExampleName, SomeExample)]

Get available typed examples.

prettyPrintByMode ∷ PrettyPlc a ⇒ PrintMode → a → Doc a

getUplcExamples ∷ IO [(ExampleName, SomeExample)]

Get available untyped examples. Currently the untyped examples are obtained by erasing typed ones, but it might be useful to have some untyped ones that can't be obtained by erasure.

helpText

Arguments

∷ String
Either "Untyped Plutus Core" or "Typed Plutus Core"

→ String

loadASTfromFlat ∷ (ProgramLike p, Flat ann) ⇒ AstNameType → Input → IO (p ann)

Read and deserialise a Flat-encoded AST

parseInput

Arguments

∷ (ProgramLike p, Rename (p SrcSpan))

⇒ Input
The source program

→ IO (Text, p SrcSpan)
The output is a program with annotation

Read and parse and check the program for UniqueError's.

parseNamedProgram ∷ ProgramLike p ⇒ String → Text → Either ParserErrorBundle (p SrcSpan)

Parse a program. The first argument (normally the file path) describes the input stream, the second is the program text.

printBudgetState ∷ PrintBudgetState cost ⇒ Term Name DefaultUni DefaultFun () → CekModel → cost → IO ()

readProgram ∷ ∀ p. (ProgramLike p, Functor p, Rename (p SrcSpan)) ⇒ Format → Input → IO (p SrcSpan)

runConvert ∷ ∀ (p ∷ Type → Type). (ProgramLike p, Functor p, Rename (p SrcSpan), PrettyBy PrettyConfigPlc (p SrcSpan)) ⇒ ConvertOptions → IO ()

Convert between textual and FLAT representations.

runDumpModel ∷ BuiltinSemanticsVariant DefaultFun → IO ()

runPrint ∷ ∀ p. (ProgramLike p, Rename (p SrcSpan), PrettyBy PrettyConfigPlc (p SrcSpan)) ⇒ PrintOptions → IO ()

runPrintBuiltinSignatures ∷ IO ()

runPrintExample ∷ IO [(ExampleName, SomeExample)] → ExampleOptions → IO ()

topSrcSpan ∷ SrcSpan

A made-up SrcSpan since there's no source locations in Flat.

writeFlat ∷ (ProgramLike p, Functor p) ⇒ Output → AstNameType → p ann → IO ()

writePrettyToOutput ∷ PrettyBy PrettyConfigPlc (p ann) ⇒ Output → PrintMode → p ann → IO ()

writeProgram ∷ (ProgramLike p, Functor p, PrettyBy PrettyConfigPlc (p ann)) ⇒ Output → Format → PrintMode → p ann → IO ()

writeToOutput ∷ Show a ⇒ Output → a → IO ()