PlutusCore.Flat.Class

Description

Generics-based generation of Flat instances

The Flat class

class Flat a where

Class of types that can be encoded/decoded

Encoding a value involves three steps:

calculate the maximum size of the serialised value, using size

preallocate a buffer of the required size

encode the value in the buffer, using encode

Minimal complete definition

Nothing

Methods

encode ∷ a → Encoding

Return the encoding corrresponding to the value

default encode ∷ (Generic a, GFlatEncode (Rep a)) ⇒ a → Encoding

decode ∷ Get a

Decode a value

default decode ∷ (Generic a, GFlatDecode (Rep a)) ⇒ Get a

size ∷ a → NumBits → NumBits

Add maximum size in bits of the value to the total count

Used to calculated maximum buffer size before encoding

default size ∷ (Generic a, GFlatSize (Rep a)) ⇒ a → NumBits → NumBits

getSize ∷ Flat a ⇒ a → NumBits

Calculate the maximum size in bits of the serialisation of the value

module GHC.Generics

class GFlatEncode f

Generic Encoder

Minimal complete definition

gencode

class GFlatDecode f

Generic Decoding

Minimal complete definition

gget

class GFlatSize f

Calculate the number of bits required for the serialisation of a value Implemented as a function that adds the maximum size to a running total

Minimal complete definition

gsize