PlutusCore.Flat.Instances.Base

Description

Flat instances for the base library

Orphan instances

Flat All
Since: 0.4.4

Instance details

Methods

encode ∷ All → Encoding

decode ∷ Get All

size ∷ All → NumBits → NumBits

Flat Any
Since: 0.4.4

Instance details

Methods

encode ∷ Any → Encoding

decode ∷ Get Any

size ∷ Any → NumBits → NumBits

Flat Int16
>>> test (0::Int16)
(True,8,"00000000")

>>> test (1::Int16)
(True,8,"00000010")

>>> test (-1::Int16)
(True,8,"00000001")

>>> test (minBound::Int16)
(True,24,"11111111 11111111 00000011")

equivalent to 0b1111111111111111

>>> test (maxBound::Int16)
(True,24,"11111110 11111111 00000011")

equivalent to 0b1111111111111110

Instance details

Methods

encode ∷ Int16 → Encoding

decode ∷ Get Int16

size ∷ Int16 → NumBits → NumBits

Flat Int32
>>> test (0::Int32)
(True,8,"00000000")

>>> test (minBound::Int32)
(True,40,"11111111 11111111 11111111 11111111 00001111")

>>> test (maxBound::Int32)
(True,40,"11111110 11111111 11111111 11111111 00001111")

Instance details

Methods

encode ∷ Int32 → Encoding

decode ∷ Get Int32

size ∷ Int32 → NumBits → NumBits

Flat Int64
>>> test (0::Int64)
(True,8,"00000000")

>>> test (minBound::Int64)
(True,80,"11111111 11111111 11111111 11111111 11111111 11111111 11111111 11111111 11111111 00000001")

>>> test (maxBound::Int64)
(True,80,"11111110 11111111 11111111 11111111 11111111 11111111 11111111 11111111 11111111 00000001")

Instance details

Methods

encode ∷ Int64 → Encoding

decode ∷ Get Int64

size ∷ Int64 → NumBits → NumBits

Flat Int8
>>> test (0::Int8)
(True,8,"00000000")

>>> test (127::Int8)
(True,8,"11111110")

>>> test (-128::Int8)
(True,8,"11111111")

Instance details

Methods

encode ∷ Int8 → Encoding

decode ∷ Get Int8

size ∷ Int8 → NumBits → NumBits

Flat Word16

Instance details

Methods

encode ∷ Word16 → Encoding

decode ∷ Get Word16

size ∷ Word16 → NumBits → NumBits

Flat Word32

Instance details

Methods

encode ∷ Word32 → Encoding

decode ∷ Get Word32

size ∷ Word32 → NumBits → NumBits

Flat Word64

Instance details

Methods

encode ∷ Word64 → Encoding

decode ∷ Get Word64

size ∷ Word64 → NumBits → NumBits

Flat Word8
Word8 always take 8 bits.

>>> test (0::Word8)
(True,8,"00000000")

>>> test (255::Word8)
(True,8,"11111111")

Instance details

Methods

encode ∷ Word8 → Encoding

decode ∷ Get Word8

size ∷ Word8 → NumBits → NumBits

Flat Integer
Integers are encoded just as the fixed size Ints.

>>> test (0::Integer)
(True,8,"00000000")

>>> test (-1::Integer)
(True,8,"00000001")

>>> test (1::Integer)
(True,8,"00000010")

>>> test (-(2^4)::Integer)
(True,8,"00011111")

>>> test (2^4::Integer)
(True,8,"00100000")

>>> test (-(2^120)::Integer)
(True,144,"11111111 11111111 11111111 11111111 11111111 11111111 11111111 11111111 11111111 11111111 11111111 11111111 11111111 11111111 11111111 11111111 11111111 00000011")

>>> test (2^120::Integer)
(True,144,"10000000 10000000 10000000 10000000 10000000 10000000 10000000 10000000 10000000 10000000 10000000 10000000 10000000 10000000 10000000 10000000 10000000 00000100")

Instance details

Methods

encode ∷ Integer → Encoding

decode ∷ Get Integer

size ∷ Integer → NumBits → NumBits

Flat Natural
Naturals are encoded just as the fixed size Words.

>>> test (0::Natural)
(True,8,"00000000")

>>> test (2^120::Natural)
(True,144,"10000000 10000000 10000000 10000000 10000000 10000000 10000000 10000000 10000000 10000000 10000000 10000000 10000000 10000000 10000000 10000000 10000000 00000010")

Instance details

Methods

encode ∷ Natural → Encoding

decode ∷ Get Natural

size ∷ Natural → NumBits → NumBits

Flat ()
(), as all data types with a single constructor, has a zero-length encoding.

>>> test ()
(True,0,"")

Instance details

Methods

encode ∷ () → Encoding

decode ∷ Get ()

size ∷ () → NumBits → NumBits

Flat Bool
One bit is plenty for a Bool.

>>> test False
(True,1,"0")

>>> test True
(True,1,"1")

Instance details

Methods

encode ∷ Bool → Encoding

decode ∷ Get Bool

size ∷ Bool → NumBits → NumBits

Flat Char
Char's are mapped to Word32 and then encoded.

For ascii characters, the encoding is standard ascii.

>>> test 'a'
(True,8,"01100001")

For unicode characters, the encoding is non standard.

>>> test 'È'
(True,16,"11001000 00000001")

>>> test '不'
(True,24,"10001101 10011100 00000001")

>>> test "\x1F600"
(True,26,"11000000 01110110 00000011 10")

Instance details

Methods

encode ∷ Char → Encoding

decode ∷ Get Char

size ∷ Char → NumBits → NumBits

Flat Double
Doubles are encoded as standard IEEE binary64 values:

IEEE_754_binary64 ≡ IEEE_754_binary64 {sign :: Sign,
                                        exponent :: MostSignificantFirst Bits11,
                                        fraction :: MostSignificantFirst Bits52}

Instance details

Methods

encode ∷ Double → Encoding

decode ∷ Get Double

size ∷ Double → NumBits → NumBits

Flat Float
Floats are encoded as standard IEEE binary32 values:

IEEE_754_binary32 ≡ IEEE_754_binary32 {sign :: Sign,
                                        exponent :: MostSignificantFirst Bits8,
                                        fraction :: MostSignificantFirst Bits23}

>>> test (0::Float)
(True,32,"00000000 00000000 00000000 00000000")

>>> test (1.4012984643E-45::Float)
(True,32,"00000000 00000000 00000000 00000001")

>>> test (1.1754942107E-38::Float)
(True,32,"00000000 01111111 11111111 11111111")

Instance details

Methods

encode ∷ Float → Encoding

decode ∷ Get Float

size ∷ Float → NumBits → NumBits

Flat Int
Integer, Int, Int16, Int32 and Int64 are defined as the ZigZag encoded version of the equivalent unsigned Word:

Int   ≡  Int   (ZigZag Word)

Int64 ≡  Int64 (ZigZag Word64)

Int32 ≡  Int32 (ZigZag Word32)

Int16 ≡  Int16 (ZigZag Word16)

Int8  ≡  Int8  (ZigZag Word8)

ZigZag a ≡ ZigZag a

ZigZag encoding alternates between positive and negative numbers, so that numbers whose absolute value is small can be encoded efficiently:

>>> test (0::Int)
(True,8,"00000000")

>>> test (-1::Int)
(True,8,"00000001")

>>> test (1::Int)
(True,8,"00000010")

>>> test (-2::Int)
(True,8,"00000011")

>>> test (2::Int)
(True,8,"00000100")

Instance details

Methods

encode ∷ Int → Encoding

decode ∷ Get Int

size ∷ Int → NumBits → NumBits

Flat Word
Natural, Word, Word16, Word32 and Word64 are encoded as a non empty list of 7 bits chunks (least significant chunk first and most significant bit first in every chunk).

Words are always encoded in a whole number of bytes, as every chunk is 8 bits long (1 bit for the List constructor, plus 7 bits for the value).

The actual definition is:

Word64 ≡   Word64 Word

Word32 ≡   Word32 Word

Word16 ≡   Word16 Word

Word ≡   Word (LeastSignificantFirst (NonEmptyList (MostSignificantFirst Word7)))

LeastSignificantFirst a ≡   LeastSignificantFirst a

NonEmptyList a ≡   Elem a
                 | Cons a (NonEmptyList a)

MostSignificantFirst a ≡   MostSignificantFirst a

Word7 ≡   V0
        | V1
        | V2
        ...
        | V127

Values between as 0 and 127 fit in a single byte.

127 (0b1111111) is represented as Elem V127 and encoded as: Elem=0 127=1111111

>>> test (127::Word)
(True,8,"01111111")

254 (0b11111110) is represented as Cons V126 (Elem V1) (254=128+126) and encoded as: Cons=1 V126=1111110 (Elem=0 V1=0000001):

>>> test (254::Word)
(True,16,"11111110 00000001")

Another example, 32768 (Ob1000000000000000 = 0000010 0000000 0000000):

>>> test (32768::Word32)
(True,24,"10000000 10000000 00000010")

As this is a variable length encoding, values are encoded in the same way, whatever their type:

>>> all (test (3::Word) ==) [test (3::Word16),test (3::Word32),test (3::Word64)]
True

Word/Int decoders return an error if the encoded value is outside their valid range:

>>> unflat @Word16 (flat @Word32 $ fromIntegral @Word16 maxBound)
Right 65535

>>> unflat @Word16 (flat @Word32 $ fromIntegral @Word16 maxBound + 1)
Left (BadEncoding ...

>>> unflat @Word32 (flat @Word64 $ fromIntegral @Word32 maxBound)
Right 4294967295

>>> unflat @Word32 (flat @Word64 $ fromIntegral @Word32 maxBound + 1)
Left (BadEncoding ...

>>> unflat @Word64 (flat @Natural $ fromIntegral @Word64 maxBound)
Right 18446744073709551615

>>> unflat @Word64 (flat @Natural $ fromIntegral @Word64 maxBound + 1)
Left (BadEncoding ...

>>> unflat @Int16 (flat @Int32 $ fromIntegral @Int16 maxBound)
Right 32767

>>> unflat @Int16 (flat @Int32 $ fromIntegral @Int16 maxBound + 1)
Left (BadEncoding ...

>>> unflat @Int32 (flat @Int64 $ fromIntegral @Int32 maxBound)
Right 2147483647

>>> unflat @Int32 (flat @Int64 $ fromIntegral @Int32 maxBound + 1)
Left (BadEncoding ...

>>> unflat @Int64 (flat @Integer $ fromIntegral @Int64 maxBound)
Right 9223372036854775807

>>> unflat @Int64 (flat @Integer $ fromIntegral @Int64 maxBound + 1)
Left (BadEncoding ...

>>> unflat @Int16 (flat @Int32 $ fromIntegral @Int16 minBound)
Right (-32768)

>>> unflat @Int16 (flat @Int32 $ fromIntegral @Int16 minBound - 1)
Left (BadEncoding ...

>>> unflat @Int32 (flat @Int64 $ fromIntegral @Int32 minBound)
Right (-2147483648)

>>> unflat @Int32 (flat @Int64 $ fromIntegral @Int32 minBound - 1)
Left (BadEncoding ...

>>> unflat @Int64 (flat @Integer $ fromIntegral @Int64 minBound)
Right (-9223372036854775808)

>>> unflat @Int64 (flat @Integer $ fromIntegral @Int64 minBound - 1)
Left (BadEncoding ...

Instance details

Methods

encode ∷ Word → Encoding

decode ∷ Get Word

size ∷ Word → NumBits → NumBits

Flat a ⇒ Flat (Complex a)
>>> test (4 :+ 2 :: Complex Word8)
(True,16,"00000100 00000010")

Instance details

Methods

encode ∷ Complex a → Encoding

decode ∷ Get (Complex a)

size ∷ Complex a → NumBits → NumBits

Flat a ⇒ Flat (Identity a)
Since: 0.4.4

Instance details

Methods

encode ∷ Identity a → Encoding

decode ∷ Get (Identity a)

size ∷ Identity a → NumBits → NumBits

Flat a ⇒ Flat (First a)
Since: 0.4.4

Instance details

Methods

encode ∷ First a → Encoding

decode ∷ Get (First a)

size ∷ First a → NumBits → NumBits

Flat a ⇒ Flat (Last a)
Since: 0.4.4

Instance details

Methods

encode ∷ Last a → Encoding

decode ∷ Get (Last a)

size ∷ Last a → NumBits → NumBits

Flat a ⇒ Flat (Max a)
Since: 0.4.4

Instance details

Methods

encode ∷ Max a → Encoding

decode ∷ Get (Max a)

size ∷ Max a → NumBits → NumBits

Flat a ⇒ Flat (Min a)
Since: 0.4.4

Instance details

Methods

encode ∷ Min a → Encoding

decode ∷ Get (Min a)

size ∷ Min a → NumBits → NumBits

Flat a ⇒ Flat (Dual a)
Since: 0.4.4

Instance details

Methods

encode ∷ Dual a → Encoding

decode ∷ Get (Dual a)

size ∷ Dual a → NumBits → NumBits

Flat a ⇒ Flat (Product a)
Since: 0.4.4

Instance details

Methods

encode ∷ Product a → Encoding

decode ∷ Get (Product a)

size ∷ Product a → NumBits → NumBits

Flat a ⇒ Flat (Sum a)
Since: 0.4.4

Instance details

Methods

encode ∷ Sum a → Encoding

decode ∷ Get (Sum a)

size ∷ Sum a → NumBits → NumBits

Flat a ⇒ Flat (NonEmpty a)
>>> test (B.fromList [True])
(True,2,"10")

>>> test (B.fromList [False,False])
(True,4,"0100")

Instance details

Methods

encode ∷ NonEmpty a → Encoding

decode ∷ Get (NonEmpty a)

size ∷ NonEmpty a → NumBits → NumBits

(Integral a, Flat a) ⇒ Flat (Ratio a)
Ratios are encoded as tuples of (numerator,denominator)

>>> test (3%4::Ratio Word8)
(True,16,"00000011 00000100")

Instance details

Methods

encode ∷ Ratio a → Encoding

decode ∷ Get (Ratio a)

size ∷ Ratio a → NumBits → NumBits

Flat a ⇒ Flat (Maybe a)
>>> test (Nothing::Maybe Bool)
(True,1,"0")

>>> test (Just False::Maybe Bool)
(True,2,"10")

Instance details

Methods

encode ∷ Maybe a → Encoding

decode ∷ Get (Maybe a)

size ∷ Maybe a → NumBits → NumBits

Flat a ⇒ Flat [a]
>>> test ([]::[Bool])
(True,1,"0")

>>> test [False,False]
(True,5,"10100")

This instance and other similar ones are declared as OVERLAPPABLE, because for better encoding/decoding performance it can be useful to declare instances of concrete types, such as [Char] (not provided out of the box).

Instance details

Methods

encode ∷ [a] → Encoding

decode ∷ Get [a]

size ∷ [a] → NumBits → NumBits

(Flat a, Flat b) ⇒ Flat (Either a b)
>>> test (Left False::Either Bool ())
(True,2,"00")

>>> test (Right ()::Either Bool ())
(True,1,"1")

Instance details

Methods

encode ∷ Either a b → Encoding

decode ∷ Get (Either a b)

size ∷ Either a b → NumBits → NumBits

Flat (Fixed a)
>>> test (MkFixed 123 :: Fixed E0)
(True,16,"11110110 00000001")

>>> test (MkFixed 123 :: Fixed E0) == test (MkFixed 123 :: Fixed E2)
True

Instance details

Methods

encode ∷ Fixed a → Encoding

decode ∷ Get (Fixed a)

size ∷ Fixed a → NumBits → NumBits

(Flat a, Flat b) ⇒ Flat (a, b)
Tuples are supported up to 7 elements.

>>> test (False,())
(True,1,"0")

>>> test ((),())
(True,0,"")

"7 elements tuples ought to be enough for anybody" (Bill Gates - apocryphal)

>>> test (False,True,True,True,False,True,True)
(True,7,"0111011")

tst (1::Int,"2","3","4","5","6","7","8") ...error

Instance details

Methods

encode ∷ (a, b) → Encoding

decode ∷ Get (a, b)

size ∷ (a, b) → NumBits → NumBits

Flat (f a) ⇒ Flat (Alt f a)
>>> let w = Just (11::Word8); a = Alt w <> Alt (Just 24) in tst a == tst w
True

>>> let w = Just (11::Word8); a = Alt Nothing <> Alt w in tst a == tst w
True

Since: 0.4.4

Instance details

Methods

encode ∷ Alt f a → Encoding

decode ∷ Get (Alt f a)

size ∷ Alt f a → NumBits → NumBits

(Flat a, Flat b, Flat c) ⇒ Flat (a, b, c)

Instance details

Methods

encode ∷ (a, b, c) → Encoding

decode ∷ Get (a, b, c)

size ∷ (a, b, c) → NumBits → NumBits

(Flat a, Flat b, Flat c, Flat d) ⇒ Flat (a, b, c, d)

Instance details

Methods

encode ∷ (a, b, c, d) → Encoding

decode ∷ Get (a, b, c, d)

size ∷ (a, b, c, d) → NumBits → NumBits

(Flat a, Flat b, Flat c, Flat d, Flat e) ⇒ Flat (a, b, c, d, e)

Instance details

Methods

encode ∷ (a, b, c, d, e) → Encoding

decode ∷ Get (a, b, c, d, e)

size ∷ (a, b, c, d, e) → NumBits → NumBits

(Flat a, Flat b, Flat c, Flat d, Flat e, Flat f) ⇒ Flat (a, b, c, d, e, f)

Instance details

Methods

encode ∷ (a, b, c, d, e, f) → Encoding

decode ∷ Get (a, b, c, d, e, f)

size ∷ (a, b, c, d, e, f) → NumBits → NumBits

(Flat a, Flat b, Flat c, Flat d, Flat e, Flat f, Flat g) ⇒ Flat (a, b, c, d, e, f, g)

Instance details

Methods

encode ∷ (a, b, c, d, e, f, g) → Encoding

decode ∷ Get (a, b, c, d, e, f, g)

size ∷ (a, b, c, d, e, f, g) → NumBits → NumBits