PlutusCore.Flat.Memory

Description

Memory access primitives.

Includes code from the store-core package.

Documentation

chunksToByteString ∷ (Ptr Word8, [Int]) → ByteString

chunksToByteArray ∷ (Ptr Word8, [Int]) → (ByteArray, Int)

data ByteArray

Lifted wrapper for ByteArray#.

Since ByteArray# is an unlifted type and not a member of kind Type, things like [ByteArray#] or IO ByteArray# are ill-typed. To work around this inconvenience this module provides a standard lifted wrapper, inhabiting Type. Clients are expected to use ByteArray in higher-level APIs, but wrap and unwrap ByteArray internally as they please and use functions from GHC.Exts.

Since: base-4.17.0.0

pokeByteArray ∷ ByteArray# → Int → Int → Ptr Word8 → IO (Ptr Word8)

Copy ByteArray to given pointer, returns new pointer

pokeByteString ∷ ByteString → Ptr Word8 → IO (Ptr Word8)

Copy bytestring to given pointer, returns new pointer

unsafeCreateUptoN' ∷ Int → (Ptr Word8 → IO (Int, a)) → (ByteString, a)

minusPtr ∷ Ptr a → Ptr b → Int

Computes the offset required to get from the second to the first argument. We have

p2 == p1 `plusPtr` (p2 `minusPtr` p1)

peekByteString

Arguments

∷ Ptr Word8
sourcePtr

→ Int
sourceLen

→ ByteString

Create a new bytestring, copying sourceLen bytes from sourcePtr

Since: 0.6