PlutusCore.Flat.Types

Description

Common Types

Documentation

type NumBits = Int

Number of bits

module Data.Word

module Data.Int

data Natural

Natural number

Invariant: numbers <= 0xffffffffffffffff use the NS constructor

data ShortByteString

A compact representation of a Word8 vector.

It has a lower memory overhead than a ByteString and does not contribute to heap fragmentation. It can be converted to or from a ByteString (at the cost of copying the string data). It supports very few other operations.

data Text

A space efficient, packed, unboxed Unicode text type.