PlutusCore.Generators.Hedgehog.Builtin

Documentation

class GenTypedTerm uni where

This class exists so we can provide an ad-hoc typed term generator for various universes. We usually rely-on a universe-specific generator for well-typed constants within that universe.

TODO: Move this to PlutusIR.Generators.AST, and merge genConstant with genConstant.

Methods

genTypedTerm ∷ ∀ (a ∷ Type) fun. KnownTypeAst TyName uni a ⇒ TypeRep a → Gen (Term TyName Name uni fun ())

Generate a Term in uni with the given type.

class GenArbitraryTerm uni where

This class exists so we can provide an ad-hoc arbitrary term generator for various universes.

Methods

genArbitraryTerm ∷ ∀ fun. (Bounded fun, Enum fun) ⇒ Gen (Term TyName Name uni fun ())

Generate an arbitrary Term in uni.

genConstant ∷ ∀ (a ∷ Type). TypeRep a → SomeGen DefaultUni