PlutusCore.Generators.Hedgehog.Denotation

Description

This module defines tools for associating PLC terms with their corresponding Haskell values.

Documentation

type KnownType val a = (KnownTypeAst TyName (UniOf val) a, MakeKnown val a)

data Denotation term object res

Haskell denotation of a PLC object. An object can be a Builtin or a variable for example.

Constructors

∀ args. Denotation

Fields

_denotationObject ∷ object

A PLC object.

_denotationToTerm ∷ object → term

How to embed the object into a term.

_denotationItself ∷ FoldArgs args res

The denotation of the object. E.g. the denotation of AddInteger is (+).

_denotationScheme ∷ TypeScheme term args res

The TypeScheme of the object.

data DenotationContextMember term res

A member of a DenotationContext. object is existentially quantified, so the only thing that can be done with it, is turning it into a Term using _denotationToTerm.

Constructors

∀ object. DenotationContextMember (Denotation term object res)

newtype DenotationContext term

A context of DenotationContextMembers. Each row is a mapping from a type to a list of things that can return that type. For example it can contain a mapping from integer to 1. a bound variable of type integer 2. a bound variable of functional type with the result being integer 3. the AddInteger Builtin or any other Builtin which returns an integer.

Constructors

DenotationContext

Fields

unDenotationContext ∷ DMap TypeRep (Compose List (DenotationContextMember term))

lookupInContext ∷ ∀ a term. TypeRep a → DenotationContext term → [DenotationContextMember term a]

Look up a list of Denotations from DenotationContext each of which has a type that ends in the same type as the one that the TypeRep argument represents.

denoteVariable ∷ KnownType (Term TyName Name uni fun ()) res ⇒ Name → TypeRep res → res → Denotation (Term TyName Name uni fun ()) Name res

Get the Denotation of a variable.

insertVariable ∷ KnownType (Term TyName Name uni fun ()) a ⇒ Name → TypeRep a → a → DenotationContext (Term TyName Name uni fun ()) → DenotationContext (Term TyName Name uni fun ())

Insert a variable into a DenotationContext.

insertBuiltin ∷ DefaultFun → DenotationContext (Term TyName Name DefaultUni DefaultFun ()) → DenotationContext (Term TyName Name DefaultUni DefaultFun ())

Insert a builtin into a DenotationContext.

typedBuiltins ∷ DenotationContext (Term TyName Name DefaultUni DefaultFun ())

A DenotationContext that consists of DefaultFuns.

DEPRECATED: No need to update for a new builtin. Outdated, since we moved to quickcheck generators.