PlutusCore.Generators.Hedgehog.Test

Description

This module defines functions useful for testing.

Documentation

data TypeEvalCheckError uni fun

The type of errors that can occur during type-eval checking.

Constructors

TypeEvalCheckErrorIllFormed !(Error uni fun ())

TypeEvalCheckErrorIllTyped !(Normalized (Type TyName uni ())) !(Normalized (Type TyName uni ()))

TypeEvalCheckErrorException !String

TypeEvalCheckErrorIllEvaled !(EvaluationResult (Term TyName Name uni fun ())) !(EvaluationResult (Term TyName Name uni fun ()))
The former is an expected result of evaluation, the latter -- is an actual one.

data TypeEvalCheckResult uni fun

Type-eval checking of a term results in a value of this type.

Constructors

TypeEvalCheckResult

Fields

_termCheckResultType ∷ Normalized (Type TyName uni ())

The type of the term.

_termCheckResultValue ∷ EvaluationResult (Term TyName Name uni fun ())

The result of evaluation of the term.

type TypeEvalCheckM uni fun = Either (TypeEvalCheckError uni fun)

The monad type-eval checking runs in.

typeEvalCheckBy

Arguments

∷ (uni ~ DefaultUni, fun ~ DefaultFun, KnownTypeAst TyName uni a, MakeKnown (Term TyName Name uni fun ()) a, PrettyPlc structural)

⇒ (Term TyName Name uni fun () → Either (EvaluationException structural operational (Term TyName Name uni fun ())) (Term TyName Name uni fun ()))
An evaluator.

→ TermOf (Term TyName Name uni fun ()) a

→ TypeEvalCheckM uni fun (TermOf (Term TyName Name uni fun ()) (TypeEvalCheckResult uni fun))

Type check and evaluate a term and check that the expected result is equal to the actual one.

unsafeTypeEvalCheck ∷ (uni ~ DefaultUni, fun ~ DefaultFun, KnownTypeAst TyName uni a, MakeKnown (Term TyName Name uni fun ()) a) ⇒ TermOf (Term TyName Name uni fun ()) a → TermOf (Term TyName Name uni fun ()) (EvaluationResult (Term TyName Name uni fun ()))

Type check and evaluate a term and check that the expected result is equal to the actual one. Throw an error in case something goes wrong.

getSampleTermValue ∷ (uni ~ DefaultUni, fun ~ DefaultFun, KnownTypeAst TyName uni a, MakeKnown (Term TyName Name uni fun ()) a) ⇒ TermGen a → IO (TermOf (Term TyName Name uni fun ()) (EvaluationResult (Term TyName Name uni fun ())))

Generate a term using a given generator and check that it's well-typed and evaluates correctly.

getSampleProgramAndValue ∷ (uni ~ DefaultUni, fun ~ DefaultFun, KnownTypeAst TyName uni a, MakeKnown (Term TyName Name uni fun ()) a) ⇒ TermGen a → IO (Program TyName Name uni fun (), EvaluationResult (Term TyName Name uni fun ()))

Generate a program using a given generator and check that it's well-typed and evaluates correctly.

printSampleProgramAndValue ∷ (uni ~ DefaultUni, fun ~ DefaultFun, KnownTypeAst TyName uni a, MakeKnown (Term TyName Name uni fun ()) a) ⇒ TermGen a → IO ()

Generate a program using a given generator, check that it's well-typed and evaluates correctly and pretty-print it to stdout using the default pretty-printing mode.

sampleProgramValueGolden

Arguments

∷ (uni ~ DefaultUni, fun ~ DefaultFun, KnownTypeAst TyName uni a, MakeKnown (Term TyName Name uni fun ()) a)

⇒ String
folder

→ String
name

→ TermGen a
A term generator.

→ IO ()

Generate a pair of files: folder.name.plc and folder.name.plc.golden. The first file contains a term generated by a term generator (wrapped in Program), the second file contains the result of evaluation of the term.

propEvaluate

Arguments

∷ (uni ~ DefaultUni, fun ~ DefaultFun, KnownTypeAst TyName uni a, MakeKnown (Term TyName Name uni fun ()) a, PrettyPlc structural)

⇒ (Term TyName Name uni fun () → Either (EvaluationException structural operational (Term TyName Name uni fun ())) (Term TyName Name uni fun ()))
An evaluator.

→ TermGen a
A term/value generator.

→ Property

A property-based testing procedure for evaluators. Checks whether a term generated along with the value it's supposed to compute to indeed computes to that value according to the provided evaluate.