PlutusCore.Generators.NEAT.Common

Documentation

data Z

data S n

Constructors

FZ

FS n

fromZ ∷ Z → a

Absurd for the zero type.

Namespaces

data NameState n

Constructors

NameState

Fields

nameOf ∷ n → Name

freshNameStrings ∷ Stream Text

newtype TyNameState n

Constructors

TyNameState (NameState n)

tynameOf ∷ TyNameState n → n → TyName

emptyNameState ∷ Stream Text → NameState Z

Create an empty name state from a stream of text names.

extNameState ∷ MonadQuote m ⇒ NameState n → m (NameState (S n))

Extend name state with a fresh name.

emptyTyNameState ∷ Stream Text → TyNameState Z

Create an empty name state from a stream of text names.

extTyNameState ∷ MonadQuote m ⇒ TyNameState n → m (TyNameState (S n))

Extend type name state with a fresh type name.

mkTextNameStream ∷ Text → Stream Text

Create a stream of names |x0, x1, x2, ...| from a prefix |"x"|