PlutusCore.Generators.NEAT.Spec

Description

This file contains the tests and some associated machinery but not the generators.

Documentation

tests ∷ TestTree

data Options

Options for parallel conjunction strategies. Only matters for predicates using the Cool data type instead of Bool.

Constructors

D
Sequential

O
Optimal Short-circuiting

F
Parallel (fair)

OF
Optimal Short-circuiting and fairness

OS
Optimal Short-circuiting and choice-subset detection

OSF
Subset choice short-circuiting

data TestFail

Constructors

GenError GenError

TypeError (TypeError (Term TyName Name DefaultUni DefaultFun ()) DefaultUni DefaultFun ())

AgdaErrorP ()

FVErrorP FreeVariableError

CkP (CkEvaluationException DefaultUni DefaultFun)

UCekP (CekEvaluationException Name DefaultUni DefaultFun)

Ctrex Ctrex

bigTest ∷ (Check t a, Enumerable a) ⇒ String → t → (t → a → Assertion) → TestTree

generate examples using search' and then generate one big test that applies the given test to each of them.

packAssertion ∷ Show e ⇒ (t → a → ExceptT e Quote ()) → t → a → Assertion

given a prop, generate one test

tynames ∷ Stream Text

Stream of type names t0, t1, t2, ..

names ∷ Stream Text

Stream of names x0, x1, x2, ..

throwCtrex ∷ Ctrex → ExceptT TestFail Quote ()

Throw a counter-example.

data Ctrex

Constructors

CtrexNormalizeConvertCommuteTypes (Kind ()) ClosedTypeG (Type TyName DefaultUni ()) (Type TyName DefaultUni ())

CtrexNormalTypesCannotReduce (Kind ()) ClosedTypeG

CtrexKindCheckFail (Kind ()) ClosedTypeG

CtrexKindPreservationFail (Kind ()) ClosedTypeG

CtrexKindMismatch (Kind ()) ClosedTypeG (Kind ()) (Kind ())

CtrexTypeNormalizationFail (Kind ()) ClosedTypeG

CtrexTypeNormalizationMismatch (Kind ()) ClosedTypeG (Type TyName DefaultUni ()) (Type TyName DefaultUni ())

CtrexTypeCheckFail ClosedTypeG ClosedTermG

CtrexTypePreservationFail ClosedTypeG ClosedTermG (Term TyName Name DefaultUni DefaultFun ()) (Term TyName Name DefaultUni DefaultFun ())

CtrexTermEvaluationFail String ClosedTypeG ClosedTermG

CtrexTermEvaluationMismatch ClosedTypeG ClosedTermG [(String, Term TyName Name DefaultUni DefaultFun ())]

CtrexUntypedTermEvaluationMismatch ClosedTypeG ClosedTermG [(String, Term Name DefaultUni DefaultFun ())]

handleError ∷ Type TyName DefaultUni () → ErrorWithCause (EvaluationError structural operational) term → Either (ErrorWithCause (EvaluationError structural operational) term) (Term TyName Name DefaultUni DefaultFun ())

handleUError ∷ ErrorWithCause (EvaluationError structural operational) term → Either (ErrorWithCause (EvaluationError structural operational) term) (Term Name DefaultUni DefaultFun ())

newtype GenDepth

Search depth, measured in program size

Constructors

GenDepth

Fields

unGenDepth ∷ Int

newtype GenMode

Search strategy

Constructors

GenMode

Fields

unGenMode ∷ Options