PlutusCore.Generators.NEAT.Term

Description

This file contains 1. A duplicate of the Plutus Core Abstract Syntax (types and terms) 2. A kind checker and a type checker 3. Reduction semantics for types

Documentation

data TypeBuiltinG

Constructors

TyByteStringG

TyIntegerG

TyBoolG

TyUnitG

TyStringG

TyListG TypeBuiltinG

TyDataG

data TermConstantG

Constructors

TmIntegerG Integer

TmByteStringG ByteString

TmStringG Text

TmBoolG Bool

TmUnitG ()

TmDataG Data

data TypeG n

Constructors

TyVarG n

TyFunG (TypeG n) (TypeG n)

TyIFixG (TypeG n) (Kind ()) (TypeG n)

TyForallG (Kind ()) (TypeG (S n))

TyBuiltinG TypeBuiltinG

TyLamG (TypeG (S n))

TyAppG (TypeG n) (TypeG n) (Kind ())

type ClosedTypeG = TypeG Z

convertClosedType ∷ (MonadQuote m, MonadError GenError m) ⇒ Stream Text → Kind () → ClosedTypeG → m (Type TyName DefaultUni ())

Convert generated closed types to Plutus types.

data TermG tyname name

Constructors

VarG name

LamAbsG (TermG tyname (S name))

ApplyG (TermG tyname name) (TermG tyname name) (TypeG tyname)

TyAbsG (TermG (S tyname) name)

TyInstG (TermG tyname name) (TypeG (S tyname)) (TypeG tyname) (Kind ())

ConstantG TermConstantG

BuiltinG DefaultFun

WrapG (TermG tyname name)

UnWrapG (TypeG tyname) (Kind ()) (TypeG tyname) (TermG tyname name)

ErrorG (TypeG tyname)

type ClosedTermG = TermG Z Z

convertClosedTerm ∷ (MonadQuote m, MonadError GenError m) ⇒ Stream Text → Stream Text → ClosedTypeG → ClosedTermG → m (Term TyName Name DefaultUni DefaultFun ())

Convert generated closed terms to Plutus terms.

class Check t a where

Methods

check ∷ t → a → Cool

stepTypeG ∷ TypeG n → Maybe (TypeG n)

Reduce a generated type by a single step, or fail.

normalizeTypeG ∷ TypeG n → TypeG n

Normalise a generated type.

data GenError

Constructors

∀ tyname.Show tyname ⇒ BadTypeG (Kind ()) (TypeG tyname)

∀ tyname name.(Show tyname, Show name) ⇒ BadTermG (TypeG tyname) (TermG tyname name)

newtype Neutral a

Constructors

Neutral

Fields

unNeutral ∷ a

Orphan instances

Enumerable ByteString

Instance details

Methods

enumerate ∷ ∀ (f ∷ Type → Type). (Typeable f, Sized f) ⇒ Shared f ByteString

Enumerable Data

Instance details

Methods

enumerate ∷ ∀ (f ∷ Type → Type). (Typeable f, Sized f) ⇒ Shared f Data

Enumerable DefaultFun

Instance details

Methods

enumerate ∷ ∀ (f ∷ Type → Type). (Typeable f, Sized f) ⇒ Shared f DefaultFun

Enumerable Text

Instance details

Methods

enumerate ∷ ∀ (f ∷ Type → Type). (Typeable f, Sized f) ⇒ Shared f Text

Enumerable tyname ⇒ Enumerable (Normalized (TypeG tyname))

Instance details

Methods

enumerate ∷ ∀ (f ∷ Type → Type). (Typeable f, Sized f) ⇒ Shared f (Normalized (TypeG tyname))

Enumerable tyname ⇒ Enumerable (Neutral (TypeG tyname))

Instance details

Methods

enumerate ∷ ∀ (f ∷ Type → Type). (Typeable f, Sized f) ⇒ Shared f (Neutral (TypeG tyname))