PlutusCore.Generators.QuickCheck.Common

Documentation

type TypeCtx = Map TyName (Kind ())

inferKind ∷ TypeCtx → Type TyName DefaultUni () → Either String (Kind ())

Infer the kind of a type in a given kind context

unsafeInferKind ∷ HasCallStack ⇒ TypeCtx → Type TyName DefaultUni () → Kind ()

Partial unsafeInferKind, useful for context where invariants are set up to guarantee that types are well-kinded.

checkKind ∷ TypeCtx → Type TyName DefaultUni () → Kind () → Either String ()

Check well-kindedness of a type in a context

genList ∷ Int → Int → Gen a → Gen [a]

Generate a list with the given minimum and maximum lengths. It is similar to Hedgehog.Internal.Gen.list.

Note that genList 0 n gen behaves differently than resize n (listOf gen), because

   resize m (genList 0 n gen) = genList 0 n (resize m gen)

whereas

   resize m (resize n (listOf gen)) = resize n (listOf gen)

Orphan instances

PrettyBy config i ⇒ DefaultPrettyBy config (NonNegative i)

Instance details

Methods

defaultPrettyBy ∷ config → NonNegative i → Doc ann

defaultPrettyListBy ∷ config → [NonNegative i] → Doc ann

PrettyDefaultBy config (NonNegative i) ⇒ PrettyBy config (NonNegative i)

Instance details

Methods

prettyBy ∷ config → NonNegative i → Doc ann

prettyListBy ∷ config → [NonNegative i] → Doc ann

Pretty i ⇒ Pretty (NonNegative i)

Instance details

Methods

pretty ∷ NonNegative i → Doc ann

prettyList ∷ [NonNegative i] → Doc ann

Testable (Either String ())

Instance details

Methods

property ∷ Either String () → Property

propertyForAllShrinkShow ∷ Gen a → (a → [a]) → (a → [String]) → (a → Either String ()) → Property