PlutusCore.Generators.QuickCheck.GenerateTypes

Generators for well-kinded types

genAtomicType ∷ Kind () → GenTm (Type TyName DefaultUni ())

Generate "small" types at a given kind such as builtins, bound variables, bound datatypes, and lambda abstractions t0 ... tn. T

genType ∷ Kind () → GenTm (Type TyName DefaultUni ())

Generate a type at a given kind

genClosedType ∷ Kind () → Gen (Type TyName DefaultUni ())

Generate a closed type at a given kind

genClosedTypeDebug ∷ Kind () → Gen (Type TyName DefaultUni ())

Generate a closed type at a given kind

genTypeWithCtx ∷ TypeCtx → Kind () → Gen (Type TyName DefaultUni ())

Generate a type in the given context with the given kind.

genTypeWithCtxDebug ∷ TypeCtx → Kind () → Gen (Type TyName DefaultUni ())

Generate a type in the given context with the given kind.

genKindAndTypeWithCtx ∷ TypeCtx → Gen (Kind (), Type TyName DefaultUni ())

Generate a well-kinded type and its kind in a given context

builtinKind ∷ SomeTypeIn DefaultUni → Kind ()

Get the kind of a builtin

genKindAndType ∷ Gen (Kind (), Type TyName DefaultUni ())

Generate an arbitrary kind and closed type of that kind.

genKindAndTypeDebug ∷ Gen (Kind (), Type TyName DefaultUni ())

Generate an arbitrary kind and closed type of that kind.

normalizeTy ∷ Type TyName DefaultUni () → Type TyName DefaultUni ()

Normalize a type.

genCtx ∷ Gen TypeCtx

Generate a context of free type variables with kinds