PlutusCore.Name.Unique

Description

A Name is a datatype used to identify a variable inside the Plutus Core languages. Name comparisons are a fundamental part of the domain logic, and comparing Text directly is inefficient. As a solution to this problem, we provide the Unique type which is an integer associated to the Name, unique to each instantiation of the type. We can, therefore, compare the integers instead, which is obviously much more cost-effective.

We distinguish between the names of term variables and type variables by defining the TyName wrapper over Name. Since the code we usually write is polymorphic in the name type, we want to be able to define a class of names which have an associated Unique. This class is HasUnique, see the definition below.

Types

data Name

A Name represents variables/names in Plutus Core.

Constructors

Name

Fields

_nameText ∷ Text

The identifier name, for use in error messages.

_nameUnique ∷ Unique

A Unique assigned to the name, allowing for cheap comparisons in the compiler.

isIdentifierStartingChar ∷ Char → Bool

Allowed characters in the starting position of a non-quoted identifier.

isIdentifierChar ∷ Char → Bool

Allowed characters in a non-starting position of a non-quoted identifier.

isQuotedIdentifierChar ∷ Char → Bool

Allowed characters in a quoted identifier.

isValidUnquotedName ∷ Text → Bool

toPrintedName ∷ Text → Text

Quote the name with backticks if it is not a valid unquoted name. It does not check whether the given name is a valid quoted name.

newtype TyName

We use a newtype to enforce separation between names used for types and those used for terms.

Constructors

TyName

Fields

unTyName ∷ Name

data Named a

Constructors

Named

Fields

_namedString ∷ Text

_namedValue ∷ a

newtype Unique

A unique identifier

Constructors

Unique

Fields

unUnique ∷ Int

newtype TypeUnique

The unique of a type-level name.

Constructors

TypeUnique

Fields

unTypeUnique ∷ Unique

newtype TermUnique

The unique of a term-level name.

Constructors

TermUnique

Fields

unTermUnique ∷ Unique

class HasText a where

Types which have a textual name attached to them.

Methods

theText ∷ Lens' a Text

class Coercible unique Unique ⇒ HasUnique a unique | a → unique where

Types which have a Unique attached to them, mostly names.

Minimal complete definition

Nothing

Methods

unique ∷ Lens' a unique

default unique ∷ (Wrapped a, HasUnique (Unwrapped a) unique', Coercible unique' unique) ⇒ Lens' a unique

theUnique ∷ HasUnique name unique ⇒ Lens' name Unique

A lens focused on the Unique of a name.

Functions

mapNameString ∷ (Text → Text) → Name → Name

Apply a function to the string representation of a Name.

mapTyNameString ∷ (Text → Text) → TyName → TyName

Apply a function to the string representation of a TyName.