PlutusCore.Name.UniqueMap

Description

A type for maps (key-value associations), where the key type can be identified by Uniques. In practice, these types are usually names. This approach is preferred when it is more efficient to compare the associated Uniques instead of the underlying type.

Documentation

newtype UniqueMap unique a

A mapping from Uniques to arbitrary values of type a. Since Unique is equivalent to Int (see PlutusCore.Name.Unique), we can use an IntMap representation for this type.

Constructors

UniqueMap

Fields

unUniqueMap ∷ IntMap a

insertByUnique ∷ Coercible unique Unique ⇒ unique → a → UniqueMap unique a → UniqueMap unique a

Insert a value a by a unique.

insertByName ∷ HasUnique name unique ⇒ name → a → UniqueMap unique a → UniqueMap unique a

Insert a value a by the unique of a name.

singletonByName ∷ HasUnique name unique ⇒ name → a → UniqueMap unique a

Create the singleton map of the unique of a name and a value a.

insertNamed ∷ (HasText name, HasUnique name unique) ⇒ name → a → UniqueMap unique (Named a) → UniqueMap unique (Named a)

Insert a named value a by the index of the unique of the name.

insertByNameIndex ∷ (HasUnique name unique1, Coercible unique2 Unique) ⇒ name → a → UniqueMap unique2 a → UniqueMap unique2 a

Insert a value by the index of the unique of a name. Unlike insertByUnique and insertByName, this function does not provide any static guarantees, so you can for example insert by a type-level name in a map from term-level uniques.

fromFoldable ∷ Foldable f ⇒ (i → a → UniqueMap unique a → UniqueMap unique a) → f (i, a) → UniqueMap unique a

Convert a Foldable into a UniqueMap using the given insertion function.

fromUniques ∷ Foldable f ⇒ Coercible Unique unique ⇒ f (unique, a) → UniqueMap unique a

Convert a Foldable with uniques into a UniqueMap.

fromNames ∷ Foldable f ⇒ HasUnique name unique ⇒ f (name, a) → UniqueMap unique a

Convert a Foldable with names into a UniqueMap.

lookupUnique ∷ Coercible unique Unique ⇒ unique → UniqueMap unique a → Maybe a

Look up a value by a unique.

lookupName ∷ HasUnique name unique ⇒ name → UniqueMap unique a → Maybe a

Look up a value by the unique of a name.

restrictKeys ∷ UniqueMap unique v → UniqueSet unique → UniqueMap unique v

lookupNameIndex ∷ (HasUnique name unique1, Coercible unique2 Unique) ⇒ name → UniqueMap unique2 a → Maybe a

Look up a value by the index of the unique of a name. Unlike lookupUnique and lookupName, this function does not provide any static guarantees, so you can for example look up a type-level name in a map from term-level uniques.