PlutusCore.Parser

Description

Parsers for PLC terms in DefaultUni.

Documentation

module PlutusCore.Parser.Type

module PlutusCore.Parser.ParserCommon

module PlutusCore.Parser.Builtin

program ∷ Parser (Program TyName Name DefaultUni DefaultFun SrcSpan)

Parser for PLC programs.

parseProgram ∷ (MonadError ParserErrorBundle m, MonadQuote m) ⇒ Text → m (Program TyName Name DefaultUni DefaultFun SrcSpan)

Parse a PLC program. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors. This passes "test" to the parser as the name of the input stream; to supply a name explicity, use `parse program name input`.

parseTerm ∷ (MonadError ParserErrorBundle m, MonadQuote m) ⇒ Text → m (Term TyName Name DefaultUni DefaultFun SrcSpan)

Parse a PLC term. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors.

parseType ∷ (MonadError ParserErrorBundle m, MonadQuote m) ⇒ Text → m (Type TyName DefaultUni SrcSpan)

Parse a PLC type. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors.

data SourcePos

The data type SourcePos represents source positions. It contains the name of the source file, a line number, and a column number. Source line and column positions change intensively during parsing, so we need to make them strict to avoid memory leaks.

data ParserError

An error encountered during parsing.

Constructors

BuiltinTypeNotAStar !Text !SourcePos

UnknownBuiltinFunction !Text !SourcePos ![Text]

InvalidBuiltinConstant !Text !Text !SourcePos