PlutusCore.Parser.Builtin

Documentation

cachedBuiltin ∷ Map Text DefaultFun

builtinFunction ∷ Parser DefaultFun

Parser for builtin functions. Atm the parser can only parse DefaultFun.

conInteger ∷ Parser Integer

Parser for integer constants.

hexByte ∷ Parser Word8

Parser for a pair of hex digits to a Word8.

conBS ∷ Parser ByteString

Parser for bytestring constants. They start with "#".

conText ∷ Parser Text

Parser for string constants (wrapped in double quotes). Note that Data.Text.pack "performs replacement on invalid scalar values", which means that Unicode surrogate code points (corresponding to integers in the range 0xD800-0xDFFF) are converted to the Unicode replacement character U+FFFD (decimal 65533). Thus `(con string "XxD800Z")` parses to a Text object whose second character is U+FFFD.

conUnit ∷ Parser ()

Parser for unit.

conBool ∷ Parser Bool

Parser for bool.

conList ∷ DefaultUni (Esc a) → Parser [a]

Parser for lists.

conArray ∷ DefaultUni (Esc a) → Parser (Vector a)

Parser for arrays.

conPair ∷ DefaultUni (Esc a) → DefaultUni (Esc b) → Parser (a, b)

Parser for pairs.

conDataNoParens ∷ Parser Data

conData ∷ ExpectParens → Parser Data

con0xBS ∷ Parser ByteString

conBLS12_381_G1_Element ∷ Parser Element

conBLS12_381_G2_Element ∷ Parser Element

constantOf ∷ ExpectParens → DefaultUni (Esc a) → Parser a

Parser for constants of the given type.

constant ∷ Parser (Some (ValueOf DefaultUni))

Parser of constants whose type is in DefaultUni.

data ExpectParens

Constructors

ExpectParensYes

ExpectParensNo