PlutusCore.Parser.ParserCommon

Description

Common functions for parsers of UPLC, PLC, and PIR.

Documentation

newtype ParserState

Constructors

ParserState

Fields

identifiers ∷ Map Text Unique

type Parser = ParsecT ParserError Text (StateT ParserState (ReaderT (Maybe Version) Quote))

initial ∷ ParserState

getVersion ∷ Parser (Maybe Version)

Get the version of the program being parsed, if we know it.

withVersion ∷ Version → Parser a → Parser a

Set the version of the program being parsed.

whenVersion ∷ (Version → Bool) → Parser () → Parser ()

Run an action conditionally based on a predicate on the version. If we don't know the version then the predicate is assumed to be false, i.e. we act if we _know_ the predicate is satisfied.

parse ∷ (MonadError ParserErrorBundle m, MonadQuote m) ⇒ Parser a → String → Text → m a

toErrorB ∷ Either (ParseErrorBundle Text ParserError) a → Either ParserErrorBundle a

parseGen ∷ (MonadError ParserErrorBundle m, MonadQuote m) ⇒ Parser a → Text → m a

Generic parser function in which the file path is just "test".

whitespace ∷ Parser ()

Space consumer.

leadingWhitespace ∷ Parser a → Parser a

trailingWhitespace ∷ Parser a → Parser a

withSpan' ∷ (SrcSpan → Parser a) → Parser a

Returns a parser for a by calling the supplied function on the starting and ending positions of a.

The supplied function should usually return a parser that does not consume trailing whitespaces. Otherwise, the end position will be the first character after the trailing whitespaces.

withSpan ∷ (SrcSpan → Parser a) → Parser a

Like withSpan', but the result parser consumes whitespaces.

@withSpan = (<* whitespace) . withSpan'

lexeme ∷ Parser a → Parser a

symbol ∷ Text → Parser Text

inParens ∷ Parser a → Parser a

inBrackets ∷ Parser a → Parser a

inBraces ∷ Parser a → Parser a

toSrcSpan ∷ SourcePos → SourcePos → SrcSpan

version ∷ Parser Version

name ∷ Parser Name

Parses a Name. Does not consume leading or trailing whitespaces.

Orphan instances

(Stream s, MonadQuote m) ⇒ MonadQuote (ParsecT e s m)

Instance details

Methods

liftQuote ∷ Quote a → ParsecT e s m a