PlutusCore.Pretty

Basic types and functions

data Doc ann

The abstract data type Doc ann represents pretty documents that have been annotated with data of type ann.

More specifically, a value of type Doc represents a non-empty set of possible layouts of a document. The layout functions select one of these possibilities, taking into account things like the width of the output document.

The annotation is an arbitrary piece of data associated with (part of) a document. Annotations may be used by the rendering backends in order to display output differently, such as

color information (e.g. when rendering to the terminal)

mouseover text (e.g. when rendering to rich HTML)

whether to show something or not (to allow simple or detailed versions)

The simplest way to display a Doc is via the Show class.

>>> putStrLn (show (vsep ["hello", "world"]))
hello
world

class Pretty a where

Overloaded conversion to Doc.

Laws:

output should be pretty. :-)

Minimal complete definition

pretty

Methods

pretty ∷ a → Doc ann

>>> pretty 1 <+> pretty "hello" <+> pretty 1.234
1 hello 1.234

prettyList ∷ [a] → Doc ann

prettyList is only used to define the instance Pretty a => Pretty [a]. In normal circumstances only the pretty function is used.

>>> prettyList [1, 23, 456]
[1, 23, 456]

class PrettyBy config a where

A class for pretty-printing values in a configurable manner.

A basic example:

>>> data Case = UpperCase | LowerCase
>>> data D = D
>>> instance PrettyBy Case D where prettyBy UpperCase D = "D"; prettyBy LowerCase D = "d"
>>> prettyBy UpperCase D
D
>>> prettyBy LowerCase D
d

The library provides instances for common types like Integer or Bool, so you can't define your own PrettyBy SomeConfig Integer instance. And for the same reason you should not define instances like PrettyBy SomeAnotherConfig a for universally quantified a, because such an instance would overlap with the existing ones. Take for example

>>> data ViaShow = ViaShow
>>> instance Show a => PrettyBy ViaShow a where prettyBy ViaShow = pretty . show

with such an instance prettyBy ViaShow (1 :: Int) throws an error about overlapping instances:

• Overlapping instances for PrettyBy ViaShow Int
    arising from a use of ‘prettyBy’
  Matching instances:
    instance PrettyDefaultBy config Int => PrettyBy config Int
    instance [safe] Show a => PrettyBy ViaShow a

There's a newtype provided specifically for the purpose of defining a PrettyBy instance for any a: PrettyAny. Read its docs for details on when you might want to use it.

The PrettyBy instance for common types is defined in a way that allows to override default pretty-printing behaviour, read the docs of HasPrettyDefaults for details.

Minimal complete definition

Nothing

Methods

prettyBy ∷ config → a → Doc ann

Pretty-print a value of type a the way a config specifies it. The default implementation of prettyBy is in terms of pretty, defaultPrettyFunctorBy or defaultPrettyBifunctorBy depending on the kind of the data type that you're providing an instance for. For example, the default implementation of prettyBy for a monomorphic type is going to be "ignore the config and call pretty over the value":

>>> newtype N = N Int deriving newtype (Pretty)
>>> instance PrettyBy () N
>>> prettyBy () (N 42)
42

The default implementation of prettyBy for a Functor is going to be in terms of defaultPrettyFunctorBy:

>>> newtype N a = N a deriving stock (Functor) deriving newtype (Pretty)
>>> instance PrettyBy () a => PrettyBy () (N a)
>>> prettyBy () (N (42 :: Int))
42

It's fine for the data type to have a phantom parameter as long as the data type is still a Functor (i.e. the parameter has to be of kind Type). Then defaultPrettyFunctorBy is used again:

>>> newtype N a = N Int deriving stock (Functor) deriving newtype (Pretty)
>>> instance PrettyBy () (N b)
>>> prettyBy () (N 42)
42

If the data type has a single parameter of any other kind, then it's not a functor and so like in the monomorphic case pretty is used:

>>> newtype N (b :: Bool) = N Int deriving newtype (Pretty)
>>> instance PrettyBy () (N b)
>>> prettyBy () (N 42)
42

Same applies to a data type with two parameters: if both the parameters are of kind Type, then the data type is assumed to be a Bifunctor and hence defaultPrettyBifunctorBy is used. If the right parameter is of kind Type and the left parameter is of any other kind, then we fallback to assuming the data type is a Functor and defining prettyBy as defaultPrettyFunctorBy. If both the parameters are not of kind Type, we fallback to implementing prettyBy in terms of pretty like in the monomorphic case.

Note that in all those cases a Pretty instance for the data type has to already exist, so that we can derive a PrettyBy one in terms of it. If it doesn't exist or if your data type is not supported (for example, if it has three or more parameters of kind Type), then you'll need to provide the implementation manually.

prettyListBy ∷ config → [a] → Doc ann

prettyListBy is used to define the default PrettyBy instance for [a] and NonEmpty a. In normal circumstances only the prettyBy function is used. The default implementation of prettyListBy is in terms of defaultPrettyFunctorBy.

newtype IgnorePrettyConfig a

A newtype wrapper around a whose point is to provide a PrettyBy config instance for anything that has a Pretty instance.

Constructors

IgnorePrettyConfig

Fields

unIgnorePrettyConfig ∷ a

data AttachPrettyConfig config a

A config together with some value. The point is to provide a Pretty instance for anything that has a PrettyBy config instance.

Constructors

AttachPrettyConfig !config !a

class Render str where

A class for rendering Docs as string types.

Methods

render ∷ Doc ann → str

Render a Doc as a string type.

type PrettyParens = PrettyBy RenderContext

For pretty-printing a value with a minimum amount of parens.

display ∷ ∀ str a. (Pretty a, Render str) ⇒ a → str

Pretty-print and render a value as a string type.

displayBy ∷ ∀ str a config. (PrettyBy config a, Render str) ⇒ config → a → str

Pretty-print and render a value as a string type in a configurable way.

juxtRenderContext ∷ RenderContext

An initial RenderContext. An expression printed in this context gets enclosed in parens unless its outermost operator (if any) binds even stronger than function application.

Defaults

prettyPlc ∷ PrettyPlc a ⇒ a → Doc ann

Pretty-print a value in the default mode using the classic view.

displayPlc ∷ (PrettyPlc a, Render str) ⇒ a → str

Render a value to String in the default mode using the classic view.

prettyPlcSimple ∷ PrettyPlc a ⇒ a → Doc ann

Pretty-print a value in the debug mode using the classic view.

displayPlcSimple ∷ (PrettyPlc a, Render str) ⇒ a → str

Render a value to String in the debug mode using the classic view.

Global configuration

data CondensedErrors

Whether to pretty-print PLC errors in full or with some information omitted.

Constructors

CondensedErrorsYes

CondensedErrorsNo

type DefaultPrettyPlcStrategy a = (PrettyClassic a, PrettyReadable a)

A constraint that allows to derive PrettyBy PrettyConfigPlc instances, see below.

newtype PrettyConfigPlcOptions

Options for pretty-printing PLC entities.

Constructors

PrettyConfigPlcOptions

Fields

_pcpoCondensedErrors ∷ CondensedErrors

data PrettyConfigPlcStrategy

Strategy for pretty-printing PLC entities.

Constructors

PrettyConfigPlcClassic (PrettyConfigClassic PrettyConfigName)

PrettyConfigPlcReadable (PrettyConfigReadable PrettyConfigName)

data PrettyConfigPlc

Global configuration used for pretty-printing PLC entities.

Constructors

PrettyConfigPlc

Fields

_pcpOptions ∷ PrettyConfigPlcOptions

_pcpStrategy ∷ PrettyConfigPlcStrategy

type PrettyPlc = PrettyBy PrettyConfigPlc

The "pretty-printable PLC entity" constraint.

prettyConfigPlcOptions ∷ PrettyConfigPlcOptions

The PrettyConfigPlcOptions used by default: print errors in full.

prettyConfigPlcClassic ∷ PrettyConfigPlcOptions → PrettyConfigPlc

The PrettyConfigPlc used by default: use the classic view and print neither Uniques, nor name attachments.

prettyConfigPlcClassicSimple ∷ PrettyConfigPlcOptions → PrettyConfigPlc

The PrettyConfigPlc used for debugging: use the classic view and print Uniques, but not name attachments.

prettyConfigPlcReadable ∷ PrettyConfigPlcOptions → PrettyConfigPlc

The PrettyConfigPlc used by default and for readability: use the refined view and print Uniques but not name attachments.

prettyConfigPlcReadableSimple ∷ PrettyConfigPlcOptions → PrettyConfigPlc

The PrettyConfigPlc used for debugging and readability: use the refined view and print neither Uniques nor name attachments.

Custom functions for PLC types.

prettyPlcClassic ∷ PrettyPlc a ⇒ a → Doc ann

Pretty-print a PLC value in the default mode using the classic view.

prettyPlcClassicSimple ∷ PrettyPlc a ⇒ a → Doc ann

Pretty-print a PLC value without unique indices using the classic view.

prettyPlcReadable ∷ PrettyPlc a ⇒ a → Doc ann

Pretty-print a PLC value in the default mode using the readable view.

prettyPlcReadableSimple ∷ PrettyPlc a ⇒ a → Doc ann

Pretty-print a PLC value without unique indices using the readable view.

prettyPlcCondensedErrorBy ∷ PrettyPlc a ⇒ (PrettyConfigPlcOptions → PrettyConfigPlc) → a → Doc ann

Pretty-print a PLC value using the condensed way (see CondensedErrors) of pretty-printing PLC errors (in case there are any).

displayPlcCondensedErrorClassic ∷ (PrettyPlc a, Render str) ⇒ a → str

Render an error to String in the condensed manner using the classic view.

Names

newtype PrettyConfigName

A config that determines how to pretty-print a PLC name.

Constructors

PrettyConfigName

Fields

_pcnShowsUnique ∷ Bool

Whether to show the Unique of a name or not.

class HasPrettyConfigName config where

A class of configs from which a PrettyConfigName can be extracted.

Methods

toPrettyConfigName ∷ config → PrettyConfigName

prettyConfigName ∷ PrettyConfigName

The PrettyConfigName used by default: print Unique indexes after nams.

prettyConfigNameSimple ∷ PrettyConfigName

The PrettyConfigName to be used when Unique indices don't matter. Easier to read.

Classic view

data PrettyConfigClassic configName

Configuration for the classic pretty-printing.

Constructors

PrettyConfigClassic

Fields

_pccConfigName ∷ configName

How to pretty-print names.

_pccDisplayAnn ∷ Bool

Whether to display annotations.

type PrettyClassicBy configName = PrettyBy (PrettyConfigClassic configName)

The "classically pretty-printable" constraint.

type PrettyClassic = PrettyClassicBy PrettyConfigName

consAnnIf ∷ Pretty ann ⇒ PrettyConfigClassic configName → ann → [Doc dann] → [Doc dann]

Add a pretty-printed annotation to a list of Docs if the given config enables pretty-printing of annotations.

prettyClassic ∷ PrettyClassic a ⇒ a → Doc ann

Pretty-print a value in the default mode using the classic view.

prettyClassicSimple ∷ PrettyClassic a ⇒ a → Doc ann

Pretty-print a value in the simple mode using the classic view.

Readable view

data ShowKinds

Constructors

ShowKindsYes

ShowKindsNonType

ShowKindsNo

data PrettyConfigReadable configName

Configuration for the readable pretty-printing.

Constructors

PrettyConfigReadable

Fields

_pcrConfigName ∷ configName

_pcrRenderContext ∷ RenderContext

_pcrShowKinds ∷ ShowKinds

prettyReadable ∷ PrettyReadable a ⇒ a → Doc ann

Pretty-print something with the PrettyConfigReadable config.

prettyReadableSimple ∷ PrettyReadable a ⇒ a → Doc ann

Pretty-print something with the PrettyConfigReadableSimple config.

pcrConfigName ∷ ∀ configName configName. Lens (PrettyConfigReadable configName) (PrettyConfigReadable configName) configName configName

pcrRenderContext ∷ ∀ configName. Lens' (PrettyConfigReadable configName) RenderContext

pcrShowKinds ∷ ∀ configName. Lens' (PrettyConfigReadable configName) ShowKinds

type PrettyReadableBy configName = PrettyBy (PrettyConfigReadable configName)

The "readably pretty-printable" constraint.

type PrettyReadable = PrettyReadableBy PrettyConfigName

newtype AsReadable a

For rendering things in a readable manner regardless of the pretty-printing function chosen. I.e. all of show, pretty, prettyClassic will use PrettyReadable instead of doing what they normally do. prettyBy config (AsReadable x) requires config to have a PrettyConfigName and respects it.

This wrapper can be particularly useful if you want to apply a function having a Show or Pretty or PrettyClassic or PrettyPlc or whatever constraint, but want to get the argument rendered in a readable manner instead.

Constructors

AsReadable

Fields

unAsReadable ∷ a

data Parened a

A value of type a to render in parens using the readable pretty-printer.

Constructors

Parened

Fields

parenOpening ∷ String

parenClosing ∷ String

parenedValue ∷ a

inBraces ∷ a → Parened a

Enclose the given value, so that it's rendered inside of braces with no additional parens regardless of the RenderContext.

topPrettyConfigReadable ∷ configName → ShowKinds → PrettyConfigReadable configName

A PrettyConfigReadable with the fixity specified to topFixity.

botPrettyConfigReadable ∷ configName → ShowKinds → PrettyConfigReadable configName

A PrettyConfigReadable with the fixity specified to botFixity.

binderFixity ∷ Fixity

The fixity of a binder.

arrowFixity ∷ Fixity

The fixity of (->).

iterTyForallPrettyM ∷ (MonadPrettyReadable configName env m, PrettyReadableBy configName arg, PrettyReadableBy configName body) ⇒ [arg] → body → m (Doc ann)

Lay out an iterated TyForall via iterBinderPrettyM.

iterLamAbsPrettyM ∷ (MonadPrettyReadable configName env m, PrettyReadableBy configName arg, PrettyReadableBy configName body) ⇒ [arg] → body → m (Doc ann)

Lay out an iterated LamAbs via iterBinderPrettyM.

iterTyAbsPrettyM ∷ (MonadPrettyReadable configName env m, PrettyReadableBy configName arg, PrettyReadableBy configName body) ⇒ [arg] → body → m (Doc ann)

Lay out an iterated TyAbs via iterBinderPrettyM.

iterArrowPrettyM ∷ (MonadPrettyReadable configName env m, PrettyReadableBy configName a) ⇒ [a] → a → m (Doc ann)

Lay out an iterated ->.

iterAppDocM ∷ MonadPrettyContext config env m ⇒ (AnyToDoc config ann → AnyToDoc config ann → NonEmpty (Doc ann)) → m (Doc ann)

Lay out an iteration application, providing to the caller a function to render the head of the application and a function to render each of the arguments.

iterInterAppPrettyM ∷ (MonadPrettyReadable configName env m, PrettyReadableBy configName fun, PrettyReadableBy configName ty, PrettyReadableBy configName term) ⇒ fun → [Either ty term] → m (Doc ann)

Lay out interleaved function applications either as

foo {a} x {b} y z

or as

foo
  {a}
  x
  {b}
  y
  z

Lefts are laid out in braces, Rights are laid out without braces.

iterAppPrettyM ∷ (MonadPrettyContext config env m, PrettyBy config fun, PrettyBy config term) ⇒ fun → [term] → m (Doc ann)

Lay out iterated function applications either as

foo x y z

or as

foo
  x
  y
  z

Utils

prettyBytes ∷ ByteString → Doc ann

newtype ConstConfig

The type of configs used for pretty-printing constants. Has a RenderContext inside, so that we don't add redundant parens to the output.

Constructors

ConstConfig

Fields

unConstConfig ∷ RenderContext

type PrettyConst = PrettyBy ConstConfig

type PrettyUni uni = (PrettyParens (SomeTypeIn uni), Closed uni, uni `Everywhere` PrettyConst)

The set of constraints we need to be able to print built-in types and their values.

type ThrowableBuiltins uni fun = (PrettyUni uni, Pretty fun, Typeable uni, Typeable fun)

The set of constraints we need to be able to throw exceptions with things with built-in types and functions in them.

prettyConst ∷ PrettyConst a ⇒ RenderContext → a → Doc ann

module Text.PrettyBy.Fixity