PlutusCore.Pretty.Classic

Description

A "classic" (i.e. as seen in the specification) way to pretty-print PLC entities.

Documentation

data PrettyConfigClassic configName

Configuration for the classic pretty-printing.

Constructors

PrettyConfigClassic

Fields

_pccConfigName ∷ configName

How to pretty-print names.

_pccDisplayAnn ∷ Bool

Whether to display annotations.

type PrettyClassicBy configName = PrettyBy (PrettyConfigClassic configName)

The "classically pretty-printable" constraint.

type PrettyClassic = PrettyClassicBy PrettyConfigName

type PrettyParens = PrettyBy RenderContext

For pretty-printing a value with a minimum amount of parens.

juxtRenderContext ∷ RenderContext

An initial RenderContext. An expression printed in this context gets enclosed in parens unless its outermost operator (if any) binds even stronger than function application.

consAnnIf ∷ Pretty ann ⇒ PrettyConfigClassic configName → ann → [Doc dann] → [Doc dann]

Add a pretty-printed annotation to a list of Docs if the given config enables pretty-printing of annotations.

prettyConfigClassic ∷ PrettyConfigClassic PrettyConfigName

prettyConfigClassicSimple ∷ PrettyConfigClassic PrettyConfigName

prettyClassic ∷ PrettyClassic a ⇒ a → Doc ann

Pretty-print a value in the default mode using the classic view.

prettyClassicSimple ∷ PrettyClassic a ⇒ a → Doc ann

Pretty-print a value in the simple mode using the classic view.