PlutusCore.Pretty.Default

Documentation

prettyPlc ∷ PrettyPlc a ⇒ a → Doc ann

Pretty-print a value in the default mode using the classic view.

displayPlc ∷ (PrettyPlc a, Render str) ⇒ a → str

Render a value to String in the default mode using the classic view.

prettyPlcSimple ∷ PrettyPlc a ⇒ a → Doc ann

Pretty-print a value in the debug mode using the classic view.

displayPlcSimple ∷ (PrettyPlc a, Render str) ⇒ a → str

Render a value to String in the debug mode using the classic view.

displayPlcCondensedErrorClassic ∷ (PrettyPlc a, Render str) ⇒ a → str

Render an error to String in the condensed manner using the classic view.