PlutusCore.Pretty.Extra

Description

Pretty-printing stuff, some of which should probably go into the main library.

Documentation

type PrettyParens = PrettyBy RenderContext

For pretty-printing a value with a minimum amount of parens.

juxtRenderContext ∷ RenderContext

An initial RenderContext. An expression printed in this context gets enclosed in parens unless its outermost operator (if any) binds even stronger than function application.

Orphan instances

Profunctor InContextM

Instance details

Methods

dimap ∷ (a → b) → (c → d) → InContextM b c → InContextM a d

lmap ∷ (a → b) → InContextM b c → InContextM a c

rmap ∷ (b → c) → InContextM a b → InContextM a c

(#.) ∷ ∀ a b c q. Coercible c b ⇒ q b c → InContextM a b → InContextM a c

(.#) ∷ ∀ a b c q. Coercible b a ⇒ InContextM b c → q a b → InContextM a c

PrettyDefaultBy config [a] ⇒ DefaultPrettyBy config (Set a)

Instance details

Methods

defaultPrettyBy ∷ config → Set a → Doc ann

defaultPrettyListBy ∷ config → [Set a] → Doc ann

PrettyDefaultBy config [a] ⇒ DefaultPrettyBy config (Vector a)

Instance details

Methods

defaultPrettyBy ∷ config → Vector a → Doc ann

defaultPrettyListBy ∷ config → [Vector a] → Doc ann

PrettyDefaultBy config (Set a) ⇒ PrettyBy config (Set a)

Instance details

Methods

prettyBy ∷ config → Set a → Doc ann

prettyListBy ∷ config → [Set a] → Doc ann

PrettyDefaultBy config (Vector a) ⇒ PrettyBy config (Vector a)

Instance details

Methods

prettyBy ∷ config → Vector a → Doc ann

prettyListBy ∷ config → [Vector a] → Doc ann

PrettyDefaultBy config [(k, v)] ⇒ DefaultPrettyBy config (Map k v)

Instance details

Methods

defaultPrettyBy ∷ config → Map k v → Doc ann

defaultPrettyListBy ∷ config → [Map k v] → Doc ann

PrettyDefaultBy config (Map k v) ⇒ PrettyBy config (Map k v)

Instance details

Methods

prettyBy ∷ config → Map k v → Doc ann

prettyListBy ∷ config → [Map k v] → Doc ann