PlutusCore.Pretty.Plc

Description

The global pretty-printing config used to pretty-print everything in the PLC world. This module also defines custom pretty-printing functions for PLC types as a convenience.

Global configuration

data CondensedErrors

Whether to pretty-print PLC errors in full or with some information omitted.

Constructors

CondensedErrorsYes

CondensedErrorsNo

newtype PrettyConfigPlcOptions

Options for pretty-printing PLC entities.

Constructors

PrettyConfigPlcOptions

Fields

_pcpoCondensedErrors ∷ CondensedErrors

data PrettyConfigPlcStrategy

Strategy for pretty-printing PLC entities.

Constructors

PrettyConfigPlcClassic (PrettyConfigClassic PrettyConfigName)

PrettyConfigPlcReadable (PrettyConfigReadable PrettyConfigName)

data PrettyConfigPlc

Global configuration used for pretty-printing PLC entities.

Constructors

PrettyConfigPlc

Fields

_pcpOptions ∷ PrettyConfigPlcOptions

_pcpStrategy ∷ PrettyConfigPlcStrategy

type PrettyPlc = PrettyBy PrettyConfigPlc

The "pretty-printable PLC entity" constraint.

type DefaultPrettyPlcStrategy a = (PrettyClassic a, PrettyReadable a)

A constraint that allows to derive PrettyBy PrettyConfigPlc instances, see below.

prettyConfigPlcOptions ∷ PrettyConfigPlcOptions

The PrettyConfigPlcOptions used by default: print errors in full.

prettyConfigPlcClassic ∷ PrettyConfigPlcOptions → PrettyConfigPlc

The PrettyConfigPlc used by default: use the classic view and print neither Uniques, nor name attachments.

prettyConfigPlcClassicSimple ∷ PrettyConfigPlcOptions → PrettyConfigPlc

The PrettyConfigPlc used for debugging: use the classic view and print Uniques, but not name attachments.

prettyConfigPlcReadable ∷ PrettyConfigPlcOptions → PrettyConfigPlc

The PrettyConfigPlc used by default and for readability: use the refined view and print Uniques but not name attachments.

prettyConfigPlcReadableSimple ∷ PrettyConfigPlcOptions → PrettyConfigPlc

The PrettyConfigPlc used for debugging and readability: use the refined view and print neither Uniques nor name attachments.

Custom functions for PLC types.

prettyPlcClassic ∷ PrettyPlc a ⇒ a → Doc ann

Pretty-print a PLC value in the default mode using the classic view.

prettyPlcClassicSimple ∷ PrettyPlc a ⇒ a → Doc ann

Pretty-print a PLC value without unique indices using the classic view.

prettyPlcReadable ∷ PrettyPlc a ⇒ a → Doc ann

Pretty-print a PLC value in the default mode using the readable view.

prettyPlcReadableSimple ∷ PrettyPlc a ⇒ a → Doc ann

Pretty-print a PLC value without unique indices using the readable view.

prettyPlcCondensedErrorBy ∷ PrettyPlc a ⇒ (PrettyConfigPlcOptions → PrettyConfigPlc) → a → Doc ann

Pretty-print a PLC value using the condensed way (see CondensedErrors) of pretty-printing PLC errors (in case there are any).