PlutusCore.Pretty.PrettyConst

Documentation

newtype ConstConfig

The type of configs used for pretty-printing constants. Has a RenderContext inside, so that we don't add redundant parens to the output.

Constructors

ConstConfig

Fields

unConstConfig ∷ RenderContext

type PrettyConst = PrettyBy ConstConfig

type PrettyUni uni = (PrettyParens (SomeTypeIn uni), Closed uni, uni `Everywhere` PrettyConst)

The set of constraints we need to be able to print built-in types and their values.

type ThrowableBuiltins uni fun = (PrettyUni uni, Pretty fun, Typeable uni, Typeable fun)

The set of constraints we need to be able to throw exceptions with things with built-in types and functions in them.

prettyConst ∷ PrettyConst a ⇒ RenderContext → a → Doc ann

newtype NoParens a

For rendering values without parens, i.e. in botRenderContext.

Constructors

NoParens

Fields

unNoParens ∷ a

asBytes ∷ Word8 → Doc ann

toBytes ∷ ByteString → Doc ann

Orphan instances

Pretty (SomeTypeIn uni) ⇒ Pretty (SomeTypeIn (Kinded uni))

Instance details

Methods

pretty ∷ SomeTypeIn (Kinded uni) → Doc ann

prettyList ∷ [SomeTypeIn (Kinded uni)] → Doc ann

(Closed uni, Everywhere uni PrettyConst) ⇒ Pretty (ValueOf uni a)

Instance details

Methods

pretty ∷ ValueOf uni a → Doc ann

prettyList ∷ [ValueOf uni a] → Doc ann

(Closed uni, Everywhere uni PrettyConst) ⇒ Pretty (Some (ValueOf uni))

Instance details

Methods

pretty ∷ Some (ValueOf uni) → Doc ann

prettyList ∷ [Some (ValueOf uni)] → Doc ann