PlutusCore.Pretty.Readable

Description

A "readable" Agda-like way to pretty-print PLC entities.

Documentation

data Associativity

Associativity of an operator.

Constructors

LeftAssociative

RightAssociative

NonAssociative

type Fixity = FixityOver Precedence

FixityOver instantiated at Precedence.

data RenderContextOver prec

A context that an expression is being rendered in.

Constructors

RenderContext

Fields

_renderContextDirection ∷ !Direction

_renderContextFixity ∷ !(FixityOver prec)

type RenderContext = RenderContextOver Precedence

FixityOver instantiated at Precedence.

data Direction

Direction in which pretty-printing goes. For example in x + y x is pretty-printed to the left of + and y is pretty-printed to the right of +.

Constructors

ToTheLeft

ToTheRight

data FixityOver prec

Fixity of an operator.

We allow unary operators to have associativity, because it's useful to distinguish between an expression like -(-x) (unary minus, left-associative) and ~~b (boolean NOT, right-associative).

Associativity of unary operators also matters when pretty-printing expressions like (-x) + y, which is pretty-printed as -x + y, assuming unary minus has the same fixity as + (and both the operators are left-associative). I.e. unary minus is handled just like the binary one: (0 - x) + y is pretty-printed as 0 - x + y.

Postfix operators are handled similarly. E.g. if ! is left-associative, then (x!)! is pretty-printed as x!! and if it's right-associative -- (x!)!.

The data type is parameterized, so that the user can choose precedence to be integer/fractional, bounded/unbounded, etc (we could also allows operators to be partially or totally ordered, but at the moment prec is required to implement Ord, i.e. it has to be totally ordered). By default we go with bounded fractional precedence, see the main Text.Fixity module.

Constructors

Fixity

Fields

_fixityAssociativity ∷ !Associativity

_fixityPrecedence ∷ !prec

type Precedence = Double

Fractional precedence, so that it's always possible to squeeze an operator precedence between two existing precedences. Ranges over [-20, 120]. A normal operator should have a precedence within [0, 100). It might be useful to have a negative precedence if you're trying to model some already existing system, for example in Haskell ($) has precedence 0, but clearly if b then y else f $ x should be rendered without any parens, hence the precedence of if_then_else_ is less than 0, i.e. negative.

The precedence of juxtaposition is 100. Normally you want juxtaposition to have the highest precedence, but some languages have operators that bind tighter than juxtaposition, e.g. Haskell's postfix _{_}: f z { x = y } means f (z {x = y}).

type MonadPretty config env (m ∷ Type → Type) = (MonadReader env m, HasPrettyConfig env config)

A constraint for "m is a monad that allows to pretty-print values in it by a config".

class HasPrettyConfig env config | env → config where

A constraint for "config is a part of env".

Methods

prettyConfig ∷ Lens' env config

type AnyToDoc config ann = ∀ a. PrettyBy config a ⇒ a → Doc ann

The type of a general config-based pretty-printer.

newtype InContextM config a

A monad for precedence-aware pretty-printing.

Constructors

InContextM

Fields

unInContextM ∷ Reader (Sole config) a

newtype Sole a

A newtype wrapper around a introduced for its HasPrettyConfig instance.

Constructors

Sole

Fields

unSole ∷ a

type MonadPrettyContext config env (m ∷ Type → Type) = (MonadPretty config env m, HasRenderContext config)

A constraint for "m is a Monad supporting configurable precedence-aware pretty-printing".

class HasRenderContext config where

A constraint for "RenderContext is a part of config".

Methods

renderContext ∷ Lens' config RenderContext

class Profunctor (p ∷ Type → Type → Type) where

Formally, the class Profunctor represents a profunctor from Hask -> Hask.

Intuitively it is a bifunctor where the first argument is contravariant and the second argument is covariant.

You can define a Profunctor by either defining dimap or by defining both lmap and rmap.

If you supply dimap, you should ensure that:

dimap id id ≡ id

If you supply lmap and rmap, ensure:

lmap id ≡ id
rmap id ≡ id

If you supply both, you should also ensure:

dimap f g ≡ lmap f . rmap g

These ensure by parametricity:

dimap (f . g) (h . i) ≡ dimap g h . dimap f i
lmap (f . g) ≡ lmap g . lmap f
rmap (f . g) ≡ rmap f . rmap g

Minimal complete definition

dimap | lmap, rmap

Methods

dimap ∷ (a → b) → (c → d) → p b c → p a d

Map over both arguments at the same time.

dimap f g ≡ lmap f . rmap g

lmap ∷ (a → b) → p b c → p a c

Map the first argument contravariantly.

lmap f ≡ dimap f id

rmap ∷ (b → c) → p a b → p a c

Map the second argument covariantly.

rmap ≡ dimap id

type PrettyParens = PrettyBy RenderContext

For pretty-printing a value with a minimum amount of parens.

encloseIn

Arguments

∷ Ord prec

⇒ (a → a)
Enclose a value of type a in parens.

→ RenderContextOver prec
An outer context.

→ FixityOver prec
An inner fixity.

→ a

→ a

Enclose an a (using the provided function) if required or leave it as is. The need for enclosing is determined from an outer RenderContext and the inner fixity.

botFixity ∷ Fixity

A fixity with the lowest precedence. When used as a part of an outer context, never causes addition of parens.

juxtFixity ∷ Fixity

The fixity of juxtaposition.

unitFixity ∷ Fixity

The fixity of a unitary expression which is safe to render without parens in any context.

topFixity ∷ Fixity

A fixity with the highest precedence. When used as a part of an outer context, always causes addition of parens.

botRenderContext ∷ RenderContext

An initial RenderContext. An expression printed in this context never gets enclosed in parens.

topRenderContext ∷ RenderContext

An initial RenderContext. An expression printed in this context always gets enclosed in parens.

prettyM ∷ (MonadPretty config env m, PrettyBy config a) ⇒ a → m (Doc ann)

Pretty-print a value in a configurable way in a monad holding a config.

displayM ∷ ∀ str a m env config. (MonadPretty config env m, PrettyBy config a, Render str) ⇒ a → m str

Pretty-print and render a value as a string type in a configurable way in a monad holding a config.

runInContextM ∷ config → InContextM config a → a

Run InContextM by supplying a config.

inContextM ∷ (a → InContextM config (Doc ann)) → config → a → Doc ann

Takes a monadic pretty-printer and turns it into one that receives a config explicitly. Useful for defining instances of PrettyBy monadically when writing precedence-aware pretty-printing code (and since all functions below are monadic, it's currenty the only option).

encloseM ∷ MonadPrettyContext config env m ⇒ Fixity → Doc ann → m (Doc ann)

Enclose a Doc in parentheses if required or leave it as is. The need for enclosing is determined from an outer RenderContext (stored in the environment of the monad) and the inner fixity provided as an argument.

withPrettyIn ∷ MonadPrettyContext config env m ⇒ ((∀ a. PrettyBy config a ⇒ Direction → Fixity → a → Doc ann) → m r) → m r

Instantiate a supplied continuation with a precedence-aware pretty-printer.

withPrettyAt ∷ MonadPrettyContext config env m ⇒ Direction → Fixity → (AnyToDoc config ann → m r) → m r

Instantiate a supplied continuation with a pretty-printer specialized to supplied Fixity and Direction.

unitDocM ∷ MonadPrettyContext config env m ⇒ Doc ann → m (Doc ann)

Call encloseM on unitFixity.

compoundDocM ∷ MonadPrettyContext config env m ⇒ Fixity → ((∀ a. PrettyBy config a ⇒ Direction → Fixity → a → Doc ann) → Doc ann) → m (Doc ann)

Instantiate a supplied continuation with a pretty-printer and apply encloseM, specialized to supplied Fixity, to the result.

sequenceDocM ∷ MonadPrettyContext config env m ⇒ Direction → Fixity → (AnyToDoc config ann → Doc ann) → m (Doc ann)

Instantiate a supplied continuation with a pretty-printer specialized to supplied Fixity and Direction and apply encloseM specialized to the provided fixity to the result. This can be useful for pretty-printing a sequence of values (possibly consisting of a single value).

infixDocM ∷ MonadPrettyContext config env m ⇒ Fixity → (AnyToDoc config ann → AnyToDoc config ann → Doc ann) → m (Doc ann)

Instantiate a supplied continuation with two pretty-printers (one is going in the ToTheLeft direction, the other is in the ToTheRight direction) specialized to supplied Fixity and apply encloseM, specialized to the same fixity, to the result. The idea is that to the outside an infix operator has the same inner fixity as it has the outer fixity to inner subexpressions.

juxtPrettyM ∷ (MonadPrettyContext config env m, PrettyBy config a, PrettyBy config b) ⇒ a → b → m (Doc ann)

Pretty-print two things with a space between them. The fixity of the context in which the arguments get pretty-printed is set to juxtFixity.

juxtRenderContext ∷ RenderContext

An initial RenderContext. An expression printed in this context gets enclosed in parens unless its outermost operator (if any) binds even stronger than function application.

data ShowKinds

Constructors

ShowKindsYes

ShowKindsNonType

ShowKindsNo

data PrettyConfigReadable configName

Configuration for the readable pretty-printing.

Constructors

PrettyConfigReadable

Fields

_pcrConfigName ∷ configName

_pcrRenderContext ∷ RenderContext

_pcrShowKinds ∷ ShowKinds

type PrettyReadableBy configName = PrettyBy (PrettyConfigReadable configName)

The "readably pretty-printable" constraint.

type PrettyReadable = PrettyReadableBy PrettyConfigName

newtype AsReadable a

For rendering things in a readable manner regardless of the pretty-printing function chosen. I.e. all of show, pretty, prettyClassic will use PrettyReadable instead of doing what they normally do. prettyBy config (AsReadable x) requires config to have a PrettyConfigName and respects it.

This wrapper can be particularly useful if you want to apply a function having a Show or Pretty or PrettyClassic or PrettyPlc or whatever constraint, but want to get the argument rendered in a readable manner instead.

Constructors

AsReadable

Fields

unAsReadable ∷ a

data Parened a

A value of type a to render in parens using the readable pretty-printer.

Constructors

Parened

Fields

parenOpening ∷ String

parenClosing ∷ String

parenedValue ∷ a

type MonadPrettyReadable configName env m = MonadPretty (PrettyConfigReadable configName) env m

A constraint for "m is a monad that allows for pretty-printing values via a PrettyConfigReadable.

type HasPrettyConfigReadable env configName = HasPrettyConfig env (PrettyConfigReadable configName)

type ReadableToDoc configName ann = ∀ a. PrettyReadableBy configName a ⇒ a → Doc ann

The type of a PrettyConfigReadable-based pretty-printer, similar to AnyToDoc.

prettyReadable ∷ PrettyReadable a ⇒ a → Doc ann

Pretty-print something with the PrettyConfigReadable config.

prettyReadableSimple ∷ PrettyReadable a ⇒ a → Doc ann

Pretty-print something with the PrettyConfigReadableSimple config.

pcrConfigName ∷ ∀ configName configName. Lens (PrettyConfigReadable configName) (PrettyConfigReadable configName) configName configName

pcrRenderContext ∷ ∀ configName. Lens' (PrettyConfigReadable configName) RenderContext

pcrShowKinds ∷ ∀ configName. Lens' (PrettyConfigReadable configName) ShowKinds

inBraces ∷ a → Parened a

Enclose the given value, so that it's rendered inside of braces with no additional parens regardless of the RenderContext.

topPrettyConfigReadable ∷ configName → ShowKinds → PrettyConfigReadable configName

A PrettyConfigReadable with the fixity specified to topFixity.

botPrettyConfigReadable ∷ configName → ShowKinds → PrettyConfigReadable configName

A PrettyConfigReadable with the fixity specified to botFixity.

binderFixity ∷ Fixity

The fixity of a binder.

arrowFixity ∷ Fixity

The fixity of (->).

iterTyForallPrettyM ∷ (MonadPrettyReadable configName env m, PrettyReadableBy configName arg, PrettyReadableBy configName body) ⇒ [arg] → body → m (Doc ann)

Lay out an iterated TyForall via iterBinderPrettyM.

iterLamAbsPrettyM ∷ (MonadPrettyReadable configName env m, PrettyReadableBy configName arg, PrettyReadableBy configName body) ⇒ [arg] → body → m (Doc ann)

Lay out an iterated LamAbs via iterBinderPrettyM.

iterTyAbsPrettyM ∷ (MonadPrettyReadable configName env m, PrettyReadableBy configName arg, PrettyReadableBy configName body) ⇒ [arg] → body → m (Doc ann)

Lay out an iterated TyAbs via iterBinderPrettyM.

iterArrowPrettyM ∷ (MonadPrettyReadable configName env m, PrettyReadableBy configName a) ⇒ [a] → a → m (Doc ann)

Lay out an iterated ->.

iterAppDocM ∷ MonadPrettyContext config env m ⇒ (AnyToDoc config ann → AnyToDoc config ann → NonEmpty (Doc ann)) → m (Doc ann)

Lay out an iteration application, providing to the caller a function to render the head of the application and a function to render each of the arguments.

iterInterAppPrettyM ∷ (MonadPrettyReadable configName env m, PrettyReadableBy configName fun, PrettyReadableBy configName ty, PrettyReadableBy configName term) ⇒ fun → [Either ty term] → m (Doc ann)

Lay out interleaved function applications either as

foo {a} x {b} y z

or as

foo
  {a}
  x
  {b}
  y
  z

Lefts are laid out in braces, Rights are laid out without braces.

iterAppPrettyM ∷ (MonadPrettyContext config env m, PrettyBy config fun, PrettyBy config term) ⇒ fun → [term] → m (Doc ann)

Lay out iterated function applications either as

foo x y z

or as

foo
  x
  y
  z

iterBinderPrettyM ∷ (MonadPrettyReadable configName env m, PrettyReadableBy configName arg, PrettyReadableBy configName body) ⇒ (Doc ann → Doc ann) → [arg] → body → m (Doc ann)

Lay out an iterated binder. For example, this function lays out iterated lambdas either as

\(x : a) (y : b) (z : c) -> body

or as

\(x : a)
 (y : b)
 (z : c) ->
  body