PlutusCore.Rename

Description

The user-facing API of the renamer.

Documentation

data Renamed a

A wrapper for signifying that the value inside of it satisfies global uniqueness.

It's safe to call unRenamed, it's not safe to call Renamed, hence the latter is only exported from this internal module and should not be exported from the main API.

Don't provide any instances allowing the user to create a Renamed (even out of an existing one like with Functor).

class Rename a where

The class of things that can be renamed. I.e. things that are capable of satisfying the global uniqueness condition.

Methods

rename ∷ MonadQuote m ⇒ a → m a

Rename Uniques so that they're globally unique. In case there are any free variables, they must be left untouched and bound variables must not get renamed to free ones. Must always assign new names to bound variables, so that rename can be used for alpha-renaming as well.

getRenamed ∷ (Rename a, MonadQuote m) ⇒ a → m (Renamed a)

rename a value and wrap the result in Renamed, so that it can be passed around and it's visible in the types that the thing inside satisfies global uniqueness.

data Dupable a

Dupable a is isomorphic to a, but the only way to extract the a is via liftDupable (defined in the main API module because of a constraint requirement) which renames the stored value along the way. This type is used whenever

preserving global uniqueness is required

some value may be used multiple times

so we annotate such a value with Dupable and call liftDupable at each usage, which ensures global uniqueness is preserved.

unDupable is not supposed to be exported. Don't provide any instances allowing the user to access the underlying value.

dupable ∷ a → Dupable a

Wrap a value in Dupable.

liftDupable ∷ (MonadQuote m, Rename a) ⇒ Dupable a → m a

Extract the value stored in a Dupable a and rename it.