PlutusCore.Rename.Monad

Description

The monad that the renamer runs in and related infrastructure.

Documentation

newtype RenameT ren m a

The monad the renamer runs in.

Constructors

RenameT

Fields

unRenameT ∷ ReaderT ren m a

type ScopedRenameT = RenameT ScopedRenaming

newtype Renaming unique

A renaming is a mapping from old uniques to new ones.

Constructors

Renaming

Fields

unRenaming ∷ UniqueMap unique unique

type TypeRenaming = Renaming TypeUnique

A type-level renaming. Needed for instantiating functions running over types in generic RenameT ren m to a particular type of renaming.

data ScopedRenaming

Scoping-aware mapping from locally unique uniques to globally unique uniques.

Constructors

ScopedRenaming

Fields

_scopedRenamingTypes ∷ Renaming TypeUnique

_scopedRenamingTerms ∷ Renaming TermUnique

class Coercible unique Unique ⇒ HasRenaming ren unique where

A class that specifies which Renaming a ren has inside. A ren can contain several Renamings (like Scoped, for example).

Methods

renaming ∷ Lens' ren (Renaming unique)

scopedRenamingTypes ∷ Lens' ScopedRenaming (Renaming TypeUnique)

scopedRenamingTerms ∷ Lens' ScopedRenaming (Renaming TermUnique)

runRenameT ∷ Monoid ren ⇒ RenameT ren m a → m a

Run a RenameT computation with an empty renaming.

lookupNameM ∷ (HasUnique name unique, HasRenaming ren unique, MonadReader ren m) ⇒ name → m (Maybe unique)

Look up the new unique a name got mapped to.

renameNameM ∷ (HasRenaming ren unique, HasUnique name unique, MonadReader ren m) ⇒ name → m name

Rename a name that has a unique inside.

withFreshenedName ∷ (HasRenaming ren unique, HasUnique name unique, MonadQuote m, MonadReader ren m) ⇒ name → (name → m c) → m c

Replace the unique in a name by a new unique, save the mapping from the old unique to the new one and supply the updated value to a continuation.

withRenamedName ∷ (HasRenaming ren unique, HasUnique name unique, MonadReader ren m) ⇒ name → name → m c → m c

Run a RenameT computation in the environment extended by the mapping from an old name to a new one.