PlutusCore.StdLib.Data.List

Description

Built-in list and related functions.

Documentation

list ∷ uni `HasTypeLevel` List ⇒ Type tyname uni ()

[] as a built-in PLC type.

data MatchOption

Allows one to choose which way of doing pattern matching on built-in types to use: currently only ChooseList-like builtins are supported.

Constructors

UseChoose

matchList ∷ TermLike term TyName Name DefaultUni DefaultFun ⇒ MatchOption → term ()

Pattern matching on built-in lists. matchList {a} xs on built-in lists is equivalent to unwrap xs on lists defined in PLC itself (hence why we bind r after xs).

Either

/\(a :: *) -> \(xs : list a) -> /\(r :: *) -> (z : r) (f : a -> list a -> r) ->
    matchList
        {a}
        {r}
        z
        f
        xs

or

/\(a :: *) -> \(xs : list a) -> /\(r :: *) -> (z : r) (f : a -> list a -> r) ->
    chooseList
        {a}
        {() -> r}
        xs
        (\(u : ()) -> z)
        (\(u : ()) -> f (head {a} xs) (tail {a} xs))
        ()

depending on the MatchOption argument.

foldrList ∷ TermLike term TyName Name DefaultUni DefaultFun ⇒ MatchOption → term ()

foldr over built-in lists.

/\(a :: *) (r :: *) -> \(f : a -> r -> r) (z : r) ->
    fix {list a} {r} \(rec : list a -> r) (xs : list a) ->
        matchList {a} xs {r} z \(x : a) (xs' : list a) -> f x (rec xs') 

foldList ∷ TermLike term TyName Name DefaultUni DefaultFun ⇒ MatchOption → term ()

'foldl'' as a PLC term.

/\(a :: *) (r :: *) -> \(f : r -> a -> r) ->
    fix {r} {list a -> r} \(rec : r -> list a -> r) (z : r) (xs : list a) ->
        matchList {a} xs {r} z \(x : a) (xs' : list a) -> rec (f z x) xs' 

sum ∷ TermLike term TyName Name DefaultUni DefaultFun ⇒ MatchOption → term ()

sumr ∷ TermLike term TyName Name DefaultUni DefaultFun ⇒ MatchOption → term ()

product ∷ TermLike term TyName Name DefaultUni DefaultFun ⇒ MatchOption → term ()

product as a PLC term.

foldList {integer} {integer} multiplyInteger 1