PlutusCore.StdLib.Meta

Description

Functions that generate Plutus Core terms from Haskell values and vice versa.

Documentation

metaIntegerToNat ∷ TermLike term TyName Name uni fun ⇒ Integer → term ()

Convert an Integer to a nat. TODO: convert PLC's integer to nat instead.

metaEitherToSum ∷ TermLike term TyName Name uni fun ⇒ Type TyName uni () → Type TyName uni () → Either (term ()) (term ()) → term ()

Convert a Haskell Either to a PLC sum.

metaListToScottList ∷ TermLike term TyName Name uni fun ⇒ Type TyName uni () → [term ()] → term ()

Convert a Haskell list of Terms to a PLC list.