PlutusCore.Test

Documentation

mapTestLimit ∷ (TestLimit → TestLimit) → Property → Property

Map the TestLimit of a Property with a given function.

withAtLeastTests ∷ TestLimit → Property → Property

Set the number of times a property should be executed before it is considered successful, unless it's already higher than that.

mapTestLimitAtLeast ∷ TestLimit → (TestLimit → TestLimit) → Property → Property

Set the number of times a property should be executed before it is considered successful, unless the given function scales it higher than that.

checkFails ∷ Property → IO ()

Check that the given Property fails.

isSerialisable ∷ Some (ValueOf DefaultUni) → Bool

Check whether the given constant can be serialised. Useful for tests of the parser/deserializer where we need to filter out unprintable/unserialisable terms. Technically, G1, G2 elements etc can be printed but not serialised, but here for simplicity we just assume that all unserialisable terms are unprintable too.

class ToTPlc a uni fun | a → uni fun where

Class for ad-hoc overloading of things which can be turned into a PLC program. Any errors from the process should be caught.

Methods

toTPlc ∷ a → ExceptT SomeException IO (Program TyName Name uni fun ())

class ToUPlc a uni fun | a → uni fun where

Methods

toUPlc ∷ a → ExceptT SomeException IO (Program Name uni fun ())

pureTry ∷ Exception e ⇒ a → Either e a

catchAll ∷ a → ExceptT SomeException IO a

rethrow ∷ ExceptT SomeException IO a → IO a

runTPlc ∷ ToTPlc a DefaultUni DefaultFun ⇒ [a] → ExceptT SomeException IO (EvaluationResult (Term TyName Name DefaultUni DefaultFun ()))

runUPlc ∷ ToUPlc a DefaultUni DefaultFun ⇒ [a] → ExceptT SomeException IO (Term Name DefaultUni DefaultFun ())

runUPlcBudget ∷ ToUPlc a DefaultUni DefaultFun ⇒ [a] → ExceptT SomeException IO ExBudget

runUPlcFull ∷ ToUPlc a DefaultUni DefaultFun ⇒ [a] → ExceptT SomeException IO (Term Name DefaultUni DefaultFun (), ExBudget, [Text])

runUPlcLogs ∷ ToUPlc a DefaultUni DefaultFun ⇒ [a] → ExceptT SomeException IO [Text]

ppCatch ∷ PrettyPlc a ⇒ ExceptT SomeException IO a → IO (Doc ann)

ppCatchReadable ∷ ∀ a ann. PrettyBy (PrettyConfigReadable PrettyConfigName) a ⇒ ExceptT SomeException IO a → IO (Doc ann)

goldenTPlc ∷ ToTPlc a DefaultUni DefaultFun ⇒ TestName → a → TestNested

goldenTPlcReadable ∷ ToTPlc a DefaultUni DefaultFun ⇒ TestName → a → TestNested

goldenUPlc ∷ ToUPlc a DefaultUni DefaultFun ⇒ TestName → a → TestNested

goldenUPlcReadable ∷ ToUPlc a DefaultUni DefaultFun ⇒ TestName → a → TestNested

goldenTEval ∷ ToTPlc a DefaultUni DefaultFun ⇒ TestName → [a] → TestNested

goldenUEval ∷ ToUPlc a DefaultUni DefaultFun ⇒ TestName → [a] → TestNested

goldenUEvalLogs ∷ ToUPlc a DefaultUni DefaultFun ⇒ TestName → [a] → TestNested

goldenUEvalProfile ∷ ToUPlc a DefaultUni DefaultFun ⇒ TestName → [a] → TestNested

This is mostly useful for profiling a test that is normally tested with one of the other functions, as it's a drop-in replacement and you can then pass the output into traceToStacks.

goldenUEvalProfile' ∷ ToUPlc a DefaultUni DefaultFun ⇒ TestName → [a] → TestNested

This is mostly useful for profiling a test that is normally tested with one of the other functions, as it's a drop-in replacement and you can then pass the output into traceToStacks.

goldenUEvalBudget ∷ ToUPlc a DefaultUni DefaultFun ⇒ TestName → [a] → TestNested

goldenSize ∷ ToUPlc a DefaultUni DefaultFun ⇒ TestName → a → TestNested

initialSrcSpan ∷ FilePath → SrcSpan

A made-up SrcSpan for testing.

topSrcSpan ∷ SrcSpan

newtype NoMarkRenameT ren m a

A version of RenameT that fails to take free variables into account.

Constructors

NoMarkRenameT

Fields

unNoMarkRenameT ∷ RenameT ren m a

noMarkRename ∷ Monoid ren ⇒ (t → NoMarkRenameT ren m t) → t → m t

newtype NoRenameT (ren ∷ Type) m a

A version of RenameT that does not perform any renaming at all.

Constructors

NoRenameT

Fields

unNoRenameT ∷ m a

noRename ∷ MonadQuote m ⇒ (t → m ()) → (t → NoRenameT ren m t) → t → m t

newtype BrokenRenameT ren m a

A broken version of RenameT whose local updates the scope globally (as opposed to locally).

Constructors

BrokenRenameT

Fields

unBrokenRenameT ∷ StateT ren m a

runBrokenRenameT ∷ (Monad m, Monoid ren) ⇒ BrokenRenameT ren m a → m a

brokenRename ∷ (MonadQuote m, Monoid ren) ⇒ (t → m ()) → (t → BrokenRenameT ren m t) → t → m t

data Prerename

Determines whether to perform renaming before running the scoping tests. Needed for passes that don't perform renaming themselves.

Constructors

PrerenameYes

PrerenameNo

data BindingRemoval

Whether it's OK if the pass removes bindings. A renamer isn't supposed to do that, but for example an inliner may do it, since it's basically the entire point of an inliner.

Constructors

BindingRemovalOk

BindingRemovalNotOk

prop_scopingFor

Arguments

∷ (PrettyPlc (t NameAnn), Rename (t NameAnn), Scoping t)

⇒ AstGen (t ann)
A generator of types/terms/programs.

→ BindingRemoval
Whether binding removal is expected for the pass.

→ Prerename
Whether renaming is required before running the scoping tests. Note that the scoping tests rely on global uniqueness themselves, hence for any pass that doesn't perform renaming internally this needs to be PrerenameYes.

→ (t NameAnn → Quote (t NameAnn))
The runner of the pass.

→ Property

Test scoping for a renamer.

test_scopingGood

Arguments

∷ (PrettyPlc (t NameAnn), Rename (t NameAnn), Scoping t)

⇒ String
The name of the pass we're about to test.

→ AstGen (t ann)
A generator of types/terms/programs.

→ BindingRemoval
Whether binding removal is expected for the pass.

→ Prerename
Whether renaming is required before running the scoping tests. Note that the scoping tests rely on global uniqueness themselves, hence for any pass that doesn't perform renaming internally this needs to be PrerenameYes.

→ (t NameAnn → Quote (t NameAnn))
The runner of the pass.

→ TestTree

Test that a pass does not break global uniqueness.

test_scopingBad

Arguments

∷ (PrettyPlc (t NameAnn), Rename (t NameAnn), Scoping t)

⇒ String
The name of the pass we're about to test.

→ AstGen (t ann)
A generator of types/terms/programs.

→ BindingRemoval
Whether binding removal is expected for the pass.

→ Prerename
Whether renaming is required before running the scoping tests. Note that the scoping tests rely on global uniqueness themselves, hence for any pass that doesn't perform renaming internally this needs to be PrerenameYes.

→ (t NameAnn → Quote (t NameAnn))
The runner of the pass.

→ TestTree

Test that a pass breaks global uniqueness.

test_scopingSpoilRenamer ∷ (PrettyPlc (t NameAnn), Rename (t NameAnn), Scoping t, Monoid ren) ⇒ AstGen (t ann) → (t NameAnn → Quote ()) → (∀ m. (MonadQuote m, MonadReader ren m) ⇒ t NameAnn → m (t NameAnn)) → TestTree

Test that the scoping machinery fails when the given renamer is spoiled in some way (e.g. marking is removed) to ensure that the machinery does catch bugs.

Tasty extras

module Test.Tasty.Extras

Orphan instances

Monoid SrcSpan

Instance details

Methods

mempty ∷ SrcSpan

mappend ∷ SrcSpan → SrcSpan → SrcSpan

mconcat ∷ [SrcSpan] → SrcSpan

Semigroup SrcSpan

Instance details

Methods

(<>) ∷ SrcSpan → SrcSpan → SrcSpan

sconcat ∷ NonEmpty SrcSpan → SrcSpan

stimes ∷ Integral b ⇒ b → SrcSpan → SrcSpan