PlutusCore.TypeCheck

Description

Kindtype inferencechecking.

Documentation

class ToKind (uni ∷ Type → Type)

For computing the Plutus kind of a built-in type. See kindOfBuiltinType.

Minimal complete definition

toSingKind

type MonadKindCheck err term uni fun ann m = (MonadError err m, ToKind uni)

The constraints that are required for kind checking.

type MonadTypeCheck err term uni fun ann m = (MonadKindCheck err term uni fun ann m, MonadNormalizeType uni m, AnnotateCaseBuiltin uni, GEq uni, Ix fun)

The general constraints that are required for type checking a Plutus AST.

type TypeErrorPlc uni fun ann = TypeError (Term TyName Name uni fun ()) uni fun ann

The PLC type error type.

type Typecheckable uni fun = (ToKind uni, HasUniApply uni, ToBuiltinMeaning uni fun, AnnotateCaseBuiltin uni)

The constraint for built-in types/functions are kind/type-checkable.

We keep this separate from MonadKindCheck/MonadTypeCheck, because those mainly constrain the monad and Typecheckable constraints only the builtins. In particular useful when the monad gets instantiated and builtins don't. Another reason is that Typecheckable is not required during type checking, since it's only needed for computing BuiltinTypes, which is passed as a regular argument to the worker of the type checker.

Configuration.

newtype BuiltinTypes uni fun

Mapping from Builtins to their Normalized Kinds.

Constructors

BuiltinTypes

Fields

unBuiltinTypes ∷ Array fun (Dupable (Normalized (Type TyName uni ())))

newtype KindCheckConfig

Configuration of the kind checker.

Constructors

KindCheckConfig

Fields

_kccHandleNameMismatches ∷ HandleNameMismatches

data TypeCheckConfig uni fun

Configuration of the type checker.

Constructors

TypeCheckConfig

Fields

_tccKindCheckConfig ∷ KindCheckConfig

_tccBuiltinTypes ∷ BuiltinTypes uni fun

tccBuiltinTypes ∷ HasTypeCheckConfig cfg uni fun ⇒ Lens' cfg (BuiltinTypes uni fun)

defKindCheckConfig ∷ KindCheckConfig

The default kind checking config.

builtinMeaningsToTypes ∷ (MonadKindCheck (TypeError term uni fun ann) term uni fun ann m, Typecheckable uni fun) ⇒ BuiltinSemanticsVariant fun → ann → m (BuiltinTypes uni fun)

Extract the TypeScheme from a BuiltinMeaning and convert it to the corresponding Kind for each built-in function.

getDefTypeCheckConfig ∷ (MonadKindCheck (TypeError term uni fun ann) term uni fun ann m, Typecheckable uni fun) ⇒ ann → m (TypeCheckConfig uni fun)

Get the default type checking config.

Kindtype inferencechecking.

inferKind ∷ MonadKindCheck (TypeError term uni fun ann) term uni fun ann m ⇒ KindCheckConfig → Type TyName uni ann → m (Kind ())

Infer the kind of a type.

checkKind ∷ MonadKindCheck (TypeError term uni fun ann) term uni fun ann m ⇒ KindCheckConfig → ann → Type TyName uni ann → Kind () → m ()

Check a type against a kind. Infers the kind of the type and checks that it's equal to the given kind throwing a TypeError (annotated with the value of the ann argument) otherwise.

inferType ∷ MonadTypeCheckPlc uni fun ann m ⇒ TypeCheckConfig uni fun → Term TyName Name uni fun ann → m (Normalized (Type TyName uni ()))

Infer the type of a term.

checkType ∷ MonadTypeCheckPlc uni fun ann m ⇒ TypeCheckConfig uni fun → ann → Term TyName Name uni fun ann → Normalized (Type TyName uni ()) → m ()

Check a term against a type. Infers the type of the term and checks that it's equal to the given type throwing a TypeError (annotated with the value of the ann argument) otherwise.

inferTypeOfProgram ∷ MonadTypeCheckPlc uni fun ann m ⇒ TypeCheckConfig uni fun → Program TyName Name uni fun ann → m (Normalized (Type TyName uni ()))

Infer the type of a program.

checkTypeOfProgram ∷ MonadTypeCheckPlc uni fun ann m ⇒ TypeCheckConfig uni fun → ann → Program TyName Name uni fun ann → Normalized (Type TyName uni ()) → m ()

Check a program against a type. Infers the type of the program and checks that it's equal to the given type throwing a TypeError (annotated with the value of the ann argument) otherwise.