PlutusCore.Value

Documentation

data Value

The underlying type of the UPLC built-in type Value.

type NestedMap = Map ByteString (Map ByteString Integer)

unpack ∷ Value → NestedMap

Unpack a Value into a map from (currency symbol, token name) to amount.

The map is guaranteed to not contain empty inner map or zero amount.

pack ∷ NestedMap → Value

Pack a map from (currency symbol, token name) to amount into a Value.

The map will be filtered so that it does not contain empty inner map or zero amount.

empty ∷ Value

fromList ∷ [(ByteString, [(ByteString, Integer)])] → Value

toList ∷ Value → [(ByteString, [(ByteString, Integer)])]

toFlatList ∷ Value → [(ByteString, ByteString, Integer)]

totalSize ∷ Value → Int

Total size, i.e., the number of distinct `(currency symbol, token name)` pairs contained in the Value.

maxInnerSize ∷ Value → Int

Size of the largest inner map.

insertCoin ∷ ByteString → ByteString → Integer → Value → Value

\(O(\log \max(m, k))\), where \(m\) is the size of the outer map, and \(k\) is the size of the largest inner map.

deleteCoin ∷ ByteString → ByteString → Value → Value

\(O(\log \max(m, k))\)

lookupCoin ∷ ByteString → ByteString → Value → Integer

\(O(\log \max(m, k))\)

valueContains ∷ Value → Value → Bool

\(O(n_{2}\log \max(m_{1}, k_{1}))\), where \(n_{2}\) is the total size of the second Value, \(m_{1}\) is the size of the outer map in the first Value and \(k_{1}\) is the size of the largest inner map in the first Value.

a contains b if for each (currency, token, amount) in b, if amount > 0, then lookup currency token a >= amount, and if amount < 0, then lookup currency token a == amount.

unionValue ∷ Value → Value → Value

The precise complexity is complicated, but an upper bound is \(O(n_{1} \log n_{2}) + O(m)\), where \(n_{1}\) is the total size of the smaller value, \(n_{2}\) is the total size of the bigger value, and \(m\) is the combined size of the outer maps.

valueData ∷ Value → Data

\(O(n)\). Encodes Value as Data, in the same way as non-builtin Value. This is the denotation of ValueData in Plutus V1, V2 and V3.

unValueData ∷ Data → BuiltinResult Value

\(O(n \log n)\). Decodes Data into Value, in the same way as non-builtin Value. This is the denotation of UnValueData in Plutus V1, V2 and V3.