PlutusCore.Version

Documentation

data Version

The version of Plutus Core used by this program.

The intention is to convey different levels of backwards compatibility for existing scripts: - Major version changes are backwards-incompatible - Minor version changes are backwards-compatible - Patch version changes should be entirely invisible (and we will likely not use this level)

The version used should be changed only when the language itself changes. For example, adding a new kind of term to the language would require a minor version bump; removing a kind of term would require a major version bump.

Similarly, changing the semantics of the language will require a version bump, typically a major one. This is the main reason why the version is actually tracked in the AST: we can have two language versions with identical ASTs but different semantics, so we need to track the version explicitly.

Compatibility is about compatibility for specific scripts, not about e.g. tools which consume scripts. Adding a new kind of term does not change how existing scripts behave, but does change what tools would need to do to process scripts.

Constructors

Version

Fields

_versionMajor ∷ Natural

_versionMinor ∷ Natural

_versionPatch ∷ Natural

versionMajor ∷ Lens' Version Natural

versionMinor ∷ Lens' Version Natural

versionPatch ∷ Lens' Version Natural

plcVersion100 ∷ Version

Plutus Core version 1.0.0

plcVersion110 ∷ Version

Plutus Core version 1.1.0

firstVersion ∷ Version

The first version of Plutus Core supported by this library.

latestVersion ∷ Version

The latest version of Plutus Core supported by this library.

knownVersions ∷ Set Version

The set of versions that are "known", i.e. that have been released and have actual differences associated with them.