PlutusIR.Compiler

Documentation

compileProgram ∷ Compiling m uni fun a ⇒ Program TyName Name uni fun a → m (PLCProgram uni fun a)

compileToReadable ∷ ∀ m uni fun a b. (Compiling m uni fun a, b ~ Provenance a) ⇒ Program TyName Name uni fun b → m (Program TyName Name uni fun b)

The 1st half of the PIR compiler pipeline up to floating/merging the lets. We stop momentarily here to give a chance to the tx-plugin to dump a "readable" version of pir (i.e. floated).

compileReadableToPlc ∷ ∀ m uni fun a b. (Compiling m uni fun a, b ~ Provenance a) ⇒ Program TyName Name uni fun b → m (PLCProgram uni fun a)

The 2nd half of the PIR compiler pipeline. Compiles a Term into a PLC Term, by removing/translating step-by-step the PIR's language constructs to PLC. Note: the result *does* have globally unique names.

type Compiling m uni fun a = (Monad m, MonadReader (CompilationCtx uni fun a) m, MonadError (Error uni fun (Provenance a)) m, AnnotateCaseBuiltin uni, CaseBuiltin uni, MonadQuote m, Ord a, AnnInline a, Typecheckable uni fun, GEq uni, PrettyUni uni, Pretty fun, Pretty a)

data Error uni fun a

Constructors

CompilationError !a !Text
A generic compilation error.

UnsupportedError !a !Text
An error relating specifically to an unsupported feature.

OptionsError !Text
An error relating to compilation options.

PLCError !(Error uni fun a)
An error from running some PLC function, lifted into this error type for convenience.

PLCTypeError !(TypeError (Term TyName Name uni fun ()) uni fun a)

PIRTypeError !(TypeErrorExt uni a)

data Provenance a

Indicates where a value comes from.

This is either an original annotation or a pieces of context explaining how the term relates to a previous Provenance. We also provide noProvenance for convenience.

The provenance should always be just the original annotation, if we have one. It should only be another kind of provenance if we're in the process of generating some term that doesn't correspond directly to a term in the original AST.

Constructors

Original a

LetBinding Recursivity (Provenance a)

TermBinding String (Provenance a)

TypeBinding String (Provenance a)

DatatypeComponent DatatypeComponent (Provenance a)

MultipleSources (Set (Provenance a))
Added for accumulating difference provenances when floating lets

data DatatypeComponent

Constructors

Constructor

ConstructorType

Destructor

DestructorType

DatatypeType

PatternFunctor

noProvenance ∷ Provenance a

data CompilationOpts a

Constructors

CompilationOpts

Fields

_coOptimize ∷ Bool

_coTypecheck ∷ Bool

_coPedantic ∷ Bool

_coVerbose ∷ Bool

_coDebug ∷ Bool

_coDatatypes ∷ DatatypeCompilationOpts

_coMaxSimplifierIterations ∷ Int

_coDoSimplifierUnwrapCancel ∷ Bool

_coDoSimplifierCaseReduce ∷ Bool

_coDoSimplifierRewrite ∷ Bool

_coDoSimplifierBeta ∷ Bool

_coDoSimplifierInline ∷ Bool

_coDoSimplifierKnownCon ∷ Bool

_coDoSimplifierCaseOfCase ∷ Bool

_coDoSimplifierEvaluateBuiltins ∷ Bool

_coDoSimplifierStrictifyBindings ∷ Bool

_coDoSimplifierRemoveDeadBindings ∷ Bool

_coInlineHints ∷ InlineHints Name (Provenance a)

_coInlineConstants ∷ Bool

_coInlineFix ∷ Bool

_coInlineCallsiteGrowth ∷ Size

_coProfile ∷ Bool

_coRelaxedFloatin ∷ Bool

_coCaseOfCaseConservative ∷ Bool

_coPreserveLogging ∷ Bool

Whether to try and preserve the logging beahviour of the program.

coOptimize ∷ ∀ a. Lens' (CompilationOpts a) Bool

coTypecheck ∷ ∀ a. Lens' (CompilationOpts a) Bool

coPedantic ∷ ∀ a. Lens' (CompilationOpts a) Bool

coVerbose ∷ ∀ a. Lens' (CompilationOpts a) Bool

coDebug ∷ ∀ a. Lens' (CompilationOpts a) Bool

coMaxSimplifierIterations ∷ ∀ a. Lens' (CompilationOpts a) Int

coDoSimplifierUnwrapCancel ∷ ∀ a. Lens' (CompilationOpts a) Bool

coDoSimplifierBeta ∷ ∀ a. Lens' (CompilationOpts a) Bool

coDoSimplifierInline ∷ ∀ a. Lens' (CompilationOpts a) Bool

coDoSimplifierEvaluateBuiltins ∷ ∀ a. Lens' (CompilationOpts a) Bool

coDoSimplifierStrictifyBindings ∷ ∀ a. Lens' (CompilationOpts a) Bool

coDoSimplifierRewrite ∷ ∀ a. Lens' (CompilationOpts a) Bool

coDoSimplifierKnownCon ∷ ∀ a. Lens' (CompilationOpts a) Bool

coInlineConstants ∷ ∀ a. Lens' (CompilationOpts a) Bool

coInlineFix ∷ ∀ a. Lens' (CompilationOpts a) Bool

coInlineHints ∷ ∀ a a. Lens (CompilationOpts a) (CompilationOpts a) (InlineHints Name (Provenance a)) (InlineHints Name (Provenance a))

coInlineCallsiteGrowth ∷ ∀ a. Lens' (CompilationOpts a) Size

coProfile ∷ ∀ a. Lens' (CompilationOpts a) Bool

coRelaxedFloatin ∷ ∀ a. Lens' (CompilationOpts a) Bool

coCaseOfCaseConservative ∷ ∀ a. Lens' (CompilationOpts a) Bool

coPreserveLogging ∷ ∀ a. Lens' (CompilationOpts a) Bool

coDatatypes ∷ ∀ a. Lens' (CompilationOpts a) DatatypeCompilationOpts

dcoStyle ∷ Iso' DatatypeCompilationOpts DatatypeStyle

data DatatypeStyle

What style to use when encoding datatypes. Generally, SumsOfProducts is superior, unless you are targeting an old Plutus Core language version.

See Note [Encoding of datatypes]

Constructors

ScottEncoding

SumsOfProducts

BuiltinCasing
A temporary data type style used to make a couple of V3 ledger-api-test tests pass before we can support casing on values of built-in types in newer protocol versions and merge this into SumsOfProducts (which is what controls whether Case is available or not).

defaultCompilationOpts ∷ CompilationOpts a

data CompilationCtx uni fun a

ccOpts ∷ ∀ uni fun a. Lens' (CompilationCtx uni fun a) (CompilationOpts a)

ccEnclosing ∷ ∀ uni fun a. Lens' (CompilationCtx uni fun a) (Provenance a)

ccTypeCheckConfig ∷ ∀ uni fun a. Lens' (CompilationCtx uni fun a) (PirTCConfig uni fun)

ccBuiltinsInfo ∷ ∀ uni fun a. Lens' (CompilationCtx uni fun a) (BuiltinsInfo uni fun)

ccBuiltinCostModel ∷ ∀ uni fun a. Lens' (CompilationCtx uni fun a) (CostingPart uni fun)

data PirTCConfig uni fun

extending the plc typecheck config with AllowEscape

Constructors

PirTCConfig

Fields

_pirConfigTCConfig ∷ TypeCheckConfig uni fun

_pirConfigAllowEscape ∷ AllowEscape

data AllowEscape

Extra flag to be passed in the TypeCheckM Reader context, to signal if the PIR expression currently being typechecked is at the top-level and thus its type can escape, or nested and thus not allowed to escape.

Constructors

YesEscape

NoEscape

toDefaultCompilationCtx ∷ (Default (BuiltinsInfo uni fun), Default (CostingPart uni fun), Default (RewriteRules uni fun)) ⇒ TypeCheckConfig uni fun → CompilationCtx uni fun a

runCompilerPass ∷ (Compiling m uni fun a, b ~ Provenance a) ⇒ m (Pass m tyname name uni fun b) → Term tyname name uni fun b → m (Term tyname name uni fun b)

simplifier ∷ Compiling m uni fun a ⇒ m (Pass m TyName Name uni fun (Provenance a))