PlutusIR.Compiler.Recursion

Description

Functions for compiling PIR recursive let-bound functions into PLC.

Documentation

compileRecTerms ∷ Compiling m uni fun a ⇒ PIRTerm uni fun a → NonEmpty (TermDef TyName Name uni fun (Provenance a)) → DefT SharedName uni fun (Provenance a) m (PIRTerm uni fun a)

Compile a mutually recursive list of var decls bound in a body.

mkFixpoint ∷ ∀ m uni fun a. Compiling m uni fun a ⇒ NonEmpty (TermDef TyName Name uni fun (Provenance a)) → DefT SharedName uni fun (Provenance a) m (Tuple (Term TyName Name uni fun) uni (Provenance a))

Given a list of var decls, create a tuples of values that computes their mutually recursive fixpoint.