PlutusIR.Compiler.Types

Documentation

data AllowEscape

Extra flag to be passed in the TypeCheckM Reader context, to signal if the PIR expression currently being typechecked is at the top-level and thus its type can escape, or nested and thus not allowed to escape.

Constructors

YesEscape

NoEscape

data PirTCConfig uni fun

extending the plc typecheck config with AllowEscape

Constructors

PirTCConfig

Fields

_pirConfigTCConfig ∷ TypeCheckConfig uni fun

_pirConfigAllowEscape ∷ AllowEscape

pirConfigTCConfig ∷ ∀ uni fun uni fun. Lens (PirTCConfig uni fun) (PirTCConfig uni fun) (TypeCheckConfig uni fun) (TypeCheckConfig uni fun)

pirConfigAllowEscape ∷ ∀ uni fun. Lens' (PirTCConfig uni fun) AllowEscape

data DatatypeStyle

What style to use when encoding datatypes. Generally, SumsOfProducts is superior, unless you are targeting an old Plutus Core language version.

See Note [Encoding of datatypes]

Constructors

ScottEncoding

SumsOfProducts

BuiltinCasing
A temporary data type style used to make a couple of V3 ledger-api-test tests pass before we can support casing on values of built-in types in newer protocol versions and merge this into SumsOfProducts (which is what controls whether Case is available or not).

newtype DatatypeCompilationOpts

Constructors

DatatypeCompilationOpts

Fields

_dcoStyle ∷ DatatypeStyle

dcoStyle ∷ Iso' DatatypeCompilationOpts DatatypeStyle

defaultDatatypeCompilationOpts ∷ DatatypeCompilationOpts

data CompilationOpts a

Constructors

CompilationOpts

Fields

_coOptimize ∷ Bool

_coTypecheck ∷ Bool

_coPedantic ∷ Bool

_coVerbose ∷ Bool

_coDebug ∷ Bool

_coDatatypes ∷ DatatypeCompilationOpts

_coMaxSimplifierIterations ∷ Int

_coDoSimplifierUnwrapCancel ∷ Bool

_coDoSimplifierCaseReduce ∷ Bool

_coDoSimplifierRewrite ∷ Bool

_coDoSimplifierBeta ∷ Bool

_coDoSimplifierInline ∷ Bool

_coDoSimplifierKnownCon ∷ Bool

_coDoSimplifierCaseOfCase ∷ Bool

_coDoSimplifierEvaluateBuiltins ∷ Bool

_coDoSimplifierStrictifyBindings ∷ Bool

_coDoSimplifierRemoveDeadBindings ∷ Bool

_coInlineHints ∷ InlineHints Name (Provenance a)

_coInlineConstants ∷ Bool

_coInlineFix ∷ Bool

_coInlineCallsiteGrowth ∷ Size

_coProfile ∷ Bool

_coRelaxedFloatin ∷ Bool

_coCaseOfCaseConservative ∷ Bool

_coPreserveLogging ∷ Bool

Whether to try and preserve the logging beahviour of the program.

coVerbose ∷ ∀ a. Lens' (CompilationOpts a) Bool

coTypecheck ∷ ∀ a. Lens' (CompilationOpts a) Bool

coRelaxedFloatin ∷ ∀ a. Lens' (CompilationOpts a) Bool

coProfile ∷ ∀ a. Lens' (CompilationOpts a) Bool

coPreserveLogging ∷ ∀ a. Lens' (CompilationOpts a) Bool

coPedantic ∷ ∀ a. Lens' (CompilationOpts a) Bool

coOptimize ∷ ∀ a. Lens' (CompilationOpts a) Bool

coMaxSimplifierIterations ∷ ∀ a. Lens' (CompilationOpts a) Int

coInlineHints ∷ ∀ a a. Lens (CompilationOpts a) (CompilationOpts a) (InlineHints Name (Provenance a)) (InlineHints Name (Provenance a))

coInlineFix ∷ ∀ a. Lens' (CompilationOpts a) Bool

coInlineConstants ∷ ∀ a. Lens' (CompilationOpts a) Bool

coInlineCallsiteGrowth ∷ ∀ a. Lens' (CompilationOpts a) Size

coDoSimplifierUnwrapCancel ∷ ∀ a. Lens' (CompilationOpts a) Bool

coDoSimplifierStrictifyBindings ∷ ∀ a. Lens' (CompilationOpts a) Bool

coDoSimplifierRewrite ∷ ∀ a. Lens' (CompilationOpts a) Bool

coDoSimplifierRemoveDeadBindings ∷ ∀ a. Lens' (CompilationOpts a) Bool

coDoSimplifierKnownCon ∷ ∀ a. Lens' (CompilationOpts a) Bool

coDoSimplifierInline ∷ ∀ a. Lens' (CompilationOpts a) Bool

coDoSimplifierEvaluateBuiltins ∷ ∀ a. Lens' (CompilationOpts a) Bool

coDoSimplifierCaseReduce ∷ ∀ a. Lens' (CompilationOpts a) Bool

coDoSimplifierCaseOfCase ∷ ∀ a. Lens' (CompilationOpts a) Bool

coDoSimplifierBeta ∷ ∀ a. Lens' (CompilationOpts a) Bool

coDebug ∷ ∀ a. Lens' (CompilationOpts a) Bool

coDatatypes ∷ ∀ a. Lens' (CompilationOpts a) DatatypeCompilationOpts

coCaseOfCaseConservative ∷ ∀ a. Lens' (CompilationOpts a) Bool

defaultCompilationOpts ∷ CompilationOpts a

data CompilationCtx uni fun a

Constructors

CompilationCtx

Fields

_ccOpts ∷ CompilationOpts a

_ccEnclosing ∷ Provenance a

_ccTypeCheckConfig ∷ PirTCConfig uni fun

Decide to either typecheck (passing a specific tcconfig) or not by passing Nothing.

_ccBuiltinsInfo ∷ BuiltinsInfo uni fun

_ccBuiltinCostModel ∷ CostingPart uni fun

_ccRewriteRules ∷ RewriteRules uni fun

ccTypeCheckConfig ∷ ∀ uni fun a. Lens' (CompilationCtx uni fun a) (PirTCConfig uni fun)

ccRewriteRules ∷ ∀ uni fun a. Lens' (CompilationCtx uni fun a) (RewriteRules uni fun)

ccOpts ∷ ∀ uni fun a. Lens' (CompilationCtx uni fun a) (CompilationOpts a)

ccEnclosing ∷ ∀ uni fun a. Lens' (CompilationCtx uni fun a) (Provenance a)

ccBuiltinsInfo ∷ ∀ uni fun a. Lens' (CompilationCtx uni fun a) (BuiltinsInfo uni fun)

ccBuiltinCostModel ∷ ∀ uni fun a. Lens' (CompilationCtx uni fun a) (CostingPart uni fun)

toDefaultCompilationCtx ∷ (Default (BuiltinsInfo uni fun), Default (CostingPart uni fun), Default (RewriteRules uni fun)) ⇒ TypeCheckConfig uni fun → CompilationCtx uni fun a

validateOpts ∷ Compiling m uni fun a ⇒ Version → m ()

getEnclosing ∷ MonadReader (CompilationCtx uni fun a) m ⇒ m (Provenance a)

withEnclosing ∷ MonadReader (CompilationCtx uni fun a) m ⇒ (Provenance a → Provenance a) → m b → m b

runIf ∷ MonadReader (CompilationCtx uni fun a) m ⇒ m Bool → (b → m b) → b → m b

runIfOpts ∷ MonadReader (CompilationCtx uni fun a) m ⇒ (b → m b) → b → m b

type PLCProgram uni fun a = Program TyName Name uni fun (Provenance a)

type PLCTerm uni fun a = Term TyName Name uni fun (Provenance a)

type PLCType uni a = Type TyName uni (Provenance a)

data PLCRecType uni fun a

A possibly recursive type.

Constructors

PlainType (PLCType uni a)

RecursiveType (RecursiveType uni fun (Provenance a))

getType ∷ PLCRecType uni fun a → PLCType uni a

Get the actual type inside a PLCRecType.

wrap ∷ Provenance a → PLCRecType uni fun a → [PLCType uni a] → PIRTerm uni fun a → PIRTerm uni fun a

Wrap a term appropriately for a possibly recursive type.

unwrap ∷ Provenance a → PLCRecType uni fun a → PIRTerm uni fun a → PIRTerm uni fun a

Unwrap a term appropriately for a possibly recursive type.

type PIRTerm uni fun a = Term TyName Name uni fun (Provenance a)

type PIRType uni a = Type TyName uni (Provenance a)

type Compiling m uni fun a = (Monad m, MonadReader (CompilationCtx uni fun a) m, MonadError (Error uni fun (Provenance a)) m, AnnotateCaseBuiltin uni, CaseBuiltin uni, MonadQuote m, Ord a, AnnInline a, Typecheckable uni fun, GEq uni, PrettyUni uni, Pretty fun, Pretty a)

type TermDef tyname name uni fun a = Def (VarDecl tyname name uni a) (Term tyname name uni fun a)

data SharedName

We generate some shared definitions compilation, this datatype defines the "keys" for those definitions.

Constructors

FixpointCombinator Integer

FixBy

toProgramName ∷ SharedName → Quote Name