PlutusIR.Core.Plated

Documentation

termSubterms ∷ Traversal' (Term tyname name uni fun a) (Term tyname name uni fun a)

Get all the direct child Terms of the given Term, including those within Bindings.

termSubtermsDeep ∷ Fold (Term tyname name uni fun ann) (Term tyname name uni fun ann)

Get all the transitive child Terms of the given Term.

termSubtypes ∷ Traversal' (Term tyname name uni fun a) (Type tyname uni a)

Get all the direct child Kinds of the given Term, including those within Bindings.

termSubtypesDeep ∷ Fold (Term tyname name uni fun ann) (Type tyname uni ann)

Get all the transitive child Kinds of the given Term.

termSubkinds ∷ Traversal' (Term tyname name uni fun ann) (Kind ann)

Get all the direct child Kinds of the given Term.

termBindings ∷ Traversal' (Term tyname name uni fun a) (Binding tyname name uni fun a)

Get all the direct child Bindings of the given Term.

termVars ∷ Traversal' (Term tyname name uni fun ann) name

Get all the direct child 'name a's of the given Term from Vars.

termConstants ∷ Traversal' (Term tyname name uni fun ann) (Some (ValueOf uni))

Get all the direct constants of the given Term from Constants.

termConstantsDeep ∷ Fold (Term tyname name uni fun ann) (Some (ValueOf uni))

Get all the transitive child Constants of the given Term.

typeSubtypes ∷ ∀ tyname (uni ∷ Type → Type) ann f. Applicative f ⇒ (Type tyname uni ann → f (Type tyname uni ann)) → Type tyname uni ann → f (Type tyname uni ann)

Get all the direct child Kinds of the given Kind.

typeSubtypesDeep ∷ ∀ tyname (uni ∷ Type → Type) ann f. (Contravariant f, Applicative f) ⇒ (Type tyname uni ann → f (Type tyname uni ann)) → Type tyname uni ann → f (Type tyname uni ann)

Get all the transitive child Kinds of the given Kind.

typeSubkinds ∷ ∀ tyname (uni ∷ Type → Type) ann f. Applicative f ⇒ (Kind ann → f (Kind ann)) → Type tyname uni ann → f (Type tyname uni ann)

Get all the direct child Kinds of the given Kind.

typeUniques ∷ ∀ tyname (uni ∷ Type → Type) ann. HasUniques (Type tyname uni ann) ⇒ Traversal' (Type tyname uni ann) Unique

Get all the direct child Uniques of the given Kind from binders TyVars.

typeUniquesDeep ∷ ∀ tyname (uni ∷ Type → Type) ann. HasUniques (Type tyname uni ann) ⇒ Fold (Type tyname uni ann) Unique

Get all the transitive child Uniques of the given Kind.

datatypeSubtypes ∷ Traversal' (Datatype tyname name uni a) (Type tyname uni a)

Get all the direct child Kinds of the given Datatype.

datatypeSubkinds ∷ Traversal' (Datatype tyname name uni a) (Kind a)

Get all the direct child Kinds of the given Datatype.

datatypeTyNames ∷ Traversal' (Datatype tyname name uni a) tyname

Get all the type-names introduces by a datatype

bindingSubterms ∷ Traversal' (Binding tyname name uni fun a) (Term tyname name uni fun a)

Get all the direct child Terms of the given Binding.

bindingSubtypes ∷ Traversal' (Binding tyname name uni fun a) (Type tyname uni a)

Get all the direct child Kinds of the given Binding.

bindingSubkinds ∷ Traversal' (Binding tyname name uni fun a) (Kind a)

Get all the direct child Kinds of the given Binding.

bindingNames ∷ Traversal' (Binding tyname name uni fun a) name

Get all the names introduces by a binding

bindingTyNames ∷ Traversal' (Binding tyname name uni fun a) tyname

Get all the type-names introduces by a binding

bindingIds ∷ (HasUnique tyname TypeUnique, HasUnique name TermUnique) ⇒ Traversal1' (Binding tyname name uni fun a) Unique

All the identifiers/names introduced by this binding In case of a datatype-binding it has multiple identifiers: the type, constructors, match function

termUniques ∷ HasUniques (Term tyname name uni fun ann) ⇒ Traversal' (Term tyname name uni fun ann) Unique

Get all the direct child Uniques of the given Term (including the type-level ones).

termUniquesDeep ∷ HasUniques (Term tyname name uni fun ann) ⇒ Fold (Term tyname name uni fun ann) Unique

Get all the transitive child Uniques of the given Term (including the type-level ones).

varDeclSubtypes ∷ ∀ tyname name (uni ∷ Type → Type) a f. Applicative f ⇒ (Type tyname uni a → f (Type tyname uni a)) → VarDecl tyname name uni a → f (VarDecl tyname name uni a)

Get all the direct child Kinds of the given VarDecl.

underBinders ∷ Arity → Traversal' (Term tyname name uni fun a) (Term tyname name uni fun a)

Focus on the term under the binders corresponding to the given arity. e.g. for arity [TermParam, TermParam] and term x y -> t it focusses on t.

_Constant ∷ Prism' (Term tyname name uni fun a) (a, Some (ValueOf uni))

View a term as a constant.