PlutusIR.MkPir

Documentation

module PlutusCore.MkPlc

type DatatypeDef tyname name uni a = Def (TyVarDecl tyname a) (Datatype tyname name uni a)

A datatype definition as a type variable.

mkLet ∷ a → Recursivity → [Binding tyname name uni fun a] → Term tyname name uni fun a → Term tyname name uni fun a

Make a let binding, unless the list of bindings is empty, in which case just return the underlying term.