PlutusIR.Test

Documentation

ppCatch ∷ (a → Doc ann) → ExceptT SomeException IO a → IO Text

asIfThrown ∷ Exception e ⇒ Except e a → ExceptT SomeException IO a

Adapt an computation that keeps its errors in an Except into one that looks as if it caught them in IO.

compileWithOpts ∷ (GEq uni, Typecheckable uni fun, CaseBuiltin uni, Ord a, AnnInline a, PrettyUni uni, Pretty fun, Pretty a, Default (BuiltinsInfo uni fun), Default (CostingPart uni fun), Default (RewriteRules uni fun)) ⇒ (CompilationCtx uni fun a → CompilationCtx uni fun a) → Program TyName Name uni fun a → Except (Error uni fun (Provenance a)) (Program TyName Name uni fun (Provenance a))

pTermAsProg ∷ Parser (Program TyName Name DefaultUni DefaultFun SrcSpan)

withGoldenFileM ∷ String → (Text → IO Text) → TestNested

goldenPir ∷ PrettyPlc b ⇒ (a → b) → Parser a → String → TestNested

goldenPirM ∷ ∀ a b. PrettyPlc b ⇒ (a → IO b) → Parser a → String → TestNested

goldenPirUnique ∷ Pretty b ⇒ (a → b) → Parser a → String → TestNested

goldenPirMUnique ∷ ∀ a b. Pretty b ⇒ (a → IO b) → Parser a → String → TestNested

goldenPirDoc ∷ (a → Doc ann) → Parser a → String → TestNested

goldenPirDocM ∷ ∀ a ann. (a → IO (Doc ann)) → Parser a → String → TestNested

goldenPlcFromPir ∷ ToTPlc a DefaultUni DefaultFun ⇒ Parser a → String → TestNested

goldenPlcFromPirScott ∷ (Ord a, Typeable a, Pretty a, AnnInline a, prog ~ Program TyName Name DefaultUni DefaultFun a) ⇒ Parser prog → String → TestNested

goldenNamedUPlcFromPir ∷ ToUPlc a DefaultUni DefaultFun ⇒ Parser a → String → TestNested

goldenEvalPir ∷ ToUPlc a DefaultUni DefaultFun ⇒ Parser a → String → TestNested

goldenTypeFromPir ∷ ∀ a. (Pretty a, Typeable a) ⇒ a → Parser (Term TyName Name DefaultUni DefaultFun a) → String → TestNested

initialSrcSpan ∷ FilePath → SrcSpan

A made-up SrcSpan for testing.

topSrcSpan ∷ SrcSpan

rethrow ∷ ExceptT SomeException IO a → IO a

prettyPlcClassicSimple ∷ PrettyPlc a ⇒ a → Doc ann

Pretty-print a PLC value without unique indices using the classic view.

Orphan instances

AnnInline SrcSpan

Instance details

Methods

annAlwaysInline ∷ SrcSpan

annSafeToInline ∷ SrcSpan

annMayInline ∷ SrcSpan

AnnInline SrcSpans

Instance details

Methods

annAlwaysInline ∷ SrcSpans

annSafeToInline ∷ SrcSpans

annMayInline ∷ SrcSpans

(GEq uni, Typecheckable uni fun, CaseBuiltin uni, PrettyUni uni, Pretty fun, Pretty a, Typeable a, Ord a, AnnInline a, Default (CostingPart uni fun), Default (BuiltinsInfo uni fun), Default (RewriteRules uni fun)) ⇒ ToTPlc (Program TyName Name uni fun a) uni fun

Instance details

Methods

toTPlc ∷ Program TyName Name uni fun a → ExceptT SomeException IO (Program0 TyName Name uni fun ())

(GEq uni, Everywhere uni Eq, Typecheckable uni fun, CaseBuiltin uni, PrettyUni uni, Pretty fun, Hashable fun, Pretty a, Typeable a, Ord a, AnnInline a, Default (CostingPart uni fun), Default (BuiltinsInfo uni fun), Default (RewriteRules uni fun)) ⇒ ToUPlc (Program TyName Name uni fun a) uni fun

Instance details

Methods

toUPlc ∷ Program TyName Name uni fun a → ExceptT SomeException IO (Program0 Name uni fun ())