PlutusIR.Transform.RewriteRules

Documentation

rewriteWith ∷ (Monoid a, t ~ Term tyname Name uni fun a, HasUniques t, MonadQuote m) ⇒ RewriteRules uni fun → t → m t

Rewrite a Term using the given RewriteRules (similar to functions of Term -> Term) Normally the rewrite rules are configured at entrypoint time of the compiler.

It goes without saying that the supplied rewrite rules must be type-preserving. MAYBE: enforce this with a `through typeCheckTerm`?

rewritePass ∷ ∀ m uni fun a. (Typecheckable uni fun, GEq uni, Ord a, MonadQuote m, Monoid a) ⇒ PirTCConfig uni fun → RewriteRules uni fun → Pass m TyName Name uni fun a

rewritePassSC ∷ ∀ m uni fun a. (Typecheckable uni fun, GEq uni, Ord a, MonadQuote m, Monoid a) ⇒ PirTCConfig uni fun → RewriteRules uni fun → Pass m TyName Name uni fun a

data RewriteRules uni fun

A bundle of composed RewriteRules, to be passed at entrypoint of the compiler.

unRewriteRules ∷ RewriteRules uni fun → ∀ tyname m a. (MonadQuote m, Monoid a) ⇒ VarsInfo tyname Name uni a → Term tyname Name uni fun a → m (Term tyname Name uni fun a)

defaultUniRewriteRules ∷ RewriteRules DefaultUni DefaultFun

The rules for the Default Universe/Builtin.