PlutusIR.TypeCheck.Internal

Description

The internal module of the type checker that defines the actual algorithms, but not the user-facing API.

Documentation

newtype BuiltinTypes (uni ∷ Type → Type) fun

Mapping from Builtins to their Normalized Kinds.

Constructors

BuiltinTypes

Fields

unBuiltinTypes ∷ Array fun (Dupable (Normalized (Type TyName uni ())))

data TypeCheckConfig (uni ∷ Type → Type) fun

Configuration of the type checker.

Constructors

TypeCheckConfig

Fields

_tccKindCheckConfig ∷ KindCheckConfig

_tccBuiltinTypes ∷ BuiltinTypes uni fun

type TypeCheckT (uni ∷ Type → Type) fun cfg (m ∷ Type → Type) = ReaderT (TypeCheckEnv uni fun cfg) m

The type checking monad that the type checker runs in. In contains a TypeCheckEnv and allows to throw TypeErrors.

type MonadKindCheck err term (uni ∷ Type → Type) fun ann (m ∷ Type → Type) = (MonadError err m, ToKind uni)

The constraints that are required for kind checking.

type MonadTypeCheck err term (uni ∷ Type → Type) fun ann (m ∷ Type → Type) = (MonadKindCheck err term uni fun ann m, MonadNormalizeType uni m, AnnotateCaseBuiltin uni, GEq uni, Ix fun)

The general constraints that are required for type checking a Plutus AST.

type MonadTypeCheckPir uni fun ann m = MonadTypeCheck (Error uni fun ann) (Term TyName Name uni fun ()) uni fun ann m

The constraints that are required for type checking Plutus IR.

tccBuiltinTypes ∷ HasTypeCheckConfig cfg uni fun ⇒ Lens' cfg (BuiltinTypes uni fun)

data PirTCConfig uni fun

extending the plc typecheck config with AllowEscape

Constructors

PirTCConfig

Fields

_pirConfigTCConfig ∷ TypeCheckConfig uni fun

_pirConfigAllowEscape ∷ AllowEscape

data AllowEscape

Extra flag to be passed in the TypeCheckM Reader context, to signal if the PIR expression currently being typechecked is at the top-level and thus its type can escape, or nested and thus not allowed to escape.

Constructors

YesEscape

NoEscape

inferTypeM ∷ ∀ m uni fun ann. MonadTypeCheckPir uni fun ann m ⇒ Term TyName Name uni fun ann → PirTCEnv uni fun m (Normalized (Type TyName uni ()))

Synthesize the type of a term, returning a normalized type.

checkTypeM ∷ MonadTypeCheckPir uni fun ann m ⇒ ann → Term TyName Name uni fun ann → Normalized (Type TyName uni ()) → PirTCEnv uni fun m ()

Check a Term against a NormalizedType.

runTypeCheckM ∷ PirTCConfig uni fun → PirTCEnv uni fun m a → m a

Run a TypeCheckM computation by supplying a TypeCheckConfig to it. Differs from its PLC version in that is passes an extra env flag YesEscape.