PlutusLedgerApi.Common

Description

The types and functions that are common among all ledger Plutus versions.

Script (de)serialization

type SerialisedScript = ShortByteString

Scripts to the ledger are serialised bytestrings.

data ScriptForEvaluation

A Plutus script ready to be evaluated on-chain, via evaluateScriptRestricting.

serialisedScript ∷ ScriptForEvaluation → SerialisedScript

Get a SerialisedScript from a ScriptForEvaluation. O(1).

deserialisedScript ∷ ScriptForEvaluation → ScriptNamedDeBruijn

Get a ScriptNamedDeBruijn from a ScriptForEvaluation. O(1).

serialiseCompiledCode ∷ ∀ a. CompiledCode a → SerialisedScript

Turns a program which was compiled using the 'PlutusTx' toolchain into a binary format that is understood by the network and can be stored on-chain.

serialiseUPLC ∷ Program DeBruijn DefaultUni DefaultFun () → SerialisedScript

Turns a program's AST (most likely manually constructed) into a binary format that is understood by the network and can be stored on-chain.

deserialiseScript

Arguments

∷ ∀ m. MonadError ScriptDecodeError m

⇒ PlutusLedgerLanguage
the Plutus ledger language of the script.

→ MajorProtocolVersion
which major protocol version the script was submitted in.

→ SerialisedScript
the script to deserialise.

→ m ScriptForEvaluation

The deserialization from a serialised script into a ScriptForEvaluation, ready to be evaluated on-chain. Called inside phase-1 validation (i.e., deserialisation error is a phase-1 error).

uncheckedDeserialiseUPLC ∷ SerialisedScript → Program DeBruijn DefaultUni DefaultFun ()

Deserialises a SerialisedScript back into an AST. Does *not* do ledger-language-version-specific checks like for allowable builtins.

data ScriptDecodeError

An error that occurred during script deserialization.

Constructors

CBORDeserialiseError !DeserialiseFailureInfo
an error from the underlying CBOR/serialise library

RemainderError !ByteString
Script was successfully parsed, but more (runaway) bytes encountered after script's position

LedgerLanguageNotAvailableError
the plutus version of the given script is not enabled yet

Fields

sdeAffectedLang ∷ !PlutusLedgerLanguage

the script's ledger language

sdeIntroPv ∷ !MajorProtocolVersion

the major protocol version that will first introduce/enable the ledger language

sdeThisPv ∷ !MajorProtocolVersion

the current protocol version

PlutusCoreLanguageNotAvailableError

Fields

sdeAffectedVersion ∷ !Version

the Plutus Core language of the script under execution.

sdeThisLang ∷ !PlutusLedgerLanguage

the Plutus ledger language of the script under execution.

sdeThisPv ∷ !MajorProtocolVersion

the current protocol version

newtype ScriptNamedDeBruijn

A script with named de-bruijn indices.

Constructors

ScriptNamedDeBruijn (Program NamedDeBruijn DefaultUni DefaultFun ())

Script evaluation

evaluateScriptCounting

Arguments

∷ PlutusLedgerLanguage
The Plutus ledger language of the script under execution.

→ MajorProtocolVersion
Which major protocol version to run the operation in

→ VerboseMode
Whether to produce log output

→ EvaluationContext
Includes the cost model to use for tallying up the execution costs

→ ScriptForEvaluation
The script to evaluate

→ [Data]
The arguments to the script

→ (LogOutput, Either EvaluationError ExBudget)

Evaluates a script, returning the minimum budget that the script would need to evaluate successfully. This will take as long as the script takes, if you need to limit the execution time of the script also, you can use evaluateScriptRestricting, which also returns the used budget.

Note: Parameterized over the ledger-plutus-version since the builtins allowed (during decoding) differs.

evaluateScriptRestricting

Arguments

∷ PlutusLedgerLanguage
The Plutus ledger language of the script under execution.

→ MajorProtocolVersion
Which major protocol version to run the operation in

→ VerboseMode
Whether to produce log output

→ EvaluationContext
Includes the cost model to use for tallying up the execution costs

→ ExBudget
The resource budget which must not be exceeded during evaluation

→ ScriptForEvaluation
The script to evaluate

→ [Data]
The arguments to the script

→ (LogOutput, Either EvaluationError ExBudget)

Evaluates a script, with a cost model and a budget that restricts how many resources it can use according to the cost model. Also returns the budget that was actually used.

Can be used to calculate budgets for scripts, but even in this case you must give a limit to guard against scripts that run for a long time or loop.

Note: Parameterized over the LedgerPlutusVersion since 1. The builtins allowed (during decoding) differ, and 2. The Plutus language versions allowed differ.

evaluateTerm ∷ ExBudgetMode cost DefaultUni DefaultFun → MajorProtocolVersion → VerboseMode → EvaluationContext → Term NamedDeBruijn DefaultUni DefaultFun () → CekReport cost NamedDeBruijn DefaultUni DefaultFun

Evaluate a fully-applied term using the CEK machine. Useful for mimicking the behaviour of the on-chain evaluator.

data VerboseMode

A simple toggle indicating whether or not we should accumulate logs during script execution.

Constructors

Verbose
accumulate all traces

Quiet
don't accumulate anything

type LogOutput = [Text]

The type of the executed script's accumulated log output: a list of Text.

It will be an empty list if the VerboseMode is set to Quiet.

data EvaluationError

Errors that can be thrown when evaluating a Plutus script.

Constructors

CekError !(CekEvaluationException NamedDeBruijn DefaultUni DefaultFun)
An error from the evaluator itself

DeBruijnError !FreeVariableError
An error in the pre-evaluation step of converting from de-Bruijn indices

CodecError !ScriptDecodeError
A deserialisation error TODO: make this error more informative when we have more information about what went wrong

CostModelParameterMismatch
An error indicating that the cost model parameters didn't match what we expected

InvalidReturnValue
The script evaluated to a value that is not a valid return value.

Network's versioning

The network's behaviour (and plutus's by extension) can change via hard forks, which directly correspond to major-number protocol version bumps.

newtype MajorProtocolVersion

This represents the major component of the Cardano protocol version. The ledger can only supply the major component of the protocol version, not the minor component, and Plutus should only need to care about the major component anyway. This relies on careful understanding between us and the ledger as to what this means.

Constructors

MajorProtocolVersion

Fields

getMajorProtocolVersion ∷ Int

data PlutusLedgerLanguage

The Plutus ledger language. These are entirely different script languages from the ledger's perspective, which on our side are interpreted in very similar ways.

It is a simple enumerated datatype (there is no major and minor components as in protocol version) and the ordering of constructors is essential for deriving Enum,Ord,Bounded.

IMPORTANT: this is different from the Plutus Core language version, Version

Constructors

PlutusV1
introduced in Alonzo HF

PlutusV2
introduced in Vasil HF

PlutusV3
introduced in Chang HF

data Version

The version of Plutus Core used by this program.

The intention is to convey different levels of backwards compatibility for existing scripts: - Major version changes are backwards-incompatible - Minor version changes are backwards-compatible - Patch version changes should be entirely invisible (and we will likely not use this level)

The version used should be changed only when the language itself changes. For example, adding a new kind of term to the language would require a minor version bump; removing a kind of term would require a major version bump.

Similarly, changing the semantics of the language will require a version bump, typically a major one. This is the main reason why the version is actually tracked in the AST: we can have two language versions with identical ASTs but different semantics, so we need to track the version explicitly.

Compatibility is about compatibility for specific scripts, not about e.g. tools which consume scripts. Adding a new kind of term does not change how existing scripts behave, but does change what tools would need to do to process scripts.

Constructors

Version

Fields

_versionMajor ∷ Natural

_versionMinor ∷ Natural

_versionPatch ∷ Natural

builtinsIntroducedIn ∷ PlutusLedgerLanguage → Map MajorProtocolVersion (Set DefaultFun)

Given a ledger language, return a map indicating which builtin functions were introduced in which MajorProtocolVersion. This must be updated when new builtins are added. It is not necessary to add entries for protocol versions where no new builtins are added. See Note [New builtins/language versions and protocol versions]

builtinsAvailableIn ∷ PlutusLedgerLanguage → MajorProtocolVersion → Set DefaultFun

Return a set containing the builtins which are available in a given LL in a given PV. All builtins are available in all LLs from pv11PV onwards.

ledgerLanguageIntroducedIn ∷ PlutusLedgerLanguage → MajorProtocolVersion

Query the protocol version that a specific Plutus ledger language was first introduced in.

ledgerLanguagesAvailableIn ∷ MajorProtocolVersion → Set PlutusLedgerLanguage

Which Plutus language versions are available in the given MajorProtocolVersion? See Note [New builtins/language versions and protocol versions]. This function (and others in this module) assumes that once a LL is available it remains available in all later PVs and that if m <= n, PlutusVm is introduced no later than PlutusVn.

Protocol Versions

shelleyPV ∷ MajorProtocolVersion

Shelley era was introduced in protocol version 2.0

allegraPV ∷ MajorProtocolVersion

Allegra era was introduced in protocol version 3.0

maryPV ∷ MajorProtocolVersion

Mary era was introduced in protocol version 4.0

alonzoPV ∷ MajorProtocolVersion

Alonzo era was introduced in protocol version 5.0

vasilPV ∷ MajorProtocolVersion

The Vasil HF introduced the Babbage era and Plutus V2

valentinePV ∷ MajorProtocolVersion

Valentine was an intra-era HF where builtin functions VerifyEcdsaSecp256k1Signature and VerifySchnorrSecp256k1Signature were enabled.

changPV ∷ MajorProtocolVersion

The Chang HF introduced the Conway era and Plutus V3

plominPV ∷ MajorProtocolVersion

The Plomin HF was an intra-era HF where some new builtin functions were introduced in Plutus V2 and V3.

pv11PV ∷ MajorProtocolVersion

Not sure what this is going to be called yet

knownPVs ∷ [MajorProtocolVersion]

The set of protocol versions that are "known", i.e. that have been released and have actual differences associated with them. This is currently only used for testing, so efficiency is not parmount and a list is fine.

Costing-related types

data ExBudget

Constructors

ExBudget

Fields

exBudgetCPU ∷ ExCPU

exBudgetMemory ∷ ExMemory

newtype ExCPU

Counts CPU units in picoseconds: maximum value for SatInt is 2^63 ps, or appproximately 106 days.

Constructors

ExCPU CostingInteger

newtype ExMemory

Counts size in machine words.

Constructors

ExMemory CostingInteger

data SatInt

fromSatInt ∷ Num a ⇒ SatInt → a

An optimized version of fromIntegral . unSatInt.

Network's costing parameters

A less drastic approach (that does not rely on a HF) to affect the network's (and plutus's by extension) behaviour is by tweaking the values of the cost model parameters.

The network does not associate names to cost model parameters; Plutus attaches names to the network's cost model parameters (values) either in a raw textual form or typed by a specific plutus version.

See Note [Cost model parameters]

type CostModelParams = Map Text Int64

A raw representation of the ledger's cost model parameters.

The associated keys/names to the parameter values are arbitrarily set by the plutus team; the ledger does not hold any such names.

See Note [Cost model parameters]

toCostModelParams ∷ IsParamName p ⇒ [(p, Int64)] → CostModelParams

Untags the plutus version from the typed cost model parameters and returns their raw textual form (internally used by CostModelInterface).

assertWellFormedCostModelParams ∷ MonadError CostModelApplyError m ⇒ CostModelParams → m ()

class (Enum a, Bounded a) ⇒ IsParamName a where

A parameter name for different plutus versions.

Each Plutus version should expose such an enumeration as an ADT and create an instance of ParamName out of it.

A valid parameter name has to be enumeration, bounded, ordered, and prettyprintable to a "lower-Kebab" string.

Minimal complete definition

showParamName

Methods

showParamName ∷ a → Text

Produce the raw textual form for a given typed-by-plutus-version cost model parameter Any implementation *must be* an injective function. The GIsParamName generic implementation guarantees injectivity.

readParamName ∷ Text → Maybe a

default implementation that inverts the showParamName operation (not very efficient)

data GenericParamName a

A Generic wrapper for use with deriving via

data CostModelApplyError

A fatal error when trying to create a cost given some plain costmodel parameters.

Constructors

CMUnknownParamError !Text
a costmodel parameter with the give name does not exist in the costmodel to be applied upon

CMInternalReadError
internal error when we are transforming the applyParams' input to json (should not happen)

CMInternalWriteError !String
internal error when we are transforming the applied params from json with given jsonstring error (should not happen)

data CostModelApplyWarn

A non-fatal warning when trying to create a cost given some plain costmodel parameters.

Constructors

CMTooManyParamsWarn
See Note [Cost model parameters from the ledger's point of view]

Fields

cmExpected ∷ !Int

cmActual ∷ !Int

CMTooFewParamsWarn
See Note [Cost model parameters from the ledger's point of view]

Fields

cmExpected ∷ !Int

cmActual ∷ !Int

Evaluation context

data EvaluationContext

An opaque type that contains all the static parameters that the evaluator needs to evaluate a script. This is so that they can be computed once and cached, rather than being recomputed on every evaluation.

Different protocol versions may require different bundles of machine parameters, which allows us for example to tweak the shape of the costing function of a builtin, so that the builtin costs less. Currently this means that we have to create multiple DefaultMachineParameters per language version, which we put into a cache (represented by an association list) in order to avoid costly recomputation of machine parameters.

In order to get the appropriate DefaultMachineParameters at validation time we look it up in the cache using a semantics variant as a key. We compute the semantics variant from the protocol version using the stored function. Note that the semantics variant depends on the language version too, but the latter is known statically (because each language version has its own evaluation context), hence there's no reason to require it to be provided at runtime.

To say it differently, there's a matrix of semantics variants indexed by (LL, PV) pairs and we cache its particular row corresponding to the statically given LL in an EvaluationContext.

The reason why we associate a DefaultMachineParameters with a semantics variant rather than a protocol version are

generally there are far more protocol versions than semantics variants supported by a specific language version, so we save on pointless duplication of bundles of machine parameters

builtins don't know anything about protocol versions, only semantics variants. It is therefore more semantically precise to associate bundles of machine parameters with semantics variants than with protocol versions

Constructors

EvaluationContext

Fields

_evalCtxLedgerLang ∷ PlutusLedgerLanguage

Specifies what language versions the EvaluationContext is for.

_evalCtxCaserBuiltin ∷ MajorProtocolVersion → CaserBuiltin DefaultUni

Specifies how 'case' on values of built-in types works: fails evaluation for older protocol versions and defers to caseBuiltin for newer ones. Note that this function doesn't depend on the PlutusLedgerLanguage or the AST version: deserialisation of a 1.0.0 AST fails upon encountering a Case node anyway, so we can safely assume here that 'case' is available. FIXME: do we need to test that it fails for older PVs? We can't submit transactions in old PVs, so maybe it doesn't matter.

_evalCtxToSemVar ∷ MajorProtocolVersion → BuiltinSemanticsVariant DefaultFun

Specifies how to get a semantics variant for this ledger language given a MajorProtocolVersion.

_evalCtxMachParsCache ∷ [(BuiltinSemanticsVariant DefaultFun, DefaultMachineVariantParameters)]

The cache of DefaultMachineParameters for each semantics variant supported by the current language version.

mkDynEvaluationContext ∷ MonadError CostModelApplyError m ⇒ PlutusLedgerLanguage → (MajorProtocolVersion → CaserBuiltin DefaultUni) → [BuiltinSemanticsVariant DefaultFun] → (MajorProtocolVersion → BuiltinSemanticsVariant DefaultFun) → CostModelParams → m EvaluationContext

Create an EvaluationContext given all builtin semantics variants supported by the provided language version.

The input is a Map of Texts to cost integer values (aka CostModelParams, CostModel) See Note [Inlining meanings of builtins].

IMPORTANT: the toSemVar argument computes the semantics variant for each MajorProtocolVersion and it must only return semantics variants from the semVars list, as well as cover ANY MajorProtocolVersion, including those that do not exist yet (i.e. toSemVar must never fail).

IMPORTANT: The evaluation context of every Plutus version must be recreated upon a protocol update with the updated cost model parameters.

toMachineParameters ∷ MajorProtocolVersion → EvaluationContext → DefaultMachineParameters

mkTermToEvaluate

Arguments

∷ MonadError EvaluationError m

⇒ PlutusLedgerLanguage
the Plutus ledger language of the script under execution.

→ MajorProtocolVersion
which major protocol version to run the operation in

→ ScriptForEvaluation
the script to evaluate

→ [Data]
the arguments that the script's underlying term will be applied to

→ m (Term NamedDeBruijn DefaultUni DefaultFun ())

Shared helper for the evaluation functions: evaluateScriptCounting and evaluateScriptRestricting,

Given a ScriptForEvaluation:

1) applies the term to a list of Data arguments (e.g. Datum, Redeemer, ScriptContext) 2) checks that the applied-term is well-scoped 3) returns the applied-term

Supporting types used in the context types

Builtins

data BuiltinByteString

An opaque type representing Plutus Core ByteStrings.

toBuiltin ∷ HasToBuiltin a ⇒ a → ToBuiltin a

fromBuiltin ∷ HasFromBuiltin arep ⇒ arep → FromBuiltin arep

toOpaque ∷ HasToOpaque a arep ⇒ a → arep

fromOpaque ∷ HasFromOpaque arep a ⇒ arep → a

Data

data Data

A generic "data" type.

The main constructor Constr represents a datatype value in sum-of-products form: Constr i args represents a use of the ith constructor along with its arguments.

The other constructors are various primitives.

Constructors

Constr Integer [Data]

Map [(Data, Data)]

List [Data]

I Integer

B ByteString

data BuiltinData

A type corresponding to the Plutus Core builtin equivalent of Data.

The point of this type is to be an opaque equivalent of Data, so as to ensure that it is only used in ways that the compiler can handle.

As such, you should use this type in your on-chain code, and in any data structures that you want to be representable on-chain.

For off-chain usage, there are conversion functions builtinDataToData and dataToBuiltinData, but note that these will not work on-chain.

Constructors

BuiltinData ~Data

class ToData a where

A typeclass for types that can be converted to and from BuiltinData.

Methods

toBuiltinData ∷ a → BuiltinData

Convert a value to BuiltinData.

class FromData a where

Methods

fromBuiltinData ∷ BuiltinData → Maybe a

Convert a value from BuiltinData, returning Nothing if this fails.

class UnsafeFromData a where

Methods

unsafeFromBuiltinData ∷ BuiltinData → a

Convert a value from BuiltinData, calling error if this fails. This is typically much faster than fromBuiltinData.

When implementing this function, make sure to call unsafeFromBuiltinData rather than fromBuiltinData when converting substructures!

This is a simple type without any validation, use with caution.

toData ∷ ToData a ⇒ a → Data

Convert a value to Data.

fromData ∷ FromData a ⇒ Data → Maybe a

Convert a value from Data, returning Nothing if this fails.

unsafeFromData ∷ UnsafeFromData a ⇒ Data → a

Convert a value from Data, throwing if this fails.

dataToBuiltinData ∷ Data → BuiltinData

NOT a builtin. Converts a Data into a BuiltinData. Only works off-chain.

builtinDataToData ∷ BuiltinData → Data

NOT a builtin. Converts a BuiltinData into a Data. Only works off-chain.

Misc

class Monad m ⇒ MonadError e (m ∷ Type → Type) | m → e

The strategy of combining computations that can throw exceptions by bypassing bound functions from the point an exception is thrown to the point that it is handled.

Is parameterized over the type of error information and the monad type constructor. It is common to use Either String as the monad type constructor for an error monad in which error descriptions take the form of strings. In that case and many other common cases the resulting monad is already defined as an instance of the MonadError class. You can also define your own error type and/or use a monad type constructor other than Either String or Either IOError. In these cases you will have to explicitly define instances of the MonadError class. (If you are using the deprecated Control.Monad.Error or Control.Monad.Trans.Error, you may also have to define an Error instance.)

Minimal complete definition

throwError, catchError