PlutusLedgerApi.Common.ParamName

Documentation

class (Enum a, Bounded a) ⇒ IsParamName a where

A parameter name for different plutus versions.

Each Plutus version should expose such an enumeration as an ADT and create an instance of ParamName out of it.

A valid parameter name has to be enumeration, bounded, ordered, and prettyprintable to a "lower-Kebab" string.

Minimal complete definition

showParamName

Methods

showParamName ∷ a → Text

Produce the raw textual form for a given typed-by-plutus-version cost model parameter Any implementation *must be* an injective function. The GIsParamName generic implementation guarantees injectivity.

readParamName ∷ Text → Maybe a

default implementation that inverts the showParamName operation (not very efficient)

newtype GenericParamName a

A Generic wrapper for use with deriving via

Constructors

GenericParamName a

toCostModelParams ∷ IsParamName p ⇒ [(p, Int64)] → CostModelParams

Untags the plutus version from the typed cost model parameters and returns their raw textual form (internally used by CostModelInterface).

tagWithParamNames ∷ ∀ k m. (Enum k, Bounded k, MonadError CostModelApplyError m, MonadWriter [CostModelApplyWarn] m) ⇒ [Int64] → m [(k, Int64)]

Given an ordered list of parameter values, tag them with their parameter names. If the passed parameter values are more than expected: the function will ignore the extraneous values at the tail of the list, if the passed values are less than expected: the function will throw an error; for more information, see Note [Cost model parameters from the ledger's point of view]

data CostModelApplyError

A fatal error when trying to create a cost given some plain costmodel parameters.

Constructors

CMUnknownParamError !Text
a costmodel parameter with the give name does not exist in the costmodel to be applied upon

CMInternalReadError
internal error when we are transforming the applyParams' input to json (should not happen)

CMInternalWriteError !String
internal error when we are transforming the applied params from json with given jsonstring error (should not happen)

data CostModelApplyWarn

A non-fatal warning when trying to create a cost given some plain costmodel parameters.

Constructors

CMTooManyParamsWarn
See Note [Cost model parameters from the ledger's point of view]

Fields

cmExpected ∷ !Int

cmActual ∷ !Int

CMTooFewParamsWarn
See Note [Cost model parameters from the ledger's point of view]

Fields

cmExpected ∷ !Int

cmActual ∷ !Int