PlutusLedgerApi.Common.ProtocolVersions

Documentation

newtype MajorProtocolVersion

This represents the major component of the Cardano protocol version. The ledger can only supply the major component of the protocol version, not the minor component, and Plutus should only need to care about the major component anyway. This relies on careful understanding between us and the ledger as to what this means.

Constructors

MajorProtocolVersion

Fields

getMajorProtocolVersion ∷ Int

Protocol Version aliases

Based on https://github.com/IntersectMBO/cardano-ledger/wiki/First-Block-of-Each-Era

shelleyPV ∷ MajorProtocolVersion

Shelley era was introduced in protocol version 2.0

allegraPV ∷ MajorProtocolVersion

Allegra era was introduced in protocol version 3.0

maryPV ∷ MajorProtocolVersion

Mary era was introduced in protocol version 4.0

alonzoPV ∷ MajorProtocolVersion

Alonzo era was introduced in protocol version 5.0

vasilPV ∷ MajorProtocolVersion

The Vasil HF introduced the Babbage era and Plutus V2

valentinePV ∷ MajorProtocolVersion

Valentine was an intra-era HF where builtin functions VerifyEcdsaSecp256k1Signature and VerifySchnorrSecp256k1Signature were enabled.

changPV ∷ MajorProtocolVersion

The Chang HF introduced the Conway era and Plutus V3

plominPV ∷ MajorProtocolVersion

The Plomin HF was an intra-era HF where some new builtin functions were introduced in Plutus V2 and V3.

pv11PV ∷ MajorProtocolVersion

Not sure what this is going to be called yet

newestPV ∷ MajorProtocolVersion

knownPVs ∷ [MajorProtocolVersion]

The set of protocol versions that are "known", i.e. that have been released and have actual differences associated with them. This is currently only used for testing, so efficiency is not parmount and a list is fine.

futurePV ∷ MajorProtocolVersion

This is a placeholder for when we don't yet know what protocol version will be used for something. It's a very high protocol version that should never appear in reality.

We should not assign names to future protocol versions until it's confirmed that they are correct, otherwise we could accidentally associate something with the wrong protocol version.