PlutusLedgerApi.Common.SerialisedScript

Documentation

type SerialisedScript = ShortByteString

Scripts to the ledger are serialised bytestrings.

serialiseCompiledCode ∷ ∀ a. CompiledCode a → SerialisedScript

Turns a program which was compiled using the 'PlutusTx' toolchain into a binary format that is understood by the network and can be stored on-chain.

serialiseUPLC ∷ Program DeBruijn DefaultUni DefaultFun () → SerialisedScript

Turns a program's AST (most likely manually constructed) into a binary format that is understood by the network and can be stored on-chain.

uncheckedDeserialiseUPLC ∷ SerialisedScript → Program DeBruijn DefaultUni DefaultFun ()

Deserialises a SerialisedScript back into an AST. Does *not* do ledger-language-version-specific checks like for allowable builtins.

scriptCBORDecoder ∷ PlutusLedgerLanguage → MajorProtocolVersion → Decoder s ScriptNamedDeBruijn

This decoder decodes the names directly into NamedDeBruijns rather than DeBruijns. This is needed because the CEK machine expects NameDeBruijns, but there are obviously no names in the serialised form of a Script. Rather than traversing the term and inserting fake names after deserialising, this lets us do at the same time as deserialising.

newtype ScriptNamedDeBruijn

A script with named de-bruijn indices.

Constructors

ScriptNamedDeBruijn (Program NamedDeBruijn DefaultUni DefaultFun ())

data ScriptForEvaluation

A Plutus script ready to be evaluated on-chain, via evaluateScriptRestricting.

data ScriptDecodeError

An error that occurred during script deserialization.

Constructors

CBORDeserialiseError !DeserialiseFailureInfo
an error from the underlying CBOR/serialise library

RemainderError !ByteString
Script was successfully parsed, but more (runaway) bytes encountered after script's position

LedgerLanguageNotAvailableError
the plutus version of the given script is not enabled yet

Fields

sdeAffectedLang ∷ !PlutusLedgerLanguage

the script's ledger language

sdeIntroPv ∷ !MajorProtocolVersion

the major protocol version that will first introduce/enable the ledger language

sdeThisPv ∷ !MajorProtocolVersion

the current protocol version

PlutusCoreLanguageNotAvailableError

Fields

sdeAffectedVersion ∷ !Version

the Plutus Core language of the script under execution.

sdeThisLang ∷ !PlutusLedgerLanguage

the Plutus ledger language of the script under execution.

sdeThisPv ∷ !MajorProtocolVersion

the current protocol version

class AsScriptDecodeError r where

Minimal complete definition

_ScriptDecodeError

Methods

_ScriptDecodeError ∷ Prism' r ScriptDecodeError

_CBORDeserialiseError ∷ Prism' r DeserialiseFailureInfo

_RemainderError ∷ Prism' r ByteString

_LedgerLanguageNotAvailableError ∷ Prism' r (PlutusLedgerLanguage, MajorProtocolVersion, MajorProtocolVersion)

_PlutusCoreLanguageNotAvailableError ∷ Prism' r (Version, PlutusLedgerLanguage, MajorProtocolVersion)

data DeserialiseFailureInfo

Similar to DeserialiseFailure, with the difference that plain string reason messages are turned into the datatype: DeserialiseFailureReason.

Constructors

DeserialiseFailureInfo

Fields

dfOffset ∷ ByteOffset

dfReason ∷ DeserialiseFailureReason

data DeserialiseFailureReason

The reason of the cbor failure as a datatype, not as a plain string.

Constructors

EndOfInput
Not enough input provided

ExpectedBytes
The bytes inside the input are malformed.

OtherReason String
This is either a cbor failure that we (plutus) are not aware of, or an underlying flat failure. We use whatever message cborg or flat returns.

deserialiseScript

Arguments

∷ ∀ m. MonadError ScriptDecodeError m

⇒ PlutusLedgerLanguage
the Plutus ledger language of the script.

→ MajorProtocolVersion
which major protocol version the script was submitted in.

→ SerialisedScript
the script to deserialise.

→ m ScriptForEvaluation

The deserialization from a serialised script into a ScriptForEvaluation, ready to be evaluated on-chain. Called inside phase-1 validation (i.e., deserialisation error is a phase-1 error).

serialisedScript ∷ ScriptForEvaluation → SerialisedScript

Get a SerialisedScript from a ScriptForEvaluation. O(1).

deserialisedScript ∷ ScriptForEvaluation → ScriptNamedDeBruijn

Get a ScriptNamedDeBruijn from a ScriptForEvaluation. O(1).