PlutusLedgerApi.Common.Versions

Description

This module contains the code for handling the various kinds of version that we care about:

Protocol versions

Plutus ledger languages

Plutus Core language versions

Cardano Protocol versions

module PlutusLedgerApi.Common.ProtocolVersions

Plutus ledger languages

data PlutusLedgerLanguage

The Plutus ledger language. These are entirely different script languages from the ledger's perspective, which on our side are interpreted in very similar ways.

It is a simple enumerated datatype (there is no major and minor components as in protocol version) and the ordering of constructors is essential for deriving Enum,Ord,Bounded.

IMPORTANT: this is different from the Plutus Core language version, Version

Constructors

PlutusV1
introduced in Alonzo HF

PlutusV2
introduced in Vasil HF

PlutusV3
introduced in Chang HF

Plutus Core language versions

data Version

The version of Plutus Core used by this program.

The intention is to convey different levels of backwards compatibility for existing scripts: - Major version changes are backwards-incompatible - Minor version changes are backwards-compatible - Patch version changes should be entirely invisible (and we will likely not use this level)

The version used should be changed only when the language itself changes. For example, adding a new kind of term to the language would require a minor version bump; removing a kind of term would require a major version bump.

Similarly, changing the semantics of the language will require a version bump, typically a major one. This is the main reason why the version is actually tracked in the AST: we can have two language versions with identical ASTs but different semantics, so we need to track the version explicitly.

Compatibility is about compatibility for specific scripts, not about e.g. tools which consume scripts. Adding a new kind of term does not change how existing scripts behave, but does change what tools would need to do to process scripts.

Constructors

Version

Fields

_versionMajor ∷ Natural

_versionMinor ∷ Natural

_versionPatch ∷ Natural

Version-testing functions

ledgerLanguageIntroducedIn ∷ PlutusLedgerLanguage → MajorProtocolVersion

Query the protocol version that a specific Plutus ledger language was first introduced in.

ledgerLanguagesAvailableIn ∷ MajorProtocolVersion → Set PlutusLedgerLanguage

Which Plutus language versions are available in the given MajorProtocolVersion? See Note [New builtins/language versions and protocol versions]. This function (and others in this module) assumes that once a LL is available it remains available in all later PVs and that if m <= n, PlutusVm is introduced no later than PlutusVn.

plcVersionsIntroducedIn ∷ PlutusLedgerLanguage → Map MajorProtocolVersion (Set Version)

A map indicating which Plutus Core versions were introduced in which MajorProtocolVersion and PlutusLedgerLanguage. Each version should appear at most once. This must be updated when new versions are added. See Note [New builtins/language versions and protocol versions]

plcVersionsAvailableIn ∷ PlutusLedgerLanguage → MajorProtocolVersion → Set Version

Which Plutus Core language versions are available in the given PlutusLedgerLanguage and MajorProtocolVersion?

builtinsIntroducedIn ∷ PlutusLedgerLanguage → Map MajorProtocolVersion (Set DefaultFun)

Given a ledger language, return a map indicating which builtin functions were introduced in which MajorProtocolVersion. This must be updated when new builtins are added. It is not necessary to add entries for protocol versions where no new builtins are added. See Note [New builtins/language versions and protocol versions]

builtinsAvailableIn ∷ PlutusLedgerLanguage → MajorProtocolVersion → Set DefaultFun

Return a set containing the builtins which are available in a given LL in a given PV. All builtins are available in all LLs from pv11PV onwards.

batch1 ∷ [DefaultFun]

batch2 ∷ [DefaultFun]

batch3 ∷ [DefaultFun]

batch4a ∷ [DefaultFun]

batch4b ∷ [DefaultFun]

batch5 ∷ [DefaultFun]

batch6 ∷ [DefaultFun]