PlutusLedgerApi.Data.V4

Description

The data-backed type interface to Plutus V4 for the ledger.

Accounts

newtype AccountId

Constructors

AccountId Credential

data AccountBalanceInterval

pattern AccountBalanceLowerBound ∷ Lovelace → AccountBalanceInterval

pattern AccountBalanceUpperBound ∷ Lovelace → AccountBalanceInterval

pattern AccountBalanceBothBounds ∷ Lovelace → Lovelace → AccountBalanceInterval

pattern AccountBalanceExact ∷ Lovelace → AccountBalanceInterval

newtype AccountBalanceIntervals

Constructors

AccountBalanceIntervals (Map AccountId AccountBalanceInterval)

Governance

newtype ColdCommitteeCredential

Constructors

ColdCommitteeCredential Credential

newtype HotCommitteeCredential

Constructors

HotCommitteeCredential Credential

newtype DRepCredential

Constructors

DRepCredential Credential

data DRep

pattern DRep ∷ DRepCredential → DRep

pattern DRepAlwaysAbstain ∷ DRep

pattern DRepAlwaysNoConfidence ∷ DRep

data Delegatee

pattern DelegStake ∷ PubKeyHash → Delegatee

pattern DelegVote ∷ DRep → Delegatee

pattern DelegStakeVote ∷ PubKeyHash → DRep → Delegatee

data TxCert

pattern TxCertRegAccount ∷ AccountId → Lovelace → TxCert

pattern TxCertUnRegAccount ∷ AccountId → Lovelace → TxCert

pattern TxCertDelegAccount ∷ AccountId → Delegatee → TxCert

pattern TxCertRegAccountDeleg ∷ AccountId → Delegatee → Lovelace → TxCert

pattern TxCertRegDRep ∷ DRepCredential → Lovelace → TxCert

pattern TxCertUpdateDRep ∷ DRepCredential → TxCert

pattern TxCertUnRegDRep ∷ DRepCredential → Lovelace → TxCert

pattern TxCertPoolRegister ∷ PubKeyHash → PubKeyHash → TxCert

pattern TxCertPoolRetire ∷ PubKeyHash → Integer → TxCert

pattern TxCertAuthHotCommittee ∷ ColdCommitteeCredential → HotCommitteeCredential → TxCert

pattern TxCertResignColdCommittee ∷ ColdCommitteeCredential → TxCert

data Voter

pattern CommitteeVoter ∷ HotCommitteeCredential → Voter

pattern DRepVoter ∷ DRepCredential → Voter

pattern StakePoolVoter ∷ PubKeyHash → Voter

data Vote

A vote. The optional anchor is omitted.

pattern VoteNo ∷ Vote

pattern VoteYes ∷ Vote

pattern Abstain ∷ Vote

data GovernanceActionId

Similar to TxOutRef, but for GovActions

pattern GovernanceActionId ∷ TxId → Integer → GovernanceActionId

gaidTxId ∷ GovernanceActionId → TxId

gaidGovActionIx ∷ GovernanceActionId → Integer

data Committee

pattern Committee ∷ Map ColdCommitteeCredential Integer → Rational → Committee

committeeMembers ∷ Committee → Map ColdCommitteeCredential Integer

committeeQuorum ∷ Committee → Rational

newtype Constitution

A constitution. The optional anchor is omitted.

Constructors

Constitution

Fields

constitutionScript ∷ Maybe ScriptHash

data ProtocolVersion

pattern ProtocolVersion ∷ Integer → Integer → ProtocolVersion

pvMajor ∷ ProtocolVersion → Integer

pvMinor ∷ ProtocolVersion → Integer

data GovernanceAction

pattern ParameterChange ∷ Maybe GovernanceActionId → ChangedParameters → Maybe ScriptHash → GovernanceAction

pattern HardForkInitiation ∷ Maybe GovernanceActionId → ProtocolVersion → GovernanceAction

pattern TreasuryWithdrawals ∷ Map Credential Lovelace → Maybe ScriptHash → GovernanceAction

pattern NoConfidence ∷ Maybe GovernanceActionId → GovernanceAction

pattern UpdateCommittee ∷ Maybe GovernanceActionId → List ColdCommitteeCredential → Map ColdCommitteeCredential Integer → Rational → GovernanceAction

pattern NewConstitution ∷ Maybe GovernanceActionId → Constitution → GovernanceAction

pattern InfoAction ∷ GovernanceAction

newtype ChangedParameters

A Plutus Data object containing proposed parameter changes. The Data object contains a Map with one entry per changed parameter, from the parameter ID to the new value. Unchanged parameters are not included.

The mapping from parameter IDs to parameters can be found in conway.cddl. -- editorconfig-checker-disable-file

Invariant: This map is non-empty, and the keys are stored in ascending order.

Constructors

ChangedParameters

Fields

getChangedParameters ∷ BuiltinData

data ProposalProcedure

A proposal procedure. The optional anchor is omitted.

pattern ProposalProcedure ∷ Lovelace → Credential → GovernanceAction → ProposalProcedure

ppDeposit ∷ ProposalProcedure → Lovelace

ppReturnAddr ∷ ProposalProcedure → Credential

ppGovernanceAction ∷ ProposalProcedure → GovernanceAction

Context types

data ScriptContext

pattern ScriptContext ∷ TxInfo → Redeemer → ScriptInfo → ScriptHash → ScriptContext

scriptContextTxInfo ∷ ScriptContext → TxInfo

scriptContextRedeemer ∷ ScriptContext → Redeemer

scriptContextScriptInfo ∷ ScriptContext → ScriptInfo

scriptContextScriptHash ∷ ScriptContext → ScriptHash

data ScriptPurpose

pattern Minting ∷ ScriptHash → CurrencySymbol → ScriptPurpose

pattern Spending ∷ ScriptHash → TxOutRef → ScriptPurpose

pattern Withdrawing ∷ ScriptHash → Credential → ScriptPurpose

pattern Certifying ∷ ScriptHash → Integer → TxCert → ScriptPurpose

pattern Voting ∷ ScriptHash → Voter → ScriptPurpose

pattern Proposing ∷ ScriptHash → Integer → ProposalProcedure → ScriptPurpose

pattern Guarding ∷ ScriptHash → Integer → ScriptPurpose

data ScriptInfo

pattern MintingScript ∷ CurrencySymbol → ScriptInfo

pattern SpendingScript ∷ TxOutRef → Maybe Datum → ScriptInfo

pattern WithdrawingScript ∷ AccountId → ScriptInfo

pattern CertifyingScript ∷ Integer → TxCert → ScriptInfo

pattern VotingScript ∷ Voter → ScriptInfo

pattern ProposingScript ∷ Integer → ProposalProcedure → ScriptInfo

pattern GuardingScript ∷ Integer → Maybe TopTxInfo → ScriptInfo

data TopTxInfo

pattern TopTxInfo ∷ List TxInfo → Map Integer Datum → AccountBalanceIntervals → TopTxInfoSimplified → TopTxInfo

topTxInfoSubTransactions ∷ TopTxInfo → List TxInfo

topTxInfoDatums ∷ TopTxInfo → Map Integer Datum

topTxInfoStartingBalanceIntervals ∷ TopTxInfo → AccountBalanceIntervals

topTxInfoSimplified ∷ TopTxInfo → TopTxInfoSimplified

data TopTxInfoSimplified

pattern TopTxInfoSimplified ∷ List TxId → List TxInInfo → List TxInInfo → List TxOut → MintValue → MintValue → List TxCert → Map AccountId Lovelace → Map AccountId Lovelace → POSIXTimeRange → Map Credential () → Map ScriptPurpose () → Map DatumHash Datum → Map Voter (Map GovernanceActionId Vote) → List ProposalProcedure → Maybe Lovelace → Lovelace → TopTxInfoSimplified

ttisIds ∷ TopTxInfoSimplified → List TxId

ttisInputs ∷ TopTxInfoSimplified → List TxInInfo

ttisReferenceInputs ∷ TopTxInfoSimplified → List TxInInfo

ttisOutputs ∷ TopTxInfoSimplified → List TxOut

ttisMints ∷ TopTxInfoSimplified → MintValue

ttisBurns ∷ TopTxInfoSimplified → MintValue

ttisTxCerts ∷ TopTxInfoSimplified → List TxCert

ttisWithdrawals ∷ TopTxInfoSimplified → Map AccountId Lovelace

ttisDirectDeposits ∷ TopTxInfoSimplified → Map AccountId Lovelace

ttisValidRange ∷ TopTxInfoSimplified → POSIXTimeRange

ttisGuards ∷ TopTxInfoSimplified → Map Credential ()

ttisScriptPurposes ∷ TopTxInfoSimplified → Map ScriptPurpose ()

ttisData ∷ TopTxInfoSimplified → Map DatumHash Datum

ttisVotes ∷ TopTxInfoSimplified → Map Voter (Map GovernanceActionId Vote)

ttisProposalProcedures ∷ TopTxInfoSimplified → List ProposalProcedure

ttisCurrentTreasuryAmount ∷ TopTxInfoSimplified → Maybe Lovelace

ttisTreasuryDonations ∷ TopTxInfoSimplified → Lovelace

Supporting types used in the context types

Builtins

data BuiltinByteString

An opaque type representing Plutus Core ByteStrings.

toBuiltin ∷ HasToBuiltin a ⇒ a → ToBuiltin a

fromBuiltin ∷ HasFromBuiltin arep ⇒ arep → FromBuiltin arep

toOpaque ∷ HasToOpaque a arep ⇒ a → arep

fromOpaque ∷ HasFromOpaque arep a ⇒ arep → a

Bytes

newtype LedgerBytes

Constructors

LedgerBytes

Fields

getLedgerBytes ∷ BuiltinByteString

fromBytes ∷ ByteString → LedgerBytes

Lift a Haskell bytestring to the Plutus abstraction LedgerBytes

Credentials

data StakingCredential

Staking credential used to assign rewards.

The staking hash constructor is the Credential required to unlock a transaction output. Either a public key credential (PubKeyHash) or a script credential (ScriptHash). Both are hashed with BLAKE2b-244. 28 byte.

The StakingPtr constructor is the certificate pointer, constructed by the given slot number, transaction and certificate indices. NB: The fields should really be all Word64, as they are implemented in Word64, but Integer is our only integral type so we need to use it instead.

pattern StakingHash ∷ Credential → StakingCredential

pattern StakingPtr ∷ Integer → Integer → Integer → StakingCredential

data Credential

Credentials required to unlock a transaction output.

The PubKeyCredential constructor represents the transaction that spends this output and must be signed by the private key. See PubKeyHash.

The ScriptCredential constructor represents the transaction that spends this output must include the validator script and be accepted by the validator. See ScriptHash.

pattern PubKeyCredential ∷ PubKeyHash → Credential

pattern ScriptCredential ∷ ScriptHash → Credential

Value

newtype Value

The Value type represents a collection of amounts of different currencies. We can think of Value as a vector space whose dimensions are currencies.

Operations on currencies are usually implemented pointwise. That is, we apply the operation to the quantities for each currency in turn. So when we add two Values the resulting Value has, for each currency, the sum of the quantities of that particular currency in the argument Value. The effect of this is that the currencies in the Value are "independent", and are operated on separately.

Whenever we need to get the quantity of a currency in a Value where there is no explicit quantity of that currency in the Value, then the quantity is taken to be zero.

There is no 'Ord Value' instance since Value is only a partial order, so compare can't do the right thing in some cases.

Constructors

Value

Fields

getValue ∷ Map CurrencySymbol (Map TokenName Integer)

newtype CurrencySymbol

ByteString representing the currency, hashed with BLAKE2b-224. It is empty for Ada, 28 bytes for MintingPolicyHash. Forms an AssetClass along with TokenName. A Value is a map from CurrencySymbol's to a map from TokenName to an Integer.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification. -- editorconfig-checker-disable-file

Constructors

CurrencySymbol

Fields

unCurrencySymbol ∷ BuiltinByteString

newtype TokenName

ByteString of a name of a token. Shown as hex-encoded bytes. Should be no longer than 32 bytes, empty for Ada. Forms an AssetClass along with a CurrencySymbol.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification. -- editorconfig-checker-disable-file

Constructors

TokenName

Fields

unTokenName ∷ BuiltinByteString

singleton ∷ CurrencySymbol → TokenName → Integer → Value

Make a Value containing only the given quantity of the given currency.

unionWith ∷ (Integer → Integer → Integer) → Value → Value → Value

Combine two Value maps with the argument function. Assumes the well-definedness of the two maps.

adaSymbol ∷ CurrencySymbol

The CurrencySymbol of the Ada currency.

adaToken ∷ TokenName

The TokenName of the Ada currency.

newtype Lovelace

Constructors

Lovelace

Fields

getLovelace ∷ Integer

newtype AssetClass

An asset class, identified by a CurrencySymbol and a TokenName.

Constructors

AssetClass

Fields

unAssetClass ∷ (CurrencySymbol, TokenName)

assetClass ∷ CurrencySymbol → TokenName → AssetClass

The curried version of AssetClass constructor

assetClassValue ∷ AssetClass → Integer → Value

A Value containing the given amount of the asset class.

assetClassValueOf ∷ Value → AssetClass → Integer

Get the quantity of the given AssetClass class in the Value.

currencySymbol ∷ ByteString → CurrencySymbol

Creates CurrencySymbol from raw ByteString.

currencySymbolValueOf ∷ Value → CurrencySymbol → Integer

Get the total value of the currency symbol in the Value map. Assumes that the underlying map doesn't contain duplicate keys.

Note that each token of the currency symbol may have a value that is positive, zero or negative.

flattenValue ∷ Value → [(CurrencySymbol, TokenName, Integer)]

Convert a Value to a simple list, keeping only the non-zero amounts. Note that the result isn't sorted, meaning v1 == v2 doesn't generally imply flattenValue v1 == flattenValue v2. Also assumes that there are no duplicate keys in the Value Map.

geq ∷ Value → Value → Bool

Check whether one Value is greater than or equal to another. See Value for an explanation of how operations on Values work.

gt ∷ Value → Value → Bool

Check whether one Value is strictly greater than another. This is *not* a pointwise operation. gt l r means geq l r && not (eq l r).

isZero ∷ Value → Bool

Check whether a Value is zero.

leq ∷ Value → Value → Bool

Check whether one Value is less than or equal to another. See Value for an explanation of how operations on Values work.

lovelaceValue ∷ Lovelace → Value

A Value containing the given quantity of Lovelace.

lovelaceValueOf ∷ Value → Lovelace

Get the quantity of Lovelace in the Value.

lt ∷ Value → Value → Bool

Check whether one Value is strictly less than another. This is *not* a pointwise operation. lt l r means leq l r && not (eq l r).

scale ∷ Module s v ⇒ s → v → v

split ∷ Value → (Value, Value)

Split a Value into its positive and negative parts. The first element of the tuple contains the negative parts of the Value, the second element contains the positive parts.

negate (fst (split a)) plus (snd (split a)) == a

symbols ∷ Value → List CurrencySymbol

The list of CurrencySymbols of a Value.

tokenName ∷ ByteString → TokenName

Creates TokenName from raw ByteString.

unsafeLovelaceValueOf ∷ Value → Lovelace

Assumes that the first token of the first currency in the Value represents lovelace. Returns a silently wrong answer or fails if that's not the case. It is thus much faster than lovelaceValueOf.

valueOf ∷ Value → CurrencySymbol → TokenName → Integer

Get the quantity of the given currency in the Value. Assumes that the underlying map doesn't contain duplicate keys.

withCurrencySymbol ∷ CurrencySymbol → Value → a → (Map TokenName Integer → a) → a

Apply a continuation function to the token quantities of the given currency symbol in the value or return a default value if the currency symbol is not present in the value.

Mint Value

data MintValue

A MintValue represents assets that are minted and burned in a transaction.

emptyMintValue ∷ MintValue

mintValueToMap ∷ MintValue → Map CurrencySymbol (Map TokenName Integer)

mintValueMinted ∷ MintValue → Value

Get the Value minted by the MintValue.

mintValueBurned ∷ MintValue → Value

Get the Value burned by the MintValue. All the negative quantities in the MintValue become positive in the resulting Value.

Time

newtype POSIXTime

POSIX time is measured as the number of milliseconds since 1970-01-01T00:00:00Z. This is not the same as Haskell's POSIXTime

Constructors

POSIXTime

Fields

getPOSIXTime ∷ Integer

type POSIXTimeRange = Interval POSIXTime

An Interval of POSIXTimes.

Types for representing transactions

data Address

An address may contain two credentials, the payment credential and optionally a StakingCredential.

pattern Address ∷ Credential → Maybe StakingCredential → Address

addressCredential ∷ Address → Credential

addressStakingCredential ∷ Address → Maybe StakingCredential

newtype PubKeyHash

The hash of a public key. This is frequently used to identify the public key, rather than the key itself. Hashed with BLAKE2b-224. 28 bytes.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

PubKeyHash

Fields

getPubKeyHash ∷ BuiltinByteString

newtype TxId

A transaction ID, i.e. the hash of a transaction. Hashed with BLAKE2b-256. 32 byte.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

TxId

Fields

getTxId ∷ BuiltinByteString

data TxInfo

pattern TxInfo ∷ TxId → Maybe Integer → List TxInInfo → List TxInInfo → List TxOut → Lovelace → MintValue → List TxCert → Map AccountId Lovelace → Map AccountId Lovelace → AccountBalanceIntervals → POSIXTimeRange → List Credential → Map Credential (Maybe Datum) → Map ScriptPurpose Redeemer → Map DatumHash Datum → Map Voter (Map GovernanceActionId Vote) → List ProposalProcedure → Maybe Lovelace → Lovelace → TxInfo

txInfoId ∷ TxInfo → TxId

txInfoSubTxIx ∷ TxInfo → Maybe Integer

txInfoInputs ∷ TxInfo → List TxInInfo

txInfoReferenceInputs ∷ TxInfo → List TxInInfo

txInfoOutputs ∷ TxInfo → List TxOut

txInfoFee ∷ TxInfo → Lovelace

txInfoMint ∷ TxInfo → MintValue

txInfoTxCerts ∷ TxInfo → List TxCert

txInfoWithdrawals ∷ TxInfo → Map AccountId Lovelace

txInfoDirectDeposits ∷ TxInfo → Map AccountId Lovelace

txInfoAccountBalanceIntervals ∷ TxInfo → AccountBalanceIntervals

txInfoValidRange ∷ TxInfo → POSIXTimeRange

txInfoGuards ∷ TxInfo → List Credential

txInfoRequiredTopLevelGuards ∷ TxInfo → Map Credential (Maybe Datum)

txInfoRedeemers ∷ TxInfo → Map ScriptPurpose Redeemer

txInfoData ∷ TxInfo → Map DatumHash Datum

txInfoVotes ∷ TxInfo → Map Voter (Map GovernanceActionId Vote)

txInfoProposalProcedures ∷ TxInfo → List ProposalProcedure

txInfoCurrentTreasuryAmount ∷ TxInfo → Maybe Lovelace

txInfoTreasuryDonation ∷ TxInfo → Lovelace

data TxOut

A transaction output, consisting of a target address, a value, optionally a datum/datum hash, and optionally a reference script.

pattern TxOut ∷ Address → Value → OutputDatum → Maybe ScriptHash → TxOut

txOutAddress ∷ TxOut → Address

txOutValue ∷ TxOut → Value

txOutDatum ∷ TxOut → OutputDatum

txOutReferenceScript ∷ TxOut → Maybe ScriptHash

data TxOutRef

A reference to a transaction output. This is a pair of a transaction ID (TxId), and an index indicating which of the outputs of that transaction we are referring to.

pattern TxOutRef ∷ TxId → Integer → TxOutRef

txOutRefId ∷ TxOutRef → TxId

txOutRefIdx ∷ TxOutRef → Integer

data TxInInfo

An input of a pending transaction.

pattern TxInInfo ∷ TxOutRef → TxOut → TxInInfo

txInInfoOutRef ∷ TxInInfo → TxOutRef

txInInfoResolved ∷ TxInInfo → TxOut

data OutputDatum

The datum attached to an output: either nothing; a datum hash; or the datum itself (an "inline datum").

pattern NoOutputDatum ∷ OutputDatum

pattern OutputDatum ∷ Datum → OutputDatum

pattern OutputDatumHash ∷ DatumHash → OutputDatum

Intervals

data Interval a

An interval of as.

The interval may be either closed or open at either end, meaning that the endpoints may or may not be included in the interval.

The interval can also be unbounded on either side.

The Eq instance gives equality of the intervals, not structural equality. There is no Ord instance, but contains gives a partial order.

Note that some of the functions on Interval rely on Enum in order to handle non-inclusive endpoints. For this reason, it may not be safe to use Intervals with non-inclusive endpoints on types whose Enum instances have partial methods.

pattern Interval ∷ ∀ a. (ToData a, UnsafeFromData a) ⇒ LowerBound a → UpperBound a → Interval a

ivFrom ∷ (ToData a, UnsafeFromData a) ⇒ Interval a → LowerBound a

ivTo ∷ (ToData a, UnsafeFromData a) ⇒ Interval a → UpperBound a

data Extended a

A set extended with a positive and negative infinity.

pattern NegInf ∷ ∀ a. Extended a

pattern PosInf ∷ ∀ a. Extended a

pattern Finite ∷ ∀ a. (ToData a, UnsafeFromData a) ⇒ a → Extended a

type Closure = Bool

Whether a bound is inclusive or not.

data UpperBound a

The upper bound of an interval.

pattern UpperBound ∷ ∀ a. (ToData a, UnsafeFromData a) ⇒ Extended a → Closure → UpperBound a

data LowerBound a

The lower bound of an interval.

pattern LowerBound ∷ ∀ a. (ToData a, UnsafeFromData a) ⇒ Extended a → Closure → LowerBound a

always ∷ (ToData a, UnsafeFromData a) ⇒ Interval a

An Interval that covers every slot. In math. notation [-∞,+∞]

from ∷ (ToData a, UnsafeFromData a) ⇒ a → Interval a

from a is an Interval that includes all values that are greater than or equal to a. In math. notation: [a,+∞]

to ∷ (ToData a, UnsafeFromData a) ⇒ a → Interval a

to a is an Interval that includes all values that are smaller than or equal to a. In math. notation: [-∞,a]

lowerBound ∷ (ToData a, UnsafeFromData a) ⇒ a → LowerBound a

Construct a lower bound from a value. The resulting bound includes all values that are equal or greater than the input value.

upperBound ∷ (ToData a, UnsafeFromData a) ⇒ a → UpperBound a

Construct an upper bound from a value. The resulting bound includes all values that are equal or smaller than the input value.

strictLowerBound ∷ (ToData a, UnsafeFromData a) ⇒ a → LowerBound a

Construct a strict lower bound from a value. The resulting bound includes all values that are (strictly) greater than the input value.

strictUpperBound ∷ (ToData a, UnsafeFromData a) ⇒ a → UpperBound a

Construct a strict upper bound from a value. The resulting bound includes all values that are (strictly) smaller than the input value.

inclusiveLowerBound ∷ (Enum a, ToData a, UnsafeFromData a) ⇒ LowerBound a → Extended a

For an enumerable type, turn an lower bound into a single inclusive bounding value.

Since the type is enumerable, non-inclusive bounds are equivalent to inclusive bounds on the successor.

See Note [Enumerable Intervals]

inclusiveUpperBound ∷ (Enum a, ToData a, UnsafeFromData a) ⇒ UpperBound a → Extended a

For an enumerable type, turn an upper bound into a single inclusive bounding value.

Since the type is enumerable, non-inclusive bounds are equivalent to inclusive bounds on the predecessor.

See Note [Enumerable Intervals]

Ratio

data Rational

Represents an arbitrary-precision ratio.

The following two invariants are maintained:

The denominator is greater than zero.

The numerator and denominator are coprime.

ratio ∷ Integer → Integer → Maybe Rational

Safely constructs a Rational from a numerator and a denominator. Returns Nothing if given a zero denominator.

unsafeRatio ∷ Integer → Integer → Rational

Makes a Rational from a numerator and a denominator.

Important note

If given a zero denominator, this function will error. If you don't mind a size increase, and care about safety, use ratio instead.

numerator ∷ Rational → Integer

Returns the numerator of its argument.

Note

It is not true in general that numerator <$> ratio x y = x; this will only hold if x and y are coprime. This is due to Rational normalizing the numerator and denominator.

denominator ∷ Rational → Integer

Returns the denominator of its argument. This will always be greater than, or equal to, 1, although the type does not describe this.

Note

It is not true in general that denominator <$> ratio x y = y; this will only hold if x and y are coprime. This is due to Rational normalizing the numerator and denominator.

fromHaskellRatio ∷ Rational → Rational

Converts a GHC Rational, preserving value.

Note: Does not work on-chain.

toHaskellRatio ∷ Rational → Rational

Converts a Rational to a GHC Rational, preserving value.

Note: Does not work on-chain.

fromGHC ∷ Rational → Rational

toGHC ∷ Rational → Rational

Association maps

data Map k a

A map associating keys and values backed by BuiltinData.

This implementation has the following characteristics:

The toBuiltinData and unsafeFromBuiltinData operations are no-op.

Other operations are slower than PlutusTx.AssocMap.Map, although equality checks on keys can be faster due to equalsData.

Many operations involve converting the keys and/or values to/from BuiltinData.

Therefore this implementation is likely a better choice than PlutusTx.AssocMap.Map if it is part of a data type defined using asData, and the key and value types have efficient toBuiltinData and unsafeFromBuiltinData operations (e.g., they are primitive types or types defined using asData).

A Map is considered well-defined if it has no duplicate keys. Most operations preserve the definedness of the resulting Map unless otherwise noted. It is important to observe that, in comparison to standard map implementations, this implementation provides slow lookup and update operations because it is based on a list representation.

unsafeFromSOPList ∷ (ToData k, ToData a) ⇒ [(k, a)] → Map k a

Unsafely create an Map from a sums of products list of pairs. This should _only_ be applied to lists which have been checked to not contain duplicate keys, otherwise the resulting Map will contain conflicting entries (two entries sharing the same key), and therefore be ill-defined. Warning: this requires traversing the list and encoding the keys and values, so it should be avoided in favor of unsafeFromBuiltinList if the input is already in BuiltinData form.

Newtypes and hash types

newtype ScriptHash

Type representing the BLAKE2b-224 hash of a script. 28 bytes.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

ScriptHash

Fields

getScriptHash ∷ BuiltinByteString

newtype Redeemer

Redeemer is a wrapper around Data values that are used as redeemers in transaction inputs.

Constructors

Redeemer

Fields

getRedeemer ∷ BuiltinData

newtype RedeemerHash

Type representing the BLAKE2b-256 hash of a redeemer. 32 bytes.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

RedeemerHash BuiltinByteString

newtype Datum

Datum is a wrapper around Data values which are used as data in transaction outputs.

Constructors

Datum

Fields

getDatum ∷ BuiltinData

newtype DatumHash

Type representing the BLAKE2b-256 hash of a datum. 32 bytes.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

DatumHash BuiltinByteString

Data

data Data

A generic "data" type.

The main constructor Constr represents a datatype value in sum-of-products form: Constr i args represents a use of the ith constructor along with its arguments.

The other constructors are various primitives.

Constructors

Constr Integer [Data]

Map [(Data, Data)]

List [Data]

I Integer

B ByteString

data BuiltinData

A type corresponding to the Plutus Core builtin equivalent of Data.

The point of this type is to be an opaque equivalent of Data, so as to ensure that it is only used in ways that the compiler can handle.

As such, you should use this type in your on-chain code, and in any data structures that you want to be representable on-chain.

For off-chain usage, there are conversion functions builtinDataToData and dataToBuiltinData, but note that these will not work on-chain.

Constructors

BuiltinData ~Data

class ToData a where

A typeclass for types that can be converted to and from BuiltinData.

Methods

toBuiltinData ∷ a → BuiltinData

Convert a value to BuiltinData.

class FromData a where

Methods

fromBuiltinData ∷ BuiltinData → Maybe a

Convert a value from BuiltinData, returning Nothing if this fails.

class UnsafeFromData a where

Methods

unsafeFromBuiltinData ∷ BuiltinData → a

Convert a value from BuiltinData, calling error if this fails. This is typically much faster than fromBuiltinData.

When implementing this function, make sure to call unsafeFromBuiltinData rather than fromBuiltinData when converting substructures!

This is a simple type without any validation, use with caution.

toData ∷ ToData a ⇒ a → Data

Convert a value to Data.

Note: off-chain only.

fromData ∷ FromData a ⇒ Data → Maybe a

Convert a value from Data, returning Nothing if this fails.

Note: off-chain only.

unsafeFromData ∷ UnsafeFromData a ⇒ Data → a

Convert a value from Data, throwing if this fails.

Note: off-chain only.

dataToBuiltinData ∷ Data → BuiltinData

NOT a builtin. Converts a Data into a BuiltinData. Only works off-chain.

builtinDataToData ∷ BuiltinData → Data

NOT a builtin. Converts a BuiltinData into a Data. Only works off-chain.