PlutusLedgerApi.Test.EvaluationEvent

Documentation

data ScriptEvaluationEvents

This type contains a list of on-chain script evaluation events. All PlutusV1 evaluations (if any) share the same cost parameters. Same with PlutusV2.

Sharing the cost parameters lets us avoid creating a new EvaluationContext for each ScriptEvaluationEvent.

Constructors

ScriptEvaluationEvents

Fields

eventsCostParamsV1 ∷ Maybe [Int64]

Cost parameters shared by all PlutusV1 evaluation events in eventsEvents, if any.

eventsCostParamsV2 ∷ Maybe [Int64]

Cost parameters shared by all PlutusV2 evaluation events in eventsEvents, if any.

eventsEvents ∷ NonEmpty ScriptEvaluationEvent

data ScriptEvaluationEvent

Information about an on-chain script evaluation event, specifically the information needed to evaluate the script, and the expected result.

Constructors

PlutusEvent PlutusLedgerLanguage ScriptEvaluationData ScriptEvaluationResult

data ScriptEvaluationData

All the data needed to evaluate a script using the ledger API, except for the cost model parameters, as these are tracked separately.

Constructors

ScriptEvaluationData

Fields

dataProtocolVersion ∷ MajorProtocolVersion

dataBudget ∷ ExBudget

dataScript ∷ SerialisedScript

dataInputs ∷ [Data]

data ScriptEvaluationResult

Constructors

ScriptEvaluationSuccess

ScriptEvaluationFailure

data UnexpectedEvaluationResult

Error type when re-evaluating a ScriptEvaluationEvent.

Constructors

UnexpectedEvaluationSuccess

Fields

ScriptEvaluationEvent

[Int64]

Cost parameters

ExBudget

Actual budget consumed

UnexpectedEvaluationFailure

Fields

ScriptEvaluationEvent

[Int64]

Cost parameters

EvaluationError

DecodeError ScriptDecodeError

data TestFailure

Constructors

InvalidResult UnexpectedEvaluationResult

MissingCostParametersFor PlutusLedgerLanguage

renderTestFailure ∷ TestFailure → String

renderTestFailures ∷ NonEmpty TestFailure → String

checkEvaluationEvent

Arguments

∷ EvaluationContext

→ [Int64]
Cost parameters

→ ScriptEvaluationEvent

→ Maybe UnexpectedEvaluationResult

Re-evaluate an on-chain script evaluation event.