PlutusLedgerApi.Test.Examples

Description

This module contains example values to be used for testing. These should NOT be used in non-test code!

Documentation

alwaysSucceedingNAryFunction ∷ Natural → SerialisedScript

Creates a script which has N arguments, and always succeeds.

alwaysFailingNAryFunction ∷ Natural → SerialisedScript

Creates a script which has N arguments, and always fails.

summingFunction ∷ SerialisedScript

saltFunction ∷ Integer → SerialisedScript → SerialisedScript

Wrap a script with lambda/app so that, for instance, it has a different hash but the same behavior.

integerToByteStringFunction ∷ SerialisedScript