PlutusLedgerApi.Test.V1.EvaluationContext

Documentation

costModelParamsForTesting ∷ [(ParamName, Int64)]

Example values of costs for PlutusV1, in expected ledger order. Suitable to be used in testing.

mCostModel ∷ MCostModel

The PlutusV1 "cost model" is constructed by the v2 "cost model", by clearing v2 introductions.

clearMachineCostModel ∷ m ~ MCekMachineCosts ⇒ m → m

Assign to mempty those CEK constructs that PlutusV1 introduces (indirectly by introducing a ledger language version with those CEK constructs).

This can be used to generate a (machine) cost model of the previous plutus version, by omitting the generation of the costs concerning the missing PlutusV1 CEK constructs.

clearBuiltinCostModel ∷ m ~ MBuiltinCostModel ⇒ m → m

Assign to mempty those builtins that the PlutusV1 introduces.

This can be used to generate a (builtin) cost model of the previous version by omitting the generation of the costs concerning the missing PlutusV1 builtins.