PlutusLedgerApi.V1.Bytes

Documentation

newtype LedgerBytes

Constructors

LedgerBytes

Fields

getLedgerBytes ∷ BuiltinByteString

data LedgerBytesError

An error that is encountered when converting a ByteString to a LedgerBytes.

Constructors

UnpairedDigit
Odd number of bytes in the original bytestring.

NotHexit !Char
A non-hex digit character ([^A-Fa-f0-9]) encountered during decoding.

fromHex ∷ ByteString → Either LedgerBytesError LedgerBytes

Convert a hex-encoded (Base16) ByteString to a LedgerBytes. May return an error (LedgerBytesError).

bytes ∷ LedgerBytes → ByteString

Extract the Haskell bytestring from inside the Plutus opaque LedgerBytes.

fromBytes ∷ ByteString → LedgerBytes

Lift a Haskell bytestring to the Plutus abstraction LedgerBytes

encodeByteString ∷ ByteString → Text

Encode a ByteString value to Base16 (i.e. hexadecimal), then decode with UTF-8 to a Text.