PlutusLedgerApi.V1.Crypto

Documentation

newtype PubKeyHash

The hash of a public key. This is frequently used to identify the public key, rather than the key itself. Hashed with BLAKE2b-224. 28 bytes.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

PubKeyHash

Fields

getPubKeyHash ∷ BuiltinByteString