PlutusLedgerApi.V1.Data.Tx

Transactions

newtype TxId

A transaction ID, i.e. the hash of a transaction. Hashed with BLAKE2b-256. 32 byte.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification. -- editorconfig-checker-disable-file

Constructors

TxId

Fields

getTxId ∷ BuiltinByteString

data ScriptTag

A tag indicating the type of script that we are pointing to.

See also ScriptPurpose

Constructors

Spend

Mint

Cert

Reward

data RedeemerPtr

A redeemer pointer is a pair of a script type tag (ScriptTag) t and an index i, picking out the i-th script of type t in the transaction.

Constructors

RedeemerPtr ScriptTag Integer

type Redeemers = Map RedeemerPtr Redeemer

Redeemers is a Map of redeemer pointer (RedeemerPtr) and its Redeemer.

Transaction outputs

data TxOut

A transaction output, consisting of a target address (Address), a value (Value), and optionally a datum hash (DatumHash).

pattern TxOut ∷ Address → Value → Maybe DatumHash → TxOut

txOutAddress ∷ TxOut → Address

txOutValue ∷ TxOut → Value

txOutDatumHash ∷ TxOut → Maybe DatumHash

data TxOutRef

A reference to a transaction output. This is a pair of a transaction ID (TxId), and an index indicating which of the outputs of that transaction we are referring to.

pattern TxOutRef ∷ TxId → Integer → TxOutRef

txOutRefId ∷ TxOutRef → TxId

txOutRefIdx ∷ TxOutRef → Integer

isPubKeyOut ∷ TxOut → Bool

Whether the output is a pay-to-pubkey output.

isPayToScriptOut ∷ TxOut → Bool

Whether the output is a pay-to-script output.

outAddress ∷ Lens' TxOut Address

The address of a transaction output.

outValue ∷ Lens' TxOut Value

The value of a transaction output. | TODO: Compute address again

txOutPubKey ∷ TxOut → Maybe PubKeyHash

The public key attached to a TxOut, if there is one.

txOutDatum ∷ TxOut → Maybe DatumHash

The datum attached to a TxOut, if there is one.

pubKeyHashTxOut ∷ Value → PubKeyHash → TxOut

Create a transaction output locked by a public key.