PlutusLedgerApi.V1.Scripts

Description

Functions for working with scripts on the ledger.

Documentation

data ScriptError

A higher-level evaluation error.

Constructors

EvaluationError ![Text] !String
Expected behavior of the engine (e.g. user-provided error)

EvaluationException !String !String
Unexpected behavior of the engine (a bug)

newtype Redeemer

Redeemer is a wrapper around Data values that are used as redeemers in transaction inputs.

Constructors

Redeemer

Fields

getRedeemer ∷ BuiltinData

newtype Datum

Datum is a wrapper around Data values which are used as data in transaction outputs.

Constructors

Datum

Fields

getDatum ∷ BuiltinData

newtype Context

Information about the state of the blockchain and about the transaction that is currently being validated, represented as a value in Data.

Constructors

Context BuiltinData

newtype DatumHash

Type representing the BLAKE2b-256 hash of a datum. 32 bytes.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

DatumHash BuiltinByteString

newtype RedeemerHash

Type representing the BLAKE2b-256 hash of a redeemer. 32 bytes.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

RedeemerHash BuiltinByteString

newtype ScriptHash

Type representing the BLAKE2b-224 hash of a script. 28 bytes.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

ScriptHash

Fields

getScriptHash ∷ BuiltinByteString