PlutusLedgerApi.V3.Contexts

Documentation

newtype ColdCommitteeCredential

Constructors

ColdCommitteeCredential Credential

newtype HotCommitteeCredential

Constructors

HotCommitteeCredential Credential

newtype DRepCredential

Constructors

DRepCredential Credential

data DRep

Constructors

DRep DRepCredential

DRepAlwaysAbstain

DRepAlwaysNoConfidence

data Delegatee

Constructors

DelegStake PubKeyHash

DelegVote DRep

DelegStakeVote PubKeyHash DRep

data TxCert

Constructors

TxCertRegStaking Credential (Maybe Lovelace)
Register staking credential with an optional deposit amount

TxCertUnRegStaking Credential (Maybe Lovelace)
Un-Register staking credential with an optional refund amount

TxCertDelegStaking Credential Delegatee
Delegate staking credential to a Delegatee

TxCertRegDeleg Credential Delegatee Lovelace
Register and delegate staking credential to a Delegatee in one certificate. Note that deposit is mandatory.

TxCertRegDRep DRepCredential Lovelace
Register a DRep with a deposit value. The optional anchor is omitted.

TxCertUpdateDRep DRepCredential
Update a DRep. The optional anchor is omitted.

TxCertUnRegDRep DRepCredential Lovelace
UnRegister a DRep with mandatory refund value

TxCertPoolRegister
A digest of the PoolParams

Fields

PubKeyHash

poolId

PubKeyHash

pool VFR

TxCertPoolRetire PubKeyHash Integer
The retirement certificate and the Epoch in which the retirement will take place

TxCertAuthHotCommittee ColdCommitteeCredential HotCommitteeCredential
Authorize a Hot credential for a specific Committee member's cold credential

TxCertResignColdCommittee ColdCommitteeCredential

data Voter

Constructors

CommitteeVoter HotCommitteeCredential

DRepVoter DRepCredential

StakePoolVoter PubKeyHash

data Vote

A vote. The optional anchor is omitted.

Constructors

VoteNo

VoteYes

Abstain

data GovernanceActionId

Similar to TxOutRef, but for GovActions

Constructors

GovernanceActionId

Fields

gaidTxId ∷ TxId

gaidGovActionIx ∷ Integer

data Committee

Constructors

Committee

Fields

committeeMembers ∷ Map ColdCommitteeCredential Integer

Committee members with epoch number when each of them expires

committeeQuorum ∷ Rational

Quorum of the committee that is necessary for a successful vote

newtype Constitution

A constitution. The optional anchor is omitted.

Constructors

Constitution

Fields

constitutionScript ∷ Maybe ScriptHash

data ProtocolVersion

Constructors

ProtocolVersion

Fields

pvMajor ∷ Integer

pvMinor ∷ Integer

newtype ChangedParameters

A Plutus Data object containing proposed parameter changes. The Data object contains a Map with one entry per changed parameter, from the parameter ID to the new value. Unchanged parameters are not included.

The mapping from parameter IDs to parameters can be found in conway.cddl.

Invariant: This map is non-empty, and the keys are stored in ascending order.

This Data object has the following format (in pseudocode):

ChangedParametersData = Map ChangedIdData ChangedManyValueData ChangedIdData = I Integer ChangedManyValueData = ChangedSingleValueData | List[ChangedSingleValueData...] -- ^ an arbitrary-length, heterogeneous (integer or ratio) list of values (to support sub-parameters)

ChangedSingleValueData = I Integer -- a proposed integer value | List[I Integer, I Integer] -- a proposed numerator,denominator (ratio value) -- ^ a 2-exact element list; *BE CAREFUL* because this can be alternatively (ambiguously) interpreted -- as a many-value data (sub-parameter) of two integer single-value data.

, where Map,I,List are the constructors of Data and Integer is the usual arbitrary-precision PlutusTx/Haskell Integer.

Constructors

ChangedParameters

Fields

getChangedParameters ∷ BuiltinData

data GovernanceAction

Constructors

ParameterChange (Maybe GovernanceActionId) ChangedParameters (Maybe ScriptHash)
Hash of the constitution script

HardForkInitiation (Maybe GovernanceActionId) ProtocolVersion
proposal to update protocol version

TreasuryWithdrawals (Map Credential Lovelace) (Maybe ScriptHash)
Hash of the constitution script

NoConfidence (Maybe GovernanceActionId)

UpdateCommittee

Fields

(Maybe GovernanceActionId)

[ColdCommitteeCredential]

Committee members to be removed

(Map ColdCommitteeCredential Integer)

Committee members to be added

Rational

New quorum

NewConstitution (Maybe GovernanceActionId) Constitution

InfoAction

data ProposalProcedure

A proposal procedure. The optional anchor is omitted.

Constructors

ProposalProcedure

Fields

ppDeposit ∷ Lovelace

ppReturnAddr ∷ Credential

ppGovernanceAction ∷ GovernanceAction

data ScriptPurpose

A ScriptPurpose uniquely identifies a Plutus script within a transaction.

Constructors

Minting CurrencySymbol

Spending TxOutRef

Rewarding Credential

Certifying

Fields

Integer

0-based index of the given TxCert in txInfoTxCerts

TxCert

Voting Voter

Proposing

Fields

Integer

0-based index of the given ProposalProcedure in txInfoProposalProcedures

ProposalProcedure

data ScriptInfo

Like ScriptPurpose but with an optional datum for spending scripts.

Constructors

MintingScript CurrencySymbol

SpendingScript TxOutRef (Maybe Datum)

RewardingScript Credential

CertifyingScript

Fields

Integer

0-based index of the given TxCert in txInfoTxCerts

TxCert

VotingScript Voter

ProposingScript

Fields

Integer

0-based index of the given ProposalProcedure in txInfoProposalProcedures

ProposalProcedure

data TxInInfo

An input of a pending transaction.

Constructors

TxInInfo

Fields

txInInfoOutRef ∷ TxOutRef

txInInfoResolved ∷ TxOut

data TxInfo

TxInfo for PlutusV3

Constructors

TxInfo

Fields

txInfoInputs ∷ [TxInInfo]

txInfoReferenceInputs ∷ [TxInInfo]

txInfoOutputs ∷ [TxOut]

txInfoFee ∷ Lovelace

txInfoMint ∷ MintValue

The Value minted by this transaction.

Invariant: This field does not contain Ada with zero quantity, unlike their namesakes in Plutus V1 and V2's ScriptContexts.

txInfoTxCerts ∷ [TxCert]

txInfoWdrl ∷ Map Credential Lovelace

txInfoValidRange ∷ POSIXTimeRange

txInfoSignatories ∷ [PubKeyHash]

txInfoRedeemers ∷ Map ScriptPurpose Redeemer

txInfoData ∷ Map DatumHash Datum

txInfoId ∷ TxId

txInfoVotes ∷ Map Voter (Map GovernanceActionId Vote)

txInfoProposalProcedures ∷ [ProposalProcedure]

txInfoCurrentTreasuryAmount ∷ Maybe Lovelace

txInfoTreasuryDonation ∷ Maybe Lovelace

data ScriptContext

The context that the currently-executing script can access.

Constructors

ScriptContext

Fields

scriptContextTxInfo ∷ TxInfo

information about the transaction the currently-executing script is included in

scriptContextRedeemer ∷ Redeemer

Redeemer for the currently-executing script

scriptContextScriptInfo ∷ ScriptInfo

the purpose of the currently-executing script, along with information associated with the purpose

findOwnInput ∷ ScriptContext → Maybe TxInInfo

Find the input currently being validated.

findDatum ∷ DatumHash → TxInfo → Maybe Datum

Find the data corresponding to a data hash, if there is one

findDatumHash ∷ Datum → TxInfo → Maybe DatumHash

Find the hash of a datum, if it is part of the pending transaction's hashes

findTxInByTxOutRef ∷ TxOutRef → TxInfo → Maybe TxInInfo

Given a UTXO reference and a transaction (TxInfo), resolve it to one of the transaction's inputs (TxInInfo).

Note: this only searches the true transaction inputs and not the referenced transaction inputs.

findContinuingOutputs ∷ ScriptContext → [Integer]

Find the indices of all the outputs that pay to the same script address we are currently spending from, if any.

getContinuingOutputs ∷ ScriptContext → [TxOut]

Get all the outputs that pay to the same script address we are currently spending from, if any.

txSignedBy ∷ TxInfo → PubKeyHash → Bool

Check if a transaction was signed by the given public key.

Validator functions

pubKeyOutputsAt ∷ PubKeyHash → TxInfo → [Value]

Get the values paid to a public key address by a pending transaction.

valuePaidTo ∷ TxInfo → PubKeyHash → Value

Get the total value paid to a public key address by a pending transaction.

valueSpent ∷ TxInfo → Value

Get the total value of inputs spent by this transaction.

valueProduced ∷ TxInfo → Value

Get the total value of outputs produced by this transaction.

ownCurrencySymbol ∷ ScriptContext → CurrencySymbol

The CurrencySymbol of the current validator script.

spendsOutput ∷ TxInfo → TxId → Integer → Bool

Check if the pending transaction spends a specific transaction output (identified by the hash of a transaction and an index into that transactions' outputs)