PlutusLedgerApi.V3.Data.Tx

Documentation

newtype TxId

A transaction ID, i.e. the hash of a transaction. Hashed with BLAKE2b-256. 32 byte.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

TxId

Fields

getTxId ∷ BuiltinByteString

data TxOutRef

A reference to a transaction output. This is a pair of a transaction ID (TxId), and an index indicating which of the outputs of that transaction we are referring to.

pattern TxOutRef ∷ TxId → Integer → TxOutRef

txOutRefId ∷ TxOutRef → TxId

txOutRefIdx ∷ TxOutRef → Integer