PlutusLedgerApi.V4.Contexts

Documentation

newtype AccountId

Constructors

AccountId Credential

data AccountBalanceInterval

Constructors

AccountBalanceLowerBound Lovelace

AccountBalanceUpperBound Lovelace

AccountBalanceBothBounds Lovelace Lovelace

AccountBalanceExact Lovelace

newtype AccountBalanceIntervals

Constructors

AccountBalanceIntervals (Map AccountId AccountBalanceInterval)

newtype ColdCommitteeCredential

Constructors

ColdCommitteeCredential Credential

newtype HotCommitteeCredential

Constructors

HotCommitteeCredential Credential

newtype DRepCredential

Constructors

DRepCredential Credential

data DRep

Constructors

DRep DRepCredential

DRepAlwaysAbstain

DRepAlwaysNoConfidence

data Delegatee

Constructors

DelegStake PubKeyHash

DelegVote DRep

DelegStakeVote PubKeyHash DRep

data TxCert

Constructors

TxCertRegAccount AccountId Lovelace

TxCertUnRegAccount AccountId Lovelace

TxCertDelegAccount AccountId Delegatee

TxCertRegAccountDeleg AccountId Delegatee Lovelace

TxCertRegDRep DRepCredential Lovelace

TxCertUpdateDRep DRepCredential

TxCertUnRegDRep DRepCredential Lovelace

TxCertPoolRegister PubKeyHash PubKeyHash

TxCertPoolRetire PubKeyHash Integer

TxCertAuthHotCommittee ColdCommitteeCredential HotCommitteeCredential

TxCertResignColdCommittee ColdCommitteeCredential

data Voter

Constructors

CommitteeVoter HotCommitteeCredential

DRepVoter DRepCredential

StakePoolVoter PubKeyHash

data Vote

A vote. The optional anchor is omitted.

Constructors

VoteNo

VoteYes

Abstain

data GovernanceActionId

Similar to TxOutRef, but for GovActions

Constructors

GovernanceActionId

Fields

gaidTxId ∷ TxId

gaidGovActionIx ∷ Integer

data Committee

Constructors

Committee

Fields

committeeMembers ∷ Map ColdCommitteeCredential Integer

Committee members with epoch number when each of them expires

committeeQuorum ∷ Rational

Quorum of the committee that is necessary for a successful vote

newtype Constitution

A constitution. The optional anchor is omitted.

Constructors

Constitution

Fields

constitutionScript ∷ Maybe ScriptHash

data ProtocolVersion

Constructors

ProtocolVersion

Fields

pvMajor ∷ Integer

pvMinor ∷ Integer

newtype ChangedParameters

A Plutus Data object containing proposed parameter changes. The Data object contains a Map with one entry per changed parameter, from the parameter ID to the new value. Unchanged parameters are not included.

The mapping from parameter IDs to parameters can be found in conway.cddl.

Invariant: This map is non-empty, and the keys are stored in ascending order.

This Data object has the following format (in pseudocode):

ChangedParametersData = Map ChangedIdData ChangedManyValueData ChangedIdData = I Integer ChangedManyValueData = ChangedSingleValueData | List[ChangedSingleValueData...] -- ^ an arbitrary-length, heterogeneous (integer or ratio) list of values (to support sub-parameters)

ChangedSingleValueData = I Integer -- a proposed integer value | List[I Integer, I Integer] -- a proposed numerator,denominator (ratio value) -- ^ a 2-exact element list; *BE CAREFUL* because this can be alternatively (ambiguously) interpreted -- as a many-value data (sub-parameter) of two integer single-value data.

, where Map,I,List are the constructors of Data and Integer is the usual arbitrary-precision PlutusTx/Haskell Integer.

Constructors

ChangedParameters

Fields

getChangedParameters ∷ BuiltinData

data GovernanceAction

Constructors

ParameterChange (Maybe GovernanceActionId) ChangedParameters (Maybe ScriptHash)
Hash of the constitution script

HardForkInitiation (Maybe GovernanceActionId) ProtocolVersion
proposal to update protocol version

TreasuryWithdrawals (Map Credential Lovelace) (Maybe ScriptHash)
Hash of the constitution script

NoConfidence (Maybe GovernanceActionId)

UpdateCommittee

Fields

(Maybe GovernanceActionId)

[ColdCommitteeCredential]

Committee members to be removed

(Map ColdCommitteeCredential Integer)

Committee members to be added

Rational

New quorum

NewConstitution (Maybe GovernanceActionId) Constitution

InfoAction

data ProposalProcedure

A proposal procedure. The optional anchor is omitted.

Constructors

ProposalProcedure

Fields

ppDeposit ∷ Lovelace

ppReturnAddr ∷ Credential

ppGovernanceAction ∷ GovernanceAction

data ScriptPurpose

Constructors

Minting ScriptHash CurrencySymbol

Spending ScriptHash TxOutRef

Withdrawing ScriptHash Credential

Certifying ScriptHash Integer TxCert

Voting ScriptHash Voter

Proposing ScriptHash Integer ProposalProcedure

Guarding ScriptHash Integer

data ScriptInfo

Constructors

MintingScript CurrencySymbol

SpendingScript TxOutRef (Maybe Datum)

WithdrawingScript AccountId

CertifyingScript Integer TxCert

VotingScript Voter

ProposingScript Integer ProposalProcedure

GuardingScript Integer (Maybe TopTxInfo)

data TxInInfo

An input of a pending transaction.

Constructors

TxInInfo

Fields

txInInfoOutRef ∷ TxOutRef

txInInfoResolved ∷ TxOut

data TxInfo

Constructors

TxInfo

Fields

txInfoId ∷ TxId

txInfoSubTxIx ∷ Maybe Integer

txInfoInputs ∷ [TxInInfo]

txInfoReferenceInputs ∷ [TxInInfo]

txInfoOutputs ∷ [TxOut]

txInfoFee ∷ Lovelace

txInfoMint ∷ MintValue

txInfoTxCerts ∷ [TxCert]

txInfoWithdrawals ∷ Map AccountId Lovelace

txInfoDirectDeposits ∷ Map AccountId Lovelace

txInfoAccountBalanceIntervals ∷ AccountBalanceIntervals

txInfoValidRange ∷ POSIXTimeRange

txInfoGuards ∷ [Credential]

txInfoRequiredTopLevelGuards ∷ Map Credential (Maybe Datum)

txInfoRedeemers ∷ Map ScriptPurpose Redeemer

txInfoData ∷ Map DatumHash Datum

txInfoVotes ∷ Map Voter (Map GovernanceActionId Vote)

txInfoProposalProcedures ∷ [ProposalProcedure]

txInfoCurrentTreasuryAmount ∷ Maybe Lovelace

txInfoTreasuryDonation ∷ Lovelace

data TopTxInfoSimplified

Constructors

TopTxInfoSimplified

Fields

ttisIds ∷ [TxId]

ttisInputs ∷ [TxInInfo]

ttisReferenceInputs ∷ [TxInInfo]

ttisOutputs ∷ [TxOut]

ttisMints ∷ MintValue

ttisBurns ∷ MintValue

ttisTxCerts ∷ [TxCert]

ttisWithdrawals ∷ Map AccountId Lovelace

ttisDirectDeposits ∷ Map AccountId Lovelace

ttisValidRange ∷ POSIXTimeRange

ttisGuards ∷ Map Credential ()

ttisScriptPurposes ∷ Map ScriptPurpose ()

ttisData ∷ Map DatumHash Datum

ttisVotes ∷ Map Voter (Map GovernanceActionId Vote)

ttisProposalProcedures ∷ [ProposalProcedure]

ttisCurrentTreasuryAmount ∷ Maybe Lovelace

ttisTreasuryDonations ∷ Lovelace

data TopTxInfo

Constructors

TopTxInfo

Fields

topTxInfoSubTransactions ∷ [TxInfo]

topTxInfoDatums ∷ Map Integer Datum

topTxInfoStartingBalanceIntervals ∷ AccountBalanceIntervals

topTxInfoSimplified ∷ TopTxInfoSimplified

data ScriptContext

Constructors

ScriptContext

Fields

scriptContextTxInfo ∷ TxInfo

scriptContextRedeemer ∷ Redeemer

scriptContextScriptInfo ∷ ScriptInfo

scriptContextScriptHash ∷ ScriptHash

findOwnInput ∷ ScriptContext → Maybe TxInInfo

findDatum ∷ DatumHash → TxInfo → Maybe Datum

findDatumHash ∷ Datum → TxInfo → Maybe DatumHash

findTxInByTxOutRef ∷ TxOutRef → TxInfo → Maybe TxInInfo

findContinuingOutputs ∷ ScriptContext → [Integer]

Find the indices of outputs in the current sub-transaction or top-level transaction that pay to the same script address we are currently spending from. This does not search the outputs of the whole transaction.

getContinuingOutputs ∷ ScriptContext → [TxOut]

Get the outputs in the current sub-transaction or top-level transaction that pay to the same script address we are currently spending from. This does not search the outputs of the whole transaction.

txSignedBy ∷ TxInfo → PubKeyHash → Bool

pubKeyOutputsAt ∷ PubKeyHash → TxInfo → [Value]

valuePaidTo ∷ TxInfo → PubKeyHash → Value

valueSpent ∷ TxInfo → Value

valueProduced ∷ TxInfo → Value

ownCurrencySymbol ∷ ScriptContext → CurrencySymbol

spendsOutput ∷ TxInfo → TxId → Integer → Bool