PlutusLedgerApi.V4.Data.Tx

Transactions

newtype TxId

A transaction ID, i.e. the hash of a transaction. Hashed with BLAKE2b-256. 32 byte.

This is a simple type without any validation, use with caution. You may want to add checks for its invariants. See the Shelley ledger specification.

Constructors

TxId

Fields

getTxId ∷ BuiltinByteString

Transaction outputs

data TxOutRef

A reference to a transaction output. This is a pair of a transaction ID (TxId), and an index indicating which of the outputs of that transaction we are referring to.

pattern TxOutRef ∷ TxId → Integer → TxOutRef

matchTxOutRef ∷ ∀ r. TxOutRef → (TxId → Integer → r) → r

txOutRefId ∷ TxOutRef → TxId

txOutRefIdx ∷ TxOutRef → Integer

data OutputDatum

The datum attached to an output: either nothing; a datum hash; or the datum itself (an "inline datum").

matchOutputDatum ∷ ∀ r. OutputDatum → r → (DatumHash → r) → (Datum → r) → r

pattern NoOutputDatum ∷ OutputDatum

pattern OutputDatumHash ∷ DatumHash → OutputDatum

pattern OutputDatum ∷ Datum → OutputDatum

data TxOut

Transaction output for Plutus V4.

pattern TxOut ∷ Address → Value → OutputDatum → Maybe ScriptHash → TxOut

matchTxOut ∷ ∀ r. TxOut → (Address → Value → OutputDatum → Maybe ScriptHash → r) → r

txOutAddress ∷ TxOut → Address

txOutValue ∷ TxOut → Value

txOutDatum ∷ TxOut → OutputDatum

txOutReferenceScript ∷ TxOut → Maybe ScriptHash

txOutPubKey ∷ TxOut → Maybe PubKeyHash

The public key attached to a TxOut, if there is one.

txOutScriptHash ∷ TxOut → Maybe ScriptHash

The validator hash attached to a TxOut, if there is one.

isPubKeyOut ∷ TxOut → Bool

Whether the output is a pay-to-pubkey output.

isPayToScriptOut ∷ TxOut → Bool

Whether the output is a pay-to-script output.

pubKeyHashTxOut ∷ Value → PubKeyHash → TxOut

Create a transaction output locked by a public key.