PlutusTx.AssocMap

Description

A map represented as an "association list" of key-value pairs.

Documentation

data Map k v

A Map of key-value pairs. A Map is considered well-defined if there are no key collisions, meaning that each value is uniquely identified by a key.

Use safeFromList to create well-defined Maps from arbitrary lists of pairs.

If cost minimisation is required, then you can use unsafeFromList but you must be certain that the list you are converting to a Map abides by the well-definedness condition.

Most operations on Maps are definedness-preserving, meaning that for the resulting Map to be well-defined then the input Map(s) have to also be well-defined. This is not checked explicitly unless mentioned in the documentation.

Take care when using fromBuiltinData and unsafeFromBuiltinData, as neither function performs deduplication of the input collection and may create invalid Maps!

singleton ∷ k → v → Map k v

A singleton map.

empty ∷ Map k v

An empty Map.

null ∷ Map k v → Bool

Is the map empty?

unsafeFromList ∷ [(k, v)] → Map k v

Unsafely create a Map from a list of pairs. This should _only_ be applied to lists which have been checked to not contain duplicate keys, otherwise the resulting Map will contain conflicting entries (two entries sharing the same key). As usual, the "keys" are considered to be the first element of the pair.

safeFromList ∷ Eq k ⇒ [(k, v)] → Map k v

In case of duplicates, this function will keep only one entry (the one that precedes). In other words, this function de-duplicates the input list.

toList ∷ Map k v → [(k, v)]

keys ∷ Map k v → [k]

The keys of a Map. Semantically, the resulting list is only a set if the Map didn't contain duplicate keys.

elems ∷ Map k v → [v]

Return all elements of the map.

lookup ∷ ∀ k v. Eq k ⇒ k → Map k v → Maybe v

Find an entry in a Map. If the Map is not well-formed (it contains duplicate keys) then this will return the value of the left-most pair in the underlying list of pairs.

member ∷ ∀ k v. Eq k ⇒ k → Map k v → Bool

Is the key a member of the map?

insert ∷ ∀ k v. Eq k ⇒ k → v → Map k v → Map k v

If a key already exists in the map, its entry will be replaced with the new value.

delete ∷ ∀ k v. Eq k ⇒ k → Map k v → Map k v

Delete an entry from the Map. Assumes that the Map is well-formed, i.e. if the underlying list of pairs contains pairs with duplicate keys then only the left-most pair will be removed.

union ∷ ∀ k v r. Eq k ⇒ Map k v → Map k r → Map k (These v r)

Combine two Maps. Keeps both values on key collisions. Note that well-formedness is only preserved if the two input maps are also well-formed. Also, as an implementation detail, in the case that the right map contains duplicate keys, and there exists a collision between the two maps, then only the left-most value of the right map will be kept.

unionWith ∷ ∀ k a. Eq k ⇒ (a → a → a) → Map k a → Map k a → Map k a

Combine two Maps with the given combination function. Note that well-formedness of the resulting map depends on the two input maps being well-formed. Also, as an implementation detail, in the case that the right map contains duplicate keys, and there exists a collision between the two maps, then only the left-most value of the right map will be kept.

filter ∷ (v → Bool) → Map k v → Map k v

Filter all values that satisfy the predicate.

mapWithKey ∷ (k → a → b) → Map k a → Map k b

Map a function over all values in the map.

mapMaybe ∷ (a → Maybe b) → Map k a → Map k b

Map keys/values and collect the Just results.

mapMaybeWithKey ∷ (k → a → Maybe b) → Map k a → Map k b

Map keys/values and collect the Just results.

all ∷ (a → Bool) → Map k a → Bool

Determines whether all elements in the map satisfy the predicate.

mapThese ∷ (v → These a b) → Map k v → (Map k a, Map k b)

A version of mapEither that works with These.