PlutusTx.Blueprint.Argument

Documentation

data ArgumentBlueprint (referencedTypes ∷ [Type])

Blueprint that defines a validator's runtime argument: datum or redeemer.

Constructors

MkArgumentBlueprint

Fields

argumentTitle ∷ Maybe Text

A short and descriptive name for the redeemer or datum.

argumentDescription ∷ Maybe Text

An informative description of the redeemer or datum.

argumentPurpose ∷ Set Purpose

A possibly empty set of purposes for the redeemer or datum.

argumentSchema ∷ Schema referencedTypes

A Plutus Data Schema.