PlutusTx.Blueprint.Definition.Derive

Documentation

deriveDefinitions ∷ ∀ ts. DefinitionsFor (UnrollAll ts) ⇒ Definitions (UnrollAll ts)

Derive a Definitions value for a list of types.

definitionRef ∷ ∀ t ts. HasBlueprintDefinition t ⇒ Schema ts

Construct a Schema that is a reference to a schema definition.

type DefinitionsFor ts = DefinitionsFor' ts ts

This class and its two instances are used internally to derive Definitions for a given list of types.

definitionsFor ∷ ∀ ts. DefinitionsFor ts ⇒ Definitions ts

class DefinitionsFor' referencedTypes acc where

Methods

definitionsFor' ∷ Definitions referencedTypes