PlutusTx.Blueprint.Definition.Unroll

Documentation

class HasBlueprintDefinition (t ∷ Type) where

Designates a class of types that could be used as a Blueprint Definition. Each such type: - could be unrolled to a list of all nested types (including the type itself). - has a unique DefinitionId.

Minimal complete definition

Nothing

Associated Types

type Unroll t ∷ [Type]

type Unroll t = Insert t (GUnroll (IfStuckRep (RepIsStuckError t) (Rep t)))

Methods

definitionId ∷ DefinitionId

default definitionId ∷ Typeable t ⇒ DefinitionId

type family UnrollIsStuckError x where ...

Compile-time error that happens when a type couldn't be unrolled (Unroll TF is "stuck")

Equations

UnrollIsStuckError x = TypeError (Text "No instance: " :<>: ShowType (HasBlueprintDefinition x))

type family RepIsStuckError x where ...

Compile-time error that happens when type's generic representation is not defined (Rep TF is "stuck")

Equations

RepIsStuckError x = TypeError (Text "No instance: " :<>: ShowType (Generic x))

type Unrolled t = Reverse (IfStuckUnroll (UnrollIsStuckError t) (Unroll t))

Same as Unroll but with a nicer error message

type family UnrollAll xs ∷ [Type] where ...

Unrolls all types in the list xs

Equations

UnrollAll '[] = '[]

UnrollAll (x ': xs) = Nub (Concat (Unrolled x) (UnrollAll xs))

type family GUnroll (t ∷ Type → Type) ∷ [Type] where ...

Unroll a generic representation of a type into a list of all nested types.

Equations

GUnroll (M1 _ _ f) = GUnroll f

GUnroll (f :*: g) = GUnroll f ++ GUnroll g

GUnroll (f :+: g) = GUnroll f ++ GUnroll g

GUnroll (K1 _ c) = Unrolled c

GUnroll U1 = '[]