PlutusTx.Blueprint.Parameter

Documentation

data ParameterBlueprint (referencedTypes ∷ [Type])

Blueprint that defines validator's compile-time parameter.

The referencedTypes phantom type parameter is used to track the types used in the contract making sure their schemas are included in the blueprint and that they are referenced in a type-safe way.

Constructors

MkParameterBlueprint

Fields

parameterTitle ∷ Maybe Text

A short and descriptive name for the parameter.

parameterDescription ∷ Maybe Text

An informative description of the parameter.

parameterPurpose ∷ Set Purpose

One of "spend", "mint", "withdraw" or "publish", or a oneOf applicator of those.

parameterSchema ∷ Schema referencedTypes

A Plutus Data Schema.

oneOfASet ∷ ToJSON a ⇒ Set a → Maybe Value