PlutusTx.Blueprint.Preamble

Documentation

data Preamble

Meta-information about the contract

Constructors

MkPreamble

Fields

preambleTitle ∷ Text

A short and descriptive title of the contract application

preambleDescription ∷ Maybe Text

A more elaborate description

preambleVersion ∷ Text

A version number for the project.

preamblePlutusVersion ∷ PlutusVersion

The Plutus version assumed for all validators

preambleLicense ∷ Maybe Text

A license under which the specification and contract code is distributed