PlutusTx.Blueprint.Schema

Documentation

data Schema (referencedTypes ∷ [Type])

Blueprint schema definition, as defined by the CIP-0057: https://github.com/cardano-foundation/CIPs/tree/master/CIP-0057#core-vocabulary

The referencedTypes phantom type parameter is used to track the types used in the contract making sure their schemas are included in the blueprint and that they are referenced in a type-safe way.

Constructors

SchemaInteger SchemaInfo IntegerSchema

SchemaBytes SchemaInfo BytesSchema

SchemaList SchemaInfo (ListSchema referencedTypes)

SchemaMap SchemaInfo (MapSchema referencedTypes)

SchemaConstructor SchemaInfo (ConstructorSchema referencedTypes)

SchemaBuiltInData SchemaInfo

SchemaBuiltInUnit SchemaInfo

SchemaBuiltInBoolean SchemaInfo

SchemaBuiltInInteger SchemaInfo

SchemaBuiltInBytes SchemaInfo

SchemaBuiltInString SchemaInfo

SchemaBuiltInPair SchemaInfo (PairSchema referencedTypes)

SchemaBuiltInList SchemaInfo (Schema referencedTypes)

SchemaOneOf (NonEmpty (Schema referencedTypes))

SchemaAnyOf (NonEmpty (Schema referencedTypes))

SchemaAllOf (NonEmpty (Schema referencedTypes))

SchemaNot (Schema referencedTypes)

SchemaDefinitionRef DefinitionId

withSchemaInfo ∷ (SchemaInfo → SchemaInfo) → Schema referencedTypes → Schema referencedTypes

data IntegerSchema

Constructors

MkIntegerSchema

Fields

multipleOf ∷ Maybe Integer

An instance is valid if division by this value results in an integer.

minimum ∷ Maybe Integer

An instance is valid only if it is greater than or exactly equal to "minimum".

maximum ∷ Maybe Integer

An instance is valid only if it is less than or exactly equal to "maximum".

exclusiveMinimum ∷ Maybe Integer

An instance is valid only if it is strictly greater than "exclusiveMinimum".

exclusiveMaximum ∷ Maybe Integer

An instance is valid only if it is strictly less than "exclusiveMaximum".

emptyIntegerSchema ∷ IntegerSchema

data BytesSchema

Constructors

MkBytesSchema

Fields

enum ∷ [ByteString]

An instance validates successfully if once hex-encoded, its value matches one of the specified values.

minLength ∷ Maybe Natural

An instance is valid if its length is greater than, or equal to, this value.

maxLength ∷ Maybe Natural

An instance is valid if its length is less than, or equal to, this value.

emptyBytesSchema ∷ BytesSchema

data ListSchema (referencedTypes ∷ [Type])

Constructors

MkListSchema

Fields

itemSchema ∷ Schema referencedTypes

Element schema

minItems ∷ Maybe Natural

An array instance is valid if its size is greater than, or equal to, this value.

maxItems ∷ Maybe Natural

An array instance is valid if its size is less than, or equal to, this value.

uniqueItems ∷ Maybe Bool

If this value is false, the instance validates successfully. If it is set to True, the instance validates successfully if all of its elements are unique.

mkListSchema ∷ Schema referencedTypes → ListSchema referencedTypes

data MapSchema (referencedTypes ∷ [Type])

Constructors

MkMapSchema

Fields

keySchema ∷ Schema referencedTypes

Key schema

valueSchema ∷ Schema referencedTypes

Value schema

minItems ∷ Maybe Natural

A map instance is valid if its size is greater than, or equal to, this value.

maxItems ∷ Maybe Natural

A map instance is valid if its size is less than, or equal to, this value.

data ConstructorSchema (referencedTypes ∷ [Type])

Constructors

MkConstructorSchema

Fields

index ∷ Natural

Constructor index

fieldSchemas ∷ [Schema referencedTypes]

Field schemas

data PairSchema (referencedTypes ∷ [Type])

Constructors

MkPairSchema

Fields

left ∷ Schema referencedTypes

Schema of the first element

right ∷ Schema referencedTypes

Schema of the second element