PlutusTx.Blueprint.Schema.Annotation

Documentation

data SchemaInfo

Additional information optionally attached to any datatype schema definition.

Constructors

MkSchemaInfo

Fields

title ∷ Maybe String

description ∷ Maybe String

comment ∷ Maybe String

emptySchemaInfo ∷ SchemaInfo

annotationsToSchemaInfo ∷ [SchemaAnn] → Either SchemaInfoError SchemaInfo

data SchemaAnn

Annotation that can be attached to a schema definition.

Constructors

MkSchemaAnnTitle SchemaTitle

MkSchemaAnnDescription SchemaDescription

MkSchemaAnnComment SchemaComment

newtype SchemaTitle

An annotation for the "title" schema attribute.

This annotation could be attached to a type or constructor: {-# ANN type MyFoo (SchemaTitle "My Foo Title") #-} {-# ANN MkMyFoo (SchemaTitle Title) #-} newtype MyFoo = MkMyFoo Int

Constructors

SchemaTitle

Fields

schemaTitleToString ∷ String

newtype SchemaDescription

An annotation for the "description" schema attribute.

This annotation could be attached to a type or constructor: {-# ANN type MyFoo (SchemaDescription "My Foo Description") #-} {-# ANN MkMyFoo (SchemaDescription Description) #-} newtype MyFoo = MkMyFoo Int

Constructors

SchemaDescription

Fields

schemaDescriptionToString ∷ String

newtype SchemaComment

An annotation for the "$comment" schema attribute.

This annotation could be attached to a type or constructor: {-# ANN type MyFoo (SchemaComment "My Foo Comment") #-} {-# ANN MkMyFoo (SchemaComment Comment) #-} newtype MyFoo = MkMyFoo Int

Constructors

SchemaComment

Fields

schemaCommentToString ∷ String