PlutusTx.Blueprint.TH

Documentation

makeIsDataSchemaIndexed ∷ Name → [(Name, Natural)] → Q [InstanceDec]

Generate a ToData, FromData, UnsafeFromData, HasBlueprintSchema instances for a type, using an explicit mapping of constructor names to indices. Use this for types where you need to keep the representation stable.

unstableMakeIsDataSchema ∷ Name → Q [Dec]

makeHasSchemaInstance ∷ Name → [(Name, Natural)] → Q [InstanceDec]

mkSchemaClause

Arguments

∷ Type
The type for the HasBlueprintSchema instance.

→ [(ConstructorInfo, SchemaInfo, Natural)]
The constructors of the type with their schema infos and indices.

→ ClauseQ
The clause for the schema function.

Make a clause for the schema function.

deriveParameterBlueprint ∷ Name → Set Purpose → ExpQ

deriveArgumentBlueprint ∷ Name → Set Purpose → ExpQ

lookupAnn ∷ Data a ⇒ Name → Q [a]

lookupSchemaTitle ∷ Name → Q (Maybe SchemaTitle)

lookupSchemaDescription ∷ Name → Q (Maybe SchemaDescription)