PlutusTx.Blueprint.Validator

Documentation

data ValidatorBlueprint (referencedTypes ∷ [Type])

A blueprint of a validator, as defined by the CIP-0057

The referencedTypes phantom type parameter is used to track the types used in the contract making sure their schemas are included in the blueprint and that they are referenced in a type-safe way.

Constructors

MkValidatorBlueprint

Fields

validatorTitle ∷ Text

A short and descriptive name for the validator.

validatorDescription ∷ Maybe Text

An informative description of the validator.

validatorRedeemer ∷ ArgumentBlueprint referencedTypes

A description of the redeemer format expected by this validator.

validatorDatum ∷ Maybe (ArgumentBlueprint referencedTypes)

A description of the datum format expected by this validator.

validatorParameters ∷ [ParameterBlueprint referencedTypes]

A list of parameters required by the script.

validatorCompiled ∷ Maybe CompiledValidator

A full compiled and CBOR-encoded serialized flat script together with its hash.

data CompiledValidator

Constructors

MkCompiledValidator

Fields

compiledValidatorCode ∷ ByteString

compiledValidatorHash ∷ ByteString

compiledValidator ∷ PlutusVersion → ByteString → CompiledValidator