PlutusTx.Compiler.Binders

Description

Convenient functions for compiling binders.

Documentation

withVarScoped ∷ CompilingDefault uni fun m ann ⇒ Var → Ann → Maybe (PIRTerm uni fun) → (VarDecl TyName Name uni Ann → m a) → m a

withVarTyScoped ∷ CompilingDefault uni fun m ann ⇒ Var → PIRType uni → (VarDecl TyName Name uni Ann → m a) → m a

Like withVarScoped, but takes a PIRType, and uses it for the type of the compiled Expr.

withVarsScoped ∷ CompilingDefault uni fun m ann ⇒ [(Var, Maybe (PIRTerm uni fun))] → ([VarDecl TyName Name uni Ann] → m a) → m a

withTyVarScoped ∷ Compiling uni fun m ann ⇒ Var → (TyVarDecl TyName Ann → m a) → m a

withTyVarsScoped ∷ Compiling uni fun m ann ⇒ [Var] → ([TyVarDecl TyName Ann] → m a) → m a

mkLamAbsScoped ∷ CompilingDefault uni fun m ann ⇒ Ann → Var → m (PIRTerm uni fun) → m (PIRTerm uni fun)

Builds a lambda, binding the given variable to a name that will be in scope when running the second argument.

mkTyAbsScoped ∷ Compiling uni fun m ann ⇒ Var → m (PIRTerm uni fun) → m (PIRTerm uni fun)

Builds a type abstraction, binding the given variable to a name that will be in scope when running the second argument.

mkIterTyAbsScoped ∷ Compiling uni fun m ann ⇒ [Var] → m (PIRTerm uni fun) → m (PIRTerm uni fun)

mkTyForallScoped ∷ Compiling uni fun m ann ⇒ Var → m (PIRType uni) → m (PIRType uni)

Builds a forall, binding the given variable to a name that will be in scope when running the second argument.

mkIterTyForallScoped ∷ Compiling uni fun m ann ⇒ [Var] → m (PIRType uni) → m (PIRType uni)

mkTyLamScoped ∷ Compiling uni fun m ann ⇒ Var → m (PIRType uni) → m (PIRType uni)

Builds a type lambda, binding the given variable to a name that will be in scope when running the second argument.

mkIterTyLamScoped ∷ Compiling uni fun m ann ⇒ [Var] → m (PIRType uni) → m (PIRType uni)