PlutusTx.Compiler.Builtins

Description

Functions for compiling Plutus Core builtins.

Documentation

builtinNames ∷ [Name]

The Names for which NameInfo needs to be provided.

defineBuiltinTypes ∷ CompilingDefault uni fun m ann ⇒ m ()

defineBuiltinTerms ∷ CompilingDefault uni fun m ann ⇒ m ()

Add definitions for all the builtin terms to the environment.

defineBoolType ∷ ∀ uni fun m ann. CompilingDefault uni fun m ann ⇒ m ()

lookupBuiltinTerm ∷ Compiling uni fun m ann ⇒ Name → m (PIRTerm uni fun)

Lookup a builtin term by its TH name. These are assumed to be present, so fails if it cannot find it.

lookupBuiltinType ∷ Compiling uni fun m ann ⇒ Name → m (PIRType uni)

Lookup a builtin type by its TH name. These are assumed to be present, so fails if it is cannot find it.

errorFunc ∷ Compiling uni fun m ann ⇒ m (PIRTerm uni fun)

The function 'error :: forall a . a'.