PlutusTx.Compiler.Trace

Documentation

traceCompilation

Arguments

∷ (MonadReader (CompileContext uni fun) m, MonadState CompileState m, MonadError (WithContext Text e) m)

⇒ Int
Context level

→ SDoc
The thing (expr, type, kind, etc.) being compiled

→ m a
The compilation action

→ m a

A combination of withContextM and traceCompilationStep.

withContextM emits a stack trace when the compilation fails, and can be turned on via `-fcontext-level=level`.

traceCompilationStep dumps the full compilation trace, and can be turned on via `-fdump-compilation-trace`.

traceCompilationStep

Arguments

∷ (MonadReader (CompileContext uni fun) m, MonadState CompileState m)

⇒ SDoc
The thing (expr, type, kind, etc.) being compiled

→ m a
The compilation action

→ m a