PlutusTx.Compiler.Types

Documentation

data Scope uni fun

Constructors

Scope (Map Name (PLCVar uni, Maybe (PIRTerm uni fun))) (Map Name PLCTyVar)

data Verbosity

Verbosity level of the Plutus Tx compiler.

Constructors

Quiet

Verbose

Debug

type Compiling uni fun m ann = (MonadError (CompileError uni fun ann) m, MonadQuote m, MonadReader (CompileContext uni fun) m, MonadState CompileState m, MonadDefs LexName uni fun Ann m, MonadWriter CoverageIndex m)

type NameInfo = Map Name TyThing

data CompileOptions

Compilation options.

Constructors

CompileOptions

Fields

coProfile ∷ ProfileOpts

coCoverage ∷ CoverageOpts

coDatatypeStyle ∷ DatatypeStyle

coRemoveTrace ∷ Bool

coInlineFix ∷ Bool

data ProfileOpts

Profiling options. All profiles everything. None is the default.

Constructors

All

None

data CoverageOpts

Coverage options See Note [Coverage annotations]

Constructors

CoverageOpts

Fields

unCoverageOpts ∷ Set CoverageType

data CompileContext uni fun

Constructors

CompileContext

Fields

ccOpts ∷ CompileOptions

ccFlags ∷ DynFlags

ccFamInstEnvs ∷ FamInstEnvs

ccNameInfo ∷ NameInfo

ccScope ∷ Scope uni fun

ccBlackholed ∷ Set Name

ccCurDef ∷ Maybe LexName

ccModBreaks ∷ Maybe ModBreaks

ccBuiltinsInfo ∷ BuiltinsInfo uni fun

ccBuiltinCostModel ∷ CostingPart uni fun

ccDebugTraceOn ∷ Bool

ccRewriteRules ∷ RewriteRules uni fun

ccSafeToInline ∷ Bool

newtype LexName

A wrapper around Name with a stable Ord instance. Use this where the ordering will affect the output of the compiler, i.e. when sorting or so on. It's fine to use Name if we're just putting them in a Set, for example.

The Eq instance we derive - it's also not stable across builds, but I believe this is only a problem if you compare things from different builds, which we don't do.

Constructors

LexName Name

data CompileState

Constructors

CompileState

Fields

csNextStep ∷ Int

The ID of the next step to be taken by the PlutusTx compiler. This is used when generating debug traces.

csPreviousSteps ∷ [Int]

The IDs of the previous steps taken by the PlutusTx compiler leading up to the current point. This is used when generating debug traces.

data CoverageType

Option `{-# OPTIONS_GHC -fplugin-opt PlutusTx.Plugin:coverage-all #-}` enables all these See Note [Adding more coverage annotations]. See Note [Coverage order]

Constructors

LocationCoverage
Check that all source locations that we can identify in GHC Core have been covered. For this to work at all we need `{-# OPTIONS_GHC -g #-}` turn on with `{-# OPTIONS_GHC -fplugin-opt PlutusTx.Plugin:coverage-location #-}`

BooleanCoverage
Check that every boolean valued expression that isn't True or False for which we know the source location have been covered. For this to work at all we need `{-# OPTIONS_GHC -g #-}` turn on with `{-# OPTIONS_GHC -fplugin-opt PlutusTx.Plugin:coverage-boolean #-}`

type CompilingDefault uni fun m ann = (uni ~ DefaultUni, fun ~ DefaultFun, Compiling uni fun m ann)

stableModuleCmp ∷ Module → Module → Ordering

Our own version of stableModuleCmp.

stableNameCmp ∷ Name → Name → Ordering

Our own version of stableNameCmp.

activeCoverageTypes ∷ CompileOptions → Set CoverageType

Get the coverage types we are using

blackhole ∷ MonadReader (CompileContext uni fun) m ⇒ Name → m a → m a

blackholed ∷ MonadReader (CompileContext uni fun) m ⇒ Name → m Bool

initialScope ∷ Scope uni fun

withCurDef ∷ Compiling uni fun m ann ⇒ LexName → m a → m a

modifyCurDeps ∷ Compiling uni fun m ann ⇒ (Set LexName → Set LexName) → m ()

module PlutusCore.Annotation