PlutusTx.Compiler.Utils

Documentation

lookupGhcTyCon ∷ Compiling uni fun m ann ⇒ Name → m TyCon

Get the TyCon for a given Name stored in the builtin name info, failing if it is missing.

lookupGhcName ∷ Compiling uni fun m ann ⇒ Name → m Name

Get the Name for a given Name stored in the builtin name info, failing if it is missing.

lookupGhcId ∷ Compiling uni fun m ann ⇒ Name → m Id

Get the Id for a given Name stored in the builtin name info, failing if it is missing.

sdToStr ∷ MonadReader (CompileContext uni fun) m ⇒ SDoc → m String

sdToTxt ∷ MonadReader (CompileContext uni fun) m ⇒ SDoc → m Text

throwSd ∷ (MonadError (CompileError uni fun ann) m, MonadReader (CompileContext uni fun) m) ⇒ (Text → Error uni fun ann) → SDoc → m a

tyConsOfExpr ∷ CoreExpr → UniqSet TyCon

tyConsOfBndr ∷ CoreBndr → UniqSet TyCon

tyConsOfBind ∷ Bind CoreBndr → UniqSet TyCon

tyConsOfAlt ∷ CoreAlt → UniqSet TyCon