PlutusTx.Coverage

Documentation

data CoverageAnnotation

Constructors

CoverLocation CovLoc

CoverBool CovLoc Bool

newtype CoverageIndex

This type keeps track of all coverage annotations and where they have been inserted / what annotations are expected to be found when executing a piece of code.

Constructors

CoverageIndex

Fields

_coverageMetadata ∷ Map CoverageAnnotation CoverageMetadata

newtype CoverageMetadata

Constructors

CoverageMetadata

Fields

_metadataSet ∷ Set Metadata

data Metadata

Constructors

ApplicationHeadSymbol String

IgnoredAnnotation
Location that is not interesting to cover. This is not generated by the compiler, but can be added later using addCoverageMetadata.

newtype CoverageData

Constructors

CoverageData

Fields

_coveredAnnotations ∷ Set CoverageAnnotation

data CoverageReport

Constructors

CoverageReport

Fields

_coverageIndex ∷ CoverageIndex

_coverageData ∷ CoverageData

data CovLoc

A source location for coverage

Constructors

CovLoc

Fields

_covLocFile ∷ String

_covLocStartLine ∷ Int

_covLocEndLine ∷ Int

_covLocStartCol ∷ Int

_covLocEndCol ∷ Int

covLocFile ∷ Lens' CovLoc String

covLocStartLine ∷ Lens' CovLoc Int

covLocEndLine ∷ Lens' CovLoc Int

covLocStartCol ∷ Lens' CovLoc Int

covLocEndCol ∷ Lens' CovLoc Int

metadataSet ∷ Iso' CoverageMetadata (Set Metadata)

coverageAnnotations ∷ Getter CoverageIndex (Set CoverageAnnotation)

ignoredAnnotations ∷ Getter CoverageIndex (Set CoverageAnnotation)

coverageMetadata ∷ Iso' CoverageIndex (Map CoverageAnnotation CoverageMetadata)

coveredAnnotations ∷ Iso' CoverageData (Set CoverageAnnotation)

addCoverageMetadata ∷ CoverageAnnotation → Metadata → CoverageIndex → CoverageIndex

Add metadata to a coverage annotation. Does nothing if the annotation is not in the index.

addLocationToCoverageIndex ∷ MonadWriter CoverageIndex m ⇒ CovLoc → m CoverageAnnotation

Include a location coverage annotation in the index

addBoolCaseToCoverageIndex ∷ MonadWriter CoverageIndex m ⇒ CovLoc → Bool → CoverageMetadata → m CoverageAnnotation

Include a boolean coverage annotation in the index

coverageDataFromLogMsg ∷ String → CoverageData