PlutusTx.Data.List

Documentation

data List a

A list type backed directly by Data. It is meant to be used whenever fast encodingdecoding tofrom Data is needed.

caseList

Arguments

∷ ∀ a r. UnsafeFromData a

⇒ (() → r)
Nil case

→ (a → List a → r)
Cons case

→ List a

→ r

Matching on the given List.

caseList'

Arguments

∷ ∀ a r. UnsafeFromData a

⇒ r
Nil case

→ (a → List a → r)
Cons case

→ List a

→ r

Like caseList, except the nil case takes an r directly, which is evaluated strictly. If r is an error or expensive computation, consider using caseList instead.

null ∷ List a → Bool

append ∷ List a → List a → List a

find ∷ UnsafeFromData a ⇒ (a → Bool) → List a → Maybe a

Find the first element that satisfies a predicate. Warning: this function can be very inefficient if the list contains elements that are expensive to decode from BuiltinData.

findIndices ∷ UnsafeFromData a ⇒ (a → Bool) → List a → List Integer

Find the indices of all elements that satisfy a predicate. Warning: this function can be very inefficient if the list contains elements that are expensive to decode from BuiltinData.

filter ∷ (UnsafeFromData a, ToData a) ⇒ (a → Bool) → List a → List a

Filter a list using a predicate. Warning: this function can be very inefficient if the list contains elements that are expensive to decode from BuiltinData.

mapMaybe ∷ (UnsafeFromData a, ToData b) ⇒ (a → Maybe b) → List a → List b

Map a function over a list and discard the results that are Nothing. Warning: this function can be very inefficient if the list contains elements that are expensive to decode from BuiltinData, or if the result of applying f is expensive to encode to BuiltinData.

any ∷ UnsafeFromData a ⇒ (a → Bool) → List a → Bool

Check if any element in the list satisfies a predicate. Warning: this function can be very inefficient if the list contains elements that are expensive to decode from BuiltinData.

all ∷ UnsafeFromData a ⇒ (a → Bool) → List a → Bool

Check if all elements in the list satisfy a predicate. Warning: this function can be very inefficient if the list contains elements that are expensive to decode from BuiltinData.

foldMap ∷ (UnsafeFromData a, Monoid m) ⇒ (a → m) → List a → m

Fold a list using a monoid. Warning: this function can be very inefficient if the list contains elements that are expensive to decode from BuiltinData.

map ∷ (UnsafeFromData a, ToData b) ⇒ (a → b) → List a → List b

Map a function over a list. Warning: this function can be very inefficient if the list contains elements that are expensive to decode from BuiltinData, or if the result of applying f is expensive to encode to BuiltinData.

length ∷ List a → Integer

Get the length of a list.

mconcat ∷ (Monoid a, UnsafeFromData a) ⇒ List a → a

Concatenate a list of monoids. Warning: this function can be very inefficient if the list contains elements that are expensive to decode from BuiltinData.

(<|) ∷ ToData a ⇒ a → List a → List a infixr 5

Prepend an element to the list.

cons ∷ ToData a ⇒ a → List a → List a

Synonym for <|.

nil ∷ List a

Construct an empty list.

singleton ∷ ToData a ⇒ a → List a

Create a list from a single element.

uncons ∷ UnsafeFromData a ⇒ List a → Maybe (a, List a)

Get the first element of a list and the rest of the list.

and ∷ List Bool → Bool

Check if all elements in the list are True. Warning: this function can be very inefficient if the list contains elements that are expensive to decode from BuiltinData.

or ∷ List Bool → Bool

Check if any element in the list is True. Warning: this function can be very inefficient if the list contains elements that are expensive to decode from BuiltinData.

elem ∷ ToData a ⇒ a → List a → Bool

Check if an element is in the list. Note: this function can leverage the better performance of equality checks for BuiltinData.

notElem ∷ ToData a ⇒ a → List a → Bool

Check if an element is not in the list.

foldr ∷ UnsafeFromData a ⇒ (a → b → b) → b → List a → b

Fold a list from the right. Warning: this function can be very inefficient if the list contains elements that are expensive to decode from BuiltinData.

foldl ∷ UnsafeFromData a ⇒ (b → a → b) → b → List a → b

Fold a list from the left. Warning: this function can be very inefficient if the list contains elements that are expensive to decode from BuiltinData.

concat ∷ List (List a) → List a

Flatten a list of lists into a single list.

concatMap ∷ UnsafeFromData a ⇒ (a → List b) → List a → List b

Map a function over a list and concatenate the results.

listToMaybe ∷ UnsafeFromData a ⇒ List a → Maybe a

Get the first element of a list if it is not empty.

uniqueElement ∷ UnsafeFromData a ⇒ List a → Maybe a

Get the element of a list if it has exactly one element.

(!!) ∷ UnsafeFromData a ⇒ List a → Integer → a infixl 9

Get the element at a given index. Warning: this is a partial function and will fail if the index is negative or greater than the length of the list. Note: this function has the same precedence as (!!) from List.

revAppend ∷ List a → List a → List a

Append two lists in reverse order.

reverse ∷ List a → List a

Reverse a list.

replicate ∷ ToData a ⇒ Integer → a → List a

Replicate a value n times.

findIndex ∷ UnsafeFromData a ⇒ (a → Bool) → List a → Maybe Integer

Find the index of the first element that satisfies a predicate. Warning: this function can be very inefficient if the list contains elements that are expensive to decode from BuiltinData.

unzip ∷ ∀ a b. List (a, b) → (List a, List b)

Split a list of pairs into a pair of lists.

zipWith ∷ (UnsafeFromData a, UnsafeFromData b, ToData c) ⇒ (a → b → c) → List a → List b → List c

Zip two lists together using a function. Warning: this function can be very inefficient if the lists contain elements that are expensive to decode from BuiltinData, or if the result of applying f is expensive to encode to BuiltinData.

head ∷ ∀ a. UnsafeFromData a ⇒ List a → a

Return the head of a list. Warning: this is a partial function and will fail if the list is empty.

last ∷ ∀ a. UnsafeFromData a ⇒ List a → a

Return the last element of a list. Warning: this is a partial function and will fail if the list is empty.

tail ∷ ∀ a. List a → List a

Return the tail of a list. Warning: this is a partial function and will fail if the list is empty.

take ∷ ∀ a. Integer → List a → List a

Take the first n elements from the list.

drop ∷ ∀ a. Integer → List a → List a

Drop the first n elements from the list.

dropWhile ∷ ∀ a. UnsafeFromData a ⇒ (a → Bool) → List a → List a

Drop elements from the list while the predicate holds. Warning: this function can be very inefficient if the list contains elements that are expensive to decode from BuiltinData.

splitAt ∷ ∀ a. Integer → List a → (List a, List a)

Split a list at a given index.

elemBy ∷ UnsafeFromData a ⇒ (a → a → Bool) → a → List a → Bool

Check if an element satisfying a binary predicate is in the list. Warning: this function can be very inefficient if the list contains elements that are expensive to decode from BuiltinData.

nubBy ∷ ∀ a. UnsafeFromData a ⇒ (a → a → Bool) → List a → List a

Removes elements from the list that satisfy a binary predicate. Warning: this function can be very inefficient if the list contains elements that are expensive to decode from BuiltinData.

nub ∷ (Eq a, UnsafeFromData a) ⇒ List a → List a

Removes duplicate elements from the list. Warning: this function can be very inefficient if the list contains elements that are expensive to decode from BuiltinData.

partition ∷ UnsafeFromData a ⇒ (a → Bool) → List a → (List a, List a)

Partition a list into two lists based on a predicate. Warning: this function can be very inefficient if the list contains elements that are expensive to decode from BuiltinData.

toBuiltinList ∷ List a → BuiltinList BuiltinData

fromBuiltinList ∷ BuiltinList BuiltinData → List a

toSOP ∷ ∀ a. UnsafeFromData a ⇒ List a → [a]

Convert a data-backed list to a sums of products list. Warning: this function can be very inefficient if the list contains elements that are expensive to decode from BuiltinData.

fromSOP ∷ ∀ a. ToData a ⇒ [a] → List a

Convert a sums of products list to a data-backed list. Warning: this function can be very inefficient if the list contains elements that are expensive to encode to BuiltinData.