PlutusTx.Lattice

Documentation

class JoinSemiLattice a where

A join semi-lattice, i.e. a partially ordered set equipped with a binary operation (\/).

Note that the mathematical definition would require an ordering constraint - we omit that so we can define instances for e.g. (->).

Methods

(\/) ∷ a → a → a

class MeetSemiLattice a where

A meet semi-lattice, i.e. a partially ordered set equipped with a binary operation (/\).

Note that the mathematical definition would require an ordering constraint - we omit that so we can define instances for e.g. (->).

Methods

(/\) ∷ a → a → a

type Lattice a = (JoinSemiLattice a, MeetSemiLattice a)

A lattice.

class JoinSemiLattice a ⇒ BoundedJoinSemiLattice a where

A bounded join semi-lattice, i.e. a join semi-lattice augmented with a distinguished element bottom which is the unit of (\/).

Methods

bottom ∷ a

class MeetSemiLattice a ⇒ BoundedMeetSemiLattice a where

A bounded meet semi-lattice, i.e. a meet semi-lattice augmented with a distinguished element top which is the unit of (/\).

Methods

top ∷ a

type BoundedLattice a = (BoundedJoinSemiLattice a, BoundedMeetSemiLattice a)

A bounded lattice.

newtype Join a

A wrapper witnessing that a join semi-lattice is a monoid with (\/) and bottom.

Constructors

Join a

newtype Meet a

A wrapper witnessing that a meet semi-lattice is a monoid with (/\) and top.

Constructors

Meet a