PlutusTx.Lift

Documentation

safeLiftWith

Arguments

∷ ∀ a uni fun m. (Lift uni a, GEq uni, Everywhere uni Eq, MonadError (Error uni fun (Provenance ())) m, MonadQuote m, Typecheckable uni fun, CaseBuiltin uni, PrettyUni uni, Pretty fun, Default (CostingPart uni fun), Default (BuiltinsInfo uni fun), Default (RewriteRules uni fun), Hashable fun)

⇒ (CompilationOpts () → CompilationOpts ())
Modifier of PIR compilation options

→ (CompilationOpts Name fun (Provenance ()) → CompilationOpts Name fun (Provenance ()))
Modifier of UPLC compilation options

→ Version

→ a

→ m (Term TyName Name uni fun (), Term NamedDeBruijn uni fun ())

Get a Plutus Core term corresponding to the given value. Allows configuring PIR and UPLC optimization options.

safeLift ∷ ∀ a uni fun m. (Lift uni a, GEq uni, Everywhere uni Eq, MonadError (Error uni fun (Provenance ())) m, MonadQuote m, Typecheckable uni fun, CaseBuiltin uni, PrettyUni uni, Pretty fun, Default (CostingPart uni fun), Default (BuiltinsInfo uni fun), Default (RewriteRules uni fun), Hashable fun) ⇒ Version → a → m (Term TyName Name uni fun (), Term NamedDeBruijn uni fun ())

Get a Plutus Core term corresponding to the given value, applying default PIR/UPLC optimizations.

safeLiftUnopt ∷ ∀ a uni fun m. (Lift uni a, GEq uni, Everywhere uni Eq, MonadError (Error uni fun (Provenance ())) m, MonadQuote m, Typecheckable uni fun, CaseBuiltin uni, PrettyUni uni, Pretty fun, Default (CostingPart uni fun), Default (BuiltinsInfo uni fun), Default (RewriteRules uni fun), Hashable fun) ⇒ Version → a → m (Term TyName Name uni fun (), Term NamedDeBruijn uni fun ())

Like safeLift but does not apply PIR/UPLC optimizations. Use this option where lifting speed is more important than optimal code.

safeLiftProgram ∷ (Lift uni a, GEq uni, Everywhere uni Eq, MonadError (Error uni fun (Provenance ())) m, MonadQuote m, Typecheckable uni fun, CaseBuiltin uni, PrettyUni uni, Pretty fun, Default (CostingPart uni fun), Default (BuiltinsInfo uni fun), Default (RewriteRules uni fun), Hashable fun) ⇒ Version → a → m (Program TyName Name uni fun (), Program NamedDeBruijn uni fun ())

Get a Plutus Core program corresponding to the given value, applying default PIR/UPLC optimizations.

safeLiftProgramUnopt ∷ (Lift uni a, GEq uni, Everywhere uni Eq, MonadError (Error uni fun (Provenance ())) m, MonadQuote m, Typecheckable uni fun, CaseBuiltin uni, PrettyUni uni, Pretty fun, Default (CostingPart uni fun), Default (BuiltinsInfo uni fun), Default (RewriteRules uni fun), Hashable fun) ⇒ Version → a → m (Program TyName Name uni fun (), Program NamedDeBruijn uni fun ())

Like safeLiftProgram but does not apply PIR/UPLC optimizations. Use this option where lifting speed is more important than optimal code.

safeLiftCode ∷ (Lift uni a, GEq uni, Everywhere uni Eq, MonadError (Error uni fun (Provenance ())) m, MonadQuote m, Typecheckable uni fun, CaseBuiltin uni, PrettyUni uni, Pretty fun, Default (CostingPart uni fun), Default (BuiltinsInfo uni fun), Default (RewriteRules uni fun), Hashable fun) ⇒ Version → a → m (CompiledCodeIn uni fun a)

safeLiftCodeUnopt ∷ (Lift uni a, GEq uni, Everywhere uni Eq, MonadError (Error uni fun (Provenance ())) m, MonadQuote m, Typecheckable uni fun, CaseBuiltin uni, PrettyUni uni, Pretty fun, Default (CostingPart uni fun), Default (BuiltinsInfo uni fun), Default (RewriteRules uni fun), Hashable fun) ⇒ Version → a → m (CompiledCodeIn uni fun a)

Like safeLiftCode but does not apply PIR/UPLC optimizations. Use this option where lifting speed is more important than optimal code.

lift ∷ (Lift uni a, ThrowableBuiltins uni fun, Typecheckable uni fun, GEq uni, Everywhere uni Eq, CaseBuiltin uni, Default (CostingPart uni fun), Default (BuiltinsInfo uni fun), Default (RewriteRules uni fun), Hashable fun) ⇒ Version → a → (Term TyName Name uni fun (), Term NamedDeBruijn uni fun ())

Get a Plutus Core term corresponding to the given value, throwing any errors that occur as exceptions and ignoring fresh names. The default PIR/UPLC optimizations are applied.

liftUnopt ∷ (Lift uni a, ThrowableBuiltins uni fun, Typecheckable uni fun, GEq uni, Everywhere uni Eq, CaseBuiltin uni, Default (CostingPart uni fun), Default (BuiltinsInfo uni fun), Default (RewriteRules uni fun), Hashable fun) ⇒ Version → a → (Term TyName Name uni fun (), Term NamedDeBruijn uni fun ())

Like lift but does not apply PIR/UPLC optimizations. Use this option where lifting speed is more important than optimal code.

liftProgram ∷ (Lift uni a, ThrowableBuiltins uni fun, Typecheckable uni fun, GEq uni, Everywhere uni Eq, CaseBuiltin uni, Default (CostingPart uni fun), Default (BuiltinsInfo uni fun), Default (RewriteRules uni fun), Hashable fun) ⇒ Version → a → (Program TyName Name uni fun (), Program NamedDeBruijn uni fun ())

Get a Plutus Core program corresponding to the given value, throwing any errors that occur as exceptions and ignoring fresh names.

liftProgramUnopt ∷ (Lift uni a, ThrowableBuiltins uni fun, Typecheckable uni fun, GEq uni, Everywhere uni Eq, CaseBuiltin uni, Default (CostingPart uni fun), Default (BuiltinsInfo uni fun), Default (RewriteRules uni fun), Hashable fun) ⇒ Version → a → (Program TyName Name uni fun (), Program NamedDeBruijn uni fun ())

Like liftProgram but does not apply PIR/UPLC optimizations. Use this option where lifting speed is more important than optimal code.

liftProgramDef ∷ Lift DefaultUni a ⇒ a → (Program TyName Name DefaultUni DefaultFun (), Program NamedDeBruijn DefaultUni DefaultFun ())

Get a Plutus Core program in the default universe with the default version, corresponding to the given value, throwing any errors that occur as exceptions and ignoring fresh names.

liftProgramDefUnopt ∷ Lift DefaultUni a ⇒ a → (Program TyName Name DefaultUni DefaultFun (), Program NamedDeBruijn DefaultUni DefaultFun ())

Like liftProgramDef but does not apply PIR/UPLC optimizations. Use this option where lifting speed is more important than optimal code.

liftCode ∷ (Lift uni a, GEq uni, Everywhere uni Eq, ThrowableBuiltins uni fun, Typecheckable uni fun, CaseBuiltin uni, Default (CostingPart uni fun), Default (BuiltinsInfo uni fun), Default (RewriteRules uni fun), Hashable fun) ⇒ Version → a → CompiledCodeIn uni fun a

Get a Plutus Core program corresponding to the given value as a CompiledCodeIn, throwing any errors that occur as exceptions and ignoring fresh names.

liftCodeUnopt ∷ (Lift uni a, GEq uni, Everywhere uni Eq, ThrowableBuiltins uni fun, Typecheckable uni fun, CaseBuiltin uni, Default (CostingPart uni fun), Default (BuiltinsInfo uni fun), Default (RewriteRules uni fun), Hashable fun) ⇒ Version → a → CompiledCodeIn uni fun a

Like liftCode but does not apply PIR/UPLC optimizations. Use this option where lifting speed is more important than optimal code.

liftCodeDef ∷ (Lift uni a, GEq uni, Everywhere uni Eq, ThrowableBuiltins uni fun, Typecheckable uni fun, CaseBuiltin uni, Default (CostingPart uni fun), Default (BuiltinsInfo uni fun), Default (RewriteRules uni fun), Hashable fun) ⇒ a → CompiledCodeIn uni fun a

Get a Plutus Core program with the default version, corresponding to the given value as a CompiledCodeIn, throwing any errors that occur as exceptions and ignoring fresh names.

liftCodeDefUnopt ∷ (Lift uni a, GEq uni, Everywhere uni Eq, ThrowableBuiltins uni fun, Typecheckable uni fun, CaseBuiltin uni, Default (CostingPart uni fun), Default (BuiltinsInfo uni fun), Default (RewriteRules uni fun), Hashable fun) ⇒ a → CompiledCodeIn uni fun a

Like liftCodeDef but does not apply PIR/UPLC optimizations. Use this option where lifting speed is more important than optimal code.

typeCheckAgainst ∷ ∀ a uni fun m. (Typeable uni a, MonadError (Error uni fun (Provenance ())) m, MonadQuote m, GEq uni, Typecheckable uni fun, CaseBuiltin uni, PrettyUni uni, Pretty fun, Default (CostingPart uni fun), Default (BuiltinsInfo uni fun), Default (RewriteRules uni fun)) ⇒ Proxy a → Program TyName Name uni fun () → m ()

Check that PLC term has the given type.

typeCode ∷ ∀ a uni fun m. (Typeable uni a, MonadError (Error uni fun (Provenance ())) m, MonadQuote m, GEq uni, Everywhere uni Eq, Typecheckable uni fun, CaseBuiltin uni, PrettyUni uni, Pretty fun, Default (CostingPart uni fun), Default (BuiltinsInfo uni fun), Default (RewriteRules uni fun), Hashable fun) ⇒ Proxy a → Program TyName Name uni fun () → m (CompiledCodeIn uni fun a)

Try to interpret a PLC program as a CompiledCodeIn of the given type. Returns successfully iff the program has the right type.

makeTypeable ∷ Type → Name → Q [Dec]

Given a universe, derive a Typeable instance for a given Plinth datatype

NOTE: it requires MultiParamTypeClasses language extension

makeLift ∷ Name → Q [Dec]

Derive Lift and Typeable instances for Plinth types

NOTE: it requires MultiParamTypeClasses,FlexibleContexts language extensions

data LiftError

Constructors

UnsupportedLiftKind !Kind

UnsupportedLiftType !Type

UserLiftError !Text

LiftMissingDataCons !Name

LiftMissingVar !Name