PlutusTx.Lift.THUtils

Documentation

showName ∷ Name → Text

Very nearly the same as showName, but doesn't print uniques, since we don't need to incorporate them into our names.

normalizeType ∷ Type → Type

Normalize a type, in particular getting rid of things like ListT in favour of applications of the actual name.

requireExtension ∷ Extension → Q ()

mkTyVarDecl ∷ MonadQuote m ⇒ Name → Kind () → m (Name, TyVarDecl TyName ())

isNewtype ∷ DatatypeInfo → Bool

tyListE ∷ [TExpQ a] → TExpQ [a]

Safe wrapper around listE for typed TH.