PlutusTx.Numeric

Type classes

class AdditiveSemigroup a where

A Semigroup that it is sensible to describe using addition.

Methods

(+) ∷ a → a → a infixl 6

class AdditiveSemigroup a ⇒ AdditiveMonoid a where

A Monoid that it is sensible to describe using addition and zero.

Methods

zero ∷ a

class AdditiveMonoid a ⇒ AdditiveGroup a where

A Group that it is sensible to describe using addition, zero, and subtraction.

Methods

(-) ∷ a → a → a infixl 6

class MultiplicativeSemigroup a where

A Semigroup that it is sensible to describe using multiplication.

Methods

(*) ∷ a → a → a infixl 7

class MultiplicativeSemigroup a ⇒ MultiplicativeMonoid a where

A Semigroup that it is sensible to describe using multiplication and one.

Methods

one ∷ a

type Semiring a = (AdditiveMonoid a, MultiplicativeMonoid a)

A semiring.

type Ring a = (AdditiveGroup a, MultiplicativeMonoid a)

A ring.

class (Ring s, AdditiveGroup v) ⇒ Module s v | v → s where

A module, with a type of scalars which can be used to scale the values.

Methods

scale ∷ s → v → v

Helper newtypes

newtype Additive a

A newtype wrapper to derive Additive classes via.

Constructors

Additive a

newtype Multiplicative a

A newtype wrapper to derive Multiplicative classes via.

Constructors

Multiplicative a

Helper functions

negate ∷ AdditiveGroup a ⇒ a → a

divMod ∷ Integer → Integer → (Integer, Integer)

Simultaneous div and mod.

quotRem ∷ Integer → Integer → (Integer, Integer)

Simultaneous quot and rem.

abs ∷ (Ord n, AdditiveGroup n) ⇒ n → n

Absolute value for any AdditiveGroup.

Orphan instances

MultiplicativeMonoid a ⇒ Monoid (Product a)

Instance details

Methods

mempty ∷ Product a

AdditiveMonoid a ⇒ Monoid (Sum a)

Instance details

Methods

mempty ∷ Sum a

MultiplicativeSemigroup a ⇒ Semigroup (Product a)

Instance details

Methods

(<>) ∷ Product a → Product a → Product a

AdditiveSemigroup a ⇒ Semigroup (Sum a)

Instance details

Methods

(<>) ∷ Sum a → Sum a → Sum a