PlutusTx.Options

Documentation

data PluginOptions

Constructors

PluginOptions

Fields

_posPlcTargetVersion ∷ Version

_posDoTypecheck ∷ Bool

_posDeferErrors ∷ Bool

_posConservativeOpts ∷ Bool

_posContextLevel ∷ Int

_posDumpPir ∷ Bool

_posDumpPlc ∷ Bool

_posDumpUPlc ∷ Bool

_posOptimize ∷ Bool

_posPedantic ∷ Bool

_posVerbosity ∷ Verbosity

_posDatatypes ∷ DatatypeCompilationOpts

_posMaxSimplifierIterationsPir ∷ Int

_posMaxSimplifierIterationsUPlc ∷ Int

_posMaxCseIterations ∷ Int

_posDoSimplifierUnwrapCancel ∷ Bool

_posDoSimplifierBeta ∷ Bool

_posDoSimplifierInline ∷ Bool

_posDoSimplifierEvaluateBuiltins ∷ Bool

_posDoSimplifierStrictifyBindings ∷ Bool

_posDoSimplifierRemoveDeadBindings ∷ Bool

_posProfile ∷ ProfileOpts

_posCoverageAll ∷ Bool

_posCoverageLocation ∷ Bool

_posCoverageBoolean ∷ Bool

_posRelaxedFloatin ∷ Bool

_posCaseOfCaseConservative ∷ Bool

_posInlineCallsiteGrowth ∷ Int

_posInlineConstants ∷ Bool

_posInlineFix ∷ Bool

_posPreserveLogging ∷ Bool

Whether to try and retain the logging behaviour of the program.

_posRemoveTrace ∷ Bool

_posDumpCompilationTrace ∷ Bool

_posCertify ∷ Maybe String

posVerbosity ∷ Lens' PluginOptions Verbosity

posRemoveTrace ∷ Lens' PluginOptions Bool

posRelaxedFloatin ∷ Lens' PluginOptions Bool

posProfile ∷ Lens' PluginOptions ProfileOpts

posPreserveLogging ∷ Lens' PluginOptions Bool

posPlcTargetVersion ∷ Lens' PluginOptions Version

posPedantic ∷ Lens' PluginOptions Bool

posOptimize ∷ Lens' PluginOptions Bool

posMaxSimplifierIterationsUPlc ∷ Lens' PluginOptions Int

posMaxSimplifierIterationsPir ∷ Lens' PluginOptions Int

posMaxCseIterations ∷ Lens' PluginOptions Int

posInlineFix ∷ Lens' PluginOptions Bool

posInlineConstants ∷ Lens' PluginOptions Bool

posInlineCallsiteGrowth ∷ Lens' PluginOptions Int

posDumpUPlc ∷ Lens' PluginOptions Bool

posDumpPlc ∷ Lens' PluginOptions Bool

posDumpPir ∷ Lens' PluginOptions Bool

posDumpCompilationTrace ∷ Lens' PluginOptions Bool

posDoTypecheck ∷ Lens' PluginOptions Bool

posDoSimplifierUnwrapCancel ∷ Lens' PluginOptions Bool

posDoSimplifierStrictifyBindings ∷ Lens' PluginOptions Bool

posDoSimplifierRemoveDeadBindings ∷ Lens' PluginOptions Bool

posDoSimplifierInline ∷ Lens' PluginOptions Bool

posDoSimplifierEvaluateBuiltins ∷ Lens' PluginOptions Bool

posDoSimplifierBeta ∷ Lens' PluginOptions Bool

posDeferErrors ∷ Lens' PluginOptions Bool

posDatatypes ∷ Lens' PluginOptions DatatypeCompilationOpts

posCoverageLocation ∷ Lens' PluginOptions Bool

posCoverageBoolean ∷ Lens' PluginOptions Bool

posCoverageAll ∷ Lens' PluginOptions Bool

posContextLevel ∷ Lens' PluginOptions Int

posConservativeOpts ∷ Lens' PluginOptions Bool

posCertify ∷ Lens' PluginOptions (Maybe String)

posCaseOfCaseConservative ∷ Lens' PluginOptions Bool

type OptionKey = Text

type OptionValue = Text

data Implication a

A data type representing option a implying option b.

Constructors

∀ b. Implication (a → Bool) (Lens' PluginOptions b) b

data PluginOption

A plugin option definition for a PluginOptions field of type a.

Constructors

∀ a.Pretty a ⇒ PluginOption

Fields

poTypeRep ∷ TypeRep a

TypeRep used for pretty printing the option.

poFun ∷ Maybe OptionValue → Validation ParseError (a → a)

Consumes an optional value, and either updates the field or reports an error.

poLens ∷ Lens' PluginOptions a

Lens focusing on the field. This is for modifying the field, as well as getting the field value from defaultPluginOptions for pretty printing.

poDescription ∷ Text

A description of the option.

poImplications ∷ [Implication a]

Implications of this option being set to a particular value. An option should not imply itself.

data ParseError

Constructors

CannotParseValue !OptionKey !OptionValue !SomeTypeRep

UnexpectedValue !OptionKey !OptionValue

MissingValue !OptionKey

UnrecognisedOption !OptionKey ![OptionKey]

newtype ParseErrors

Constructors

ParseErrors (NonEmpty ParseError)

renderParseError ∷ ParseError → Text

pluginOptions ∷ Map OptionKey PluginOption

Definition of plugin options.

flag ∷ (a → a) → OptionKey → Maybe OptionValue → Validation ParseError (a → a)

setTrue ∷ OptionKey → Maybe OptionValue → Validation ParseError (Bool → Bool)

plcParserOption ∷ Parser a → OptionKey → Maybe OptionValue → Validation ParseError (a → a)

readOption ∷ Read a ⇒ OptionKey → Maybe OptionValue → Validation ParseError (a → a)

fromReadOption ∷ Read a ⇒ OptionKey → (a → Validation ParseError b) → Maybe OptionValue → Validation ParseError (b → b)

Obtain an option value of type a from an Int.

defaultPluginOptions ∷ PluginOptions

processOne ∷ OptionKey → Maybe OptionValue → Validation ParseError (PluginOptions → PluginOptions)

applyImplications ∷ Lens' PluginOptions a → [Implication a] → PluginOptions → PluginOptions

processAll ∷ [(OptionKey, Maybe OptionValue)] → Validation ParseErrors [PluginOptions → PluginOptions]

toKeyValue ∷ CommandLineOption → (OptionKey, Maybe OptionValue)

parsePluginOptions ∷ [CommandLineOption] → Validation ParseErrors PluginOptions

Parses the arguments that were given to ghc at commandline as "-fplugin-opt PlutusTx.Plugin:opt" or "-fplugin-opt PlutusTx.Plugin:opt=val"