PlutusTx.Ord

Documentation

class Eq a ⇒ Ord a where

The Ord class is used for totally ordered datatypes.

Minimal complete definition: either compare or <=. Using compare can be more efficient for complex types.

Minimal complete definition

compare | (<=)

Methods

compare ∷ a → a → Ordering

(<) ∷ a → a → Bool infix 4

(<=) ∷ a → a → Bool infix 4

(>) ∷ a → a → Bool infix 4

(>=) ∷ a → a → Bool infix 4

max ∷ a → a → a

min ∷ a → a → a

data Ordering

Constructors

LT

EQ

GT

deriveOrd ∷ Name → Q [Dec]

derive a PlutusTx.Ord instance for a datatype/newtype, similar to Haskell's `deriving stock Ord`.

One shortcoming compared to Haskell's deriving is that you cannot deriveOrd for polymorphic phantom types.

Orphan instances

Ord Ordering

Instance details

Methods

compare ∷ Ordering → Ordering → Ordering

(<) ∷ Ordering → Ordering → Bool

(<=) ∷ Ordering → Ordering → Bool

(>) ∷ Ordering → Ordering → Bool

(>=) ∷ Ordering → Ordering → Bool

max ∷ Ordering → Ordering → Ordering

min ∷ Ordering → Ordering → Ordering

Ord ()

Instance details

Methods

compare ∷ () → () → Ordering

(<) ∷ () → () → Bool

(<=) ∷ () → () → Bool

(>) ∷ () → () → Bool

(>=) ∷ () → () → Bool

max ∷ () → () → ()

min ∷ () → () → ()

Ord Bool

Instance details

Methods

compare ∷ Bool → Bool → Ordering

(<) ∷ Bool → Bool → Bool

(<=) ∷ Bool → Bool → Bool

(>) ∷ Bool → Bool → Bool

(>=) ∷ Bool → Bool → Bool

max ∷ Bool → Bool → Bool

min ∷ Bool → Bool → Bool

Ord a ⇒ Ord (Maybe a)

Instance details

Methods

compare ∷ Maybe a → Maybe a → Ordering

(<) ∷ Maybe a → Maybe a → Bool

(<=) ∷ Maybe a → Maybe a → Bool

(>) ∷ Maybe a → Maybe a → Bool

(>=) ∷ Maybe a → Maybe a → Bool

max ∷ Maybe a → Maybe a → Maybe a

min ∷ Maybe a → Maybe a → Maybe a

Ord a ⇒ Ord [a]

Instance details

Methods

compare ∷ [a] → [a] → Ordering

(<) ∷ [a] → [a] → Bool

(<=) ∷ [a] → [a] → Bool

(>) ∷ [a] → [a] → Bool

(>=) ∷ [a] → [a] → Bool

max ∷ [a] → [a] → [a]

min ∷ [a] → [a] → [a]

(Ord a, Ord b) ⇒ Ord (Either a b)

Instance details

Methods

compare ∷ Either a b → Either a b → Ordering

(<) ∷ Either a b → Either a b → Bool

(<=) ∷ Either a b → Either a b → Bool

(>) ∷ Either a b → Either a b → Bool

(>=) ∷ Either a b → Either a b → Bool

max ∷ Either a b → Either a b → Either a b

min ∷ Either a b → Either a b → Either a b

(Ord a, Ord b) ⇒ Ord (a, b)

Instance details

Methods

compare ∷ (a, b) → (a, b) → Ordering

(<) ∷ (a, b) → (a, b) → Bool

(<=) ∷ (a, b) → (a, b) → Bool

(>) ∷ (a, b) → (a, b) → Bool

(>=) ∷ (a, b) → (a, b) → Bool

max ∷ (a, b) → (a, b) → (a, b)

min ∷ (a, b) → (a, b) → (a, b)

(Ord a, Ord b, Ord c) ⇒ Ord (a, b, c)

Instance details

Methods

compare ∷ (a, b, c) → (a, b, c) → Ordering

(<) ∷ (a, b, c) → (a, b, c) → Bool

(<=) ∷ (a, b, c) → (a, b, c) → Bool

(>) ∷ (a, b, c) → (a, b, c) → Bool

(>=) ∷ (a, b, c) → (a, b, c) → Bool

max ∷ (a, b, c) → (a, b, c) → (a, b, c)

min ∷ (a, b, c) → (a, b, c) → (a, b, c)

(Ord a, Ord b, Ord c, Ord d) ⇒ Ord (a, b, c, d)

Instance details

Methods

compare ∷ (a, b, c, d) → (a, b, c, d) → Ordering

(<) ∷ (a, b, c, d) → (a, b, c, d) → Bool

(<=) ∷ (a, b, c, d) → (a, b, c, d) → Bool

(>) ∷ (a, b, c, d) → (a, b, c, d) → Bool

(>=) ∷ (a, b, c, d) → (a, b, c, d) → Bool

max ∷ (a, b, c, d) → (a, b, c, d) → (a, b, c, d)

min ∷ (a, b, c, d) → (a, b, c, d) → (a, b, c, d)

(Ord a, Ord b, Ord c, Ord d, Ord e) ⇒ Ord (a, b, c, d, e)

Instance details

Methods

compare ∷ (a, b, c, d, e) → (a, b, c, d, e) → Ordering

(<) ∷ (a, b, c, d, e) → (a, b, c, d, e) → Bool

(<=) ∷ (a, b, c, d, e) → (a, b, c, d, e) → Bool

(>) ∷ (a, b, c, d, e) → (a, b, c, d, e) → Bool

(>=) ∷ (a, b, c, d, e) → (a, b, c, d, e) → Bool

max ∷ (a, b, c, d, e) → (a, b, c, d, e) → (a, b, c, d, e)

min ∷ (a, b, c, d, e) → (a, b, c, d, e) → (a, b, c, d, e)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f) ⇒ Ord (a, b, c, d, e, f)

Instance details

Methods

compare ∷ (a, b, c, d, e, f) → (a, b, c, d, e, f) → Ordering

(<) ∷ (a, b, c, d, e, f) → (a, b, c, d, e, f) → Bool

(<=) ∷ (a, b, c, d, e, f) → (a, b, c, d, e, f) → Bool

(>) ∷ (a, b, c, d, e, f) → (a, b, c, d, e, f) → Bool

(>=) ∷ (a, b, c, d, e, f) → (a, b, c, d, e, f) → Bool

max ∷ (a, b, c, d, e, f) → (a, b, c, d, e, f) → (a, b, c, d, e, f)

min ∷ (a, b, c, d, e, f) → (a, b, c, d, e, f) → (a, b, c, d, e, f)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f, Ord g) ⇒ Ord (a, b, c, d, e, f, g)

Instance details

Methods

compare ∷ (a, b, c, d, e, f, g) → (a, b, c, d, e, f, g) → Ordering

(<) ∷ (a, b, c, d, e, f, g) → (a, b, c, d, e, f, g) → Bool

(<=) ∷ (a, b, c, d, e, f, g) → (a, b, c, d, e, f, g) → Bool

(>) ∷ (a, b, c, d, e, f, g) → (a, b, c, d, e, f, g) → Bool

(>=) ∷ (a, b, c, d, e, f, g) → (a, b, c, d, e, f, g) → Bool

max ∷ (a, b, c, d, e, f, g) → (a, b, c, d, e, f, g) → (a, b, c, d, e, f, g)

min ∷ (a, b, c, d, e, f, g) → (a, b, c, d, e, f, g) → (a, b, c, d, e, f, g)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f, Ord g, Ord h) ⇒ Ord (a, b, c, d, e, f, g, h)

Instance details

Methods

compare ∷ (a, b, c, d, e, f, g, h) → (a, b, c, d, e, f, g, h) → Ordering

(<) ∷ (a, b, c, d, e, f, g, h) → (a, b, c, d, e, f, g, h) → Bool

(<=) ∷ (a, b, c, d, e, f, g, h) → (a, b, c, d, e, f, g, h) → Bool

(>) ∷ (a, b, c, d, e, f, g, h) → (a, b, c, d, e, f, g, h) → Bool

(>=) ∷ (a, b, c, d, e, f, g, h) → (a, b, c, d, e, f, g, h) → Bool

max ∷ (a, b, c, d, e, f, g, h) → (a, b, c, d, e, f, g, h) → (a, b, c, d, e, f, g, h)

min ∷ (a, b, c, d, e, f, g, h) → (a, b, c, d, e, f, g, h) → (a, b, c, d, e, f, g, h)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f, Ord g, Ord h, Ord i) ⇒ Ord (a, b, c, d, e, f, g, h, i)

Instance details

Methods

compare ∷ (a, b, c, d, e, f, g, h, i) → (a, b, c, d, e, f, g, h, i) → Ordering

(<) ∷ (a, b, c, d, e, f, g, h, i) → (a, b, c, d, e, f, g, h, i) → Bool

(<=) ∷ (a, b, c, d, e, f, g, h, i) → (a, b, c, d, e, f, g, h, i) → Bool

(>) ∷ (a, b, c, d, e, f, g, h, i) → (a, b, c, d, e, f, g, h, i) → Bool

(>=) ∷ (a, b, c, d, e, f, g, h, i) → (a, b, c, d, e, f, g, h, i) → Bool

max ∷ (a, b, c, d, e, f, g, h, i) → (a, b, c, d, e, f, g, h, i) → (a, b, c, d, e, f, g, h, i)

min ∷ (a, b, c, d, e, f, g, h, i) → (a, b, c, d, e, f, g, h, i) → (a, b, c, d, e, f, g, h, i)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f, Ord g, Ord h, Ord i, Ord j) ⇒ Ord (a, b, c, d, e, f, g, h, i, j)

Instance details

Methods

compare ∷ (a, b, c, d, e, f, g, h, i, j) → (a, b, c, d, e, f, g, h, i, j) → Ordering

(<) ∷ (a, b, c, d, e, f, g, h, i, j) → (a, b, c, d, e, f, g, h, i, j) → Bool

(<=) ∷ (a, b, c, d, e, f, g, h, i, j) → (a, b, c, d, e, f, g, h, i, j) → Bool

(>) ∷ (a, b, c, d, e, f, g, h, i, j) → (a, b, c, d, e, f, g, h, i, j) → Bool

(>=) ∷ (a, b, c, d, e, f, g, h, i, j) → (a, b, c, d, e, f, g, h, i, j) → Bool

max ∷ (a, b, c, d, e, f, g, h, i, j) → (a, b, c, d, e, f, g, h, i, j) → (a, b, c, d, e, f, g, h, i, j)

min ∷ (a, b, c, d, e, f, g, h, i, j) → (a, b, c, d, e, f, g, h, i, j) → (a, b, c, d, e, f, g, h, i, j)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f, Ord g, Ord h, Ord i, Ord j, Ord k) ⇒ Ord (a, b, c, d, e, f, g, h, i, j, k)

Instance details

Methods

compare ∷ (a, b, c, d, e, f, g, h, i, j, k) → (a, b, c, d, e, f, g, h, i, j, k) → Ordering

(<) ∷ (a, b, c, d, e, f, g, h, i, j, k) → (a, b, c, d, e, f, g, h, i, j, k) → Bool

(<=) ∷ (a, b, c, d, e, f, g, h, i, j, k) → (a, b, c, d, e, f, g, h, i, j, k) → Bool

(>) ∷ (a, b, c, d, e, f, g, h, i, j, k) → (a, b, c, d, e, f, g, h, i, j, k) → Bool

(>=) ∷ (a, b, c, d, e, f, g, h, i, j, k) → (a, b, c, d, e, f, g, h, i, j, k) → Bool

max ∷ (a, b, c, d, e, f, g, h, i, j, k) → (a, b, c, d, e, f, g, h, i, j, k) → (a, b, c, d, e, f, g, h, i, j, k)

min ∷ (a, b, c, d, e, f, g, h, i, j, k) → (a, b, c, d, e, f, g, h, i, j, k) → (a, b, c, d, e, f, g, h, i, j, k)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f, Ord g, Ord h, Ord i, Ord j, Ord k, Ord l) ⇒ Ord (a, b, c, d, e, f, g, h, i, j, k, l)

Instance details

Methods

compare ∷ (a, b, c, d, e, f, g, h, i, j, k, l) → (a, b, c, d, e, f, g, h, i, j, k, l) → Ordering

(<) ∷ (a, b, c, d, e, f, g, h, i, j, k, l) → (a, b, c, d, e, f, g, h, i, j, k, l) → Bool

(<=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l) → (a, b, c, d, e, f, g, h, i, j, k, l) → Bool

(>) ∷ (a, b, c, d, e, f, g, h, i, j, k, l) → (a, b, c, d, e, f, g, h, i, j, k, l) → Bool

(>=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l) → (a, b, c, d, e, f, g, h, i, j, k, l) → Bool

max ∷ (a, b, c, d, e, f, g, h, i, j, k, l) → (a, b, c, d, e, f, g, h, i, j, k, l) → (a, b, c, d, e, f, g, h, i, j, k, l)

min ∷ (a, b, c, d, e, f, g, h, i, j, k, l) → (a, b, c, d, e, f, g, h, i, j, k, l) → (a, b, c, d, e, f, g, h, i, j, k, l)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f, Ord g, Ord h, Ord i, Ord j, Ord k, Ord l, Ord m) ⇒ Ord (a, b, c, d, e, f, g, h, i, j, k, l, m)

Instance details

Methods

compare ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m) → (a, b, c, d, e, f, g, h, i, j, k, l, m) → Ordering

(<) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m) → (a, b, c, d, e, f, g, h, i, j, k, l, m) → Bool

(<=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m) → (a, b, c, d, e, f, g, h, i, j, k, l, m) → Bool

(>) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m) → (a, b, c, d, e, f, g, h, i, j, k, l, m) → Bool

(>=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m) → (a, b, c, d, e, f, g, h, i, j, k, l, m) → Bool

max ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m) → (a, b, c, d, e, f, g, h, i, j, k, l, m) → (a, b, c, d, e, f, g, h, i, j, k, l, m)

min ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m) → (a, b, c, d, e, f, g, h, i, j, k, l, m) → (a, b, c, d, e, f, g, h, i, j, k, l, m)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f, Ord g, Ord h, Ord i, Ord j, Ord k, Ord l, Ord m, Ord n) ⇒ Ord (a, b, c, d, e, f, g, h, i, j, k, l, m, n)

Instance details

Methods

compare ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n) → Ordering

(<) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n) → Bool

(<=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n) → Bool

(>) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n) → Bool

(>=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n) → Bool

max ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n)

min ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f, Ord g, Ord h, Ord i, Ord j, Ord k, Ord l, Ord m, Ord n, Ord o) ⇒ Ord (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o)

Instance details

Methods

compare ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o) → Ordering

(<) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o) → Bool

(<=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o) → Bool

(>) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o) → Bool

(>=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o) → Bool

max ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o)

min ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f, Ord g, Ord h, Ord i, Ord j, Ord k, Ord l, Ord m, Ord n, Ord o, Ord p) ⇒ Ord (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p)

Instance details

Methods

compare ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p) → Ordering

(<) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p) → Bool

(<=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p) → Bool

(>) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p) → Bool

(>=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p) → Bool

max ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p)

min ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f, Ord g, Ord h, Ord i, Ord j, Ord k, Ord l, Ord m, Ord n, Ord o, Ord p, Ord q) ⇒ Ord (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q)

Instance details

Methods

compare ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q) → Ordering

(<) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q) → Bool

(<=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q) → Bool

(>) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q) → Bool

(>=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q) → Bool

max ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q)

min ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f, Ord g, Ord h, Ord i, Ord j, Ord k, Ord l, Ord m, Ord n, Ord o, Ord p, Ord q, Ord r) ⇒ Ord (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r)

Instance details

Methods

compare ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r) → Ordering

(<) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r) → Bool

(<=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r) → Bool

(>) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r) → Bool

(>=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r) → Bool

max ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r)

min ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f, Ord g, Ord h, Ord i, Ord j, Ord k, Ord l, Ord m, Ord n, Ord o, Ord p, Ord q, Ord r, Ord s) ⇒ Ord (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s)

Instance details

Methods

compare ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s) → Ordering

(<) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s) → Bool

(<=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s) → Bool

(>) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s) → Bool

(>=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s) → Bool

max ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s)

min ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f, Ord g, Ord h, Ord i, Ord j, Ord k, Ord l, Ord m, Ord n, Ord o, Ord p, Ord q, Ord r, Ord s, Ord t) ⇒ Ord (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t)

Instance details

Methods

compare ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t) → Ordering

(<) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t) → Bool

(<=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t) → Bool

(>) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t) → Bool

(>=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t) → Bool

max ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t)

min ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f, Ord g, Ord h, Ord i, Ord j, Ord k, Ord l, Ord m, Ord n, Ord o, Ord p, Ord q, Ord r, Ord s, Ord t, Ord u) ⇒ Ord (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u)

Instance details

Methods

compare ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u) → Ordering

(<) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u) → Bool

(<=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u) → Bool

(>) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u) → Bool

(>=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u) → Bool

max ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u)

min ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f, Ord g, Ord h, Ord i, Ord j, Ord k, Ord l, Ord m, Ord n, Ord o, Ord p, Ord q, Ord r, Ord s, Ord t, Ord u, Ord v) ⇒ Ord (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v)

Instance details

Methods

compare ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v) → Ordering

(<) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v) → Bool

(<=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v) → Bool

(>) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v) → Bool

(>=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v) → Bool

max ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v)

min ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f, Ord g, Ord h, Ord i, Ord j, Ord k, Ord l, Ord m, Ord n, Ord o, Ord p, Ord q, Ord r, Ord s, Ord t, Ord u, Ord v, Ord w) ⇒ Ord (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w)

Instance details

Methods

compare ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w) → Ordering

(<) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w) → Bool

(<=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w) → Bool

(>) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w) → Bool

(>=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w) → Bool

max ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w)

min ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f, Ord g, Ord h, Ord i, Ord j, Ord k, Ord l, Ord m, Ord n, Ord o, Ord p, Ord q, Ord r, Ord s, Ord t, Ord u, Ord v, Ord w, Ord x) ⇒ Ord (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x)

Instance details

Methods

compare ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x) → Ordering

(<) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x) → Bool

(<=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x) → Bool

(>) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x) → Bool

(>=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x) → Bool

max ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x)

min ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f, Ord g, Ord h, Ord i, Ord j, Ord k, Ord l, Ord m, Ord n, Ord o, Ord p, Ord q, Ord r, Ord s, Ord t, Ord u, Ord v, Ord w, Ord x, Ord y) ⇒ Ord (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y)

Instance details

Methods

compare ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y) → Ordering

(<) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y) → Bool

(<=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y) → Bool

(>) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y) → Bool

(>=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y) → Bool

max ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y)

min ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f, Ord g, Ord h, Ord i, Ord j, Ord k, Ord l, Ord m, Ord n, Ord o, Ord p, Ord q, Ord r, Ord s, Ord t, Ord u, Ord v, Ord w, Ord x, Ord y, Ord z) ⇒ Ord (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z)

Instance details

Methods

compare ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z) → Ordering

(<) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z) → Bool

(<=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z) → Bool

(>) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z) → Bool

(>=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z) → Bool

max ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z)

min ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z)

(Ord a, Ord b, Ord c, Ord d, Ord e, Ord f, Ord g, Ord h, Ord i, Ord j, Ord k, Ord l, Ord m, Ord n, Ord o, Ord p, Ord q, Ord r, Ord s, Ord t, Ord u, Ord v, Ord w, Ord x, Ord y, Ord z, Ord t26) ⇒ Ord (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26)

Instance details

Methods

compare ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26) → Ordering

(<) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26) → Bool

(<=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26) → Bool

(>) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26) → Bool

(>=) ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26) → Bool

max ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26)

min ∷ (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26) → (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26)