PlutusTx.Plugin.Boilerplate

Description

Insert necessary or recommended GHC flags and extensions via a driver plugin. See https://plutus.cardano.intersectmbo.org/docs/using-plinth/extensions-flags-pragmas

Documentation

optNoStrict ∷ CommandLineOption

Unfortunately, it seems like the Strict extension set by the driver plugin cannot be unset via LANGUAGE NoStrict. So we add a plugin flag to allow users to do so.

boilerplateOpts ∷ [CommandLineOption]

removeBoilerplateOpts ∷ [CommandLineOption] → [CommandLineOption]

addFlagsAndExts ∷ [CommandLineOption] → HscEnv → IO HscEnv

addInlineables ∷ TcGblEnv → TcM TcGblEnv

Add INLINEABLE to all bindings that carry no user-written inline pragma.

addInlineableABE ∷ ABExport → ABExport

addInlineable ∷ LHsBindLR GhcTc GhcTc → LHsBindLR GhcTc GhcTc

needsInlineable ∷ Id → Bool

Return true if the Id carries no user-written inline pragma.

inlineable ∷ InlinePragma