PlutusTx.Plugin.Common

Documentation

data PluginCtx

Constructors

PluginCtx

Fields

pcOpts ∷ PluginOptions

pcFamEnvs ∷ FamInstEnvs

pcMarkerName ∷ Name

pcAnchorName ∷ Name

pcModuleName ∷ ModuleName

pcModuleModBreaks ∷ Maybe ModBreaks

installCorePlugin ∷ Name → [CommandLineOption] → [CoreToDo] → CoreM [CoreToDo]

plinthcModName ∷ String

anchorName ∷ String

injectAnchors ∷ TcGblEnv → TcM TcGblEnv

Wrap certain HsExprs in the typed checked module with anchor.

anchorExpr ∷ Id → LHsExpr GhcTc → LHsExpr GhcTc

Wrap an HsExpr with anchor.

isAnchorWorthy ∷ Id → HsExpr GhcTc → Bool

isTick ∷ HsExpr GhcTc → Bool

isAnchorApp ∷ Id → HsExpr GhcTc → Bool

stripGuardAnchors ∷ Id → GRHS GhcTc (LHsExpr GhcTc) → GRHS GhcTc (LHsExpr GhcTc)

Remove anchors in guards. We can't wrap otherwise with an anchor, since it would cause GHC to emit a "non-exhuastive" warning. In general, anchors in guards are probably not very useful, so we remove all anchors within guards.

stripHsAnchor ∷ Id → LHsExpr GhcTc → LHsExpr GhcTc

mkSimplPass ∷ DynFlags → CoreToDo

A simplifier pass, implemented by GHC

mkPluginPass ∷ Name → PluginOptions → CoreToDo

Our plugin works at haskell-module level granularity; the plugin looks at the module's top-level bindings for markers and compiles their right-hand-side core expressions.

type PluginM uni fun = ReaderT PluginCtx (ExceptT (CompileError uni fun Ann) CoreM)

The monad where the plugin runs in for each module. It is a core->core compiler monad, called PluginM, augmented with pure errors.

runPluginM ∷ ∀ uni fun a. (PrettyUni uni, Pretty fun) ⇒ PluginCtx → PluginM uni fun a → CoreM a

Runs the plugin monad in a given context; throws a Ghc.Exception when compilation fails.

getSourceSnippet ∷ RealSrcSpan → IO (Maybe (Doc ann))

formatSourceSnippet ∷ Int → Int → Int → String → Doc ann

compileBind ∷ CoreBind → PluginM DefaultUni DefaultFun CoreBind

Compiles all the marked expressions in the given binder into PLC literals.

compileMarkedExprs ∷ CoreExpr → PluginM DefaultUni DefaultFun CoreExpr

Compiles all the core-expressions surrounded by the marker in the given expression into PLC literals.

compileMarkedExprOrDefer ∷ String → Type → CoreExpr → PluginM DefaultUni DefaultFun CoreExpr

Behaves the same as compileMarkedExpr, unless a compilation error occurs ; if a compilation error happens and the 'defer-errors' option is turned on, the compilation error is suppressed and the original hs expression is replaced with a haskell runtime-error expression.

emitRuntimeError ∷ (PrettyUni uni, Pretty fun) ⇒ Type → CompileError uni fun Ann → PluginM uni fun CoreExpr

Given an expected Haskell type a, it generates Haskell code which throws a GHC runtime error "as" 'CompiledCode a'.

compileMarkedExpr ∷ String → Type → CoreExpr → PluginM DefaultUni DefaultFun CoreExpr

Compile the core expression that is surrounded by a plc marker, and return a core expression which evaluates to the compiled plc AST as a serialized bytestring, to be injected back to the Haskell program.

runCompiler ∷ ∀ uni fun m. (uni ~ DefaultUni, fun ~ DefaultFun, MonadReader (CompileContext uni fun) m, MonadState CompileState m, MonadWriter CoverageIndex m, MonadQuote m, MonadError (CompileError uni fun Ann) m, MonadIO m) ⇒ String → PluginOptions → CoreExpr → m (PIRProgram uni fun, UPLCProgram uni fun)

The GHC.Core to PIR to PLC compiler pipeline. Returns both the PIR and PLC output. It invokes the whole compiler chain: Core expr -> PIR expr -> PLC expr -> UPLC expr.

thNameToGhcNameOrFail ∷ Name → PluginM uni fun Name

Get the Name corresponding to the given Name, or throw an error if we can't get it.

makeByteStringLiteral ∷ ByteString → PluginM uni fun CoreExpr

Create a GHC Core expression that will evaluate to the given ByteString at runtime.

stripTicks ∷ CoreExpr → CoreExpr

Strips all enclosing Ticks off an expression.

stripCoreAnchors ∷ Name → CoreExpr → CoreExpr

mkCompiledCode ∷ ∀ a. ByteString → ByteString → ByteString → CompiledCode a

Helper to avoid doing too much construction of Core ourselves

makePrimitiveNameInfo ∷ [Name] → PluginM uni fun NameInfo

Make a NameInfo mapping the given set of TH names to their TyThings for later reference.

makeRewriteRules ∷ PluginOptions → RewriteRules DefaultUni DefaultFun