PlutusTx.Prelude

Description

The PlutusTx Prelude is a replacement for the Haskell Prelude that works better with Plutus Tx. You should use it if you're writing code that will be compiled with the Plutus Tx compiler.

    :set -XNoImplicitPrelude
    import PlutusTx.Prelude

Classes

module PlutusTx.Eq

module PlutusTx.Enum

module PlutusTx.Ord

module PlutusTx.Semigroup

module PlutusTx.Monoid

module PlutusTx.Numeric

module PlutusTx.Function

module PlutusTx.Functor

module PlutusTx.Applicative

module PlutusTx.Lattice

Monad

(>>=) ∷ Monad m ⇒ m a → (a → m b) → m b infixl 1

Sequentially compose two actions, passing any value produced by the first as an argument to the second.

'as >>= bs' can be understood as the do expression

do a <- as
   bs a

(=<<) ∷ Monad m ⇒ (a → m b) → m a → m b infixr 1

Same as >>=, but with the arguments interchanged.

(>>) ∷ Monad m ⇒ m a → m b → m b infixl 1

Sequentially compose two actions, discarding any value produced by the first, like sequencing operators (such as the semicolon) in imperative languages.

'as >> bs' can be understood as the do expression

do as
   bs

return ∷ Monad m ⇒ a → m a

Inject a value into the monadic type.

Standard functions, Tuples

module PlutusTx.Base

Tracing functions

module PlutusTx.Trace

Unit

data BuiltinUnit

String

data BuiltinString

appendString ∷ BuiltinString → BuiltinString → BuiltinString

Append two Strings.

emptyString ∷ BuiltinString

An empty String.

equalsString ∷ BuiltinString → BuiltinString → Bool

Check if two strings are equal

encodeUtf8 ∷ BuiltinString → BuiltinByteString

Convert a String into a ByteString.

Error

error ∷ () → a

Aborts evaluation with an error.

check ∷ Bool → BuiltinUnit

Checks a Bool and aborts if it is false.

Booleans

module PlutusTx.Bool

Integer numbers

data Integer

Arbitrary precision integers. In contrast with fixed-size integral types such as Int, the Integer type represents the entire infinite range of integers.

Integers are stored in a kind of sign-magnitude form, hence do not expect two's complement form when using bit operations.

If the value is small (fit into an Int), IS constructor is used. Otherwise Integer and IN constructors are used to store a BigNat representing respectively the positive or the negative value magnitude.

Invariant: Integer and IN are used iff value doesn't fit in IS

divide ∷ Integer → Integer → Integer

Integer division, rounding downwards

>>> divide (-41) 5
-9

modulo ∷ Integer → Integer → Integer

Integer remainder, always positive for a positive divisor

>>> modulo (-41) 5
4

quotient ∷ Integer → Integer → Integer

Integer division, rouding towards zero

>>> quotient (-41) 5
-8

remainder ∷ Integer → Integer → Integer

Integer remainder, same sign as dividend

>>> remainder (-41) 5
-1

even ∷ Integer → Bool

odd ∷ Integer → Bool

expMod ∷ Integer → Integer → Integer → Integer

FIXME

Maybe

module PlutusTx.Maybe

Either

module PlutusTx.Either

ByteStrings

data BuiltinByteString

An opaque type representing Plutus Core ByteStrings.

appendByteString ∷ BuiltinByteString → BuiltinByteString → BuiltinByteString

Concatenates two ByteStrings.

consByteString ∷ Integer → BuiltinByteString → BuiltinByteString

Adds a byte to the front of a ByteString.

takeByteString ∷ Integer → BuiltinByteString → BuiltinByteString

Returns the n length prefix of a ByteString.

dropByteString ∷ Integer → BuiltinByteString → BuiltinByteString

Returns the suffix of a ByteString after n elements.

sliceByteString ∷ Integer → Integer → BuiltinByteString → BuiltinByteString

Returns the substring of a ByteString from index start of length n.

lengthOfByteString ∷ BuiltinByteString → Integer

Returns the length of a ByteString.

indexByteString ∷ BuiltinByteString → Integer → Integer

Returns the byte of a ByteString at index.

emptyByteString ∷ BuiltinByteString

An empty ByteString.

decodeUtf8 ∷ BuiltinByteString → BuiltinString

Converts a ByteString to a String.

newtype BuiltinByteStringUtf8

Constructors

BuiltinByteStringUtf8

Fields

unBuiltinByteStringUtf8 ∷ BuiltinByteString

andByteString ∷ Bool → BuiltinByteString → BuiltinByteString → BuiltinByteString

Perform logical AND on two BuiltinByteString arguments, as described in CIP-122.

The first argument indicates whether padding semantics should be used or not; if False, truncation semantics will be used instead.

See also

Padding and truncation semantics

Bit indexing scheme

orByteString ∷ Bool → BuiltinByteString → BuiltinByteString → BuiltinByteString

Perform logical OR on two BuiltinByteString arguments, as described here.

The first argument indicates whether padding semantics should be used or not; if False, truncation semantics will be used instead.

See also

Padding and truncation semantics

Bit indexing scheme

xorByteString ∷ Bool → BuiltinByteString → BuiltinByteString → BuiltinByteString

Perform logical XOR on two BuiltinByteString arguments, as described here.

The first argument indicates whether padding semantics should be used or not; if False, truncation semantics will be used instead.

See also

Padding and truncation semantics

Bit indexing scheme

complementByteString ∷ BuiltinByteString → BuiltinByteString

Perform logical complement on a BuiltinByteString, as described here.

See also

Bit indexing scheme

Bit operations

readBit ∷ BuiltinByteString → Integer → Bool

Read a bit at the _bit_ index given by the Integer argument in the BuiltinByteString argument. The result will be True if the corresponding bit is set, and False if it is clear. Will error if given an out-of-bounds index argument; that is, if the index is either negative, or equal to or greater than the total number of bits in the BuiltinByteString argument.

See also

Bit indexing scheme

Operation description

writeBits ∷ BuiltinByteString → [Integer] → Bool → BuiltinByteString

Given a BuiltinByteString, a list of indexes to change, and a boolean value b to change those indexes to, set the bit at each of the specified index as follows:

If b is True, set that bit;

Otherwise, clear that bit.

Will error if any of the indexes are out-of-bounds: that is, if the index is either negative, or equal to or greater than the total number of bits in the BuiltinByteString argument.

Note

This differs slightly from the description of the corresponding operation in CIP-122; instead of a single changelist argument comprised of pairs, we instead pass a single list of indexes to change, and a single boolean value to change those indexes to. The original proposal allowed one to set and clear bits in a single operation, but constructing the list of boolean values for the updates was somewhat expensive. If it's really necessary to set some bits and clear others then it is easier to call the function twice, once to set bits and and once to clear them.

See also

Bit indexing scheme

Operation description

replicateByte ∷ Integer → Integer → BuiltinByteString

Given a length (first argument) and a byte (second argument), produce a BuiltinByteString of that length, with that byte in every position. Will error if given a negative length, or a second argument that isn't a byte (less than 0, greater than 255).

See also

Operation description

shiftByteString ∷ BuiltinByteString → Integer → BuiltinByteString

Shift a BuiltinByteString, as per CIP-123.

rotateByteString ∷ BuiltinByteString → Integer → BuiltinByteString

Rotate a BuiltinByteString, as per CIP-123.

countSetBits ∷ BuiltinByteString → Integer

Count the set bits in a BuiltinByteString, as per CIP-123.

findFirstSetBit ∷ BuiltinByteString → Integer

Find the lowest index of a set bit in a BuiltinByteString, as per CIP-123.

If given a BuiltinByteString which consists only of zero bytes (including the empty BuiltinByteString, this returns -1.

Hashes and Signatures

sha2_256 ∷ BuiltinByteString → BuiltinByteString

The SHA2-256 hash of a ByteString

sha3_256 ∷ BuiltinByteString → BuiltinByteString

The SHA3-256 hash of a ByteString

blake2b_224 ∷ BuiltinByteString → BuiltinByteString

The BLAKE2B-224 hash of a ByteString

blake2b_256 ∷ BuiltinByteString → BuiltinByteString

The BLAKE2B-256 hash of a ByteString

keccak_256 ∷ BuiltinByteString → BuiltinByteString

The KECCAK-256 hash of a ByteString

ripemd_160 ∷ BuiltinByteString → BuiltinByteString

The RIPEMD-160 hash of a ByteString

verifyEd25519Signature

Arguments

∷ BuiltinByteString
Public Key (32 bytes)

→ BuiltinByteString
Message (arbirtary length)

→ BuiltinByteString
Signature (64 bytes)

→ Bool

Ed25519 signature verification. Verify that the signature is a signature of the message by the public key. This will fail if key or the signature are not of the expected length.

verifyEcdsaSecp256k1Signature

Arguments

∷ BuiltinByteString
Verification key (33 bytes)

→ BuiltinByteString
Message hash (32 bytes)

→ BuiltinByteString
Signature (64 bytes)

→ Bool

Given an ECDSA SECP256k1 verification key, an ECDSA SECP256k1 signature, and an ECDSA SECP256k1 message hash (all as BuiltinByteStrings), verify the hash with that key and signature.

Note

There are additional well-formation requirements for the arguments beyond their length:

The first byte of the public key must correspond to the sign of the y coordinate: this is 0x02 if y is even, and 0x03 otherwise.

The remaining bytes of the public key must correspond to the x coordinate, as a big-endian integer.

The first 32 bytes of the signature must correspond to the big-endian integer representation of _r_.

The last 32 bytes of the signature must correspond to the big-endian integer representation of _s_.

While this primitive accepts a hash, any caller should only pass it hashes that they computed themselves: specifically, they should receive the message from a sender and hash it, rather than receiving the hash from said sender. Failure to do so can be dangerous. Other than length, we make no requirements of what hash gets used.

See also

secp256k1_ec_pubkey_serialize; this implements the format for the verification key that we accept, given a length argument of 33 and the SECP256K1_EC_COMPRESSED flag.

secp256k1_ecdsa_serialize_compact; this implements the format for the signature that we accept.

verifySchnorrSecp256k1Signature

Arguments

∷ BuiltinByteString
Verification key (32 bytes)

→ BuiltinByteString
Message (arbitrary length)

→ BuiltinByteString
Signature (64 bytes)

→ Bool

Given a Schnorr SECP256k1 verification key, a Schnorr SECP256k1 signature, and a message (all as BuiltinByteStrings), verify the message with that key and signature.

Note

There are additional well-formation requirements for the arguments beyond their length. Throughout, we refer to co-ordinates of the point R.

The bytes of the public key must correspond to the x coordinate, as a big-endian integer, as specified in BIP-340.

The first 32 bytes of the signature must correspond to the x coordinate, as a big-endian integer, as specified in BIP-340.

The last 32 bytes of the signature must correspond to the bytes of s, as a big-endian integer, as specified in BIP-340.

See also

BIP-340

secp256k1_xonly_pubkey_serialize; this implements the format for the verification key that we accept.

secp256k1_schnorrsig_sign; this implements the signing logic for signatures this builtin can verify.

Rational numbers

data Rational

Represents an arbitrary-precision ratio.

The following two invariants are maintained:

The denominator is greater than zero.

The numerator and denominator are coprime.

unsafeRatio ∷ Integer → Integer → Rational

Makes a Rational from a numerator and a denominator.

Important note

If given a zero denominator, this function will error. If you don't mind a size increase, and care about safety, use ratio instead.

ratio ∷ Integer → Integer → Maybe Rational

Safely constructs a Rational from a numerator and a denominator. Returns Nothing if given a zero denominator.

fromInteger ∷ Integer → Rational

Converts an Integer into the equivalent Rational.

round ∷ Rational → Integer

round r returns the nearest Integer value to r. If r is equidistant between two values, the even value will be given.

Other builtin Types

data BuiltinBool

data BuiltinData

A type corresponding to the Plutus Core builtin equivalent of Data.

The point of this type is to be an opaque equivalent of Data, so as to ensure that it is only used in ways that the compiler can handle.

As such, you should use this type in your on-chain code, and in any data structures that you want to be representable on-chain.

For off-chain usage, there are conversion functions builtinDataToData and dataToBuiltinData, but note that these will not work on-chain.

data BuiltinList a

data BuiltinPair a b

To/from Data

class ToData (a ∷ Type) where

A typeclass for types that can be converted to and from BuiltinData.

Methods

toBuiltinData ∷ a → BuiltinData

Convert a value to BuiltinData.

class FromData (a ∷ Type) where

Methods

fromBuiltinData ∷ BuiltinData → Maybe a

Convert a value from BuiltinData, returning Nothing if this fails.

class UnsafeFromData (a ∷ Type) where

Methods

unsafeFromBuiltinData ∷ BuiltinData → a

Convert a value from BuiltinData, calling error if this fails. This is typically much faster than fromBuiltinData.

When implementing this function, make sure to call unsafeFromBuiltinData rather than fromBuiltinData when converting substructures!

This is a simple type without any validation, use with caution.

BLS12_381

data BuiltinBLS12_381_G1_Element

bls12_381_G1_equals ∷ BuiltinBLS12_381_G1_Element → BuiltinBLS12_381_G1_Element → Bool

bls12_381_G1_add ∷ BuiltinBLS12_381_G1_Element → BuiltinBLS12_381_G1_Element → BuiltinBLS12_381_G1_Element

bls12_381_G1_neg ∷ BuiltinBLS12_381_G1_Element → BuiltinBLS12_381_G1_Element

bls12_381_G1_scalarMul ∷ Integer → BuiltinBLS12_381_G1_Element → BuiltinBLS12_381_G1_Element

bls12_381_G1_compress ∷ BuiltinBLS12_381_G1_Element → BuiltinByteString

bls12_381_G1_uncompress ∷ BuiltinByteString → BuiltinBLS12_381_G1_Element

bls12_381_G1_hashToGroup ∷ BuiltinByteString → BuiltinByteString → BuiltinBLS12_381_G1_Element

bls12_381_G1_compressed_zero ∷ BuiltinByteString

bls12_381_G1_compressed_generator ∷ BuiltinByteString

data BuiltinBLS12_381_G2_Element

bls12_381_G2_equals ∷ BuiltinBLS12_381_G2_Element → BuiltinBLS12_381_G2_Element → Bool

bls12_381_G2_add ∷ BuiltinBLS12_381_G2_Element → BuiltinBLS12_381_G2_Element → BuiltinBLS12_381_G2_Element

bls12_381_G2_neg ∷ BuiltinBLS12_381_G2_Element → BuiltinBLS12_381_G2_Element

bls12_381_G2_scalarMul ∷ Integer → BuiltinBLS12_381_G2_Element → BuiltinBLS12_381_G2_Element

bls12_381_G2_compress ∷ BuiltinBLS12_381_G2_Element → BuiltinByteString

bls12_381_G2_uncompress ∷ BuiltinByteString → BuiltinBLS12_381_G2_Element

bls12_381_G2_hashToGroup ∷ BuiltinByteString → BuiltinByteString → BuiltinBLS12_381_G2_Element

bls12_381_G2_compressed_zero ∷ BuiltinByteString

bls12_381_G2_compressed_generator ∷ BuiltinByteString

data BuiltinBLS12_381_MlResult

bls12_381_millerLoop ∷ BuiltinBLS12_381_G1_Element → BuiltinBLS12_381_G2_Element → BuiltinBLS12_381_MlResult

bls12_381_mulMlResult ∷ BuiltinBLS12_381_MlResult → BuiltinBLS12_381_MlResult → BuiltinBLS12_381_MlResult

bls12_381_finalVerify ∷ BuiltinBLS12_381_MlResult → BuiltinBLS12_381_MlResult → Bool

Conversions

fromBuiltin ∷ HasFromBuiltin arep ⇒ arep → FromBuiltin arep

toBuiltin ∷ HasToBuiltin a ⇒ a → ToBuiltin a

fromOpaque ∷ HasFromOpaque arep a ⇒ arep → a

toOpaque ∷ HasToOpaque a arep ⇒ a → arep

integerToByteString ∷ ByteOrder → Integer → Integer → BuiltinByteString

Convert a BuiltinInteger into a BuiltinByteString, as described in CIP-121. The first argument indicates the endianness of the conversion and the third argument is the integer to be converted, which must be non-negative. The second argument must also be non-negative and it indicates the required width of the output. If the width is zero then the output is the smallest bytestring which can contain the converted input (and in this case, the integer 0 encodes to the empty bytestring). If the width is nonzero then the output bytestring will be padded to the required width with 0x00 bytes (on the left for big-endian conversions and on the right for little-endian conversions); if the input integer is too big to fit into a bytestring of the specified width then the conversion will fail. Conversion will also fail if the specified width is greater than 8192 or the input integer is too big to fit into a bytestring of length 8192.

byteStringToInteger ∷ ByteOrder → BuiltinByteString → Integer

Convert a BuiltinByteString to a BuiltinInteger, as described in CIP-121. The first argument indicates the endianness of the conversion and the second is the bytestring to be converted. There is no limitation on the size of the bytestring. The empty bytestring is converted to the integer 0.