PlutusTx.Show.TH

Documentation

class Show a where

Conversion of values to BuiltinStrings. Unlike GHC.Show.Show, there is no showList method, because there is no Show instance for String.

Minimal complete definition

showsPrec | show

Methods

showsPrec ∷ Integer → a → ShowS

show ∷ a → BuiltinString

type ShowS = [BuiltinString] → [BuiltinString]

Currently the only way to concatenate BuiltinStrings is appendString, whose cost is linear in the total length of the two strings. A naive concatenation of multiple BuiltinStrings costs O(n^2) in the worst case, where n is the total length. By collecting the BuiltinStrings in a list and concatenating them in the end, the cost can be reduced to O(n*logn). If we add a concatStrings builtin function in the future, the cost can be further reduced to O(n).

Like ShowS, the purpose of the function type here is to turn list concatenation into function composition.

showString ∷ BuiltinString → ShowS

showSpace ∷ ShowS

showCommaSpace ∷ ShowS

showParen ∷ Bool → ShowS → ShowS

appPrec ∷ Integer

appPrec1 ∷ Integer

concatBuiltinStrings ∷ [BuiltinString] → BuiltinString

deriveShow ∷ Name → Q [Dec]

Derive Show instance. Adapted from Text.Show.Deriving.deriveShow.

deriveShowsPrec ∷ [ConstructorInfo] → [Q Dec]

Derive showsPrec definition for each data constructor.

deriveShowsPrecBody ∷ [ConstructorInfo] → Q Exp

deriveMatchForCon ∷ Name → ConstructorInfo → Q Match

Derive showsPrec body for a single data constructor.

deriveShowExpForArg ∷ Integer → Name → Q Exp

Derive the showsPrec expression for showing a single constructor argument.

parenInfixConName ∷ Name → String

Add parens if it is an infix data constructor.