PlutusTx.TH

Documentation

compile ∷ Code Q a → Code Q (CompiledCode a)

Compile a quoted Haskell expression into a corresponding Plutus Core program.

compileUntyped ∷ Q Exp → Q Exp

Compile a quoted Haskell expression into a corresponding Plutus Core program.

loadFromFile ∷ FilePath → Code Q (CompiledCode a)

Load a CompiledCode from a file. Drop-in replacement for compile.