PlutusTx.Test

Size tests

goldenSize ∷ ToUPlc a DefaultUni DefaultFun ⇒ String → a → TestNested

fitsUnder ∷ ∀ (a ∷ Type). Typeable a ⇒ String → (String, CompiledCode a) → (String, CompiledCode a) → TestTree

Compilation testing

goldenPir ∷ (PrettyUni uni, Pretty fun, uni `Everywhere` Flat, Flat fun) ⇒ String → CompiledCodeIn uni fun a → TestNested

Does not print uniques.

goldenPirReadable ∷ (PrettyUni uni, Pretty fun, uni `Everywhere` Flat, Flat fun) ⇒ String → CompiledCodeIn uni fun a → TestNested

Does not print uniques.

goldenPirBy ∷ (PrettyUni uni, Pretty fun, uni `Everywhere` Flat, Flat fun) ⇒ PrettyConfigClassic PrettyConfigName → String → CompiledCodeIn uni fun a → TestNested

goldenTPlc ∷ ToTPlc a DefaultUni DefaultFun ⇒ String → a → TestNested

goldenUPlc ∷ ToUPlc a DefaultUni DefaultFun ⇒ String → a → TestNested

goldenUPlcReadable ∷ ToUPlc a DefaultUni DefaultFun ⇒ String → a → TestNested

Evaluation testing

goldenEvalCek ∷ ToUPlc a DefaultUni DefaultFun ⇒ String → [a] → TestNested

goldenEvalCekCatch ∷ ToUPlc a DefaultUni DefaultFun ⇒ String → [a] → TestNested

goldenEvalCekLog ∷ ToUPlc a DefaultUni DefaultFun ⇒ String → [a] → TestNested

Budget testing

goldenBudget ∷ TestName → CompiledCode a → TestNested

Orphan instances

(PrettyParens (SomeTypeIn uni), GEq uni, Typecheckable uni fun, Closed uni, Everywhere uni PrettyConst, Pretty fun, Everywhere uni Flat, Flat fun, Default (CostingPart uni fun), Default (BuiltinsInfo uni fun), Default (RewriteRules uni fun)) ⇒ ToTPlc (CompiledCodeIn uni fun a) uni fun

Instance details

Methods

toTPlc ∷ CompiledCodeIn uni fun a → ExceptT SomeException IO (Program TyName Name uni fun ())

(Closed uni, Everywhere uni Flat, Flat fun) ⇒ ToUPlc (CompiledCodeIn uni fun a) uni fun

Instance details

Methods

toUPlc ∷ CompiledCodeIn uni fun a → ExceptT SomeException IO (Program Name uni fun ())