PlutusTx.Test.Golden

TH CodGen

goldenCodeGen ∷ Ppr a ⇒ TestName → Q a → ExpQ

Compilation testing

goldenPir ∷ (PrettyUni uni, Pretty fun, uni `Everywhere` Flat, Flat fun) ⇒ TestName → CompiledCodeIn uni fun a → TestNested

Does not print uniques.

goldenPirReadable ∷ (PrettyUni uni, Pretty fun, uni `Everywhere` Flat, Flat fun) ⇒ TestName → CompiledCodeIn uni fun a → TestNested

Does not print uniques.

goldenPirReadableU ∷ (PrettyUni uni, Pretty fun, uni `Everywhere` Flat, Flat fun) ⇒ TestName → CompiledCodeIn uni fun a → TestNested

Prints uniques. This should be used sparingly: a simple change to a script or a compiler pass may change all uniques, making it difficult to see the actual change if all uniques are printed. It is nonetheless useful sometimes.

goldenPirBy ∷ (PrettyUni uni, Pretty fun, uni `Everywhere` Flat, Flat fun) ⇒ PrettyConfigClassic PrettyConfigName → TestName → CompiledCodeIn uni fun a → TestNested

goldenTPlc ∷ ToTPlc a DefaultUni DefaultFun ⇒ TestName → a → TestNested

goldenUPlc ∷ ToUPlc a DefaultUni DefaultFun ⇒ TestName → a → TestNested

goldenUPlcReadable ∷ ToUPlc a DefaultUni DefaultFun ⇒ TestName → a → TestNested

goldenBudget ∷ TestName → CompiledCode a → TestNested

goldenAstSize ∷ ToUPlc a DefaultUni DefaultFun ⇒ TestName → a → TestNested

Golden evaluation testing

goldenEvalCek ∷ ToUPlc a DefaultUni DefaultFun ⇒ TestName → a → TestNested

goldenEvalCekCatch ∷ ToUPlc a DefaultUni DefaultFun ⇒ TestName → a → TestNested

goldenEvalCekCatchBudget ∷ TestName → CompiledCode a → TestNested

goldenEvalCekLog ∷ ToUPlc a DefaultUni DefaultFun ⇒ TestName → a → TestNested

Combined testing

goldenBundle ∷ TestName → CompiledCodeIn DefaultUni DefaultFun a → CompiledCodeIn DefaultUni DefaultFun b → TestNested

goldenBundle' ∷ TestName → CompiledCodeIn DefaultUni DefaultFun a → TestNested

Pretty-printing

prettyBudget ∷ ExBudget → Doc ann

Pretty-print an Execution Budget

prettyCodeSize ∷ CompiledCodeIn DefaultUni DefaultFun a → Doc ann

Pretty-print compiled code size

Given a UPLC program, there are two quantification of "size": AST Size and Flat Size. AST Size measures AST nodes of the given UPLC program. Flat Size measures the number of bytes when the given program serialized into bytestring using binary flat encoding format.

Cost of storing smart contract onchain is partially determined by the Flat size. So it is useful to have Flat size measurement in case we adopt new or introduce optimizations to the flat encoding format.