PlutusTx.Trace

Documentation

trace ∷ BuiltinString → a → a

Emit the given string as a trace message before evaluating the argument.

traceError ∷ BuiltinString → a

Log a message and then terminate the evaluation with an error.

traceIfFalse ∷ BuiltinString → Bool → Bool

Emit the given BuiltinString only if the argument evaluates to False.

traceIfTrue ∷ BuiltinString → Bool → Bool

Emit the given BuiltinString only if the argument evaluates to True.

traceBool ∷ BuiltinString → BuiltinString → Bool → Bool

Emit one of two BuiltinString depending on whether or not the argument evaluates to True or False.