Prettyprinter.Custom

Documentation

brackets' ∷ Doc a → Doc a

This prints a document enclosed by brackets, possibly indenting the output on a new line if it does not fit.

braces' ∷ Doc a → Doc a

This prints a document enclosed by braces, possibly indenting the output on a new line if it does not fit.

parens' ∷ Doc a → Doc a

This prints a document enclosed by parentheses, aligning the opening and closing parentheses.

sexp ∷ Doc a → [Doc a] → Doc a

Print a "sexp", i.e. something like "(keyword arg1 ... argN)".

(<?>) ∷ Doc ann → Doc ann → Doc ann infixr 6

Separate two documents p and q and increase indentation if q has to be put on a new line. Useful to e.g. pretty-print function application like `fun ? sep arguments`.

vcatHard ∷ [Doc ann] → Doc ann

Lay out a sequence of documents vertically with forced lines between documents. Useful for prettyprinting layout-sensitive things like let-bindings.