Test.Tasty.Extras

Documentation

newtype Layer a r

A monad allowing one to emit elements of type a. Semantically equivalent to Writer (DList a) r, but:

is faster, being based on the Church-encoded free monad

implements Monoid, so that all the Data.Foldable convenience is supported

has better ergonomics as it doesn't require the user to wrap a values into DLists

This type is also semantically equivalent to Stream (Of a) Identity r.

Useful for monadically creating tree-like structures, for example the following

import Data.Tree
yield = embed . pure
main = putStrLn . drawTree . Node "a" . toList $ do
    yield "b"
    nestWith (Node "c") $ do
        yield "d"
        yield "e"
    yield "f"

will produce

-a
|
+- b
|
+- c
|  |
|  +- d
|  |
|  `- e
|
`- f

Constructors

Layer

Fields

unLayer ∷ F ((,) a) r

embed ∷ MonadFree ((,) a) m ⇒ a → m ()

Embed the given value into a Layer-like type (either Layer itself or a monad transformer stack with Layer at the bottom).

nestWith ∷ ([a] → a) → Layer a () → Layer a ()

Collapse the given Layer into a single element by converting it into a list, applying the given function to the result and embedding it back.

newtype TestNestedM r

Constructors

TestNestedM

Fields

unTestNestedM ∷ ReaderT [FilePath] (Layer TestTree) r

type TestNested = TestNestedM ()

A TestTree of tests under some name prefix.

runTestNestedM ∷ [String] → TestNested → TestTree

Run a TestNested computation to produce a TestTree (without actually executing the tests).

testNestedNamedM

Arguments

∷ FilePath
The name of the folder.

→ String
The name of the test group to render in CLI.

→ TestNested

→ TestNested

Descend into a folder.

testNestedM ∷ FilePath → TestNested → TestNested

Descend into a folder for a TestNested computation.

testNestedGhcM ∷ TestNested → TestNested

Like testNestedM but adds a subdirectory corresponding to the GHC version being used.

runTestNested ∷ [String] → [TestNested] → TestTree

Run a list of TestNested computation to produce a TestTree (without actually executing the tests).

testNestedNamed

Arguments

∷ FilePath
The name of the folder.

→ String
The name of the test group to render in CLI.

→ [TestNested]

→ TestNested

Descend into a folder for a list of tests.

testNested ∷ FilePath → [TestNested] → TestNested

Descend into a folder for a list of TestNested computations.

testNestedGhc ∷ [TestNested] → TestNested

Like testNested but adds a subdirectory corresponding to the GHC version being used.

goldenVsText ∷ TestName → FilePath → Text → TestTree

Check the contents of a file against a Text.

goldenVsTextM ∷ TestName → FilePath → IO Text → TestTree

Check the contents of a file against a Text.

goldenVsDoc ∷ TestName → FilePath → Doc ann → TestTree

Check the contents of a file against a Doc.

goldenVsDocM ∷ TestName → FilePath → IO (Doc ann) → TestTree

Check the contents of a file against a Doc.

nestedGoldenVsText ∷ TestName → FilePath → Text → TestNested

Check the contents of a file under a name prefix against a Text.

nestedGoldenVsTextPredM

Arguments

∷ TestName
The name of the test

→ FilePath
The file extension

→ IO Text
The text-producing action to execute

→ (Text → Text → Bool)
How to compare golden file contents with the produced text

→ TestNested

Compare the contents of a file under a name prefix against a Text using a predicate.

nestedGoldenVsTextM ∷ TestName → FilePath → IO Text → TestNested

Check the contents of a file under a name prefix against a Text.

nestedGoldenVsDoc ∷ TestName → FilePath → Doc ann → TestNested

Check the contents of a file under a name prefix against a Text.

nestedGoldenVsDocM ∷ TestName → FilePath → IO (Doc ann) → TestNested

Check the contents of a file under a name prefix against a Text.

makeVersionedFilePath ∷ [FilePath] → FilePath → FilePath

Given a lists of FilePaths and a filename, concatenate the members of the list together, append the GHC version number, then append the filename. We use this to create GHC-version-dependent golden files.

Prettier equality assertions

assertEqualPretty

Arguments

∷ (Eq a, Show a, HasCallStack)

⇒ String
The message prefix

→ a
The expected value

→ a
The actual value

→ Assertion

Asserts that the specified actual value is equal to the expected value. The output message will contain the prefix, the expected value, and the actual value.

If the prefix is the empty string (i.e., ""), then the prefix is omitted and only the expected and actual values are output.

(%=?) infix 1

Arguments

∷ (Eq a, Show a, HasCallStack)

⇒ a
The expected value

→ a
The actual value

→ Assertion

Asserts that the specified actual value is equal to the expected value (with the expected value on the left-hand side).

(%?=) infix 1

Arguments

∷ (Eq a, Show a, HasCallStack)

⇒ a
The actual value

→ a
The expected value

→ Assertion

Asserts that the specified actual value is equal to the expected value (with the actual value on the left-hand side).

ignoreTestWhenHpcEnabled ∷ TestTree → TestTree

Some tests inspect GHC code, but GHC code gets instrumented when using HPC (Haskell Program Coverage), which causes those tests to fail. This function disables those tests when the custom HPC_ENABLED flag is defined.