Text.PrettyBy.Default

Description

Default rendering to string types.

Documentation

layoutDef ∷ Doc ann → SimpleDocStream ann

A default layout for default rendering.

class Render str where

A class for rendering Docs as string types.

Methods

render ∷ Doc ann → str

Render a Doc as a string type.

display ∷ ∀ str a. (Pretty a, Render str) ⇒ a → str

Pretty-print and render a value as a string type.

displayBy ∷ ∀ str a config. (PrettyBy config a, Render str) ⇒ config → a → str

Pretty-print and render a value as a string type in a configurable way.