Text.PrettyBy.Fixity

Description

Configurable precedence-aware pretty-printing.

Look into test/Expr.hs for an extended example.

Documentation

module Text.PrettyBy.Monad

module Text.Fixity

class HasRenderContext config where

A constraint for "RenderContext is a part of config".

Methods

renderContext ∷ Lens' config RenderContext

type MonadPrettyContext config env m = (MonadPretty config env m, HasRenderContext config)

A constraint for "m is a Monad supporting configurable precedence-aware pretty-printing".

newtype Sole a

A newtype wrapper around a introduced for its HasPrettyConfig instance.

Constructors

Sole

Fields

unSole ∷ a

newtype InContextM config a

A monad for precedence-aware pretty-printing.

Constructors

InContextM

Fields

unInContextM ∷ Reader (Sole config) a

type AnyToDoc config ann = ∀ a. PrettyBy config a ⇒ a → Doc ann

The type of a general config-based pretty-printer.

runInContextM ∷ config → InContextM config a → a

Run InContextM by supplying a config.

inContextM ∷ (a → InContextM config (Doc ann)) → config → a → Doc ann

Takes a monadic pretty-printer and turns it into one that receives a config explicitly. Useful for defining instances of PrettyBy monadically when writing precedence-aware pretty-printing code (and since all functions below are monadic, it's currenty the only option).

unitDocM ∷ MonadPrettyContext config env m ⇒ Doc ann → m (Doc ann)

Call encloseM on unitFixity.

encloseM ∷ MonadPrettyContext config env m ⇒ Fixity → Doc ann → m (Doc ann)

Enclose a Doc in parentheses if required or leave it as is. The need for enclosing is determined from an outer RenderContext (stored in the environment of the monad) and the inner fixity provided as an argument.

withPrettyIn ∷ MonadPrettyContext config env m ⇒ ((∀ a. PrettyBy config a ⇒ Direction → Fixity → a → Doc ann) → m r) → m r

Instantiate a supplied continuation with a precedence-aware pretty-printer.

withPrettyAt ∷ MonadPrettyContext config env m ⇒ Direction → Fixity → (AnyToDoc config ann → m r) → m r

Instantiate a supplied continuation with a pretty-printer specialized to supplied Fixity and Direction.

compoundDocM ∷ MonadPrettyContext config env m ⇒ Fixity → ((∀ a. PrettyBy config a ⇒ Direction → Fixity → a → Doc ann) → Doc ann) → m (Doc ann)

Instantiate a supplied continuation with a pretty-printer and apply encloseM, specialized to supplied Fixity, to the result.

sequenceDocM ∷ MonadPrettyContext config env m ⇒ Direction → Fixity → (AnyToDoc config ann → Doc ann) → m (Doc ann)

Instantiate a supplied continuation with a pretty-printer specialized to supplied Fixity and Direction and apply encloseM specialized to the provided fixity to the result. This can be useful for pretty-printing a sequence of values (possibly consisting of a single value).

infixDocM ∷ MonadPrettyContext config env m ⇒ Fixity → (AnyToDoc config ann → AnyToDoc config ann → Doc ann) → m (Doc ann)

Instantiate a supplied continuation with two pretty-printers (one is going in the ToTheLeft direction, the other is in the ToTheRight direction) specialized to supplied Fixity and apply encloseM, specialized to the same fixity, to the result. The idea is that to the outside an infix operator has the same inner fixity as it has the outer fixity to inner subexpressions.

juxtPrettyM ∷ (MonadPrettyContext config env m, PrettyBy config a, PrettyBy config b) ⇒ a → b → m (Doc ann)

Pretty-print two things with a space between them. The fixity of the context in which the arguments get pretty-printed is set to juxtFixity.