Text.PrettyBy.Monad

Description

A monadic interface to configurable pretty-printing.

Documentation

class HasPrettyConfig env config | env → config where

A constraint for "config is a part of env".

Methods

prettyConfig ∷ Lens' env config

type MonadPretty config env m = (MonadReader env m, HasPrettyConfig env config)

A constraint for "m is a monad that allows to pretty-print values in it by a config".

prettyM ∷ (MonadPretty config env m, PrettyBy config a) ⇒ a → m (Doc ann)

Pretty-print a value in a configurable way in a monad holding a config.

displayM ∷ ∀ str a m env config. (MonadPretty config env m, PrettyBy config a, Render str) ⇒ a → m str

Pretty-print and render a value as a string type in a configurable way in a monad holding a config.