UntypedPlutusCore.Analysis.Builtins

Documentation

data BuiltinsInfo (uni ∷ Type → Type) fun

All non-static information about builtins that the compiler might want.

Constructors

BuiltinsInfo

Fields

_biSemanticsVariant ∷ BuiltinSemanticsVariant fun

_biUnserializableConstants ∷ Some (ValueOf uni) → Bool

biSemanticsVariant ∷ ∀ uni fun fun. Lens (BuiltinsInfo uni fun) (BuiltinsInfo uni fun) (BuiltinSemanticsVariant fun) (BuiltinSemanticsVariant fun)

biUnserializableConstants ∷ ∀ uni fun uni. Lens (BuiltinsInfo uni fun) (BuiltinsInfo uni fun) (Some (ValueOf uni) → Bool) (Some (ValueOf uni) → Bool)

builtinArityInfo ∷ ∀ uni fun. ToBuiltinMeaning uni fun ⇒ BuiltinsInfo uni fun → fun → Arity

Get the arity of a builtin function from the BuiltinInfo.

constantIsSerializable ∷ ∀ uni fun. BuiltinsInfo uni fun → Some (ValueOf uni) → Bool

termIsSerializable ∷ BuiltinsInfo uni fun → Term name uni fun a → Bool

defaultUniUnserializableConstants ∷ Some (ValueOf DefaultUni) → Bool