UntypedPlutusCore.Core.Instance.Eq

Orphan instances

Hashable a ⇒ Hashable (Vector a)

Instance details

Methods

hashWithSalt ∷ Int → Vector a → Int

hash ∷ Vector a → Int

(GEq uni, Closed uni, Everywhere uni Eq, Eq fun, Eq ann, Eq (Term name uni fun ann)) ⇒ Eq (Program name uni fun ann)

Instance details

Methods

(==) ∷ Program name uni fun ann → Program name uni fun ann → Bool

(/=) ∷ Program name uni fun ann → Program name uni fun ann → Bool

(GEq uni, Closed uni, Everywhere uni Eq, Eq fun, Eq ann) ⇒ Eq (Term DeBruijn uni fun ann)

Instance details

Methods

(==) ∷ Term DeBruijn uni fun ann → Term DeBruijn uni fun ann → Bool

(/=) ∷ Term DeBruijn uni fun ann → Term DeBruijn uni fun ann → Bool

(GEq uni, Closed uni, Everywhere uni Eq, Eq fun, Eq ann) ⇒ Eq (Term FakeNamedDeBruijn uni fun ann)

Instance details

Methods

(==) ∷ Term FakeNamedDeBruijn uni fun ann → Term FakeNamedDeBruijn uni fun ann → Bool

(/=) ∷ Term FakeNamedDeBruijn uni fun ann → Term FakeNamedDeBruijn uni fun ann → Bool

(GEq uni, Closed uni, Everywhere uni Eq, Eq fun, Eq ann) ⇒ Eq (Term NamedDeBruijn uni fun ann)

Instance details

Methods

(==) ∷ Term NamedDeBruijn uni fun ann → Term NamedDeBruijn uni fun ann → Bool

(/=) ∷ Term NamedDeBruijn uni fun ann → Term NamedDeBruijn uni fun ann → Bool

(GEq uni, Closed uni, Everywhere uni Eq, Eq fun, Eq ann) ⇒ Eq (Term Name uni fun ann)

Instance details

Methods

(==) ∷ Term Name uni fun ann → Term Name uni fun ann → Bool

(/=) ∷ Term Name uni fun ann → Term Name uni fun ann → Bool

HashableTermConstraints uni fun ann ⇒ Hashable (Term DeBruijn uni fun ann)

Instance details

Methods

hashWithSalt ∷ Int → Term DeBruijn uni fun ann → Int

hash ∷ Term DeBruijn uni fun ann → Int

HashableTermConstraints uni fun ann ⇒ Hashable (Term FakeNamedDeBruijn uni fun ann)

Instance details

Methods

hashWithSalt ∷ Int → Term FakeNamedDeBruijn uni fun ann → Int

hash ∷ Term FakeNamedDeBruijn uni fun ann → Int

HashableTermConstraints uni fun ann ⇒ Hashable (Term NamedDeBruijn uni fun ann)

Instance details

Methods

hashWithSalt ∷ Int → Term NamedDeBruijn uni fun ann → Int

hash ∷ Term NamedDeBruijn uni fun ann → Int

HashableTermConstraints uni fun ann ⇒ Hashable (Term Name uni fun ann)

Instance details

Methods

hashWithSalt ∷ Int → Term Name uni fun ann → Int

hash ∷ Term Name uni fun ann → Int