UntypedPlutusCore.Evaluation.Machine.Cek

Description

The API to the CEK machine.

Running the machine

runCek ∷ ThrowableBuiltins uni fun ⇒ MachineParameters CekMachineCosts fun (CekValue uni fun ann) → ExBudgetMode cost uni fun → EmitterMode uni fun → Term Name uni fun ann → (Either (CekEvaluationException Name uni fun) (Term Name uni fun ()), cost, [Text])

Evaluate a term using the CEK machine with logging enabled and keep track of costing. A wrapper around the internal runCek to debruijn input and undebruijn output. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

runCekDeBruijn ∷ ThrowableBuiltins uni fun ⇒ MachineParameters CekMachineCosts fun (CekValue uni fun ann) → ExBudgetMode cost uni fun → EmitterMode uni fun → NTerm uni fun ann → (Either (CekEvaluationException NamedDeBruijn uni fun) (NTerm uni fun ()), cost, [Text])

Evaluate a term using the CEK machine and keep track of costing, logging is optional.

runCekNoEmit ∷ ThrowableBuiltins uni fun ⇒ MachineParameters CekMachineCosts fun (CekValue uni fun ann) → ExBudgetMode cost uni fun → Term Name uni fun ann → (Either (CekEvaluationException Name uni fun) (Term Name uni fun ()), cost)

Evaluate a term using the CEK machine with logging disabled and keep track of costing. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

evaluateCek ∷ ThrowableBuiltins uni fun ⇒ EmitterMode uni fun → MachineParameters CekMachineCosts fun (CekValue uni fun ann) → Term Name uni fun ann → (Either (CekEvaluationException Name uni fun) (Term Name uni fun ()), [Text])

Evaluate a term using the CEK machine with logging enabled. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

evaluateCekNoEmit ∷ ThrowableBuiltins uni fun ⇒ MachineParameters CekMachineCosts fun (CekValue uni fun ann) → Term Name uni fun ann → Either (CekEvaluationException Name uni fun) (Term Name uni fun ())

Evaluate a term using the CEK machine with logging disabled. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

data EvaluationResult a

The parameterized type of results various evaluation engines return. On the PLC side this becomes (via makeKnown) either a call to Error or a value of the PLC counterpart of type a.

Constructors

EvaluationSuccess !a

EvaluationFailure

extractEvaluationResult ∷ Either (EvaluationException operational structural term) a → Either (ErrorWithCause structural term) (EvaluationResult a)

Preserve the contents of an StructuralEvaluationError as a Left and turn an OperationalEvaluationError into a Right EvaluationFailure.

unsafeToEvaluationResult ∷ (PrettyPlc internal, PrettyPlc term, Typeable internal, Typeable term) ⇒ Either (EvaluationException user internal term) a → EvaluationResult a

Throw on a StructuralEvaluationError and turn an OperationalEvaluationError into an EvaluationFailure.

Errors

data CekUserError

Constructors

CekOutOfExError !ExRestrictingBudget
The final overspent (i.e. negative) budget.

CekEvaluationFailure
Error has been called or a builtin application has failed

data ErrorWithCause err cause

An error and (optionally) what caused it.

Constructors

ErrorWithCause

Fields

_ewcError ∷ !err

_ewcCause ∷ !(Maybe cause)

type CekEvaluationException name uni fun = EvaluationException CekUserError (MachineError fun) (Term name uni fun ())

The CEK machine-specific EvaluationException.

data EvaluationError operational structural

The type of errors that can occur during evaluation. There are two kinds of errors:

Operational ones -- these are errors that are indicative of the _logic_ of the program being wrong. For example, Error was executed, tailList was applied to an empty list or evaluation ran out of gas.

Structural ones -- these are errors that are indicative of the _structure_ of the program being wrong. For example, a free variable was encountered during evaluation, or a non-function was applied to an argument.

On the chain both of these are just regular failures and we don't distinguish between them there: if a script fails, it fails, it doesn't matter what the reason was. However in the tests it does matter why the failure occurred: a structural error may indicate that the test was written incorrectly while an operational error may be entirely expected.

In other words, operational errors are regular runtime errors and structural errors are "runtime type errors". Which means that evaluating an (erased) well-typed program should never produce a structural error, only an operational one. This creates a sort of "runtime type system" for UPLC and it would be great to stick to it and enforce in tests etc, but we currently don't. For example, a built-in function expecting a list but getting something else should throw a structural error, but currently it'll throw an operational one. This is something that we plan to improve upon in future.

Constructors

OperationalEvaluationError !operational

StructuralEvaluationError !structural

Costing

data ExBudgetCategory fun

Constructors

BStep StepKind

BBuiltinApp fun

BStartup

newtype CekBudgetSpender uni fun s

The CEK machine is parameterized over a spendBudget function. This makes the budgeting machinery extensible and allows us to separate budgeting logic from evaluation logic and avoid branching on the union of all possible budgeting state types during evaluation.

Constructors

CekBudgetSpender

Fields

unCekBudgetSpender ∷ ExBudgetCategory fun → ExBudget → CekM uni fun s ()

newtype ExBudgetMode cost uni fun

A budgeting mode to execute the CEK machine in.

Constructors

ExBudgetMode

Fields

unExBudgetMode ∷ ∀ s. ST s (ExBudgetInfo cost uni fun s)

data StepKind

Constructors

BConst

BVar

BLamAbs

BApply

BDelay

BForce

BBuiltin

BConstr

BCase

newtype CekExTally fun

For a detailed report on what costs how much + the same overall budget that Counting gives. The (derived) Monoid instance of CekExTally is the main piece of the machinery.

Constructors

CekExTally (MonoidalHashMap (ExBudgetCategory fun) ExBudget)

newtype CountingSt

For calculating the cost of execution by counting up using the Monoid instance of ExBudget.

Constructors

CountingSt ExBudget

data TallyingSt fun

Constructors

TallyingSt (CekExTally fun) ExBudget

newtype RestrictingSt

Constructors

RestrictingSt ExRestrictingBudget

type CekMachineCosts = CekMachineCostsBase Identity

Costs for evaluating AST nodes. Times should be specified in picoseconds, memory sizes in bytes.

Costing modes

monoidalBudgeting ∷ Monoid cost ⇒ (ExBudgetCategory fun → ExBudget → cost) → ExBudgetMode cost uni fun

Construct an ExBudgetMode out of a function returning a value of the budgeting state type. The value then gets added to the current state via (<>).

counting ∷ ExBudgetMode CountingSt uni fun

For calculating the cost of execution.

tallying ∷ Hashable fun ⇒ ExBudgetMode (TallyingSt fun) uni fun

For a detailed report on what costs how much + the same overall budget that Counting gives.

restricting ∷ ThrowableBuiltins uni fun ⇒ ExRestrictingBudget → ExBudgetMode RestrictingSt uni fun

For execution, to avoid overruns.

restrictingEnormous ∷ ThrowableBuiltins uni fun ⇒ ExBudgetMode RestrictingSt uni fun

restricting instantiated at enormousBudget.

enormousBudget ∷ ExRestrictingBudget

When we want to just evaluate the program we use the Restricting mode with an enormous budget, so that evaluation costs of on-chain budgeting are reflected accurately in benchmarks.

Emitter modes

noEmitter ∷ EmitterMode uni fun

No emitter.

logEmitter ∷ EmitterMode uni fun

Emits log only.

logWithTimeEmitter ∷ EmitterMode uni fun

Emits log with timestamp.

logWithBudgetEmitter ∷ EmitterMode uni fun

Emits log with the budget.

Misc

data BuiltinsRuntime fun val

A data wrapper around a function returning the BuiltinRuntime of a built-in function. We use data rather than newtype, because GHC is able to see through newtypes and may break carefully set up optimizations, see https://github.com/IntersectMBO/plutus/pull/4914#issuecomment-1396306606

Using data may make things more expensive, however it was verified at the time of writing that the wrapper is removed before the CEK machine starts, leaving the stored function to be used directly.

In order for lookups to be efficient the BuiltinRuntimes need to be cached, i.e. pulled out of the function statically. See makeBuiltinMeaning for how we achieve that.

Constructors

BuiltinsRuntime

Fields

unBuiltinsRuntime ∷ fun → BuiltinRuntime val

data CekValue uni fun ann

Constructors

VCon !(Some (ValueOf uni))

VDelay !(NTerm uni fun ann) !(CekValEnv uni fun ann)

VLamAbs !NamedDeBruijn !(NTerm uni fun ann) !(CekValEnv uni fun ann)

VBuiltin
A partial builtin application, accumulating arguments for eventual full application. We don't need a CekValEnv here unlike in the other constructors, because VBuiltin values always store their corresponding Terms fully discharged, see the comments at the call sites (search for VBuiltin).

Fields

!fun

So that we know, for what builtin we're calculating the cost. We can sneak this into BuiltinRuntime, so that we don't need to store it here, but somehow doing so was consistently slowing evaluation down by half a percent. Might be noise, might be not, but at least we know that removing this fun is not helpful anyway. See this commit reversing the change: https://github.com/IntersectMBO/plutus/pull/4778/commits/86a3e24ca3c671cc27c6f4344da2bcd14f961706

(NTerm uni fun ())

This must be lazy. It represents the fully discharged partial application of the builtin function that we're going to run when it's fully saturated. We need the Term to be able to return it in case full saturation is never achieved and a partial application needs to be returned in the result. The laziness is important, because the arguments are discharged values and discharging is expensive, so we don't want to do it unless we really have to. Making this field strict resulted in a 3-4.5% slowdown at the time of writing.

!(BuiltinRuntime (CekValue uni fun ann))

The partial application and its costing function. Check the docs of BuiltinRuntime for details. | A constructor value, including fully computed arguments and the tag.

VConstr !Word64 !(ArgStack uni fun ann)

readKnownCek ∷ (ThrowableBuiltins uni fun, ReadKnown (Term Name uni fun ()) a) ⇒ MachineParameters CekMachineCosts fun (CekValue uni fun ann) → Term Name uni fun ann → Either (CekEvaluationException Name uni fun) a

Unlift a value using the CEK machine. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

class Eq a ⇒ Hashable a

type ThrowableBuiltins uni fun = (PrettyUni uni, Pretty fun, Typeable uni, Typeable fun)

The set of constraints we need to be able to throw exceptions with things with built-in types and functions in them.