UntypedPlutusCore.Evaluation.Machine.Cek.ExBudgetMode

Documentation

newtype ExBudgetMode cost uni fun

A budgeting mode to execute the CEK machine in.

Constructors

ExBudgetMode

Fields

unExBudgetMode ∷ ∀ s. ST s (ExBudgetInfo cost uni fun s)

newtype CountingSt

For calculating the cost of execution by counting up using the Monoid instance of ExBudget.

Constructors

CountingSt ExBudget

newtype CekExTally fun

For a detailed report on what costs how much + the same overall budget that Counting gives. The (derived) Monoid instance of CekExTally is the main piece of the machinery.

Constructors

CekExTally (MonoidalHashMap (ExBudgetCategory fun) ExBudget)

data TallyingSt fun

Constructors

TallyingSt (CekExTally fun) ExBudget

newtype RestrictingSt

Constructors

RestrictingSt ExRestrictingBudget

class Eq a ⇒ Hashable a

The class of types that can be converted to a hash value.

Minimal implementation: hashWithSalt.

Hashable is intended exclusively for use in in-memory data structures. . Hashable does not have a fixed standard. This allows it to improve over time. . Because it does not have a fixed standard, different computers or computers on different versions of the code will observe different hash values. As such, Hashable is not recommended for use other than in-memory datastructures. Specifically, Hashable is not intended for network use or in applications which persist hashed values. For stable hashing use named hashes: sha256, crc32, xxhash etc.

If you are looking for Hashable instance in time package, check time-compat

monoidalBudgeting ∷ Monoid cost ⇒ (ExBudgetCategory fun → ExBudget → cost) → ExBudgetMode cost uni fun

Construct an ExBudgetMode out of a function returning a value of the budgeting state type. The value then gets added to the current state via (<>).

counting ∷ ExBudgetMode CountingSt uni fun

For calculating the cost of execution.

enormousBudget ∷ ExRestrictingBudget

When we want to just evaluate the program we use the Restricting mode with an enormous budget, so that evaluation costs of on-chain budgeting are reflected accurately in benchmarks.

tallying ∷ Hashable fun ⇒ ExBudgetMode (TallyingSt fun) uni fun

For a detailed report on what costs how much + the same overall budget that Counting gives.

restricting ∷ ThrowableBuiltins uni fun ⇒ ExRestrictingBudget → ExBudgetMode RestrictingSt uni fun

For execution, to avoid overruns.

restrictingLarge ∷ ThrowableBuiltins uni fun ⇒ ExBudgetMode RestrictingSt uni fun

restricting instantiated at largeBudget.

restrictingEnormous ∷ ThrowableBuiltins uni fun ⇒ ExBudgetMode RestrictingSt uni fun

restricting instantiated at enormousBudget.