UntypedPlutusCore.Evaluation.Machine.Cek.StepCounter

Documentation

newtype StepCounter (n ∷ Nat) s

A set of Word8 counters that is used in the CEK machine to count steps.

Constructors

StepCounter (MutablePrimArray s Word8)

newCounter ∷ (KnownNat n, PrimMonad m) ⇒ Proxy n → m (StepCounter n (PrimState m))

Make a new StepCounter with the given number of counters.

resetCounter ∷ ∀ n m. (KnownNat n, PrimMonad m) ⇒ StepCounter n (PrimState m) → m ()

Reset all the counters in the given StepCounter to zero.

readCounter ∷ ∀ m n. PrimMonad m ⇒ StepCounter n (PrimState m) → Int → m Word8

Read the value of a counter.

writeCounter ∷ ∀ m n. PrimMonad m ⇒ StepCounter n (PrimState m) → Int → Word8 → m ()

Write to a counter.

modifyCounter ∷ PrimMonad m ⇒ Int → (Word8 → Word8) → StepCounter n (PrimState m) → m Word8

Modify the value of a counter. Returns the modified value.

data Peano

The type of natural numbers in Peano form.

Constructors

Z

S Peano

type family NatToPeano n where ...

Equations

NatToPeano 0 = 'Z

NatToPeano n = 'S (NatToPeano (n - 1))

class Applicative f ⇒ UpwardsM f n where

Methods

upwardsM ∷ Int → (Int → f ()) → f ()

upwardsM i k means k i *> k (i + 1) *> ... *> k (i + n - 1). We use this function in order to statically unroll a loop in itraverseCounter_ through instance resolution. This makes validation benchmarks a couple of percent faster.

itraverseCounter_ ∷ ∀ n m. (UpwardsM m (NatToPeano n), PrimMonad m) ⇒ (Int → Word8 → m ()) → StepCounter n (PrimState m) → m ()

Traverse the counters with an effectful function.

iforCounter_ ∷ (UpwardsM m (NatToPeano n), PrimMonad m) ⇒ StepCounter n (PrimState m) → (Int → Word8 → m ()) → m ()

Traverse the counters with an effectful function.