UntypedPlutusCore.Rename

Description

The user-facing API of the untyped renamer. See PlutusCore.Rename for details.

Documentation

class Rename a where

The class of things that can be renamed. I.e. things that are capable of satisfying the global uniqueness condition.

Methods

rename ∷ MonadQuote m ⇒ a → m a

Rename Uniques so that they're globally unique. In case there are any free variables, they must be left untouched and bound variables must not get renamed to free ones. Must always assign new names to bound variables, so that rename can be used for alpha-renaming as well.

Orphan instances

HasUniques (Program name uni fun ann) ⇒ Rename (Program name uni fun ann)

Instance details

Methods

rename ∷ MonadQuote m ⇒ Program name uni fun ann → m (Program name uni fun ann)

HasUniques (Term name uni fun ann) ⇒ Rename (Term name uni fun ann)

Instance details

Methods

rename ∷ MonadQuote m ⇒ Term name uni fun ann → m (Term name uni fun ann)