UntypedPlutusCore.Subst

Documentation

substVarA ∷ Applicative f ⇒ (name → f (Maybe (Term name uni fun ann))) → Term name uni fun ann → f (Term name uni fun ann)

Applicatively replace a variable using the given function.

substVar ∷ (name → Maybe (Term name uni fun ann)) → Term name uni fun ann → Term name uni fun ann

Replace a variable using the given function.

termSubstNamesM ∷ Monad m ⇒ (name → m (Maybe (Term name uni fun ann))) → Term name uni fun ann → m (Term name uni fun ann)

Naively monadically substitute names using the given function (i.e. do not substitute binders).

termSubstNames ∷ (name → Maybe (Term name uni fun ann)) → Term name uni fun ann → Term name uni fun ann

Naively substitute names using the given function (i.e. do not substitute binders).

termMapNames ∷ ∀ name name' uni fun ann. (name → name') → Term name uni fun ann → Term name' uni fun ann

Completely replace the names with a new name type.

programMapNames ∷ ∀ name name' uni fun ann. (name → name') → Program name uni fun ann → Program name' uni fun ann

substConstantA ∷ Applicative f ⇒ (ann → Some (ValueOf uni) → f (Maybe (Term name uni fun ann))) → Term name uni fun ann → f (Term name uni fun ann)

Applicatively replace a constant using the given function.

substConstant ∷ (ann → Some (ValueOf uni) → Maybe (Term name uni fun ann)) → Term name uni fun ann → Term name uni fun ann

Replace a constant using the given function.

termSubstConstantsM ∷ Monad m ⇒ (ann → Some (ValueOf uni) → m (Maybe (Term name uni fun ann))) → Term name uni fun ann → m (Term name uni fun ann)

Monadically substitute constants using the given function.

termSubstConstants ∷ (ann → Some (ValueOf uni) → Maybe (Term name uni fun ann)) → Term name uni fun ann → Term name uni fun ann

Substitute constants using the given function.

vTerm ∷ Fold (Term name uni fun ann) name

Get all the term variables in a term.