UntypedPlutusCore.Transform.Certify.Trace

Documentation

data CertifiedOptStage

Datatype which represents optimization passes which are also certified.

This means that these passes are formalized as part of the certifier, and adding a new pass constructor to this type means that it is expected the pass will be also certified in the same PR.

Constructors

FloatDelay

ForceDelay

ForceCaseDelay

Inline

CSE

ApplyToCase

data UncertifiedOptStage

Datatype which represents optimization passes which are not yet certified.

IMPORTANT: if you add a new pass, or modify an existing pass, without also modifying the certifier in the same PR, you must add/move its corresponding constructor to this type. Please also open an issue at https://github.com/IntersectMBO/plutus/issues.

Constructors

CaseOfCase

LetFloatOut

CaseReduce

type OptStage = Either UncertifiedOptStage CertifiedOptStage

pattern FloatDelayStage ∷ OptStage

pattern ForceDelayStage ∷ OptStage

pattern ForceCaseDelayStage ∷ OptStage

pattern CaseReduceStage ∷ OptStage

pattern InlineStage ∷ OptStage

pattern CseStage ∷ OptStage

pattern ApplyToCaseStage ∷ OptStage

pattern CaseOfCaseStage ∷ OptStage

pattern LetFloatOutStage ∷ OptStage

data Optimization name uni fun a

Constructors

Optimization

Fields

beforeAST ∷ Term name uni fun a

stage ∷ OptStage

hints ∷ Hints

afterAST ∷ Term name uni fun a

newtype OptimizerTrace name uni fun a

Constructors

OptimizerTrace

Fields

optimizerTrace ∷ [Optimization name uni fun a]

initOptimizerTrace ∷ OptimizerTrace name uni fun a

allASTs ∷ OptimizerTrace name uni fun a → [Term name uni fun a]