UntypedPlutusCore.Transform.Inline

Description

An inlining pass.

This pass is essentially a copy of the PIR inliner, and should be KEPT IN SYNC with it. It's hard to do this with true abstraction, so we just have to keep two copies reasonably similar.

However, there are some differences. In the interests of making it easier to keep things in sync, these are explicitly listed in Note [Differences from PIR inliner]. If you add another difference, please note it there! Obviously fewer differences is better.

See Note [The problem of inlining destructors] for why this pass exists.

Documentation

inline

Arguments

∷ ∀ name uni fun m a. ExternalConstraints name uni fun m

⇒ AstSize
inline threshold

→ Bool
inline constants

→ Bool
preserve logging

→ InlineHints name a

→ BuiltinSemanticsVariant fun

→ Term name uni fun a

→ SimplifierT name uni fun a m (Term name uni fun a)

Inline simple bindings. Relies on global uniqueness, and preserves it. See Note [Inlining and global uniqueness]

newtype InlineHints name a

Constructors

InlineHints

Fields

shouldInline ∷ a → name → Inline

Exported for testing

type InlineM name uni fun a = ReaderT (InlineInfo name fun a) (StateT (S name uni fun a) Quote)

The monad the inliner runs in.

data S name uni fun a

UPLC inliner state

Constructors

S

Fields

_subst ∷ Subst name uni fun a

_vars ∷ UniqueMap TermUnique (VarInfo name uni fun a)

data InlineInfo name fun a

Constructors

InlineInfo

Fields

_iiUsages ∷ Usages

_iiHints ∷ InlineHints name a

_iiBuiltinSemanticsVariant ∷ BuiltinSemanticsVariant fun

_iiInlineConstants ∷ Bool

_iiInlineCallsiteGrowth ∷ AstSize

_iiPreserveLogging ∷ Bool

newtype Subst name uni fun a

Wrapper of term substitution so that it's similar to the PIR inliner. See Note [Differences from PIR inliner] 1

Constructors

Subst

Fields

_termEnv ∷ TermEnv name uni fun a

newtype TermEnv name uni fun a

Term substitution, Subst in the paper. A map of unprocessed variable and its substitution range.

Constructors

TermEnv

Fields

_unTermEnv ∷ UniqueMap TermUnique (InlineTerm name uni fun a)

isFirstVarBeforeEffects ∷ ∀ name uni fun ann. InliningConstraints name uni fun ⇒ BuiltinSemanticsVariant fun → name → Term name uni fun ann → Bool

isStrictIn ∷ ∀ name uni fun a. Eq name ⇒ name → Term name uni fun a → Bool

Check if the given name is strict in the given term. This means that at least one occurrence of the name is found outside of the following: * delay term * lambda body * case branch

effectSafe

Arguments

∷ ∀ name uni fun a. InliningConstraints name uni fun

⇒ Term name uni fun a

→ name

→ Bool
is it pure? See Note [Inlining and purity]

→ InlineM name uni fun a Bool