Evaluation.Builtins.BLS12_381.Utils

Documentation

compressionBit ∷ Word8

ByteString utilities. These are used in tests to check that the format of compressed points conforms to the specification at https://github.com/supranational/blst#serialization-format .

infinityBit ∷ Word8

signBit ∷ Word8

unsafeUnconsBS ∷ ByteString → (Word8, ByteString)

modifyMSB ∷ (Word8 → Word8) → ByteString → ByteString

Apply some function to the most significant byte of a bytestring

flipBits ∷ Word8 → ByteString → ByteString

Flip a specified set of bits in the most significant byte of a bytestring.

clearBits ∷ Word8 → ByteString → ByteString

Clear a specified set of bits in the most significant byte of a bytestring.

setBits ∷ Word8 → ByteString → ByteString

Set a specified set of bits in the most significant byte of a bytestring.

isSet ∷ Word8 → ByteString → Bool

Check that a specified set of bits is set in the most significant byte of a bytestring.