Evaluation.Builtins.Common

Documentation

unsafeSplitStructuralOperational ∷ (PrettyPlc structural, PrettyPlc term, Typeable structural, Typeable term) ⇒ Either (EvaluationException structural operational term) a → EvaluationResult a

Throw on a StructuralError and turn an OperationalError into an EvaluationFailure (thus erasing the content of the error in the latter case).

evaluateCek ∷ ∀ (uni ∷ Type → Type) fun ann. ThrowableBuiltins uni fun ⇒ EmitterMode uni fun → MachineParameters CekMachineCosts fun (CekValue uni fun ann) → Term Name uni fun ann → (Either (CekEvaluationException Name uni fun) (Term Name uni fun ()), [Text])

Evaluate a term using the CEK machine with logging enabled. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

evaluateCekNoEmit ∷ ∀ (uni ∷ Type → Type) fun ann. ThrowableBuiltins uni fun ⇒ MachineParameters CekMachineCosts fun (CekValue uni fun ann) → Term Name uni fun ann → Either (CekEvaluationException Name uni fun) (Term Name uni fun ())

Evaluate a term using the CEK machine with logging disabled. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

readKnownCek ∷ ∀ (uni ∷ Type → Type) fun a ann. (ThrowableBuiltins uni fun, ReadKnown (Term Name uni fun ()) a) ⇒ MachineParameters CekMachineCosts fun (CekValue uni fun ann) → Term Name uni fun ann → Either (CekEvaluationException Name uni fun) a

Unlift a value using the CEK machine. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

typecheckAnd ∷ (MonadError (TypeErrorPlc uni fun ()) m, Typecheckable uni fun, GEq uni, CaseBuiltin uni, Closed uni, uni `Everywhere` ExMemoryUsage) ⇒ BuiltinSemanticsVariant fun → (MachineParameters CekMachineCosts fun (CekValue uni fun ()) → Term Name uni fun () → a) → CostingPart uni fun → Term TyName Name uni fun () → m a

Type check and evaluate a term.

typecheckEvaluateCek ∷ (MonadError (TypeErrorPlc uni fun ()) m, Typecheckable uni fun, GEq uni, uni `Everywhere` ExMemoryUsage, PrettyUni uni, Pretty fun, CaseBuiltin uni) ⇒ BuiltinSemanticsVariant fun → CostingPart uni fun → Term TyName Name uni fun () → m (EvaluationResult (Term Name uni fun ()), [Text])

Type check and evaluate a term, logging enabled.

typecheckEvaluateCekNoEmit ∷ (MonadError (TypeErrorPlc uni fun ()) m, Typecheckable uni fun, GEq uni, uni `Everywhere` ExMemoryUsage, PrettyUni uni, Pretty fun, CaseBuiltin uni) ⇒ BuiltinSemanticsVariant fun → CostingPart uni fun → Term TyName Name uni fun () → m (EvaluationResult (Term Name uni fun ()))

Type check and evaluate a term, logging disabled.

typecheckReadKnownCek ∷ (MonadError (TypeErrorPlc uni fun ()) m, Typecheckable uni fun, GEq uni, uni `Everywhere` ExMemoryUsage, PrettyUni uni, Pretty fun, CaseBuiltin uni, ReadKnown (Term Name uni fun ()) a) ⇒ BuiltinSemanticsVariant fun → CostingPart uni fun → Term TyName Name uni fun () → m (Either (CekEvaluationException Name uni fun) a)

Type check and convert a Plutus Core term to a Haskell value.

type PlcType = Type TyName DefaultUni ()

type PlcTerm = Term TyName Name DefaultUni DefaultFun ()

type UplcTerm = Term Name DefaultUni DefaultFun ()

data CekResult

Constructors

TypeCheckError PlcError

CekError

CekSuccess UplcTerm

evalTerm ∷ PlcTerm → CekResult

mkApp1 ∷ DefaultFun → PlcTerm → PlcTerm

mkApp2 ∷ DefaultFun → PlcTerm → PlcTerm → PlcTerm

ok ∷ PlcTerm → Property

Term evaluates successfully

fails ∷ PlcTerm → Property

Term fails to evaluate successfully

evalOkEq ∷ PlcTerm → PlcTerm → Property

evalOkTrue ∷ PlcTerm → Property

integer ∷ Integer → PlcTerm

bytestring ∷ ByteString → PlcTerm

zero ∷ PlcTerm

one ∷ PlcTerm

true ∷ PlcTerm

false ∷ PlcTerm

cekSuccessFalse ∷ CekResult

cekSuccessTrue ∷ CekResult