PlutusCore.Bitwise

Description

Implementations for CIP-121, CIP-122 and CIP-123. Grouped because they all operate on ByteStrings, and require similar functionality.

Wrappers

integerToByteStringWrapper ∷ Bool → Integer → Integer → BuiltinResult ByteString

Wrapper for integerToByteString to make it more convenient to define as a builtin.

byteStringToIntegerWrapper ∷ Bool → ByteString → Integer

Wrapper for byteStringToInteger to make it more convenient to define as a builtin.

shiftByteStringWrapper ∷ ByteString → Integer → ByteString

Wrapper for calling shiftByteString safely. Specifically, we avoid various edge cases:

Empty ByteStrings and zero moves don't do anything

Bit moves whose absolute value is larger than the bit length produce all-zeroes

This also ensures we don't accidentally hit integer overflow issues.

rotateByteStringWrapper ∷ ByteString → Integer → ByteString

Wrapper for calling rotateByteString safely. Specifically, we avoid various edge cases:

Empty ByteStrings and zero moves don't do anything

Bit moves whose absolute value is larger than the bit length gets modulo reduced

Furthermore, we can convert all rotations into positive rotations, by noting that a rotation by b is the same as a rotation by b mod bitLen, where bitLen is the length of the ByteString argument in bits. This value is always non-negative, and if we get 0, we have nothing to do. This reduction also helps us avoid integer overflow issues.

writeBitsWrapper ∷ ByteString → [Integer] → [Bool] → BuiltinResult ByteString

Needed due to the complexities of passing lists of pairs as arguments. Effectively, we pass the second argument as required by CIP-122 in its 'unzipped' form, truncating mismatches.

Implementation details

data IntegerToByteStringError

Structured type to help indicate conversion errors.

Constructors

NegativeInput

NotEnoughDigits

integerToByteStringMaximumOutputLength ∷ Integer

integerToByteString ∷ ByteOrder → Int → Integer → Either IntegerToByteStringError ByteString

Conversion from Integer to ByteString, as per CIP-121.

For performance and clarity, the endianness argument uses ByteOrder, and the length argument is an Int.

byteStringToInteger ∷ ByteOrder → ByteString → Integer

Conversion from ByteString to Integer, as per CIP-121.

For clarity, the stated endianness argument uses ByteOrder.

andByteString ∷ Bool → ByteString → ByteString → ByteString

Bitwise logical AND, as per CIP-122.

orByteString ∷ Bool → ByteString → ByteString → ByteString

Bitwise logical OR, as per CIP-122.

xorByteString ∷ Bool → ByteString → ByteString → ByteString

Bitwise logical XOR, as per CIP-122.

complementByteString ∷ ByteString → ByteString

Bitwise logical complement, as per CIP-122.

readBit ∷ ByteString → Int → BuiltinResult Bool

Bit read at index, as per CIP-122

writeBits ∷ ByteString → [(Integer, Bool)] → BuiltinResult ByteString

Bulk bit write, as per CIP-122

replicateByte ∷ Int → Word8 → BuiltinResult ByteString

Byte replication, as per CIP-122

shiftByteString ∷ ByteString → Int → ByteString

Shifts, as per CIP-123.

rotateByteString ∷ ByteString → Int → ByteString

Rotations, as per CIP-123.

countSetBits ∷ ByteString → Int

Counting the number of set bits, as per CIP-123.

findFirstSetBit ∷ ByteString → Int

Finding the first set bit's index, as per CIP-123.