PlutusCore.Builtin.KnownType

Documentation

data BuiltinError

The type of errors that readKnown and makeKnown can return.

throwBuiltinErrorWithCause ∷ (MonadError (ErrorWithCause err cause) m, AsUnliftingEvaluationError err, AsEvaluationFailure err) ⇒ cause → BuiltinError → m void

Attach a cause to a BuiltinError and throw that. Note that an evaluator might require the cause to be computed lazily for best performance on the happy path, hence this function must not force its first argument. TODO: wrap cause in Lazy once we have it.

type KnownBuiltinTypeIn uni val a = (HasConstantIn uni val, PrettyParens (SomeTypeIn uni), GEq uni, uni `HasTermLevel` a)

A constraint for "a is a ReadKnownIn and MakeKnownIn by means of being included in uni".

type KnownBuiltinType val a = KnownBuiltinTypeIn (UniOf val) val a

A constraint for "a is a ReadKnownIn and MakeKnownIn by means of being included in UniOf term".

data BuiltinResult a

The monad that makeKnown runs in. Equivalent to ExceptT BuiltinError Emitter, except optimized in two ways:

everything is strict

has the BuiltinSuccess constructor that is used for returning a value with no logs attached, which is the most common case for us, so it helps a lot not to construct and deconstruct a redundant tuple

Moving from ExceptT BuiltinError Emitter to this data type gave us a speedup of 8% of total evaluation time.

Logs are represented as a DList, because we don't particularly care about the efficiency of logging, since there's no logging on the chain and builtins don't emit much anyway. Otherwise we'd have to use text-builder or text-builder-linear or something of this sort.

Constructors

BuiltinSuccess a

BuiltinSuccessWithLogs (DList Text) a

BuiltinFailure (DList Text) BuiltinError

type ReadKnownM = Either BuiltinError

The monad that readKnown runs in.

class uni ~ UniOf val ⇒ MakeKnownIn uni val a where

Minimal complete definition

Nothing

Methods

makeKnown ∷ a → BuiltinResult val

Convert a Haskell value to the corresponding PLC value. The inverse of readKnown.

default makeKnown ∷ KnownBuiltinType val a ⇒ a → BuiltinResult val

readKnownConstant ∷ ∀ val a. KnownBuiltinType val a ⇒ val → ReadKnownM a

Convert a constant embedded into a PLC term to the corresponding Haskell value.

type MakeKnown val = MakeKnownIn (UniOf val) val

class uni ~ UniOf val ⇒ ReadKnownIn uni val a where

Minimal complete definition

Nothing

Methods

readKnown ∷ val → ReadKnownM a

Convert a PLC value to the corresponding Haskell value. The inverse of makeKnown.

default readKnown ∷ KnownBuiltinType val a ⇒ val → ReadKnownM a

type ReadKnown val = ReadKnownIn (UniOf val) val

makeKnownOrFail ∷ MakeKnownIn uni val a ⇒ a → EvaluationResult val

Same as makeKnown, but allows for neither emitting nor storing the cause of a failure.

readKnownSelf ∷ (ReadKnown val a, AsUnliftingEvaluationError err, AsEvaluationFailure err) ⇒ val → Either (ErrorWithCause err val) a

Same as readKnown, but the cause of a potential failure is the provided term itself.