PlutusCore.Evaluation.Machine.Exception

Description

The exceptions that an abstract machine can throw.

Documentation

newtype UnliftingError

The error message part of an UnliftingEvaluationError.

Constructors

MkUnliftingError

Fields

unUnliftingError ∷ Text

class AsUnliftingError r where

Minimal complete definition

_UnliftingError

Methods

_UnliftingError ∷ Prism' r UnliftingError

_MkUnliftingError ∷ Prism' r Text

data BuiltinError

The type of errors that readKnown and makeKnown can return.

Constructors

BuiltinUnliftingEvaluationError UnliftingEvaluationError

BuiltinEvaluationFailure

data MachineError fun

Errors which can occur during a run of an abstract machine.

Constructors

NonPolymorphicInstantiationMachineError
An attempt to reduce a not immediately reducible type instantiation.

NonWrapUnwrappedMachineError
An attempt to unwrap a not wrapped term.

NonFunctionalApplicationMachineError
An attempt to reduce a not immediately reducible application.

OpenTermEvaluatedMachineError
An attempt to evaluate an open term.

UnliftingMachineError UnliftingError
An attempt to compute a constant application resulted in UnliftingError.

BuiltinTermArgumentExpectedMachineError
A builtin expected a term argument, but something else was received.

UnexpectedBuiltinTermArgumentMachineError
A builtin received a term argument when something else was expected

NonConstrScrutinized

MissingCaseBranch Word64

class AsMachineError r fun | r → fun where

Minimal complete definition

_MachineError

Methods

_MachineError ∷ Prism' r (MachineError fun)

_NonPolymorphicInstantiationMachineError ∷ Prism' r ()

_NonWrapUnwrappedMachineError ∷ Prism' r ()

_NonFunctionalApplicationMachineError ∷ Prism' r ()

_OpenTermEvaluatedMachineError ∷ Prism' r ()

_UnliftingMachineError ∷ Prism' r UnliftingError

_BuiltinTermArgumentExpectedMachineError ∷ Prism' r ()

_UnexpectedBuiltinTermArgumentMachineError ∷ Prism' r ()

_NonConstrScrutinized ∷ Prism' r ()

_MissingCaseBranch ∷ Prism' r Word64

data EvaluationError operational structural

The type of errors that can occur during evaluation. There are two kinds of errors:

Operational ones -- these are errors that are indicative of the _logic_ of the program being wrong. For example, error was executed, tailList was applied to an empty list or evaluation ran out of gas.

Structural ones -- these are errors that are indicative of the _structure_ of the program being wrong. For example, a free variable was encountered during evaluation, a non-function was applied to an argument or tailList was applied to a non-list.

On the chain both of these are just regular failures and we don't distinguish between them there: if a script fails, it fails, it doesn't matter what the reason was. However in the tests it does matter why the failure occurred: a structural error may indicate that the test was written incorrectly while an operational error may be entirely expected.

In other words, operational errors are regular runtime errors and structural errors are "runtime type errors". Which means that evaluating an (erased) well-typed program should never produce a structural error, only an operational one. This creates a sort of "runtime type system" for UPLC and it would be great to stick to it and enforce in tests etc, but we currently don't. For example, a built-in function expecting a list but getting something else should throw a structural error, but currently it'll throw an operational one. This is something that we plan to improve upon in future.

Constructors

OperationalEvaluationError !operational

StructuralEvaluationError !structural

class AsEvaluationError r operational structural | r → operational structural where

Minimal complete definition

_EvaluationError

Methods

_EvaluationError ∷ Prism' r (EvaluationError operational structural)

_OperationalEvaluationError ∷ Prism' r operational

_StructuralEvaluationError ∷ Prism' r structural

data ErrorWithCause err cause

An error and (optionally) what caused it.

Constructors

ErrorWithCause

Fields

_ewcError ∷ !err

_ewcCause ∷ !(Maybe cause)

type EvaluationException operational structural = ErrorWithCause (EvaluationError operational structural)

throwNotAConstant ∷ MonadError BuiltinError m ⇒ m void

throwing ∷ MonadError e m ⇒ AReview e t → t → m x

throwing_ ∷ MonadError e m ⇒ AReview e () → m x

throwingWithCause ∷ ∀ exc e t term m x. (exc ~ ErrorWithCause e term, MonadError exc m) ⇒ AReview e t → t → Maybe term → m x

Prismatically throw an error and its (optional) cause.

extractEvaluationResult ∷ Either (EvaluationException operational structural term) a → Either (ErrorWithCause structural term) (EvaluationResult a)

Preserve the contents of an StructuralEvaluationError as a Left and turn an OperationalEvaluationError into a Right EvaluationFailure.

unsafeToEvaluationResult ∷ (PrettyPlc internal, PrettyPlc term, Typeable internal, Typeable term) ⇒ Either (EvaluationException user internal term) a → EvaluationResult a

Throw on a StructuralEvaluationError and turn an OperationalEvaluationError into an EvaluationFailure.