PlutusCore.Executable.Parsers

Description

Common option parsers for executables

Documentation

input ∷ Parser Input

Parser for an input stream. If none is specified, default to stdin for ease of use in pipeline.

fileInput ∷ Parser Input

stdInput ∷ Parser Input

output ∷ Parser Output

Parser for an output stream. If none is specified, default to stdout for ease of use in pipeline.

fileOutput ∷ Parser Output

stdOutput ∷ Parser Output

noOutput ∷ Parser Output

formatHelp ∷ String

formatReader ∷ String → Maybe Format

inputformat ∷ Parser Format

outputformat ∷ Parser Format

tracemode ∷ Parser TraceMode

files ∷ Parser Files

applyOpts ∷ Parser ApplyOptions

printmode ∷ Parser PrintMode

nameformat ∷ Parser NameFormat

certifier ∷ Parser Certifier

printOpts ∷ Parser PrintOptions

convertOpts ∷ Parser ConvertOptions

optimiseOpts ∷ Parser OptimiseOptions

exampleMode ∷ Parser ExampleMode

exampleAvailable ∷ Parser ExampleMode

exampleName ∷ Parser ExampleName

exampleSingle ∷ Parser ExampleMode

exampleOpts ∷ Parser ExampleOptions

builtinSemanticsVariantReader ∷ String → Maybe (BuiltinSemanticsVariant DefaultFun)

showBuiltinSemanticsVariant ∷ BuiltinSemanticsVariant DefaultFun → String

builtinSemanticsVariant ∷ Parser (BuiltinSemanticsVariant DefaultFun)

pirFormatHelp ∷ String

pirFormatReader ∷ String → Maybe PirFormat

pPirInputFormat ∷ Parser PirFormat

pPirOutputFormat ∷ Parser PirFormat

languageReader ∷ String → Maybe Language

pLanguage ∷ Parser Language