PlutusCore.Generators.Hedgehog.Builtin

Documentation

class GenTypedTerm uni where

This class exists so we can provide an ad-hoc typed term generator for various universes. We usually rely-on a universe-specific generator for well-typed constants within that universe.

TODO: Move this to PlutusIR.Generators.AST, and merge genConstant with genConstant.

Methods

genTypedTerm ∷ ∀ (a ∷ Type) tyname name fun. TypeRep a → Gen (Term tyname name uni fun ())

Generate a Term in uni with the given type.

class GenArbitraryTerm uni where

This class exists so we can provide an ad-hoc arbitrary term generator for various universes.

Methods

genArbitraryTerm ∷ ∀ fun. (Bounded fun, Enum fun) ⇒ Gen (Term TyName Name uni fun ())

Generate an arbitrary Term in uni.

genConstant ∷ ∀ (a ∷ Type). TypeRep a → SomeGen DefaultUni

genInteger ∷ Gen Integer

genByteString ∷ Gen ByteString

genText ∷ Gen Text

genData ∷ Int → Gen Data

genI ∷ Gen Data

genB ∷ Gen Data

genList ∷ Int → Gen Data

genMap ∷ Int → Gen Data

genConstr ∷ Int → Gen Data

genBls12_381_G1_Element ∷ Gen Element

genBls12_381_G2_Element ∷ Gen Element

genBls12_381_MlResult ∷ Gen MlResult