PlutusCore.Value

Documentation

data Value

The underlying type of the UPLC built-in type Value.

data K

A ByteString with maximum length of maxKeyLen bytes.

k ∷ ByteString → Maybe K

unK ∷ K → ByteString

maxKeyLen ∷ Int

data Quantity

A signed 128-bit integer quantity.

quantity ∷ Integer → Maybe Quantity

Smart constructor for Quantity that validates bounds.

unQuantity ∷ Quantity → Integer

zeroQuantity ∷ Quantity

The zero quantity.

addQuantity ∷ Quantity → Quantity → Maybe Quantity

Safely add two quantities, checking for overflow.

negativeAmounts ∷ Value → Int

type NestedMap = Map K (Map K Quantity)

unpack ∷ Value → NestedMap

Unpack a Value into a map from (currency symbol, token name) to quantity.

The map is guaranteed to not contain empty inner map or zero quantity.

pack ∷ NestedMap → Value

Pack a map from (currency symbol, token name) to quantity into a Value.

The map will be filtered so that it does not contain empty inner map or zero quantity.

empty ∷ Value

fromList ∷ [(K, [(K, Quantity)])] → BuiltinResult Value

toList ∷ Value → [(K, [(K, Quantity)])]

toFlatList ∷ Value → [(K, K, Quantity)]

totalSize ∷ Value → Int

Total size, i.e., the number of distinct `(currency symbol, token name)` pairs contained in the Value.

maxInnerSize ∷ Value → Int

Size of the largest inner map.

insertCoin ∷ ByteString → ByteString → Integer → Value → BuiltinResult Value

\(O(\log \max(m, k))\), where \(m\) is the size of the outer map, and \(k\) is the size of the largest inner map.

deleteCoin ∷ ByteString → ByteString → Value → Value

\(O(\log \max(m, k))\)

scaleValue ∷ Integer → Value → BuiltinResult Value

\(O(n)\). Scale each token by the given constant factor.

lookupCoin ∷ ByteString → ByteString → Value → Integer

\(O(\log \max(m, k))\)

valueContains ∷ Value → Value → BuiltinResult Bool

\(O(n_{2}\log \max(m_{1}, k_{1}))\), where \(n_{2}\) is the total size of the second Value, \(m_{1}\) is the size of the outer map in the first Value and \(k_{1}\) is the size of the largest inner map in the first Value.

a contains b if for each (currency, token, quantity) in b, lookup currency token a >= quantity.

Both values must not contain negative amounts.

unionValue ∷ Value → Value → BuiltinResult Value

\(O(n_{1}) + O(n_{2})\), where \(n_{1}\) and \(n_{2}\) are the total sizes (i.e., sum of inner map sizes) of the two maps.

Shortcircuits if either value is empty.

Since unionValue is commutative, we switch the arguments whenever the second value is larger in total size than the first one. We have found through experimentation that this results in better performance in practice.

valueData ∷ Value → BuiltinResult Data

\(O(n)\). Encodes Value as Data, in the same way as non-builtin Value. This is the denotation of ValueData in Plutus V1, V2 and V3.

unValueData ∷ Data → BuiltinResult Value

\(O(n)\). Decodes Data into Value. This is the denotation of UnValueData in Plutus V1, V2 and V3.