PlutusIR.Subst

Documentation

fvTerm ∷ HasUnique name TermUnique ⇒ Traversal' (Term tyname name uni fun ann) name

Get all the free term variables in a PIR term.

ftvTerm ∷ HasUnique tyname TypeUnique ⇒ Traversal' (Term tyname name uni fun ann) tyname

Get all the free type variables in a PIR term.

fvBinding ∷ HasUnique name TermUnique ⇒ Traversal' (Binding tyname name uni fun ann) name

Get all the free variables in a PIR single let-binding.

ftvBinding ∷ HasUnique tyname TypeUnique ⇒ Recursivity → Traversal' (Binding tyname name uni fun ann) tyname

Get all the free type variables in a PIR single let-binding.

ftvTy ∷ ∀ tyname unique (uni ∷ Type → Type) ann. HasUnique tyname unique ⇒ Traversal' (Type tyname uni ann) tyname

Get all the free type variables in a type.

vTerm ∷ Fold (Term tyname name uni fun ann) name

Get all the term variables in a term.

tvTerm ∷ Fold (Term tyname name uni fun ann) tyname

Get all the type variables in a term.

tvTy ∷ ∀ tyname (uni ∷ Type → Type) ann f. (Contravariant f, Applicative f) ⇒ (tyname → f tyname) → Type tyname uni ann → f (Type tyname uni ann)

Get all the type variables in a type.