PlutusTx.Blueprint.Class

Documentation

class HasSchema (t ∷ Type) (referencedTypes ∷ [Type]) where

A class of types that have a Blueprint schema definition and can reference other schema definitions of other types.

Methods

schema ∷ Schema referencedTypes