PlutusTx.Blueprint.Definition.Unroll

Documentation

type family UnrollAll xs ∷ [Type] where ...

Equations

UnrollAll '[] = '[]

UnrollAll (x ': xs) = Concat (Unroll x) (UnrollAll xs)

type family Unroll (p ∷ Type) ∷ [Type] where ...

Unroll a type into a list of all nested types (including the type itself).

If a type doesn't have a generic representation, then this type family gets "stuck". The good news is that for the purpose of deriving schema definitions, we only need to consider types that are either end-user defined (and therefore have a generic representation) or built-in types that we explicitly list here as terminals in order not to get "stuck".

Equations

Unroll Int = '[Int]

Unroll Integer = '[Integer]

Unroll Text = '[Text]

Unroll BuiltinData = '[BuiltinData]

Unroll BuiltinUnit = '[BuiltinUnit]

Unroll BuiltinString = '[BuiltinString]

Unroll (BuiltinList a) = Unroll a

Unroll (BuiltinPair a b) = Unroll a ++ Unroll b

Unroll BuiltinByteString = '[BuiltinByteString]

Unroll [a] = Unroll a

Unroll (a, b) = Unroll a ++ Unroll b

Unroll (Maybe a) = Unroll a

Unroll p = Prepend p (GUnroll (Break (NoGeneric p) (Rep p)))

type family Break e (rep ∷ Type → Type) ∷ Type → Type where ...

Detect stuck type family: https://blog.csongor.co.uk/report-stuck-families/#custom-type-errors

Equations

Break _ (M1 a b c) = M1 a b c

Break _ (f :*: g) = f :*: g

Break _ (f :+: g) = f :+: g

Break _ (K1 a b) = K1 a b

Break e U1 = U1

Break e x = e

type family NoGeneric t where ...

Equations

NoGeneric x = TypeError (Text "No instance for " :<>: ShowType (Generic x))

type family GUnroll (t ∷ Type → Type) ∷ [Type] where ...

Unroll a generic representation of a type into a list of all nested types.

Equations

GUnroll (M1 _ _ f) = GUnroll f

GUnroll (f :*: g) = GUnroll f ++ GUnroll g

GUnroll (f :+: g) = GUnroll f ++ GUnroll g

GUnroll (K1 _ c) = Unroll c

GUnroll U1 = '[]

type family Insert x xs where ...

Insert x into xs unless it's already there.

Equations

Insert x '[] = '[x]

Insert x (x : xs) = x ': xs

Insert x (y : xs) = y ': Insert x xs

type family Prepend x xs where ...

Equations

Prepend x '[] = '[x]

Prepend x (x : xs) = x ': xs

Prepend x (y : xs) = x ': (y ': xs)

type family Concat (as ∷ [k]) (bs ∷ [k]) ∷ [k] where ...

Concatenates two type-level lists

Equations

Concat '[] bs = bs

Concat as '[] = as

Concat (a : as) bs = a ': Concat as bs

type family (as ∷ [k]) ++ (bs ∷ [k]) ∷ [k] where ... infixr 5

Concatenates two type-level lists removing duplicates.

Equations

'[] ++ bs = bs

as ++ '[] = as

(a : as) ++ bs = Insert a (as ++ bs)

class Unrollable ts where

This class and its two instances are used internally to derive Definitions for a given list of types.

Methods

unroll ∷ Definitions ts

deriveDefinitions ∷ ∀ ts. Unrollable (UnrollAll ts) ⇒ Definitions (UnrollAll ts)