PlutusTx.Ord

Documentation

class Eq a ⇒ Ord a where

The Ord class is used for totally ordered datatypes.

Minimal complete definition: either compare or <=. Using compare can be more efficient for complex types.

Minimal complete definition

compare | (<=)

Methods

compare ∷ a → a → Ordering

(<) ∷ a → a → Bool infix 4

(<=) ∷ a → a → Bool infix 4

(>) ∷ a → a → Bool infix 4

(>=) ∷ a → a → Bool infix 4

max ∷ a → a → a

min ∷ a → a → a

data Ordering

Constructors

LT

EQ

GT