PlutusTx.Plugin.Boilerplate

Description

Insert necessary or recommended GHC flags and extensions via a driver plugin. See https://plutus.cardano.intersectmbo.org/docs/using-plinth/extensions-flags-pragmas

Documentation

optNoStrict ∷ CommandLineOption

Unfortunately, it seems like the Strict extension set by the driver plugin cannot be unset via LANGUAGE NoStrict. So we add a plugin flag to allow users to do so.

boilerplateOpts ∷ [CommandLineOption]

removeBoilerplateOpts ∷ [CommandLineOption] → [CommandLineOption]

addFlagsAndExts ∷ [CommandLineOption] → HscEnv → IO HscEnv