PlutusTx.These

Documentation

data These a b

A These a b is either an a, or a b or an a and a b. Plutus version of These.

Constructors

This a

That b

These a b

these ∷ (a → c) → (b → c) → (a → b → c) → These a b → c

theseWithDefault ∷ a → b → (a → b → c) → These a b → c

Consume a 'These a b' value.