UntypedPlutusCore

Documentation

module PlutusCore.Name.Unique

module UntypedPlutusCore.Subst

module UntypedPlutusCore.Size

module UntypedPlutusCore.Simplify

module UntypedPlutusCore.DeBruijn

module UntypedPlutusCore.Core

module UntypedPlutusCore.Check.Scope

data Term name uni fun ann

The type of Untyped Plutus Core terms. Mirrors the type of Typed Plutus Core terms except

all types are removed

IWrap and Unwrap are removed

type abstractions are replaced with Delay

type instantiations are replaced with Force

The latter two are due to the fact that we don't have value restriction in Typed Plutus Core and hence a computation can be stuck expecting only a single type argument for the computation to become unstuck. Therefore we can't just silently remove type abstractions and instantiations and need to replace them with something else that also blocks evaluation (in order for the semantics of an erased program to match with the semantics of the original typed one). Delay and Force serve exactly this purpose.

Constructors

Var !ann !name

LamAbs !ann !name !(Term name uni fun ann)

Apply !ann !(Term name uni fun ann) !(Term name uni fun ann)

Force !ann !(Term name uni fun ann)

Delay !ann !(Term name uni fun ann)

Constant !ann !(Some (ValueOf uni))

Builtin !ann !fun

Error !ann

Constr !ann !Word64 ![Term name uni fun ann]

Case !ann !(Term name uni fun ann) !(Vector (Term name uni fun ann))

data Program name uni fun ann

A Program is simply a Term coupled with a Version of the core language.

Constructors

Program

Fields

_progAnn ∷ ann

_progVer ∷ Version

_progTerm ∷ Term name uni fun ann

applyProgram ∷ (MonadError ApplyProgramError m, Semigroup a) ⇒ Program name uni fun a → Program name uni fun a → m (Program name uni fun a)

Applies one program to another. Fails if the versions do not match and tries to merge annotations.

parseScoped ∷ (MonadError (Error uni fun SrcSpan) m, MonadQuote m) ⇒ Text → m (Program Name DefaultUni DefaultFun SrcSpan)

Parse and rewrite so that names are globally unique, not just unique within their scope.

data DefaultUni a

The universe used by default.

data DefaultFun

Default built-in functions.

When updating these, make sure to add them to the protocol version listing! See Note [New builtins/language versions and protocol versions]