UntypedPlutusCore.Evaluation.Machine.SteppableCek.DebugDriver

Documentation

class Breakpointable ann bps | ann → bps where

Leave abstract the types of annotation and breakpoints. The only thing the driver requires is an inclusion relation of breakpoints into the Annotation

Methods

hasBreakpoints ∷ ann → bps → Bool

data CekState uni fun ann

data Cmd bps

The commands that the driver may receive from the client (tui,cli,test,etc)

Constructors

Step
Instruct the driver to a *SINGLE* step. Note: No need to pass breakpoints here because the stepping granularity is *minimal*.

Continue bps
Instruct to multi-step until end-of-program or until breakpoint reached

Next bps
Instruct to multi-step over the function call at point or until breakpoint reached

Finish bps
Instruct to multi-step to end of current function or until breakpoint reached

runDriverT ∷ ∀ uni fun ann bps m. (Breakpointable ann bps, MonadFree (DebugF uni fun ann bps) m) ⇒ NTerm uni fun ann → m ()

Entrypoint of the driver

data DebugF uni fun ann bps a

The drivers's suspension functor

Constructors

InputF (Cmd bps → a)
Await for the client (e.g. TUI) to tell what to do next (Cmd).

DriverLogF Text a
The debug driver wants to log something

StepF
An enumeratee of Driver State (generator+iteratee): Yield a state before doing a step, then await for a state to resume after the step. See Note [Stepping the driver].

Fields

(CekState uni fun ann)

yield with the current driver's state before running a step

(CekState uni fun ann → a)

resume back with a state after the step interpretation | A generator of CekState to yield to client (e.g. TUI)

UpdateClientF (CekState uni fun ann) a

Reexport some functions for convenience

mkCekTrans ∷ ∀ cost uni fun ann m s. (ThrowableBuiltins uni fun, PrimMonad m, s ~ PrimState m) ⇒ MachineParameters CekMachineCosts fun (CekValue uni fun ann) → ExBudgetMode cost uni fun → EmitterMode uni fun → Slippage → m (CekTrans uni fun ann s, ExBudgetInfo cost uni fun s)

Based on the supplied arguments, initialize the CEK environment and construct a state transition function. Returns the constructed transition function paired with the methods to live access the running budget.

type CekTrans uni fun ann s = Trans (CekM uni fun s) (CekState uni fun ann)

class Monad m ⇒ MonadFree (f ∷ Type → Type) (m ∷ Type → Type) | m → f

iterM ∷ (Functor f, Monad m) ⇒ (f (m a) → m a) → Free f a → m a

iterTM ∷ ∀ f (m ∷ Type → Type) t a. (Functor f, Monad m, MonadTrans t, Monad (t m)) ⇒ (f (t m a) → t m a) → FreeT f m a → t m a

partialIterT ∷ Monad m ⇒ Integer → (∀ a. f a → m a) → FreeT f m b → FreeT f m b

cutoff ∷ ∀ (f ∷ Type → Type) (m ∷ Type → Type) a. (Functor f, Monad m) ⇒ Integer → FreeT f m a → FreeT f m (Maybe a)

data FreeT (f ∷ Type → Type) (m ∷ Type → Type) a